API.AttachmentMenuBot

Documentation

data AttachmentMenuBot

Represents a bot added to attachment menu

bot_user_id User identifier of the bot added to attachment menu

name Name for the bot in attachment menu

name_color Color to highlight selected name of the bot if appropriate; may be null

default_icon Default attachment menu icon for the bot in SVG format; may be null

ios_static_icon Attachment menu icon for the bot in SVG format for the official iOS app; may be null

ios_animated_icon Attachment menu icon for the bot in TGS format for the official iOS app; may be null

android_icon Attachment menu icon for the bot in TGS format for the official Android app; may be null

macos_icon Attachment menu icon for the bot in TGS format for the official native macOS app; may be null

icon_color Color to highlight selected icon of the bot if appropriate; may be null

Constructors

AttachmentMenuBot

Fields

icon_color :: Maybe AttachmentMenuBotColor

macos_icon :: Maybe File

android_icon :: Maybe File

ios_animated_icon :: Maybe File

ios_static_icon :: Maybe File

default_icon :: Maybe File

name_color :: Maybe AttachmentMenuBotColor

name :: Maybe String

bot_user_id :: Maybe Int