API.BackgroundFill

Documentation

data BackgroundFill

Describes a fill of a background

Constructors

BackgroundFillSolid
Describes a solid fill of a background

color A color of the background in the RGB24 format

Fields

color :: Maybe Int

BackgroundFillGradient
Describes a gradient fill of a background

top_color A top color of the background in the RGB24 format

bottom_color A bottom color of the background in the RGB24 format

rotation_angle Clockwise rotation angle of the gradient, in degrees; 0-359. Must be always divisible by 45

Fields

rotation_angle :: Maybe Int

bottom_color :: Maybe Int

top_color :: Maybe Int

BackgroundFillFreeformGradient
Describes a freeform gradient fill of a background

colors A list of 3 or 4 colors of the freeform gradients in the RGB24 format

Fields

colors :: Maybe [Int]