API.BasicGroup

Documentation

data BasicGroup

Represents a basic group of 0-200 users (must be upgraded to a supergroup to accommodate more than 200 users)

id Group identifier

member_count Number of members in the group

status Status of the current user in the group

is_active True, if the group is active

upgraded_to_supergroup_id Identifier of the supergroup to which this group was upgraded; 0 if none

Constructors

BasicGroup

Fields

upgraded_to_supergroup_id :: Maybe Int

is_active :: Maybe Bool

status :: Maybe ChatMemberStatus

member_count :: Maybe Int

_id :: Maybe Int