API.Chat

Documentation

data Chat

A chat. (Can be a private chat, basic group, supergroup, or secret chat)

id Chat unique identifier

type Type of the chat

title Chat title

photo Chat photo; may be null

permissions Actions that non-administrator chat members are allowed to take in the chat

last_message Last message in the chat; may be null

positions Positions of the chat in chat lists

message_sender_id Identifier of a user or chat that is selected to send messages in the chat; may be null if the user can't change message sender

has_protected_content True, if chat content can't be saved locally, forwarded, or copied

is_marked_as_unread True, if the chat is marked as unread

is_blocked True, if the chat is blocked by the current user and private messages from the chat can't be received

has_scheduled_messages True, if the chat has scheduled messages

can_be_deleted_only_for_self True, if the chat messages can be deleted only for the current user while other users will continue to see the messages

can_be_deleted_for_all_users True, if the chat messages can be deleted for all users

can_be_reported True, if the chat can be reported to Telegram moderators through reportChat or reportChatPhoto

default_disable_notification Default value of the disable_notification parameter, used when a message is sent to the chat

unread_count Number of unread messages in the chat

last_read_inbox_message_id Identifier of the last read incoming message

last_read_outbox_message_id Identifier of the last read outgoing message

unread_mention_count Number of unread messages with a mention/reply in the chat

unread_reaction_count Number of messages with unread reactions in the chat

notification_settings Notification settings for the chat

available_reactions List of reactions, available in the chat

message_ttl Current message Time To Live setting (self-destruct timer) for the chat; 0 if not defined. TTL is counted from the time message or its content is viewed in secret chats and from the send date in other chats

theme_name If non-empty, name of a theme, set for the chat

action_bar Information about actions which must be possible to do through the chat action bar; may be null

video_chat Information about video chat of the chat

pending_join_requests Information about pending join requests; may be null

reply_markup_message_id Identifier of the message from which reply markup needs to be used; 0 if there is no default custom reply markup in the chat

draft_message A draft of a message in the chat; may be null

client_data Application-specific data associated with the chat. (For example, the chat scroll position or local chat notification settings can be stored here.) Persistent if the message database is used

Constructors

Chat

Fields

client_data :: Maybe String

draft_message :: Maybe DraftMessage

reply_markup_message_id :: Maybe Int

pending_join_requests :: Maybe ChatJoinRequestsInfo

video_chat :: Maybe VideoChat

action_bar :: Maybe ChatActionBar

theme_name :: Maybe String

message_ttl :: Maybe Int

available_reactions :: Maybe [String]

notification_settings :: Maybe ChatNotificationSettings

unread_reaction_count :: Maybe Int

unread_mention_count :: Maybe Int

last_read_outbox_message_id :: Maybe Int

last_read_inbox_message_id :: Maybe Int

unread_count :: Maybe Int

default_disable_notification :: Maybe Bool

can_be_reported :: Maybe Bool

can_be_deleted_for_all_users :: Maybe Bool

can_be_deleted_only_for_self :: Maybe Bool

has_scheduled_messages :: Maybe Bool

is_blocked :: Maybe Bool

is_marked_as_unread :: Maybe Bool

has_protected_content :: Maybe Bool

message_sender_id :: Maybe MessageSender

positions :: Maybe [ChatPosition]

last_message :: Maybe Message

permissions :: Maybe ChatPermissions

photo :: Maybe ChatPhotoInfo

title :: Maybe String

_type :: Maybe ChatType

_id :: Maybe Int

p :: Show a => String -> Maybe a -> String

cc :: [String] -> String