API.ChatEventAction

Documentation

data ChatEventAction

Represents a chat event

Constructors

ChatEventMessageEdited
A message was edited

old_message The original message before the edit

new_message The message after it was edited

Fields

new_message :: Maybe Message

old_message :: Maybe Message

ChatEventMessageDeleted
A message was deleted

message Deleted message

Fields

message :: Maybe Message

ChatEventPollStopped
A poll in a message was stopped

message The message with the poll

Fields

message :: Maybe Message

ChatEventMessagePinned
A message was pinned

message Pinned message

Fields

message :: Maybe Message

ChatEventMessageUnpinned
A message was unpinned

message Unpinned message

Fields

message :: Maybe Message

ChatEventMemberJoined
A new member joined the chat

ChatEventMemberJoinedByInviteLink
A new member joined the chat by an invite link

invite_link Invite link used to join the chat

Fields

invite_link :: Maybe ChatInviteLink

ChatEventMemberLeft
A member left the chat

ChatEventMemberInvited
A new chat member was invited

user_id New member user identifier

status New member status

Fields

status :: Maybe ChatMemberStatus

user_id :: Maybe Int

ChatEventMemberPromoted
A chat member has gained/lost administrator status, or the list of their administrator privileges has changed

user_id Affected chat member user identifier

old_status Previous status of the chat member

new_status New status of the chat member

Fields

new_status :: Maybe ChatMemberStatus

old_status :: Maybe ChatMemberStatus

user_id :: Maybe Int

ChatEventMemberRestricted
A chat member was restrictedunrestricted or bannedunbanned, or the list of their restrictions has changed

member_id Affected chat member identifier

old_status Previous status of the chat member

new_status New status of the chat member

Fields

new_status :: Maybe ChatMemberStatus

old_status :: Maybe ChatMemberStatus

member_id :: Maybe MessageSender

ChatEventTitleChanged
The chat title was changed

old_title Previous chat title

new_title New chat title

Fields

new_title :: Maybe String

old_title :: Maybe String

ChatEventPermissionsChanged
The chat permissions was changed

old_permissions Previous chat permissions

new_permissions New chat permissions

Fields

new_permissions :: Maybe ChatPermissions

old_permissions :: Maybe ChatPermissions

ChatEventDescriptionChanged
The chat description was changed

old_description Previous chat description

new_description New chat description

Fields

new_description :: Maybe String

old_description :: Maybe String

ChatEventUsernameChanged
The chat username was changed

old_username Previous chat username

new_username New chat username

Fields

new_username :: Maybe String

old_username :: Maybe String

ChatEventPhotoChanged
The chat photo was changed

old_photo Previous chat photo value; may be null

new_photo New chat photo value; may be null

Fields

new_photo :: Maybe ChatPhoto

old_photo :: Maybe ChatPhoto

ChatEventInvitesToggled
The can_invite_users permission of a supergroup chat was toggled

can_invite_users New value of can_invite_users permission

Fields

can_invite_users :: Maybe Bool

ChatEventLinkedChatChanged
The linked chat of a supergroup was changed

old_linked_chat_id Previous supergroup linked chat identifier

new_linked_chat_id New supergroup linked chat identifier

Fields

new_linked_chat_id :: Maybe Int

old_linked_chat_id :: Maybe Int

ChatEventSlowModeDelayChanged
The slow_mode_delay setting of a supergroup was changed

old_slow_mode_delay Previous value of slow_mode_delay

new_slow_mode_delay New value of slow_mode_delay

Fields

new_slow_mode_delay :: Maybe Int

old_slow_mode_delay :: Maybe Int

ChatEventMessageTtlSettingChanged
The message TTL setting was changed

old_message_ttl_setting Previous value of message_ttl_setting

new_message_ttl_setting New value of message_ttl_setting

Fields

new_message_ttl_setting :: Maybe Int

old_message_ttl_setting :: Maybe Int

ChatEventSignMessagesToggled
The sign_messages setting of a channel was toggled

sign_messages New value of sign_messages

Fields

sign_messages :: Maybe Bool

ChatEventStickerSetChanged
The supergroup sticker set was changed

old_sticker_set_id Previous identifier of the chat sticker set; 0 if none

new_sticker_set_id New identifier of the chat sticker set; 0 if none

Fields

new_sticker_set_id :: Maybe Int

old_sticker_set_id :: Maybe Int

ChatEventLocationChanged
The supergroup location was changed

old_location Previous location; may be null

new_location New location; may be null

Fields

new_location :: Maybe ChatLocation

old_location :: Maybe ChatLocation

ChatEventIsAllHistoryAvailableToggled
The is_all_history_available setting of a supergroup was toggled

is_all_history_available New value of is_all_history_available

Fields

is_all_history_available :: Maybe Bool

ChatEventInviteLinkEdited
A chat invite link was edited

old_invite_link Previous information about the invite link

new_invite_link New information about the invite link

Fields

new_invite_link :: Maybe ChatInviteLink

old_invite_link :: Maybe ChatInviteLink

ChatEventInviteLinkRevoked
A chat invite link was revoked

invite_link The invite link

Fields

invite_link :: Maybe ChatInviteLink

ChatEventInviteLinkDeleted
A revoked chat invite link was deleted

invite_link The invite link

Fields

invite_link :: Maybe ChatInviteLink

ChatEventVoiceChatCreated
A voice chat was created

group_call_id Identifier of the voice chat. The voice chat can be received through the method getGroupCall

Fields

group_call_id :: Maybe Int

ChatEventVoiceChatDiscarded
A voice chat was discarded

group_call_id Identifier of the voice chat. The voice chat can be received through the method getGroupCall

Fields

group_call_id :: Maybe Int

ChatEventVoiceChatParticipantIsMutedToggled
A voice chat participant was muted or unmuted

participant_id Identifier of the affected group call participant

is_muted New value of is_muted

Fields

is_muted :: Maybe Bool

participant_id :: Maybe MessageSender

ChatEventVoiceChatParticipantVolumeLevelChanged
A voice chat participant volume level was changed

participant_id Identifier of the affected group call participant

volume_level New value of volume_level; 1-20000 in hundreds of percents

Fields

volume_level :: Maybe Int

participant_id :: Maybe MessageSender

ChatEventVoiceChatMuteNewParticipantsToggled
The mute_new_participants setting of a voice chat was toggled

mute_new_participants New value of the mute_new_participants setting

Fields

mute_new_participants :: Maybe Bool