API.ChatMemberStatus

Documentation

data ChatMemberStatus

Provides information about the status of a member in a chat

Constructors

ChatMemberStatusCreator
The user is the owner of the chat and has all the administrator privileges

custom_title A custom title of the owner; 0-16 characters without emojis; applicable to supergroups only

is_anonymous True, if the creator isn't shown in the chat member list and sends messages anonymously; applicable to supergroups only

is_member True, if the user is a member of the chat

Fields

is_member :: Maybe Bool

is_anonymous :: Maybe Bool

custom_title :: Maybe String

ChatMemberStatusAdministrator
The user is a member of the chat and has some additional privileges. In basic groups, administrators can edit and delete messages sent by others, add new members, ban unprivileged members, and manage voice chats. In supergroups and channels, there are more detailed options for administrator privileges

custom_title A custom title of the administrator; 0-16 characters without emojis; applicable to supergroups only

can_be_edited True, if the current user can edit the administrator privileges for the called user

can_manage_chat True, if the administrator can get chat event log, get chat statistics, get message statistics in channels, get channel members, see anonymous administrators in supergroups and ignore slow mode. Implied by any other privilege; applicable to supergroups and channels only

can_change_info True, if the administrator can change the chat title, photo, and other settings

can_post_messages True, if the administrator can create channel posts; applicable to channels only

can_edit_messages True, if the administrator can edit messages of other users and pin messages; applicable to channels only

can_delete_messages True, if the administrator can delete messages of other users

can_invite_users True, if the administrator can invite new users to the chat

can_restrict_members True, if the administrator can restrict, ban, or unban chat members; always true for channels

can_pin_messages True, if the administrator can pin messages; applicable to basic groups and supergroups only

can_promote_members True, if the administrator can add new administrators with a subset of their own privileges or demote administrators that were directly or indirectly promoted by them

can_manage_voice_chats True, if the administrator can manage voice chats

is_anonymous True, if the administrator isn't shown in the chat member list and sends messages anonymously; applicable to supergroups only

Fields

is_anonymous :: Maybe Bool

can_manage_voice_chats :: Maybe Bool

can_promote_members :: Maybe Bool

can_pin_messages :: Maybe Bool

can_restrict_members :: Maybe Bool

can_invite_users :: Maybe Bool

can_delete_messages :: Maybe Bool

can_edit_messages :: Maybe Bool

can_post_messages :: Maybe Bool

can_change_info :: Maybe Bool

can_manage_chat :: Maybe Bool

can_be_edited :: Maybe Bool

custom_title :: Maybe String

ChatMemberStatusMember
The user is a member of the chat, without any additional privileges or restrictions

ChatMemberStatusRestricted
The user is under certain restrictions in the chat. Not supported in basic groups and channels

is_member True, if the user is a member of the chat

restricted_until_date Point in time (Unix timestamp) when restrictions will be lifted from the user; 0 if never. If the user is restricted for more than 366 days or for less than 30 seconds from the current time, the user is considered to be restricted forever

permissions User permissions in the chat

Fields

permissions :: Maybe ChatPermissions

restricted_until_date :: Maybe Int

is_member :: Maybe Bool

ChatMemberStatusLeft
The user or the chat is not a chat member

ChatMemberStatusBanned
The user or the chat was banned (and hence is not a member of the chat). Implies the user can't return to the chat, view messages, or be used as a participant identifier to join a voice chat of the chat

banned_until_date Point in time (Unix timestamp) when the user will be unbanned; 0 if never. If the user is banned for more than 366 days or for less than 30 seconds from the current time, the user is considered to be banned forever. Always 0 in basic groups

Fields

banned_until_date :: Maybe Int