API.ConnectedWebsite

Documentation

data ConnectedWebsite

Contains information about one website the current user is logged in with Telegram

id Website identifier

domain_name The domain name of the website

bot_user_id User identifier of a bot linked with the website

browser The version of a browser used to log in

platform Operating system the browser is running on

log_in_date Point in time (Unix timestamp) when the user was logged in

last_active_date Point in time (Unix timestamp) when obtained authorization was last used

ip IP address from which the user was logged in, in human-readable format

location Human-readable description of a country and a region, from which the user was logged in, based on the IP address

Constructors

ConnectedWebsite

Fields

location :: Maybe String

ip :: Maybe String

last_active_date :: Maybe Int

log_in_date :: Maybe Int

platform :: Maybe String

browser :: Maybe String

bot_user_id :: Maybe Int

domain_name :: Maybe String

_id :: Maybe Int