API.FileDownload

Documentation

data FileDownload

Describes a file added to file download list

file_id File identifier

message The message with the file

add_date Point in time (Unix timestamp) when the file was added to the download list

complete_date Point in time (Unix timestamp) when the file downloading was completed; 0 if the file downloading isn't completed

is_paused True, if downloading of the file is paused

Constructors

FileDownload

Fields

is_paused :: Maybe Bool

complete_date :: Maybe Int

add_date :: Maybe Int

message :: Maybe Message

file_id :: Maybe Int