API.Functions.AddLocalMessage

Documentation

data AddLocalMessage

Adds a local message to a chat. The message is persistent across application restarts only if the message database is used. Returns the added message

chat_id Target chat

sender_user_id Identifier of the user who will be shown as the sender of the message; may be 0 for channel posts

reply_to_message_id Identifier of the message to reply to or 0

disable_notification Pass true to disable notification for the message

input_message_content The content of the message to be added

Constructors

AddLocalMessage

Fields

input_message_content :: Maybe InputMessageContent

disable_notification :: Maybe Bool

reply_to_message_id :: Maybe Int

sender_user_id :: Maybe Int

chat_id :: Maybe Int