API.Functions.ChangeImportedContacts

Documentation

data ChangeImportedContacts

Changes imported contacts using the list of contacts saved on the device. Imports newly added contacts and, if at least the file database is enabled, deletes recently deleted contacts.

Query result depends on the result of the previous query, so only one query is possible at the same time

contacts The new list of contacts, contact's vCard are ignored and are not imported

Constructors

ChangeImportedContacts

Fields

contacts :: Maybe [Contact]

p :: Show a => String -> Maybe a -> String

cc :: [String] -> String