API.Functions.DownloadFile

Documentation

data DownloadFile

Downloads a file from the cloud. Download progress and completion of the download will be notified through updateFile updates

file_id Identifier of the file to download

priority Priority of the download (1-32). The higher the priority, the earlier the file will be downloaded. If the priorities of two files are equal, then the last one for which downloadFile was called will be downloaded first

offset The starting position from which the file needs to be downloaded

limit Number of bytes which need to be downloaded starting from the "offset" position before the download will be automatically canceled; use 0 to download without a limit

synchronous If false, this request returns file state just after the download has been started. If true, this request returns file state only after

the download has succeeded, has failed, has been canceled or a new downloadFile request with different offset/limit parameters was sent

Constructors

DownloadFile

Fields

synchronous :: Maybe Bool

limit :: Maybe Int

offset :: Maybe Int

priority :: Maybe Int

file_id :: Maybe Int

p :: Show a => String -> Maybe a -> String

cc :: [String] -> String