API.Functions.GetChatMessageCount

Documentation

data GetChatMessageCount

Returns approximate number of messages of the specified type in the chat

chat_id Identifier of the chat in which to count messages

filter Filter for message content; searchMessagesFilterEmpty is unsupported in this function

return_local If true, returns count that is available locally without sending network requests, returning -1 if the number of messages is unknown

Constructors

GetChatMessageCount

Fields

return_local :: Maybe Bool

_filter :: Maybe SearchMessagesFilter

chat_id :: Maybe Int