API.Functions.GetChats

Documentation

data GetChats

Returns an ordered list of chats in a chat list. Chats are sorted by the pair (order, chat_id) in decreasing order. (For example, to get a list of chats from the beginning, the offset_order should be equal to a biggest signed 64-bit number 9223372036854775807 == 2^63 - 1).

For optimal performance the number of returned chats is chosen by the library

chat_list The chat list in which to return chats

offset_order Chat order to return chats from

offset_chat_id Chat identifier to return chats from

limit The maximum number of chats to be returned. It is possible that fewer chats than the limit are returned even if the end of the list is not reached

Constructors

GetChats

Fields

limit :: Maybe Int

offset_chat_id :: Maybe Int

offset_order :: Maybe Int

chat_list :: Maybe ChatList