API.Functions.GetMessageLink

Documentation

data GetMessageLink

Returns an HTTPS link to a message in a chat. Available only for already sent messages in supergroups and channels, or if message.can_get_media_timestamp_links and a media timestamp link is generated. This is an offline request

chat_id Identifier of the chat to which the message belongs

message_id Identifier of the message

media_timestamp If not 0, timestamp from which the videoaudiovideo note/voice note playing must start, in seconds. The media can be in the message content or in its web page preview

for_album Pass true to create a link for the whole media album

for_comment Pass true to create a link to the message as a channel post comment, or from a message thread

Constructors

GetMessageLink

Fields

for_comment :: Maybe Bool

for_album :: Maybe Bool

media_timestamp :: Maybe Int

message_id :: Maybe Int

chat_id :: Maybe Int

p :: Show a => String -> Maybe a -> String

cc :: [String] -> String