API.Functions.GetPassportAuthorizationForm

Documentation

data GetPassportAuthorizationForm

Returns a Telegram Passport authorization form for sharing data with a service

bot_user_id User identifier of the service's bot

scope Telegram Passport element types requested by the service

public_key Service's public key

nonce Unique request identifier provided by the service

Constructors

GetPassportAuthorizationForm

Fields

nonce :: Maybe String

public_key :: Maybe String

scope :: Maybe String

bot_user_id :: Maybe Int

p :: Show a => String -> Maybe a -> String

cc :: [String] -> String