API.Functions.JoinGroupCall

Documentation

data JoinGroupCall

Joins an active group call. Returns join response payload for tgcalls

group_call_id Group call identifier

participant_id Identifier of a group call participant, which will be used to join the call; pass null to join as self; voice chats only

audio_source_id Caller audio channel synchronization source identifier; received from tgcalls

payload Group call join payload; received from tgcalls

is_muted True, if the user's microphone is muted

is_my_video_enabled True, if the user's video is enabled

invite_hash If non-empty, invite hash to be used to join the group call without being muted by administrators

Constructors

JoinGroupCall

Fields

invite_hash :: Maybe String

is_my_video_enabled :: Maybe Bool

is_muted :: Maybe Bool

payload :: Maybe String

audio_source_id :: Maybe Int

participant_id :: Maybe MessageSender

group_call_id :: Maybe Int

p :: Show a => String -> Maybe a -> String

cc :: [String] -> String