API.Functions.LoadChats

Documentation

data LoadChats

Loads more chats from a chat list. The loaded chats and their positions in the chat list will be sent through updates. Chats are sorted by the pair (chat.position.order, chat.id) in descending order. Returns a 404 error if all chats have been loaded

chat_list The chat list in which to load chats; pass null to load chats from the main chat list

limit The maximum number of chats to be loaded. For optimal performance, the number of loaded chats is chosen by TDLib and can be smaller than the specified limit, even if the end of the list is not reached

Constructors

LoadChats

Fields

limit :: Maybe Int

chat_list :: Maybe ChatList

p :: Show a => String -> Maybe a -> String

cc :: [String] -> String