API.Functions.RecoverPassword

Documentation

data RecoverPassword

Recovers the 2-step verification password using a recovery code sent to an email address that was previously set up

recovery_code Recovery code to check

new_password New password of the user; may be empty to remove the password

new_hint New password hint; may be empty

Constructors

RecoverPassword

Fields

new_hint :: Maybe String

new_password :: Maybe String

recovery_code :: Maybe String

p :: Show a => String -> Maybe a -> String

cc :: [String] -> String