API.Functions.SearchChatMessages

Documentation

data SearchChatMessages

Searches for messages with given words in the chat. Returns the results in reverse chronological order, i.e. in order of decreasing message_id. Cannot be used in secret chats with a non-empty query

(searchSecretMessages must be used instead), or without an enabled message database. For optimal performance, the number of returned messages is chosen by TDLib and can be smaller than the specified limit

chat_id Identifier of the chat in which to search messages

query Query to search for

sender_id Identifier of the sender of messages to search for; pass null to search for messages from any sender. Not supported in secret chats

from_message_id Identifier of the message starting from which history must be fetched; use 0 to get results from the last message

offset Specify 0 to get results from exactly the from_message_id or a negative offset to get the specified message and some newer messages

limit The maximum number of messages to be returned; must be positive and can't be greater than 100. If the offset is negative, the limit must be greater than -offset. For optimal performance, the number of returned messages is chosen by TDLib and can be smaller than the specified limit

filter Additional filter for messages to search; pass null to search for all messages

message_thread_id If not 0, only messages in the specified thread will be returned; supergroups only

Constructors

SearchChatMessages

Fields

message_thread_id :: Maybe Int

_filter :: Maybe SearchMessagesFilter

limit :: Maybe Int

offset :: Maybe Int

from_message_id :: Maybe Int

sender_id :: Maybe MessageSender

query :: Maybe String

chat_id :: Maybe Int