API.Functions.SetChatMemberStatus

Documentation

data SetChatMemberStatus

Changes the status of a chat member, needs appropriate privileges. This function is currently not suitable for transferring chat ownership; use transferChatOwnership instead. Use addChatMember or banChatMember if some additional parameters needs to be passed

chat_id Chat identifier

member_id Member identifier. Chats can be only banned and unbanned in supergroups and channels

status The new status of the member in the chat

Constructors

SetChatMemberStatus

Fields

status :: Maybe ChatMemberStatus

member_id :: Maybe MessageSender

chat_id :: Maybe Int