API.Functions.SetCommands

Documentation

data SetCommands

Sets the list of commands supported by the bot for the given user scope and language; for bots only

scope The scope to which the commands are relevant; pass null to change commands in the default bot command scope

language_code A two-letter ISO 639-1 country code. If empty, the commands will be applied to all users from the given scope, for which language there are no dedicated commands

commands List of the bot's commands

Constructors

SetCommands

Fields

commands :: Maybe [BotCommand]

language_code :: Maybe String

scope :: Maybe BotCommandScope

p :: Show a => String -> Maybe a -> String

cc :: [String] -> String