API.Functions.SetPassword

Documentation

data SetPassword

Changes the password for the current user. If a new recovery email address is specified, then the change will not be applied until the new recovery email address is confirmed

old_password Previous password of the user

new_password New password of the user; may be empty to remove the password

new_hint New password hint; may be empty

set_recovery_email_address Pass true if the recovery email address must be changed

new_recovery_email_address New recovery email address; may be empty

Constructors

SetPassword

Fields

new_recovery_email_address :: Maybe String

set_recovery_email_address :: Maybe Bool

new_hint :: Maybe String

new_password :: Maybe String

old_password :: Maybe String

p :: Show a => String -> Maybe a -> String

cc :: [String] -> String