API.Functions.StartGroupCallRecording

Documentation

data StartGroupCallRecording

Starts recording of an active group call. Requires groupCall.can_be_managed group call flag

group_call_id Group call identifier

title Group call recording title; 0-64 characters

record_video Pass true to record a video file instead of an audio file

use_portrait_orientation Pass true to use portrait orientation for video instead of landscape one

Constructors

StartGroupCallRecording

Fields

use_portrait_orientation :: Maybe Bool

record_video :: Maybe Bool

title :: Maybe String

group_call_id :: Maybe Int