API.Functions.UploadFile

Documentation

data UploadFile

Asynchronously uploads a file to the cloud without sending it in a message. updateFile will be used to notify about upload progress and successful completion of the upload. The file will not have a persistent remote identifier until it will be sent in a message

file File to upload

file_type File type; pass null if unknown

priority Priority of the upload (1-32). The higher the priority, the earlier the file will be uploaded. If the priorities of two files are equal, then the first one for which uploadFile was called will be uploaded first

Constructors

UploadFile

Fields

priority :: Maybe Int

file_type :: Maybe FileType

file :: Maybe InputFile