API.InlineKeyboardButtonType

Documentation

data InlineKeyboardButtonType

Describes the type of an inline keyboard button

Constructors

InlineKeyboardButtonTypeUrl
A button that opens a specified URL

url HTTP or tg:// URL to open

Fields

url :: Maybe String

InlineKeyboardButtonTypeLoginUrl
A button that opens a specified URL and automatically logs in in current user if they allowed to do that

url An HTTP URL to open

id Unique button identifier

forward_text If non-empty, new text of the button in forwarded messages

Fields

forward_text :: Maybe String

_id :: Maybe Int

url :: Maybe String

InlineKeyboardButtonTypeCallback
A button that sends a special callback query to a bot

data Data to be sent to the bot via a callback query

Fields

_data :: Maybe String

InlineKeyboardButtonTypeCallbackGame
A button with a game that sends a special callback query to a bot. This button must be in the first column and row of the keyboard and can be attached only to a message with content of the type messageGame

InlineKeyboardButtonTypeSwitchInline
A button that forces an inline query to the bot to be inserted in the input field

query Inline query to be sent to the bot

in_current_chat True, if the inline query should be sent from the current chat

Fields

in_current_chat :: Maybe Bool

query :: Maybe String

InlineKeyboardButtonTypeBuy
A button to buy something. This button must be in the first column and row of the keyboard and can be attached only to a message with content of the type messageInvoice