API.InputFile

Documentation

data InputFile

Points to a file

Constructors

InputFileId
A file defined by its unique ID

id Unique file identifier

Fields

__id :: Maybe Int

InputFileRemote
A file defined by its remote ID. The remote ID is guaranteed to be usable only if the corresponding file is still accessible to the user and known to TDLib.

For example, if the file is from a message, then the message must be not deleted and accessible to the user. If the file database is disabled, then the corresponding object with the file must be preloaded by the application

id Remote file identifier

Fields

_id :: Maybe String

InputFileLocal
A file defined by a local path

path Local path to the file

Fields

path :: Maybe String

InputFileGenerated
A file generated by the application

original_path Local path to a file from which the file is generated; may be empty if there is no such file

conversion String specifying the conversion applied to the original file; must be persistent across application restarts. Conversions beginning with # are reserved for internal TDLib usage

expected_size Expected size of the generated file, in bytes; 0 if unknown

Fields

expected_size :: Maybe Int

conversion :: Maybe String

original_path :: Maybe String