API.InputInlineQueryResult

Documentation

data InputInlineQueryResult

Represents a single result of an inline query; for bots only

Constructors

InputInlineQueryResultAnimation
Represents a link to an animated GIF or an animated (i.e., without sound) H.264/MPEG-4 AVC video

id Unique identifier of the query result

title Title of the query result

thumbnail_url URL of the result thumbnail (JPEG, GIF, or MPEG4), if it exists

thumbnail_mime_type MIME type of the video thumbnail. If non-empty, must be one of "imagejpeg", "imagegif" and "video/mp4"

video_url The URL of the video file (file size must not exceed 1MB)

video_mime_type MIME type of the video file. Must be one of "imagegif" and "videomp4"

video_duration Duration of the video, in seconds

video_width Width of the video

video_height Height of the video

reply_markup The message reply markup. Must be of type replyMarkupInlineKeyboard or null

input_message_content The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageAnimation, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

Fields

input_message_content :: Maybe InputMessageContent

reply_markup :: Maybe ReplyMarkup

video_height :: Maybe Int

video_width :: Maybe Int

video_duration :: Maybe Int

video_mime_type :: Maybe String

video_url :: Maybe String

thumbnail_mime_type :: Maybe String

thumbnail_url :: Maybe String

title :: Maybe String

_id :: Maybe String

InputInlineQueryResultArticle
Represents a link to an article or web page

id Unique identifier of the query result

url URL of the result, if it exists

hide_url True, if the URL must be not shown

title Title of the result

param_description A short description of the result

thumbnail_url URL of the result thumbnail, if it exists

thumbnail_width Thumbnail width, if known

thumbnail_height Thumbnail height, if known

reply_markup The message reply markup. Must be of type replyMarkupInlineKeyboard or null

input_message_content The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

Fields

input_message_content :: Maybe InputMessageContent

reply_markup :: Maybe ReplyMarkup

thumbnail_height :: Maybe Int

thumbnail_width :: Maybe Int

thumbnail_url :: Maybe String

description :: Maybe String

title :: Maybe String

hide_url :: Maybe Bool

url :: Maybe String

_id :: Maybe String

InputInlineQueryResultAudio
Represents a link to an MP3 audio file

id Unique identifier of the query result

title Title of the audio file

performer Performer of the audio file

audio_url The URL of the audio file

audio_duration Audio file duration, in seconds

reply_markup The message reply markup. Must be of type replyMarkupInlineKeyboard or null

input_message_content The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageAudio, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

Fields

input_message_content :: Maybe InputMessageContent

reply_markup :: Maybe ReplyMarkup

audio_duration :: Maybe Int

audio_url :: Maybe String

performer :: Maybe String

title :: Maybe String

_id :: Maybe String

InputInlineQueryResultContact
Represents a user contact

id Unique identifier of the query result

contact User contact

thumbnail_url URL of the result thumbnail, if it exists

thumbnail_width Thumbnail width, if known

thumbnail_height Thumbnail height, if known

reply_markup The message reply markup. Must be of type replyMarkupInlineKeyboard or null

input_message_content The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

Fields

input_message_content :: Maybe InputMessageContent

reply_markup :: Maybe ReplyMarkup

thumbnail_height :: Maybe Int

thumbnail_width :: Maybe Int

thumbnail_url :: Maybe String

contact :: Maybe Contact

_id :: Maybe String

InputInlineQueryResultDocument
Represents a link to a file

id Unique identifier of the query result

title Title of the resulting file

param_description Short description of the result, if known

document_url URL of the file

mime_type MIME type of the file content; only "applicationpdf" and "applicationzip" are currently allowed

thumbnail_url The URL of the file thumbnail, if it exists

thumbnail_width Width of the thumbnail

thumbnail_height Height of the thumbnail

reply_markup The message reply markup. Must be of type replyMarkupInlineKeyboard or null

input_message_content The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageDocument, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

Fields

input_message_content :: Maybe InputMessageContent

reply_markup :: Maybe ReplyMarkup

thumbnail_height :: Maybe Int

thumbnail_width :: Maybe Int

thumbnail_url :: Maybe String

mime_type :: Maybe String

document_url :: Maybe String

description :: Maybe String

title :: Maybe String

_id :: Maybe String

InputInlineQueryResultGame
Represents a game

id Unique identifier of the query result

game_short_name Short name of the game

reply_markup Message reply markup. Must be of type replyMarkupInlineKeyboard or null

Fields

reply_markup :: Maybe ReplyMarkup

game_short_name :: Maybe String

_id :: Maybe String

InputInlineQueryResultLocation
Represents a point on the map

id Unique identifier of the query result

location Location result

live_period Amount of time relative to the message sent time until the location can be updated, in seconds

title Title of the result

thumbnail_url URL of the result thumbnail, if it exists

thumbnail_width Thumbnail width, if known

thumbnail_height Thumbnail height, if known

reply_markup The message reply markup. Must be of type replyMarkupInlineKeyboard or null

input_message_content The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

Fields

input_message_content :: Maybe InputMessageContent

reply_markup :: Maybe ReplyMarkup

thumbnail_height :: Maybe Int

thumbnail_width :: Maybe Int

thumbnail_url :: Maybe String

title :: Maybe String

live_period :: Maybe Int

location :: Maybe Location

_id :: Maybe String

InputInlineQueryResultPhoto
Represents link to a JPEG image

id Unique identifier of the query result

title Title of the result, if known

param_description A short description of the result, if known

thumbnail_url URL of the photo thumbnail, if it exists

photo_url The URL of the JPEG photo (photo size must not exceed 5MB)

photo_width Width of the photo

photo_height Height of the photo

reply_markup The message reply markup. Must be of type replyMarkupInlineKeyboard or null

input_message_content The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessagePhoto, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

Fields

input_message_content :: Maybe InputMessageContent

reply_markup :: Maybe ReplyMarkup

photo_height :: Maybe Int

photo_width :: Maybe Int

photo_url :: Maybe String

thumbnail_url :: Maybe String

description :: Maybe String

title :: Maybe String

_id :: Maybe String

InputInlineQueryResultSticker
Represents a link to a WEBP or TGS sticker

id Unique identifier of the query result

thumbnail_url URL of the sticker thumbnail, if it exists

sticker_url The URL of the WEBP or TGS sticker (sticker file size must not exceed 5MB)

sticker_width Width of the sticker

sticker_height Height of the sticker

reply_markup The message reply markup. Must be of type replyMarkupInlineKeyboard or null

input_message_content The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageSticker, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

Fields

input_message_content :: Maybe InputMessageContent

reply_markup :: Maybe ReplyMarkup

sticker_height :: Maybe Int

sticker_width :: Maybe Int

sticker_url :: Maybe String

thumbnail_url :: Maybe String

_id :: Maybe String

InputInlineQueryResultVenue
Represents information about a venue

id Unique identifier of the query result

venue Venue result

thumbnail_url URL of the result thumbnail, if it exists

thumbnail_width Thumbnail width, if known

thumbnail_height Thumbnail height, if known

reply_markup The message reply markup. Must be of type replyMarkupInlineKeyboard or null

input_message_content The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

Fields

input_message_content :: Maybe InputMessageContent

reply_markup :: Maybe ReplyMarkup

thumbnail_height :: Maybe Int

thumbnail_width :: Maybe Int

thumbnail_url :: Maybe String

venue :: Maybe Venue

_id :: Maybe String

InputInlineQueryResultVideo
Represents a link to a page containing an embedded video player or a video file

id Unique identifier of the query result

title Title of the result

param_description A short description of the result, if known

thumbnail_url The URL of the video thumbnail (JPEG), if it exists

video_url URL of the embedded video player or video file

mime_type MIME type of the content of the video URL, only "texthtml" or "videomp4" are currently supported

video_width Width of the video

video_height Height of the video

video_duration Video duration, in seconds

reply_markup The message reply markup. Must be of type replyMarkupInlineKeyboard or null

input_message_content The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageVideo, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

Fields

input_message_content :: Maybe InputMessageContent

reply_markup :: Maybe ReplyMarkup

video_duration :: Maybe Int

video_height :: Maybe Int

video_width :: Maybe Int

mime_type :: Maybe String

video_url :: Maybe String

thumbnail_url :: Maybe String

description :: Maybe String

title :: Maybe String

_id :: Maybe String

InputInlineQueryResultVoiceNote
Represents a link to an opus-encoded audio file within an OGG container, single channel audio

id Unique identifier of the query result

title Title of the voice note

voice_note_url The URL of the voice note file

voice_note_duration Duration of the voice note, in seconds

reply_markup The message reply markup. Must be of type replyMarkupInlineKeyboard or null

input_message_content The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageVoiceNote, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

Fields

input_message_content :: Maybe InputMessageContent

reply_markup :: Maybe ReplyMarkup

voice_note_duration :: Maybe Int

voice_note_url :: Maybe String

title :: Maybe String

_id :: Maybe String