API.InputPassportElementErrorSource

Documentation

data InputPassportElementErrorSource

Contains the description of an error in a Telegram Passport element; for bots only

Constructors

InputPassportElementErrorSourceUnspecified
The element contains an error in an unspecified place. The error will be considered resolved when new data is added

element_hash Current hash of the entire element

Fields

element_hash :: Maybe String

InputPassportElementErrorSourceDataField
A data field contains an error. The error is considered resolved when the field's value changes

field_name Field name

data_hash Current data hash

Fields

data_hash :: Maybe String

field_name :: Maybe String

InputPassportElementErrorSourceFrontSide
The front side of the document contains an error. The error is considered resolved when the file with the front side of the document changes

file_hash Current hash of the file containing the front side

Fields

file_hash :: Maybe String

InputPassportElementErrorSourceReverseSide
The reverse side of the document contains an error. The error is considered resolved when the file with the reverse side of the document changes

file_hash Current hash of the file containing the reverse side

Fields

file_hash :: Maybe String

InputPassportElementErrorSourceSelfie
The selfie contains an error. The error is considered resolved when the file with the selfie changes

file_hash Current hash of the file containing the selfie

Fields

file_hash :: Maybe String

InputPassportElementErrorSourceTranslationFile
One of the files containing the translation of the document contains an error. The error is considered resolved when the file with the translation changes

file_hash Current hash of the file containing the translation

Fields

file_hash :: Maybe String

InputPassportElementErrorSourceTranslationFiles
The translation of the document contains an error. The error is considered resolved when the list of files changes

file_hashes Current hashes of all files with the translation

Fields

file_hashes :: Maybe [String]

InputPassportElementErrorSourceFile
The file contains an error. The error is considered resolved when the file changes

file_hash Current hash of the file which has the error

Fields

file_hash :: Maybe String

InputPassportElementErrorSourceFiles
The list of attached files contains an error. The error is considered resolved when the file list changes

file_hashes Current hashes of all attached files

Fields

file_hashes :: Maybe [String]