API.InputSticker

Documentation

data InputSticker

A sticker to be added to a sticker set

sticker File with the sticker; must fit in a 512x512 square. For WEBP stickers and masks the file must be in PNG format, which will be converted to WEBP server-side. Otherwise, the file must be local or uploaded within a week. See https://core.telegram.org/animated_stickers#technical-requirements for technical requirements

emojis Emojis corresponding to the sticker

type Sticker type

Constructors

InputSticker

Fields

_type :: Maybe StickerType

emojis :: Maybe String

sticker :: Maybe InputFile

p :: Show a => String -> Maybe a -> String

cc :: [String] -> String