API.Message

Documentation

data Message

Describes a message

id Message identifier, unique for the chat to which the message belongs

sender_user_id Identifier of the user who sent the message; 0 if unknown. Currently, it is unknown for channel posts and for channel posts automatically forwarded to discussion group

chat_id Chat identifier

sending_state Information about the sending state of the message; may be null

scheduling_state Information about the scheduling state of the message; may be null

is_outgoing True, if the message is outgoing

can_be_edited True, if the message can be edited. For live location and poll messages this fields shows whether editMessageLiveLocation or stopPoll can be used with this message by the client

can_be_forwarded True, if the message can be forwarded

can_be_deleted_only_for_self True, if the message can be deleted only for the current user while other users will continue to see it

can_be_deleted_for_all_users True, if the message can be deleted for all users

is_channel_post True, if the message is a channel post. All messages to channels are channel posts, all other messages are not channel posts

contains_unread_mention True, if the message contains an unread mention for the current user

date Point in time (Unix timestamp) when the message was sent

edit_date Point in time (Unix timestamp) when the message was last edited

forward_info Information about the initial message sender; may be null

reply_to_message_id If non-zero, the identifier of the message this message is replying to; can be the identifier of a deleted message

ttl For self-destructing messages, the message's TTL (Time To Live), in seconds; 0 if none. TDLib will send updateDeleteMessages or updateMessageContent once the TTL expires

ttl_expires_in Time left before the message expires, in seconds

via_bot_user_id If non-zero, the user identifier of the bot through which this message was sent

author_signature For channel posts, optional author signature

views Number of times this message was viewed

media_album_id Unique identifier of an album this message belongs to. Only photos and videos can be grouped together in albums

restriction_reason If non-empty, contains a human-readable description of the reason why access to this message must be restricted

content Content of the message

reply_markup Reply markup for the message; may be null

Constructors

Message

Fields

reply_markup :: Maybe ReplyMarkup

content :: Maybe MessageContent

restriction_reason :: Maybe String

media_album_id :: Maybe Int

views :: Maybe Int

author_signature :: Maybe String

via_bot_user_id :: Maybe Int

ttl_expires_in :: Maybe Float

ttl :: Maybe Int

reply_to_message_id :: Maybe Int

forward_info :: Maybe MessageForwardInfo

edit_date :: Maybe Int

date :: Maybe Int

contains_unread_mention :: Maybe Bool

is_channel_post :: Maybe Bool

can_be_deleted_for_all_users :: Maybe Bool

can_be_deleted_only_for_self :: Maybe Bool

can_be_forwarded :: Maybe Bool

can_be_edited :: Maybe Bool

is_outgoing :: Maybe Bool

scheduling_state :: Maybe MessageSchedulingState

sending_state :: Maybe MessageSendingState

chat_id :: Maybe Int

sender_user_id :: Maybe Int

_id :: Maybe Int