API.MessageContent

Documentation

data MessageContent

Contains the content of a message

Constructors

MessageText
A text message

text Text of the message

web_page A preview of the web page that's mentioned in the text; may be null

Fields

web_page :: Maybe WebPage

_text :: Maybe FormattedText

MessageAnimation
An animation message (GIF-style).

animation The animation description

caption Animation caption

is_secret True, if the animation thumbnail must be blurred and the animation must be shown only while tapped

Fields

is_secret :: Maybe Bool

caption :: Maybe FormattedText

animation :: Maybe Animation

MessageAudio
An audio message

audio The audio description

caption Audio caption

Fields

caption :: Maybe FormattedText

audio :: Maybe Audio

MessageDocument
A document message (general file)

document The document description

caption Document caption

Fields

caption :: Maybe FormattedText

document :: Maybe Document

MessagePhoto
A photo message

photo The photo description

caption Photo caption

is_secret True, if the photo must be blurred and must be shown only while tapped

Fields

is_secret :: Maybe Bool

caption :: Maybe FormattedText

_photo :: Maybe Photo

MessageExpiredPhoto
An expired photo message (self-destructed after TTL has elapsed)

MessageSticker
A sticker message

sticker The sticker description

Fields

sticker :: Maybe Sticker

MessageVideo
A video message

video The video description

caption Video caption

is_secret True, if the video thumbnail must be blurred and the video must be shown only while tapped

Fields

is_secret :: Maybe Bool

caption :: Maybe FormattedText

video :: Maybe Video

MessageExpiredVideo
An expired video message (self-destructed after TTL has elapsed)

MessageVideoNote
A video note message

video_note The video note description

is_viewed True, if at least one of the recipients has viewed the video note

is_secret True, if the video note thumbnail must be blurred and the video note must be shown only while tapped

Fields

is_secret :: Maybe Bool

is_viewed :: Maybe Bool

video_note :: Maybe VideoNote

MessageVoiceNote
A voice note message

voice_note The voice note description

caption Voice note caption

is_listened True, if at least one of the recipients has listened to the voice note

Fields

is_listened :: Maybe Bool

caption :: Maybe FormattedText

voice_note :: Maybe VoiceNote

MessageLocation
A message with a location

location The location description

live_period Time relative to the message send date, for which the location can be updated, in seconds

expires_in Left time for which the location can be updated, in seconds. updateMessageContent is not sent when this field changes

heading For live locations, a direction in which the location moves, in degrees; 1-360. If 0 the direction is unknown

proximity_alert_radius For live locations, a maximum distance to another chat member for proximity alerts, in meters (0-100000). 0 if the notification is disabled. Available only for the message sender

Fields

proximity_alert_radius :: Maybe Int

heading :: Maybe Int

expires_in :: Maybe Int

live_period :: Maybe Int

location :: Maybe Location

MessageVenue
A message with information about a venue

venue The venue description

Fields

venue :: Maybe Venue

MessageContact
A message with a user contact

contact The contact description

Fields

contact :: Maybe Contact

MessageDice
A dice message. The dice value is randomly generated by the server

initial_state The animated stickers with the initial dice animation; may be null if unknown. updateMessageContent will be sent when the sticker became known

final_state The animated stickers with the final dice animation; may be null if unknown. updateMessageContent will be sent when the sticker became known

emoji Emoji on which the dice throw animation is based

value The dice value. If the value is 0, the dice don't have final state yet

success_animation_frame_number Number of frame after which a success animation like a shower of confetti needs to be shown on updateMessageSendSucceeded

Fields

success_animation_frame_number :: Maybe Int

value :: Maybe Int

emoji :: Maybe String

final_state :: Maybe DiceStickers

initial_state :: Maybe DiceStickers

MessageGame
A message with a game

game The game description

Fields

game :: Maybe Game

MessagePoll
A message with a poll

poll The poll description

Fields

poll :: Maybe Poll

MessageInvoice
A message with an invoice from a bot

title Product title

param_description Product description

photo Product photo; may be null

currency Currency for the product price

total_amount Product total price in the smallest units of the currency

start_parameter Unique invoice bot start_parameter. To share an invoice use the URL https://t.me/{bot_username}?start={start_parameter}

is_test True, if the invoice is a test invoice

need_shipping_address True, if the shipping address should be specified

receipt_message_id The identifier of the message with the receipt, after the product has been purchased

Fields

receipt_message_id :: Maybe Int

need_shipping_address :: Maybe Bool

is_test :: Maybe Bool

start_parameter :: Maybe String

total_amount :: Maybe Int

currency :: Maybe String

_photo :: Maybe Photo

description :: Maybe String

title :: Maybe String

MessageCall
A message with information about an ended call

is_video True, if the call was a video call

discard_reason Reason why the call was discarded

duration Call duration, in seconds

Fields

duration :: Maybe Int

discard_reason :: Maybe CallDiscardReason

is_video :: Maybe Bool

MessageVoiceChatScheduled
A new voice chat was scheduled

group_call_id Identifier of the voice chat. The voice chat can be received through the method getGroupCall

start_date Point in time (Unix timestamp) when the group call is supposed to be started by an administrator

Fields

start_date :: Maybe Int

group_call_id :: Maybe Int

MessageVoiceChatStarted
A newly created voice chat

group_call_id Identifier of the voice chat. The voice chat can be received through the method getGroupCall

Fields

group_call_id :: Maybe Int

MessageVoiceChatEnded
A message with information about an ended voice chat

duration Call duration, in seconds

Fields

duration :: Maybe Int

MessageInviteVoiceChatParticipants
A message with information about an invite to a voice chat

group_call_id Identifier of the voice chat. The voice chat can be received through the method getGroupCall

user_ids Invited user identifiers

Fields

user_ids :: Maybe [Int]

group_call_id :: Maybe Int

MessageBasicGroupChatCreate
A newly created basic group

title Title of the basic group

member_user_ids User identifiers of members in the basic group

Fields

member_user_ids :: Maybe [Int]

title :: Maybe String

MessageSupergroupChatCreate
A newly created supergroup or channel

title Title of the supergroup or channel

Fields

title :: Maybe String

MessageChatChangeTitle
An updated chat title

title New chat title

Fields

title :: Maybe String

MessageChatChangePhoto
An updated chat photo

photo New chat photo

Fields

photo :: Maybe ChatPhoto

MessageChatDeletePhoto
A deleted chat photo

MessageChatAddMembers
New chat members were added

member_user_ids User identifiers of the new members

Fields

member_user_ids :: Maybe [Int]

MessageChatJoinByLink
A new member joined the chat by invite link

MessageChatDeleteMember
A chat member was deleted

user_id User identifier of the deleted chat member

Fields

user_id :: Maybe Int

MessageChatUpgradeTo
A basic group was upgraded to a supergroup and was deactivated as the result

supergroup_id Identifier of the supergroup to which the basic group was upgraded

Fields

supergroup_id :: Maybe Int

MessageChatUpgradeFrom
A supergroup has been created from a basic group

title Title of the newly created supergroup

basic_group_id The identifier of the original basic group

Fields

basic_group_id :: Maybe Int

title :: Maybe String

MessagePinMessage
A message has been pinned

message_id Identifier of the pinned message, can be an identifier of a deleted message or 0

Fields

message_id :: Maybe Int

MessageScreenshotTaken
A screenshot of a message in the chat has been taken

MessageChatSetTheme
A theme in the chat has been changed

theme_name If non-empty, name of a new theme, set for the chat. Otherwise chat theme was reset to the default one

Fields

theme_name :: Maybe String

MessageChatSetTtl
The TTL (Time To Live) setting for messages in the chat has been changed

ttl New message TTL setting

Fields

ttl :: Maybe Int

MessageCustomServiceAction
A non-standard action has happened in the chat

text Message text to be shown in the chat

Fields

text :: Maybe String

MessageGameScore
A new high score was achieved in a game

game_message_id Identifier of the message with the game, can be an identifier of a deleted message

game_id Identifier of the game; may be different from the games presented in the message with the game

score New score

Fields

score :: Maybe Int

game_id :: Maybe Int

game_message_id :: Maybe Int

MessagePaymentSuccessful
A payment has been completed

invoice_chat_id Identifier of the chat, containing the corresponding invoice message; 0 if unknown

invoice_message_id Identifier of the message with the corresponding invoice; can be an identifier of a deleted message

currency Currency for the price of the product

total_amount Total price for the product, in the smallest units of the currency

Fields

total_amount :: Maybe Int

currency :: Maybe String

invoice_message_id :: Maybe Int

invoice_chat_id :: Maybe Int

MessagePaymentSuccessfulBot
A payment has been completed; for bots only

currency Currency for price of the product

total_amount Total price for the product, in the smallest units of the currency

invoice_payload Invoice payload

shipping_option_id Identifier of the shipping option chosen by the user; may be empty if not applicable

order_info Information about the order; may be null

telegram_payment_charge_id Telegram payment identifier

provider_payment_charge_id Provider payment identifier

Fields

provider_payment_charge_id :: Maybe String

telegram_payment_charge_id :: Maybe String

order_info :: Maybe OrderInfo

shipping_option_id :: Maybe String

invoice_payload :: Maybe String

total_amount :: Maybe Int

currency :: Maybe String

MessageContactRegistered
A contact has registered with Telegram

MessageWebsiteConnected
The current user has connected a website by logging in using Telegram Login Widget on it

domain_name Domain name of the connected website

Fields

domain_name :: Maybe String

MessagePassportDataSent
Telegram Passport data has been sent

types List of Telegram Passport element types sent

Fields

types :: Maybe [PassportElementType]

MessagePassportDataReceived
Telegram Passport data has been received; for bots only

elements List of received Telegram Passport elements

credentials Encrypted data credentials

Fields

credentials :: Maybe EncryptedCredentials

elements :: Maybe [EncryptedPassportElement]

MessageProximityAlertTriggered
A user in the chat came within proximity alert range

traveler The user or chat, which triggered the proximity alert

watcher The user or chat, which subscribed for the proximity alert

distance The distance between the users

Fields

distance :: Maybe Int

watcher :: Maybe MessageSender

traveler :: Maybe MessageSender

MessageUnsupported
Message content that is not supported in the current TDLib version