API.MessageSendingState

Documentation

data MessageSendingState

Contains information about the sending state of the message

Constructors

MessageSendingStatePending
The message is being sent now, but has not yet been delivered to the server

MessageSendingStateFailed
The message failed to be sent

error_code An error code; 0 if unknown

error_message Error message

can_retry True, if the message can be re-sent

need_another_sender True, if the message can be re-sent only on behalf of a different sender

retry_after Time left before the message can be re-sent, in seconds. No update is sent when this field changes

Fields

retry_after :: Maybe Float

need_another_sender :: Maybe Bool

can_retry :: Maybe Bool

error_message :: Maybe String

error_code :: Maybe Int