API.PageBlock

Documentation

data PageBlock

Describes a block of an instant view web page

Constructors

PageBlockTitle
The title of a page

title Title

Fields

_title :: Maybe RichText

PageBlockSubtitle
The subtitle of a page

subtitle Subtitle

Fields

subtitle :: Maybe RichText

PageBlockAuthorDate
The author and publishing date of a page

author Author

publish_date Point in time (Unix timestamp) when the article was published; 0 if unknown

Fields

publish_date :: Maybe Int

_author :: Maybe RichText

PageBlockHeader
A header

header Header

Fields

header :: Maybe RichText

PageBlockSubheader
A subheader

subheader Subheader

Fields

subheader :: Maybe RichText

PageBlockKicker
A kicker

kicker Kicker

Fields

kicker :: Maybe RichText

PageBlockParagraph
A text paragraph

text Paragraph text

Fields

text :: Maybe RichText

PageBlockPreformatted
A preformatted text paragraph

text Paragraph text

language Programming language for which the text needs to be formatted

Fields

language :: Maybe String

text :: Maybe RichText

PageBlockFooter
The footer of a page

footer Footer

Fields

footer :: Maybe RichText

PageBlockDivider
An empty block separating a page

PageBlockAnchor
An invisible anchor on a page, which can be used in a URL to open the page from the specified anchor

name Name of the anchor

Fields

name :: Maybe String

PageBlockList
A list of data blocks

items The items of the list

Fields

items :: Maybe [PageBlockListItem]

PageBlockBlockQuote
A block quote

text Quote text

credit Quote credit

Fields

credit :: Maybe RichText

text :: Maybe RichText

PageBlockPullQuote
A pull quote

text Quote text

credit Quote credit

Fields

credit :: Maybe RichText

text :: Maybe RichText

PageBlockAnimation
An animation

animation Animation file; may be null

caption Animation caption

need_autoplay True, if the animation must be played automatically

Fields

need_autoplay :: Maybe Bool

caption :: Maybe PageBlockCaption

animation :: Maybe Animation

PageBlockAudio
An audio file

audio Audio file; may be null

caption Audio file caption

Fields

caption :: Maybe PageBlockCaption

audio :: Maybe Audio

PageBlockPhoto
A photo

photo Photo file; may be null

caption Photo caption

url URL that needs to be opened when the photo is clicked

Fields

url :: Maybe String

caption :: Maybe PageBlockCaption

_photo :: Maybe Photo

PageBlockVideo
A video

video Video file; may be null

caption Video caption

need_autoplay True, if the video must be played automatically

is_looped True, if the video must be looped

Fields

is_looped :: Maybe Bool

need_autoplay :: Maybe Bool

caption :: Maybe PageBlockCaption

video :: Maybe Video

PageBlockVoiceNote
A voice note

voice_note Voice note; may be null

caption Voice note caption

Fields

caption :: Maybe PageBlockCaption

voice_note :: Maybe VoiceNote

PageBlockCover
A page cover

cover Cover

Fields

cover :: Maybe PageBlock

PageBlockEmbedded
An embedded web page

url Web page URL, if available

html HTML-markup of the embedded page

poster_photo Poster photo, if available; may be null

width Block width; 0 if unknown

height Block height; 0 if unknown

caption Block caption

is_full_width True, if the block must be full width

allow_scrolling True, if scrolling needs to be allowed

Fields

allow_scrolling :: Maybe Bool

is_full_width :: Maybe Bool

caption :: Maybe PageBlockCaption

height :: Maybe Int

width :: Maybe Int

poster_photo :: Maybe Photo

html :: Maybe String

url :: Maybe String

PageBlockEmbeddedPost
An embedded post

url Web page URL

author Post author

author_photo Post author photo; may be null

date Point in time (Unix timestamp) when the post was created; 0 if unknown

page_blocks Post content

caption Post caption

Fields

caption :: Maybe PageBlockCaption

page_blocks :: Maybe [PageBlock]

date :: Maybe Int

author_photo :: Maybe Photo

author :: Maybe String

url :: Maybe String

PageBlockCollage
A collage

page_blocks Collage item contents

caption Block caption

Fields

caption :: Maybe PageBlockCaption

page_blocks :: Maybe [PageBlock]

PageBlockSlideshow
A slideshow

page_blocks Slideshow item contents

caption Block caption

Fields

caption :: Maybe PageBlockCaption

page_blocks :: Maybe [PageBlock]

PageBlockChatLink
A link to a chat

title Chat title

photo Chat photo; may be null

username Chat username, by which all other information about the chat can be resolved

Fields

username :: Maybe String

photo :: Maybe ChatPhotoInfo

title :: Maybe String

PageBlockTable
A table

caption Table caption

cells Table cells

is_bordered True, if the table is bordered

is_striped True, if the table is striped

Fields

is_striped :: Maybe Bool

is_bordered :: Maybe Bool

cells :: Maybe [PageBlockTableCell]

_caption :: Maybe RichText

PageBlockDetails
A collapsible block

header Always visible heading for the block

page_blocks Block contents

is_open True, if the block is open by default

Fields

is_open :: Maybe Bool

page_blocks :: Maybe [PageBlock]

header :: Maybe RichText

PageBlockRelatedArticles
Related articles

header Block header

articles List of related articles

Fields

articles :: Maybe [PageBlockRelatedArticle]

header :: Maybe RichText

PageBlockMap
A map

location Location of the map center

zoom Map zoom level

width Map width

height Map height

caption Block caption

Fields

caption :: Maybe PageBlockCaption

height :: Maybe Int

width :: Maybe Int

zoom :: Maybe Int

location :: Maybe Location

p :: Show a => String -> Maybe a -> String

cc :: [String] -> String