API.PassportSuitableElement

Documentation

data PassportSuitableElement

Contains information about a Telegram Passport element that was requested by a service

type Type of the element

is_selfie_required True, if a selfie is required with the identity document

is_translation_required True, if a certified English translation is required with the document

is_native_name_required True, if personal details must include the user's name in the language of their country of residence

Constructors

PassportSuitableElement

Fields

is_native_name_required :: Maybe Bool

is_translation_required :: Maybe Bool

is_selfie_required :: Maybe Bool

_type :: Maybe PassportElementType