API.PhoneNumberAuthenticationSettings

Documentation

data PhoneNumberAuthenticationSettings

Contains settings for the authentication of the user's phone number

allow_flash_call Pass true if the authentication code may be sent via flash call to the specified phone number

is_current_phone_number Pass true if the authenticated phone number is used on the current device

allow_sms_retriever_api For official applications only. True, if the application can use Android SMS Retriever API (requires Google Play Services >= 10.2) to automatically receive the authentication code from the SMS. See https://developers.google.com/identity/sms-retriever/ for more details

Constructors

PhoneNumberAuthenticationSettings

Fields

allow_sms_retriever_api :: Maybe Bool

is_current_phone_number :: Maybe Bool

allow_flash_call :: Maybe Bool