API.Poll

Documentation

data Poll

Describes a poll

id Unique poll identifier

question Poll question, 1-255 characters

options List of poll answer options

total_voter_count Total number of voters, participating in the poll

recent_voter_user_ids User identifiers of recent voters, if the poll is non-anonymous

is_anonymous True, if the poll is anonymous

type Type of the poll

open_period Amount of time the poll will be active after creation, in seconds

close_date Point in time (Unix timestamp) when the poll will be automatically closed

is_closed True, if the poll is closed

Constructors

Poll

Fields

is_closed :: Maybe Bool

close_date :: Maybe Int

open_period :: Maybe Int

_type :: Maybe PollType

is_anonymous :: Maybe Bool

recent_voter_user_ids :: Maybe [Int]

total_voter_count :: Maybe Int

options :: Maybe [PollOption]

question :: Maybe String

_id :: Maybe Int