API.StickerSetInfo

Documentation

data StickerSetInfo

Represents short information about a sticker set

id Identifier of the sticker set

title Title of the sticker set

name Name of the sticker set

thumbnail Sticker set thumbnail in WEBP or TGS format with width and height 100; may be null

thumbnail_outline Sticker set thumbnail's outline represented as a list of closed vector paths; may be empty. The coordinate system origin is in the upper-left corner

is_installed True, if the sticker set has been installed by the current user

is_archived True, if the sticker set has been archived. A sticker set can't be installed and archived simultaneously

is_official True, if the sticker set is official

is_animated True, is the stickers in the set are animated

is_masks True, if the stickers in the set are masks

is_viewed True for already viewed trending sticker sets

size Total number of stickers in the set

covers Contains up to the first 5 stickers from the set, depending on the context. If the application needs more stickers the full set should be requested

Constructors

StickerSetInfo

Fields

covers :: Maybe [Sticker]

size :: Maybe Int

is_viewed :: Maybe Bool

is_masks :: Maybe Bool

is_animated :: Maybe Bool

is_official :: Maybe Bool

is_archived :: Maybe Bool

is_installed :: Maybe Bool

thumbnail_outline :: Maybe [ClosedVectorPath]

thumbnail :: Maybe Thumbnail

name :: Maybe String

title :: Maybe String

_id :: Maybe Int