API.Supergroup

Documentation

data Supergroup

Represents a supergroup or channel with zero or more members (subscribers in the case of channels). From the point of view of the system, a channel is a special kind of a supergroup: only administrators can post and see the list of members, and posts from all administrators use the name and photo of the channel instead of individual names and profile photos. Unlike supergroups, channels can have an unlimited number of subscribers

id Supergroup or channel identifier

username Username of the supergroup or channel; empty for private supergroups or channels

date Point in time (Unix timestamp) when the current user joined, or the point in time when the supergroup or channel was created, in case the user is not a member

status Status of the current user in the supergroup or channel; custom title will be always empty

member_count Number of members in the supergroup or channel; 0 if unknown. Currently it is guaranteed to be known only if the supergroup or channel was found through SearchPublicChats

has_linked_chat True, if the channel has a discussion group, or the supergroup is the designated discussion group for a channel

has_location True, if the supergroup is connected to a location, i.e. the supergroup is a location-based supergroup

sign_messages True, if messages sent to the channel should contain information about the sender. This field is only applicable to channels

is_slow_mode_enabled True, if the slow mode is enabled in the supergroup

is_channel True, if the supergroup is a channel

is_verified True, if the supergroup or channel is verified

restriction_reason If non-empty, contains a human-readable description of the reason why access to this supergroup or channel must be restricted

is_scam True, if many users reported this supergroup as a scam

Constructors

Supergroup

Fields

is_scam :: Maybe Bool

restriction_reason :: Maybe String

is_verified :: Maybe Bool

is_channel :: Maybe Bool

is_slow_mode_enabled :: Maybe Bool

sign_messages :: Maybe Bool

has_location :: Maybe Bool

has_linked_chat :: Maybe Bool

member_count :: Maybe Int

status :: Maybe ChatMemberStatus

date :: Maybe Int

username :: Maybe String

_id :: Maybe Int