API.Update

Documentation

data Update

Contains notifications about data changes

Constructors

UpdateAuthorizationState
The user authorization state has changed

authorization_state New authorization state

Fields

authorization_state :: Maybe AuthorizationState

UpdateNewMessage
A new message was received; can also be an outgoing message

message The new message

Fields

message :: Maybe Message

UpdateMessageSendAcknowledged
A request to send a message has reached the Telegram server. This doesn't mean that the message will be sent successfully or even that the send message request will be processed. This update will be sent only if the option "use_quick_ack" is set to true. This update may be sent multiple times for the same message

chat_id The chat identifier of the sent message

message_id A temporary message identifier

Fields

message_id :: Maybe Int

chat_id :: Maybe Int

UpdateMessageSendSucceeded
A message has been successfully sent

message Information about the sent message. Usually only the message identifier, date, and content are changed, but almost all other fields can also change

old_message_id The previous temporary message identifier

Fields

old_message_id :: Maybe Int

message :: Maybe Message

UpdateMessageSendFailed
A message failed to send. Be aware that some messages being sent can be irrecoverably deleted, in which case updateDeleteMessages will be received instead of this update

message Contains information about the message which failed to send

old_message_id The previous temporary message identifier

error_code An error code

error_message Error message

Fields

error_message :: Maybe String

error_code :: Maybe Int

old_message_id :: Maybe Int

message :: Maybe Message

UpdateMessageContent
The message content has changed

chat_id Chat identifier

message_id Message identifier

new_content New message content

Fields

new_content :: Maybe MessageContent

message_id :: Maybe Int

chat_id :: Maybe Int

UpdateMessageEdited
A message was edited. Changes in the message content will come in a separate updateMessageContent

chat_id Chat identifier

message_id Message identifier

edit_date Point in time (Unix timestamp) when the message was edited

reply_markup New message reply markup; may be null

Fields

reply_markup :: Maybe ReplyMarkup

edit_date :: Maybe Int

message_id :: Maybe Int

chat_id :: Maybe Int

UpdateMessageIsPinned
The message pinned state was changed

chat_id Chat identifier

message_id The message identifier

is_pinned True, if the message is pinned

Fields

is_pinned :: Maybe Bool

message_id :: Maybe Int

chat_id :: Maybe Int

UpdateMessageInteractionInfo
The information about interactions with a message has changed

chat_id Chat identifier

message_id Message identifier

interaction_info New information about interactions with the message; may be null

Fields

interaction_info :: Maybe MessageInteractionInfo

message_id :: Maybe Int

chat_id :: Maybe Int

UpdateMessageContentOpened
The message content was opened. Updates voice note messages to "listened", video note messages to "viewed" and starts the TTL timer for self-destructing messages

chat_id Chat identifier

message_id Message identifier

Fields

message_id :: Maybe Int

chat_id :: Maybe Int

UpdateMessageMentionRead
A message with an unread mention was read

chat_id Chat identifier

message_id Message identifier

unread_mention_count The new number of unread mention messages left in the chat

Fields

unread_mention_count :: Maybe Int

message_id :: Maybe Int

chat_id :: Maybe Int

UpdateMessageLiveLocationViewed
A message with a live location was viewed. When the update is received, the application is supposed to update the live location

chat_id Identifier of the chat with the live location message

message_id Identifier of the message with live location

Fields

message_id :: Maybe Int

chat_id :: Maybe Int

UpdateNewChat
A new chat has been loaded/created. This update is guaranteed to come before the chat identifier is returned to the application. The chat field changes will be reported through separate updates

chat The chat

Fields

chat :: Maybe Chat

UpdateChatTitle
The title of a chat was changed

chat_id Chat identifier

title The new chat title

Fields

title :: Maybe String

chat_id :: Maybe Int

UpdateChatPhoto
A chat photo was changed

chat_id Chat identifier

photo The new chat photo; may be null

Fields

photo :: Maybe ChatPhotoInfo

chat_id :: Maybe Int

UpdateChatPermissions
Chat permissions was changed

chat_id Chat identifier

permissions The new chat permissions

Fields

permissions :: Maybe ChatPermissions

chat_id :: Maybe Int

UpdateChatLastMessage
The last message of a chat was changed. If last_message is null, then the last message in the chat became unknown. Some new unknown messages might be added to the chat in this case

chat_id Chat identifier

last_message The new last message in the chat; may be null

positions The new chat positions in the chat lists

Fields

positions :: Maybe [ChatPosition]

last_message :: Maybe Message

chat_id :: Maybe Int

UpdateChatPosition
The position of a chat in a chat list has changed. Instead of this update updateChatLastMessage or updateChatDraftMessage might be sent

chat_id Chat identifier

position New chat position. If new order is 0, then the chat needs to be removed from the list

Fields

position :: Maybe ChatPosition

chat_id :: Maybe Int

UpdateChatIsMarkedAsUnread
A chat was marked as unread or was read

chat_id Chat identifier

is_marked_as_unread New value of is_marked_as_unread

Fields

is_marked_as_unread :: Maybe Bool

chat_id :: Maybe Int

UpdateChatIsBlocked
A chat was blocked or unblocked

chat_id Chat identifier

is_blocked New value of is_blocked

Fields

is_blocked :: Maybe Bool

chat_id :: Maybe Int

UpdateChatHasScheduledMessages
A chat's has_scheduled_messages field has changed

chat_id Chat identifier

has_scheduled_messages New value of has_scheduled_messages

Fields

has_scheduled_messages :: Maybe Bool

chat_id :: Maybe Int

UpdateChatVoiceChat
A chat voice chat state has changed

chat_id Chat identifier

voice_chat New value of voice_chat

Fields

voice_chat :: Maybe VoiceChat

chat_id :: Maybe Int

UpdateChatDefaultDisableNotification
The value of the default disable_notification parameter, used when a message is sent to the chat, was changed

chat_id Chat identifier

default_disable_notification The new default_disable_notification value

Fields

default_disable_notification :: Maybe Bool

chat_id :: Maybe Int

UpdateChatReadInbox
Incoming messages were read or number of unread messages has been changed

chat_id Chat identifier

last_read_inbox_message_id Identifier of the last read incoming message

unread_count The number of unread messages left in the chat

Fields

unread_count :: Maybe Int

last_read_inbox_message_id :: Maybe Int

chat_id :: Maybe Int

UpdateChatReadOutbox
Outgoing messages were read

chat_id Chat identifier

last_read_outbox_message_id Identifier of last read outgoing message

Fields

last_read_outbox_message_id :: Maybe Int

chat_id :: Maybe Int

UpdateChatUnreadMentionCount
The chat unread_mention_count has changed

chat_id Chat identifier

unread_mention_count The number of unread mention messages left in the chat

Fields

unread_mention_count :: Maybe Int

chat_id :: Maybe Int

UpdateChatNotificationSettings
Notification settings for a chat were changed

chat_id Chat identifier

notification_settings The new notification settings

Fields

_notification_settings :: Maybe ChatNotificationSettings

chat_id :: Maybe Int

UpdateScopeNotificationSettings
Notification settings for some type of chats were updated

scope Types of chats for which notification settings were updated

notification_settings The new notification settings

Fields

notification_settings :: Maybe ScopeNotificationSettings

scope :: Maybe NotificationSettingsScope

UpdateChatMessageTtlSetting
The message Time To Live setting for a chat was changed

chat_id Chat identifier

message_ttl_setting New value of message_ttl_setting

Fields

message_ttl_setting :: Maybe Int

chat_id :: Maybe Int

UpdateChatActionBar
The chat action bar was changed

chat_id Chat identifier

action_bar The new value of the action bar; may be null

Fields

action_bar :: Maybe ChatActionBar

chat_id :: Maybe Int

UpdateChatTheme
The chat theme was changed

chat_id Chat identifier

theme_name The new name of the chat theme; may be empty if theme was reset to default

Fields

theme_name :: Maybe String

chat_id :: Maybe Int

UpdateChatReplyMarkup
The default chat reply markup was changed. Can occur because new messages with reply markup were received or because an old reply markup was hidden by the user

chat_id Chat identifier

reply_markup_message_id Identifier of the message from which reply markup needs to be used; 0 if there is no default custom reply markup in the chat

Fields

reply_markup_message_id :: Maybe Int

chat_id :: Maybe Int

UpdateChatDraftMessage
A chat draft has changed. Be aware that the update may come in the currently opened chat but with old content of the draft. If the user has changed the content of the draft, this update shouldn't be applied

chat_id Chat identifier

draft_message The new draft message; may be null

positions The new chat positions in the chat lists

Fields

positions :: Maybe [ChatPosition]

draft_message :: Maybe DraftMessage

chat_id :: Maybe Int

UpdateChatFilters
The list of chat filters or a chat filter has changed

chat_filters The new list of chat filters

Fields

chat_filters :: Maybe [ChatFilterInfo]

UpdateChatOnlineMemberCount
The number of online group members has changed. This update with non-zero count is sent only for currently opened chats. There is no guarantee that it will be sent just after the count has changed

chat_id Identifier of the chat

online_member_count New number of online members in the chat, or 0 if unknown

Fields

online_member_count :: Maybe Int

chat_id :: Maybe Int

UpdateNotification
A notification was changed

notification_group_id Unique notification group identifier

notification Changed notification

Fields

notification :: Maybe Notification

notification_group_id :: Maybe Int

UpdateNotificationGroup
A list of active notifications in a notification group has changed

notification_group_id Unique notification group identifier

type New type of the notification group

chat_id Identifier of a chat to which all notifications in the group belong

notification_settings_chat_id Chat identifier, which notification settings must be applied to the added notifications

is_silent True, if the notifications should be shown without sound

total_count Total number of unread notifications in the group, can be bigger than number of active notifications

added_notifications List of added group notifications, sorted by notification ID

removed_notification_ids Identifiers of removed group notifications, sorted by notification ID

Fields

removed_notification_ids :: Maybe [Int]

added_notifications :: Maybe [Notification]

total_count :: Maybe Int

is_silent :: Maybe Bool

notification_settings_chat_id :: Maybe Int

chat_id :: Maybe Int

__type :: Maybe NotificationGroupType

notification_group_id :: Maybe Int

UpdateActiveNotifications
Contains active notifications that was shown on previous application launches. This update is sent only if the message database is used. In that case it comes once before any updateNotification and updateNotificationGroup update

groups Lists of active notification groups

Fields

groups :: Maybe [NotificationGroup]

UpdateHavePendingNotifications
Describes whether there are some pending notification updates. Can be used to prevent application from killing, while there are some pending notifications

have_delayed_notifications True, if there are some delayed notification updates, which will be sent soon

have_unreceived_notifications True, if there can be some yet unreceived notifications, which are being fetched from the server

Fields

have_unreceived_notifications :: Maybe Bool

have_delayed_notifications :: Maybe Bool

UpdateDeleteMessages
Some messages were deleted

chat_id Chat identifier

message_ids Identifiers of the deleted messages

is_permanent True, if the messages are permanently deleted by a user (as opposed to just becoming inaccessible)

from_cache True, if the messages are deleted only from the cache and can possibly be retrieved again in the future

Fields

from_cache :: Maybe Bool

is_permanent :: Maybe Bool

message_ids :: Maybe [Int]

chat_id :: Maybe Int

UpdateUserChatAction
User activity in the chat has changed

chat_id Chat identifier

message_thread_id If not 0, a message thread identifier in which the action was performed

user_id Identifier of a user performing an action

action The action description

Fields

action :: Maybe ChatAction

user_id :: Maybe Int

message_thread_id :: Maybe Int

chat_id :: Maybe Int

UpdateUserStatus
The user went online or offline

user_id User identifier

status New status of the user

Fields

status :: Maybe UserStatus

user_id :: Maybe Int

UpdateUser
Some data of a user has changed. This update is guaranteed to come before the user identifier is returned to the application

user New data about the user

Fields

user :: Maybe User

UpdateBasicGroup
Some data of a basic group has changed. This update is guaranteed to come before the basic group identifier is returned to the application

basic_group New data about the group

Fields

basic_group :: Maybe BasicGroup

UpdateSupergroup
Some data of a supergroup or a channel has changed. This update is guaranteed to come before the supergroup identifier is returned to the application

supergroup New data about the supergroup

Fields

supergroup :: Maybe Supergroup

UpdateSecretChat
Some data of a secret chat has changed. This update is guaranteed to come before the secret chat identifier is returned to the application

secret_chat New data about the secret chat

Fields

secret_chat :: Maybe SecretChat

UpdateUserFullInfo
Some data from userFullInfo has been changed

user_id User identifier

user_full_info New full information about the user

Fields

user_full_info :: Maybe UserFullInfo

user_id :: Maybe Int

UpdateBasicGroupFullInfo
Some data from basicGroupFullInfo has been changed

basic_group_id Identifier of a basic group

basic_group_full_info New full information about the group

Fields

basic_group_full_info :: Maybe BasicGroupFullInfo

basic_group_id :: Maybe Int

UpdateSupergroupFullInfo
Some data from supergroupFullInfo has been changed

supergroup_id Identifier of the supergroup or channel

supergroup_full_info New full information about the supergroup

Fields

supergroup_full_info :: Maybe SupergroupFullInfo

supergroup_id :: Maybe Int

UpdateServiceNotification
Service notification from the server. Upon receiving this the application must show a popup with the content of the notification

type Notification type. If type begins with AUTH_KEY_DROP_, then two buttons Cancel and "Log out" should be shown under notification; if user presses the second, all local data should be destroyed using Destroy method

content Notification content

Fields

content :: Maybe MessageContent

_type :: Maybe String

UpdateFile
Information about a file was updated

file New data about the file

Fields

file :: Maybe File

UpdateFileGenerationStart
The file generation process needs to be started by the application

generation_id Unique identifier for the generation process

original_path The path to a file from which a new file is generated; may be empty

destination_path The path to a file that should be created and where the new file should be generated

conversion String specifying the conversion applied to the original file. If conversion is "" than original_path contains an HTTP/HTTPS URL of a file, which should be downloaded by the application

Fields

conversion :: Maybe String

destination_path :: Maybe String

original_path :: Maybe String

generation_id :: Maybe Int

UpdateFileGenerationStop
File generation is no longer needed

generation_id Unique identifier for the generation process

Fields

generation_id :: Maybe Int

UpdateCall
New call was created or information about a call was updated

call New data about a call

Fields

call :: Maybe Call

UpdateGroupCall
Information about a group call was updated

group_call New data about a group call

Fields

group_call :: Maybe GroupCall

UpdateGroupCallParticipant
Information about a group call participant was changed. The updates are sent only after the group call is received through getGroupCall and only if the call is joined or being joined

group_call_id Identifier of group call

participant New data about a participant

Fields

participant :: Maybe GroupCallParticipant

group_call_id :: Maybe Int

UpdateNewCallSignalingData
New call signaling data arrived

call_id The call identifier

data The data

Fields

_data :: Maybe String

call_id :: Maybe Int

UpdateUserPrivacySettingRules
Some privacy setting rules have been changed

setting The privacy setting

rules New privacy rules

Fields

rules :: Maybe UserPrivacySettingRules

setting :: Maybe UserPrivacySetting

UpdateUnreadMessageCount
Number of unread messages in a chat list has changed. This update is sent only if the message database is used

chat_list The chat list with changed number of unread messages

unread_count Total number of unread messages

unread_unmuted_count Total number of unread messages in unmuted chats

Fields

unread_unmuted_count :: Maybe Int

unread_count :: Maybe Int

chat_list :: Maybe ChatList

UpdateUnreadChatCount
Number of unread chats, i.e. with unread messages or marked as unread, has changed. This update is sent only if the message database is used

chat_list The chat list with changed number of unread messages

total_count Approximate total number of chats in the chat list

unread_count Total number of unread chats

unread_unmuted_count Total number of unread unmuted chats

marked_as_unread_count Total number of chats marked as unread

marked_as_unread_unmuted_count Total number of unmuted chats marked as unread

Fields

marked_as_unread_unmuted_count :: Maybe Int

marked_as_unread_count :: Maybe Int

unread_unmuted_count :: Maybe Int

unread_count :: Maybe Int

total_count :: Maybe Int

chat_list :: Maybe ChatList

UpdateOption
An option changed its value

name The option name

value The new option value

Fields

value :: Maybe OptionValue

name :: Maybe String

UpdateStickerSet
A sticker set has changed

sticker_set The sticker set

Fields

sticker_set :: Maybe StickerSet

UpdateInstalledStickerSets
The list of installed sticker sets was updated

is_masks True, if the list of installed mask sticker sets was updated

sticker_set_ids The new list of installed ordinary sticker sets

Fields

sticker_set_ids :: Maybe [Int]

is_masks :: Maybe Bool

UpdateTrendingStickerSets
The list of trending sticker sets was updated or some of them were viewed

sticker_sets The prefix of the list of trending sticker sets with the newest trending sticker sets

Fields

sticker_sets :: Maybe StickerSets

UpdateRecentStickers
The list of recently used stickers was updated

is_attached True, if the list of stickers attached to photo or video files was updated, otherwise the list of sent stickers is updated

sticker_ids The new list of file identifiers of recently used stickers

Fields

sticker_ids :: Maybe [Int]

is_attached :: Maybe Bool

UpdateFavoriteStickers
The list of favorite stickers was updated

sticker_ids The new list of file identifiers of favorite stickers

Fields

sticker_ids :: Maybe [Int]

UpdateSavedAnimations
The list of saved animations was updated

animation_ids The new list of file identifiers of saved animations

Fields

animation_ids :: Maybe [Int]

UpdateSelectedBackground
The selected background has changed

for_dark_theme True, if background for dark theme has changed

background The new selected background; may be null

Fields

background :: Maybe Background

for_dark_theme :: Maybe Bool

UpdateChatThemes
The list of available chat themes has changed

chat_themes The new list of chat themes

Fields

chat_themes :: Maybe [ChatTheme]

UpdateLanguagePackStrings
Some language pack strings have been updated

localization_target Localization target to which the language pack belongs

language_pack_id Identifier of the updated language pack

strings List of changed language pack strings

Fields

strings :: Maybe [LanguagePackString]

language_pack_id :: Maybe String

localization_target :: Maybe String

UpdateConnectionState
The connection state has changed. This update must be used only to show a human-readable description of the connection state

state The new connection state

Fields

state :: Maybe ConnectionState

UpdateTermsOfService
New terms of service must be accepted by the user. If the terms of service are declined, then the deleteAccount method should be called with the reason "Decline ToS update"

terms_of_service_id Identifier of the terms of service

terms_of_service The new terms of service

Fields

terms_of_service :: Maybe TermsOfService

terms_of_service_id :: Maybe String

UpdateUsersNearby
The list of users nearby has changed. The update is guaranteed to be sent only 60 seconds after a successful searchChatsNearby request

users_nearby The new list of users nearby

Fields

users_nearby :: Maybe [ChatNearby]

UpdateDiceEmojis
The list of supported dice emojis has changed

emojis The new list of supported dice emojis

Fields

emojis :: Maybe [String]

UpdateAnimatedEmojiMessageClicked
Some animated emoji message was clicked and a big animated sticker should be played if the message is visible on the screen. chatActionWatchingAnimations with the text of the message needs to be sent if the sticker is played

chat_id Chat identifier

message_id Message identifier

sticker The animated sticker to be played

Fields

sticker :: Maybe Sticker

message_id :: Maybe Int

chat_id :: Maybe Int

UpdateAnimationSearchParameters
The parameters of animation search through GetOption("animation_search_bot_username") bot has changed

provider Name of the animation search provider

emojis The new list of emojis suggested for searching

Fields

emojis :: Maybe [String]

provider :: Maybe String

UpdateSuggestedActions
The list of suggested to the user actions has changed

added_actions Added suggested actions

removed_actions Removed suggested actions

Fields

removed_actions :: Maybe [SuggestedAction]

added_actions :: Maybe [SuggestedAction]

UpdateNewInlineQuery
A new incoming inline query; for bots only

id Unique query identifier

sender_user_id Identifier of the user who sent the query

user_location User location; may be null

chat_type Contains information about the type of the chat, from which the query originated; may be null if unknown

query Text of the query

offset Offset of the first entry to return

Fields

offset :: Maybe String

query :: Maybe String

chat_type :: Maybe ChatType

user_location :: Maybe Location

sender_user_id :: Maybe Int

_id :: Maybe Int

UpdateNewChosenInlineResult
The user has chosen a result of an inline query; for bots only

sender_user_id Identifier of the user who sent the query

user_location User location; may be null

query Text of the query

result_id Identifier of the chosen result

inline_message_id Identifier of the sent inline message, if known

Fields

inline_message_id :: Maybe String

result_id :: Maybe String

query :: Maybe String

user_location :: Maybe Location

sender_user_id :: Maybe Int

UpdateNewCallbackQuery
A new incoming callback query; for bots only

id Unique query identifier

sender_user_id Identifier of the user who sent the query

chat_id Identifier of the chat where the query was sent

message_id Identifier of the message, from which the query originated

chat_instance Identifier that uniquely corresponds to the chat to which the message was sent

payload Query payload

Fields

payload :: Maybe CallbackQueryPayload

chat_instance :: Maybe Int

message_id :: Maybe Int

chat_id :: Maybe Int

sender_user_id :: Maybe Int

_id :: Maybe Int

UpdateNewInlineCallbackQuery
A new incoming callback query from a message sent via a bot; for bots only

id Unique query identifier

sender_user_id Identifier of the user who sent the query

inline_message_id Identifier of the inline message, from which the query originated

chat_instance An identifier uniquely corresponding to the chat a message was sent to

payload Query payload

Fields

payload :: Maybe CallbackQueryPayload

chat_instance :: Maybe Int

inline_message_id :: Maybe String

sender_user_id :: Maybe Int

_id :: Maybe Int

UpdateNewShippingQuery
A new incoming shipping query; for bots only. Only for invoices with flexible price

id Unique query identifier

sender_user_id Identifier of the user who sent the query

invoice_payload Invoice payload

shipping_address User shipping address

Fields

shipping_address :: Maybe Address

invoice_payload :: Maybe String

sender_user_id :: Maybe Int

_id :: Maybe Int

UpdateNewPreCheckoutQuery
A new incoming pre-checkout query; for bots only. Contains full information about a checkout

id Unique query identifier

sender_user_id Identifier of the user who sent the query

currency Currency for the product price

total_amount Total price for the product, in the smallest units of the currency

invoice_payload Invoice payload

shipping_option_id Identifier of a shipping option chosen by the user; may be empty if not applicable

order_info Information about the order; may be null

Fields

order_info :: Maybe OrderInfo

shipping_option_id :: Maybe String

invoice_payload :: Maybe String

total_amount :: Maybe Int

currency :: Maybe String

sender_user_id :: Maybe Int

_id :: Maybe Int

UpdateNewCustomEvent
A new incoming event; for bots only

event A JSON-serialized event

Fields

event :: Maybe String

UpdateNewCustomQuery
A new incoming query; for bots only

id The query identifier

data JSON-serialized query data

timeout Query timeout

Fields

timeout :: Maybe Int

_data :: Maybe String

_id :: Maybe Int

UpdatePoll
A poll was updated; for bots only

poll New data about the poll

Fields

poll :: Maybe Poll

UpdatePollAnswer
A user changed the answer to a poll; for bots only

poll_id Unique poll identifier

user_id The user, who changed the answer to the poll

option_ids 0-based identifiers of answer options, chosen by the user

Fields

option_ids :: Maybe [Int]

user_id :: Maybe Int

poll_id :: Maybe Int

UpdateChatMember
User rights changed in a chat; for bots only

chat_id Chat identifier

actor_user_id Identifier of the user, changing the rights

date Point in time (Unix timestamp) when the user rights was changed

invite_link If user has joined the chat using an invite link, the invite link; may be null

old_chat_member Previous chat member

new_chat_member New chat member

Fields

new_chat_member :: Maybe ChatMember

old_chat_member :: Maybe ChatMember

invite_link :: Maybe ChatInviteLink

date :: Maybe Int

actor_user_id :: Maybe Int

chat_id :: Maybe Int