API.UserFullInfo

Documentation

data UserFullInfo

Contains full information about a user

photo User profile photo; may be null

is_blocked True, if the user is blocked by the current user

can_be_called True, if the user can be called

supports_video_calls True, if a video call can be created with the user

has_private_calls True, if the user can't be called due to their privacy settings

has_private_forwards True, if the user can't be linked in forwarded messages due to their privacy settings

need_phone_number_privacy_exception True, if the current user needs to explicitly allow to share their phone number with the user when the method addContact is used

bio A short user bio

group_in_common_count Number of group chats where both the other user and the current user are a member; 0 for the current user

bot_info For bots, information about the bot; may be null

Constructors

UserFullInfo

Fields

bot_info :: Maybe BotInfo

group_in_common_count :: Maybe Int

bio :: Maybe String

need_phone_number_privacy_exception :: Maybe Bool

has_private_forwards :: Maybe Bool

has_private_calls :: Maybe Bool

supports_video_calls :: Maybe Bool

can_be_called :: Maybe Bool

is_blocked :: Maybe Bool

photo :: Maybe ChatPhoto