API.VideoNote

Documentation

data VideoNote

Describes a video note. The video must be equal in width and height, cropped to a circle, and stored in MPEG4 format

duration Duration of the video, in seconds; as defined by the sender

length Video width and height; as defined by the sender

minithumbnail Video minithumbnail; may be null

thumbnail Video thumbnail in JPEG format; as defined by the sender; may be null

video File containing the video

Constructors

VideoNote

Fields

video :: Maybe File

thumbnail :: Maybe Thumbnail

minithumbnail :: Maybe Minithumbnail

_length :: Maybe Int

duration :: Maybe Int