API.WebPage

Documentation

data WebPage

Describes a web page preview

url Original URL of the link

display_url URL to display

type Type of the web page. Can be: article, photo, audio, video, document, profile, app, or something else

site_name Short name of the site (e.g., Google Docs, App Store)

title Title of the content

param_description Description of the content

photo Image representing the content; may be null

embed_url URL to show in the embedded preview

embed_type MIME type of the embedded preview, (e.g., texthtml or videomp4)

embed_width Width of the embedded preview

embed_height Height of the embedded preview

duration Duration of the content, in seconds

author Author of the content

animation Preview of the content as an animation, if available; may be null

audio Preview of the content as an audio file, if available; may be null

document Preview of the content as a document, if available (currently only available for small PDF files and ZIP archives); may be null

sticker Preview of the content as a sticker for small WEBP files, if available; may be null

video Preview of the content as a video, if available; may be null

video_note Preview of the content as a video note, if available; may be null

voice_note Preview of the content as a voice note, if available; may be null

instant_view_version Version of instant view, available for the web page (currently can be 1 or 2), 0 if none

Constructors

WebPage

Fields

instant_view_version :: Maybe Int

voice_note :: Maybe VoiceNote

video_note :: Maybe VideoNote

video :: Maybe Video

sticker :: Maybe Sticker

document :: Maybe Document

audio :: Maybe Audio

animation :: Maybe Animation

author :: Maybe String

duration :: Maybe Int

embed_height :: Maybe Int

embed_width :: Maybe Int

embed_type :: Maybe String

embed_url :: Maybe String

photo :: Maybe Photo

description :: Maybe FormattedText

title :: Maybe String

site_name :: Maybe String

_type :: Maybe String

display_url :: Maybe String

url :: Maybe String