TD.Data.AccentColor

Documentation

data AccentColor

Constructors

AccentColor
Contains information about supported accent color for user/chat name, background of empty chat photo, replies to messages and link previews

Fields

_id :: Maybe Int

Accent color identifier

built_in_accent_color_id :: Maybe Int

Identifier of a built-in color to use in places, where only one color is needed; 0-6

light_theme_colors :: Maybe [Int]

The list of 1-3 colors in RGB format, describing the accent color, as expected to be shown in light themes

dark_theme_colors :: Maybe [Int]

The list of 1-3 colors in RGB format, describing the accent color, as expected to be shown in dark themes

min_channel_chat_boost_level :: Maybe Int

The minimum chat boost level required to use the color in a channel chat