TD.Data.AlternativeVideo

Documentation

data AlternativeVideo

Constructors

AlternativeVideo
Describes an alternative re-encoded quality of a video file

Fields

_id :: Maybe Int

Unique identifier of the alternative video, which is used in the HLS file

width :: Maybe Int

Video width

height :: Maybe Int

Video height

codec :: Maybe Text

Codec used for video file encoding, for example, "h264", "h265", or "av1"

hls_file :: Maybe File

HLS file describing the video

video :: Maybe File

File containing the video