TD.Data.AnimatedEmoji

Description

Documentation

data AnimatedEmoji

Constructors

AnimatedEmoji
Describes an animated or custom representation of an emoji

Fields

sound :: Maybe File

File containing the sound to be played when the sticker is clicked; may be null. The sound is encoded with the Opus codec, and stored inside an OGG container

fitzpatrick_type :: Maybe Int

Emoji modifier fitzpatrick type; 0-6; 0 if none

sticker_height :: Maybe Int

Expected height of the sticker, which can be used if the sticker is null

sticker_width :: Maybe Int

Expected width of the sticker, which can be used if the sticker is null

sticker :: Maybe Sticker

Sticker for the emoji; may be null if yet unknown for a custom emoji. If the sticker is a custom emoji, it can have arbitrary format different from stickerFormatTgs