TD.Data.ArchiveChatListSettings

Description

Documentation

data ArchiveChatListSettings

Constructors

ArchiveChatListSettings
Contains settings for automatic moving of chats to and from the Archive chat lists

Fields

keep_chats_from_folders_archived :: Maybe Bool

True, if unmuted chats, that are always included or pinned in a folder, will be kept in the Archive chat list when they get a new message. Ignored if keep_unmuted_chats_archived == true

keep_unmuted_chats_archived :: Maybe Bool

True, if unmuted chats will be kept in the Archive chat list when they get a new message

archive_and_mute_new_chats_from_unknown_users :: Maybe Bool

True, if new chats from non-contacts will be automatically archived and muted. Can be set to true only if the option "can_archive_and_mute_new_chats_from_unknown_users" is true