TD.Data.AttachmentMenuBot

Documentation

data AttachmentMenuBot

Constructors

AttachmentMenuBot
Represents a bot, which can be added to attachment or side menu

Fields

bot_user_id :: Maybe Int

User identifier of the bot

supports_self_chat :: Maybe Bool

True, if the bot supports opening from attachment menu in the chat with the bot

supports_user_chats :: Maybe Bool

True, if the bot supports opening from attachment menu in private chats with ordinary users

supports_bot_chats :: Maybe Bool

True, if the bot supports opening from attachment menu in private chats with other bots

supports_group_chats :: Maybe Bool

True, if the bot supports opening from attachment menu in basic group and supergroup chats

supports_channel_chats :: Maybe Bool

True, if the bot supports opening from attachment menu in channel chats

request_write_access :: Maybe Bool

True, if the user must be asked for the permission to send messages to the bot

is_added :: Maybe Bool

True, if the bot was explicitly added by the user. If the bot isn't added, then on the first bot launch toggleBotIsAddedToAttachmentMenu must be called and the bot must be added or removed

show_in_attachment_menu :: Maybe Bool

True, if the bot must be shown in the attachment menu

show_in_side_menu :: Maybe Bool

True, if the bot must be shown in the side menu

show_disclaimer_in_side_menu :: Maybe Bool

True, if a disclaimer, why the bot is shown in the side menu, is needed

name :: Maybe Text

Name for the bot in attachment menu

name_color :: Maybe AttachmentMenuBotColor

Color to highlight selected name of the bot if appropriate; may be null

default_icon :: Maybe File

Default icon for the bot in SVG format; may be null

ios_static_icon :: Maybe File

Icon for the bot in SVG format for the official iOS app; may be null

ios_animated_icon :: Maybe File

Icon for the bot in TGS format for the official iOS app; may be null

ios_side_menu_icon :: Maybe File

Icon for the bot in PNG format for the official iOS app side menu; may be null

android_icon :: Maybe File

Icon for the bot in TGS format for the official Android app; may be null

android_side_menu_icon :: Maybe File

Icon for the bot in SVG format for the official Android app side menu; may be null

macos_icon :: Maybe File

Icon for the bot in TGS format for the official native macOS app; may be null

macos_side_menu_icon :: Maybe File

Icon for the bot in PNG format for the official macOS app side menu; may be null

icon_color :: Maybe AttachmentMenuBotColor

Color to highlight selected icon of the bot if appropriate; may be null

web_app_placeholder :: Maybe File

Default placeholder for opened Web Apps in SVG format; may be null