TD.Data.AuthenticationCodeInfo

Documentation

data AuthenticationCodeInfo

Constructors

AuthenticationCodeInfo
Information about the authentication code that was sent

Fields

phone_number :: Maybe Text

A phone number that is being authenticated

_type :: Maybe AuthenticationCodeType

The way the code was sent to the user

next_type :: Maybe AuthenticationCodeType

The way the next code will be sent to the user; may be null

timeout :: Maybe Int

Timeout before the code can be re-sent, in seconds