TD.Data.AuthenticationCodeType

Description

Documentation

data AuthenticationCodeType

Provides information about the method by which an authentication code is delivered to the user

Constructors

AuthenticationCodeTypeTelegramMessage
An authentication code is delivered via a private Telegram message, which can be viewed from another active session @length Length of the code

Fields

_length :: Maybe Int

AuthenticationCodeTypeSms
An authentication code is delivered via an SMS message to the specified phone number @length Length of the code

Fields

_length :: Maybe Int

AuthenticationCodeTypeCall
An authentication code is delivered via a phone call to the specified phone number @length Length of the code

Fields

_length :: Maybe Int

AuthenticationCodeTypeFlashCall
An authentication code is delivered by an immediately canceled call to the specified phone number. The phone number that calls is the code that must be entered automatically @pattern Pattern of the phone number from which the call will be made

Fields

pattern :: Maybe String

AuthenticationCodeTypeMissedCall
An authentication code is delivered by an immediately canceled call to the specified phone number. The last digits of the phone number that calls are the code that must be entered manually by the user phone_number_prefix Prefix of the phone number from which the call will be made length Number of digits in the code, excluding the prefix

Fields

_length :: Maybe Int

phone_number_prefix :: Maybe String