TD.Data.AuthorizationState

Documentation

data AuthorizationState

Represents the current authorization state of the TDLib client

Constructors

AuthorizationStateWaitTdlibParameters
Initialization parameters are needed. Call setTdlibParameters to provide them

AuthorizationStateWaitPhoneNumber
TDLib needs the user's phone number to authorize. Call setAuthenticationPhoneNumber to provide the phone number, or use requestQrCodeAuthentication or checkAuthenticationBotToken for other authentication options

AuthorizationStateWaitEmailAddress
TDLib needs the user's email address to authorize. Call setAuthenticationEmailAddress to provide the email address, or directly call checkAuthenticationEmailCode with Apple ID/Google ID token if allowed

Fields

allow_apple_id :: Maybe Bool

True, if authorization through Apple ID is allowed

allow_google_id :: Maybe Bool

True, if authorization through Google ID is allowed

AuthorizationStateWaitEmailCode
TDLib needs the user's authentication code sent to an email address to authorize. Call checkAuthenticationEmailCode to provide the code

Fields

allow_apple_id :: Maybe Bool

True, if authorization through Apple ID is allowed

allow_google_id :: Maybe Bool

True, if authorization through Google ID is allowed

code_info :: Maybe EmailAddressAuthenticationCodeInfo

Information about the sent authentication code

email_address_reset_state :: Maybe EmailAddressResetState

Reset state of the email address; may be null if the email address can't be reset

AuthorizationStateWaitCode
TDLib needs the user's authentication code to authorize. Call checkAuthenticationCode to check the code

Fields

_code_info :: Maybe AuthenticationCodeInfo

Information about the authorization code that was sent

AuthorizationStateWaitOtherDeviceConfirmation
The user needs to confirm authorization on another logged in device by scanning a QR code with the provided link

Fields

link :: Maybe Text

A tg:// URL for the QR code. The link will be updated frequently

AuthorizationStateWaitRegistration
The user is unregistered and need to accept terms of service and enter their first name and last name to finish registration. Call registerUser to accept the terms of service and provide the data

Fields

terms_of_service :: Maybe TermsOfService

Telegram terms of service

AuthorizationStateWaitPassword
The user has been authorized, but needs to enter a 2-step verification password to start using the application. Call checkAuthenticationPassword to provide the password, or requestAuthenticationPasswordRecovery to recover the password, or deleteAccount to delete the account after a week

Fields

password_hint :: Maybe Text

Hint for the password; may be empty

has_recovery_email_address :: Maybe Bool

True, if a recovery email address has been set up

has_passport_data :: Maybe Bool

True, if some Telegram Passport elements were saved

recovery_email_address_pattern :: Maybe Text

Pattern of the email address to which the recovery email was sent; empty until a recovery email has been sent

AuthorizationStateReady
The user has been successfully authorized. TDLib is now ready to answer general requests

AuthorizationStateLoggingOut
The user is currently logging out

AuthorizationStateClosing
TDLib is closing, all subsequent queries will be answered with the error 500. Note that closing TDLib can take a while. All resources will be freed only after authorizationStateClosed has been received

AuthorizationStateClosed
TDLib client is in its final state. All databases are closed and all resources are released. No other updates will be received after this. All queries will be responded to with error code 500. To continue working, one must create a new instance of the TDLib client