TD.Data.AutoDownloadSettings

Description

Documentation

data AutoDownloadSettings

Constructors

AutoDownloadSettings
Contains auto-download settings

Fields

use_less_data_for_calls :: Maybe Bool

True, if "use less data for calls" option needs to be enabled

preload_stories :: Maybe Bool

True, if stories needs to be preloaded

preload_next_audio :: Maybe Bool

True, if the next audio track needs to be preloaded while the user is listening to an audio file

preload_large_videos :: Maybe Bool

True, if the beginning of video files needs to be preloaded for instant playback

video_upload_bitrate :: Maybe Int

The maximum suggested bitrate for uploaded videos, in kbit/s

max_other_file_size :: Maybe Int

The maximum size of other file types to be auto-downloaded, in bytes

max_video_file_size :: Maybe Int

The maximum size of a video file to be auto-downloaded, in bytes

max_photo_file_size :: Maybe Int

The maximum size of a photo file to be auto-downloaded, in bytes

is_auto_download_enabled :: Maybe Bool

True, if the auto-download is enabled