TD.Data.BackgroundType

Description

Documentation

data BackgroundType

Describes the type of a background

Constructors

BackgroundTypeWallpaper
A wallpaper in JPEG format

Fields

is_moving :: Maybe Bool

True, if the background needs to be slightly moved when device is tilted

is_blurred :: Maybe Bool

True, if the wallpaper must be downscaled to fit in 450x450 square and then box-blurred with radius 12

BackgroundTypePattern
A PNG or TGV (gzipped subset of SVG with MIME type "application/x-tgwallpattern") pattern to be combined with the background fill chosen by the user

Fields

is_moving :: Maybe Bool

True, if the background needs to be slightly moved when device is tilted

is_inverted :: Maybe Bool

True, if the background fill must be applied only to the pattern itself. All other pixels are black in this case. For dark themes only

intensity :: Maybe Int

Intensity of the pattern when it is shown above the filled background; 0-100.

fill :: Maybe BackgroundFill

Fill of the background

BackgroundTypeFill
A filled background @fill The background fill

Fields

fill :: Maybe BackgroundFill

Fill of the background