TD.Data.BankCardInfo

Description

Documentation

data BankCardInfo

Constructors

BankCardInfo
Information about a bank card title Title of the bank card description actions Actions that can be done with the bank card number

Fields

actions :: Maybe [BankCardActionOpenUrl]

title :: Maybe String