TD.Data.BasicGroup

Description

Documentation

data BasicGroup

Constructors

BasicGroup
Represents a basic group of 0-200 users (must be upgraded to a supergroup to accommodate more than 200 users)

Fields

upgraded_to_supergroup_id :: Maybe Int

Identifier of the supergroup to which this group was upgraded; 0 if none

is_active :: Maybe Bool

True, if the group is active

status :: Maybe ChatMemberStatus

Status of the current user in the group

member_count :: Maybe Int

Number of members in the group

_id :: Maybe Int

Group identifier