TD.Data.BasicGroupFullInfo

Description

Documentation

data BasicGroupFullInfo

Constructors

BasicGroupFullInfo
Contains full information about a basic group

Fields

bot_commands :: Maybe [BotCommands]

List of commands of bots in the group

invite_link :: Maybe ChatInviteLink

Primary invite link for this group; may be null. For chat administrators with can_invite_users right only. Updated only after the basic group is opened

can_toggle_aggressive_anti_spam :: Maybe Bool

True, if aggressive anti-spam checks can be enabled or disabled in the supergroup after upgrading the basic group to a supergroup

can_hide_members :: Maybe Bool

True, if non-administrators and non-bots can be hidden in responses to getSupergroupMembers and searchChatMembers for non-administrators after upgrading the basic group to a supergroup

members :: Maybe [ChatMember]

Group members

creator_user_id :: Maybe Int

User identifier of the creator of the group; 0 if unknown

description :: Maybe String

photo :: Maybe ChatPhoto

Chat photo; may be null if empty or unknown. If non-null, then it is the same photo as in chat.photo