TD.Data.BlockList

Description

Documentation

data BlockList

Describes a type of a block list

Constructors

BlockListMain
The main block list that disallows writing messages to the current user, receiving their status and photo, viewing of stories, and some other actions

BlockListStories
The block list that disallows viewing of stories of the current user