TD.Data.BotInfo

Documentation

data BotInfo

Constructors

BotInfo
Contains information about a bot

Fields

short_description :: Maybe Text

The text that is shown on the bot's profile page and is sent together with the link when users share the bot

description :: Maybe Text

The text shown in the chat with the bot if the chat is empty

photo :: Maybe Photo

Photo shown in the chat with the bot if the chat is empty; may be null

animation :: Maybe Animation

Animation shown in the chat with the bot if the chat is empty; may be null

menu_button :: Maybe BotMenuButton

Information about a button to show instead of the bot commands menu button; may be null if ordinary bot commands menu must be shown

commands :: Maybe [BotCommand]

List of the bot commands

privacy_policy_url :: Maybe Text

The HTTP link to the privacy policy of the bot. If empty, then privacy command must be used if supported by the bot. If the command isn't supported, then https:telegram.orgprivacy-tpa must be opened

default_group_administrator_rights :: Maybe ChatAdministratorRights

Default administrator rights for adding the bot to basic group and supergroup chats; may be null

default_channel_administrator_rights :: Maybe ChatAdministratorRights

Default administrator rights for adding the bot to channels; may be null

has_media_previews :: Maybe Bool

True, if the bot has media previews

edit_commands_link :: Maybe InternalLinkType

The internal link, which can be used to edit bot commands; may be null

edit_description_link :: Maybe InternalLinkType

The internal link, which can be used to edit bot description; may be null

edit_description_media_link :: Maybe InternalLinkType

The internal link, which can be used to edit the photo or animation shown in the chat with the bot if the chat is empty; may be null

edit_settings_link :: Maybe InternalLinkType

The internal link, which can be used to edit bot settings; may be null