TD.Data.BusinessAwayMessageSchedule

Documentation

data BusinessAwayMessageSchedule

Describes conditions for sending of away messages by a Telegram Business account

Constructors

BusinessAwayMessageScheduleAlways
Send away messages always

BusinessAwayMessageScheduleOutsideOfOpeningHours
Send away messages outside of the business opening hours

BusinessAwayMessageScheduleCustom
Send away messages only in the specified time span

Fields

start_date :: Maybe Int

Point in time (Unix timestamp) when the away messages will start to be sent

end_date :: Maybe Int

Point in time (Unix timestamp) when the away messages will stop to be sent