TD.Data.BusinessBotRights

Documentation

data BusinessBotRights

Constructors

BusinessBotRights
Describes rights of a business bot

Fields

can_reply :: Maybe Bool

True, if the bot can send and edit messages in the private chats that had incoming messages in the last 24 hours

can_read_messages :: Maybe Bool

True, if the bot can mark incoming private messages as read

can_delete_sent_messages :: Maybe Bool

True, if the bot can delete sent messages

can_delete_all_messages :: Maybe Bool

True, if the bot can delete any message

can_edit_name :: Maybe Bool

True, if the bot can edit name of the business account

can_edit_bio :: Maybe Bool

True, if the bot can edit bio of the business account

can_edit_profile_photo :: Maybe Bool

True, if the bot can edit profile photo of the business account

can_edit_username :: Maybe Bool

True, if the bot can edit username of the business account

can_view_gifts_and_stars :: Maybe Bool

True, if the bot can view gifts and amount of Telegram Stars owned by the business account

can_sell_gifts :: Maybe Bool

True, if the bot can sell regular gifts received by the business account

can_change_gift_settings :: Maybe Bool

True, if the bot can change gift receiving settings of the business account

can_transfer_and_upgrade_gifts :: Maybe Bool

True, if the bot can transfer and upgrade gifts received by the business account

can_transfer_stars :: Maybe Bool

True, if the bot can transfer Telegram Stars received by the business account to account of the bot, or use them to upgrade and transfer gifts

can_manage_stories :: Maybe Bool

True, if the bot can post, edit and delete stories

defaultBusinessBotRights :: BusinessBotRights