TD.Data.BusinessFeature

Documentation

data BusinessFeature

Describes a feature available to Business user accounts

Constructors

BusinessFeatureLocation
The ability to set location

BusinessFeatureOpeningHours
The ability to set opening hours

BusinessFeatureQuickReplies
The ability to use quick replies

BusinessFeatureGreetingMessage
The ability to set up a greeting message

BusinessFeatureAwayMessage
The ability to set up an away message

BusinessFeatureAccountLinks
The ability to create links to the business account with predefined message text

BusinessFeatureStartPage
The ability to customize start page

BusinessFeatureBots
The ability to connect a bot to the account

BusinessFeatureEmojiStatus
The ability to show an emoji status along with the business name

BusinessFeatureChatFolderTags
The ability to display folder names for each chat in the chat list

BusinessFeatureUpgradedStories
Allowed to use many additional features for stories