TD.Data.BusinessInfo

Documentation

data BusinessInfo

Constructors

BusinessInfo
Contains information about a Telegram Business account

Fields

location :: Maybe BusinessLocation

Location of the business; may be null if none

opening_hours :: Maybe BusinessOpeningHours

Opening hours of the business; may be null if none. The hours are guaranteed to be valid and has already been split by week days

local_opening_hours :: Maybe BusinessOpeningHours

Opening hours of the business in the local time; may be null if none. The hours are guaranteed to be valid and has already been split by week days. Local time zone identifier will be empty. An updateUserFullInfo update is not triggered when value of this field changes

next_open_in :: Maybe Int

Time left before the business will open the next time, in seconds; 0 if unknown. An updateUserFullInfo update is not triggered when value of this field changes

next_close_in :: Maybe Int

Time left before the business will close the next time, in seconds; 0 if unknown. An updateUserFullInfo update is not triggered when value of this field changes

greeting_message_settings :: Maybe BusinessGreetingMessageSettings

The greeting message; may be null if none or the Business account is not of the current user

away_message_settings :: Maybe BusinessAwayMessageSettings

The away message; may be null if none or the Business account is not of the current user

start_page :: Maybe BusinessStartPage

Information about start page of the account; may be null if none