TD.Data.BusinessRecipients

Documentation

data BusinessRecipients

Constructors

BusinessRecipients
Describes private chats chosen for automatic interaction with a business

Fields

chat_ids :: Maybe [Int]

Identifiers of selected private chats

excluded_chat_ids :: Maybe [Int]

Identifiers of private chats that are always excluded; for businessConnectedBot only

select_existing_chats :: Maybe Bool

True, if all existing private chats are selected

select_new_chats :: Maybe Bool

True, if all new private chats are selected

select_contacts :: Maybe Bool

True, if all private chats with contacts are selected

select_non_contacts :: Maybe Bool

True, if all private chats with non-contacts are selected

exclude_selected :: Maybe Bool

If true, then all private chats except the selected are chosen. Otherwise, only the selected chats are chosen

defaultBusinessRecipients :: BusinessRecipients