TD.Data.CallDiscardReason

Description

Documentation

data CallDiscardReason

Describes the reason why a call was discarded

Constructors

CallDiscardReasonEmpty
The call wasn't discarded, or the reason is unknown

CallDiscardReasonMissed
The call was ended before the conversation started. It was canceled by the caller or missed by the other party

CallDiscardReasonDeclined
The call was ended before the conversation started. It was declined by the other party

CallDiscardReasonDisconnected
The call was ended during the conversation because the users were disconnected

CallDiscardReasonHungUp
The call was ended because one of the parties hung up