TD.Data.CallState

Documentation

data CallState

Describes the current call state

Constructors

CallStatePending
The call is pending, waiting to be accepted by a user

Fields

is_created :: Maybe Bool

True, if the call has already been created by the server

is_received :: Maybe Bool

True, if the call has already been received by the other party

CallStateExchangingKeys
The call has been answered and encryption keys are being exchanged

CallStateReady
The call is ready to use

Fields

protocol :: Maybe CallProtocol

Call protocols supported by the other call participant

servers :: Maybe [CallServer]

List of available call servers

config :: Maybe Text

A JSON-encoded call config

encryption_key :: Maybe ByteString

Call encryption key

emojis :: Maybe [Text]

Encryption key emojis fingerprint

allow_p2p :: Maybe Bool

True, if peer-to-peer connection is allowed by users privacy settings

custom_parameters :: Maybe Text

Custom JSON-encoded call parameters to be passed to tgcalls

CallStateHangingUp
The call is hanging up after discardCall has been called

CallStateDiscarded
The call has ended successfully

Fields

reason :: Maybe CallDiscardReason

The reason why the call has ended

need_rating :: Maybe Bool

True, if the call rating must be sent to the server

need_debug_information :: Maybe Bool

True, if the call debug information must be sent to the server

need_log :: Maybe Bool

True, if the call log must be sent to the server

CallStateError
The call has ended with an error

Fields

_error :: Maybe Error

Error. An error with the code 4005000 will be returned if an outgoing call is missed because of an expired timeout