TD.Data.CanTransferOwnershipResult

Documentation

data CanTransferOwnershipResult

Represents result of checking whether the current session can be used to transfer a chat ownership to another user

Constructors

CanTransferOwnershipResultOk
The session can be used

CanTransferOwnershipResultPasswordNeeded
The 2-step verification needs to be enabled first

CanTransferOwnershipResultPasswordTooFresh
The 2-step verification was enabled recently, user needs to wait

Fields

retry_after :: Maybe Int

Time left before the session can be used to transfer ownership of a chat, in seconds

CanTransferOwnershipResultSessionTooFresh
The session was created recently, user needs to wait

Fields

retry_after :: Maybe Int

Time left before the session can be used to transfer ownership of a chat, in seconds