TD.Data.Chat

Description

Documentation

data Chat

Constructors

Chat
A chat. (Can be a private chat, basic group, supergroup, or secret chat)

Fields

client_data :: Maybe String

Application-specific data associated with the chat. (For example, the chat scroll position or local chat notification settings can be stored here.) Persistent if the message database is used

draft_message :: Maybe DraftMessage

A draft of a message in the chat; may be null

reply_markup_message_id :: Maybe Int

Identifier of the message from which reply markup needs to be used; 0 if there is no default custom reply markup in the chat

pending_join_requests :: Maybe ChatJoinRequestsInfo

Information about pending join requests; may be null

video_chat :: Maybe VideoChat

Information about video chat of the chat

action_bar :: Maybe ChatActionBar

Information about actions which must be possible to do through the chat action bar; may be null

theme_name :: Maybe String

If non-empty, name of a theme, set for the chat

message_ttl :: Maybe Int

Current message Time To Live setting (self-destruct timer) for the chat; 0 if not defined. TTL is counted from the time message or its content is viewed in secret chats and from the send date in other chats

available_reactions :: Maybe ChatAvailableReactions

Types of reaction, available in the chat

notification_settings :: Maybe ChatNotificationSettings

Notification settings for the chat

unread_reaction_count :: Maybe Int

Number of messages with unread reactions in the chat

unread_mention_count :: Maybe Int

Number of unread messages with a mention/reply in the chat

last_read_outbox_message_id :: Maybe Int

Identifier of the last read outgoing message

last_read_inbox_message_id :: Maybe Int

Identifier of the last read incoming message

unread_count :: Maybe Int

Number of unread messages in the chat

default_disable_notification :: Maybe Bool

Default value of the disable_notification parameter, used when a message is sent to the chat

can_be_reported :: Maybe Bool

True, if the chat can be reported to Telegram moderators through reportChat or reportChatPhoto

can_be_deleted_for_all_users :: Maybe Bool

True, if the chat messages can be deleted for all users

can_be_deleted_only_for_self :: Maybe Bool

True, if the chat messages can be deleted only for the current user while other users will continue to see the messages

has_scheduled_messages :: Maybe Bool

True, if the chat has scheduled messages

is_blocked :: Maybe Bool

True, if the chat is blocked by the current user and private messages from the chat can't be received

is_marked_as_unread :: Maybe Bool

True, if the chat is marked as unread

has_protected_content :: Maybe Bool

True, if chat content can't be saved locally, forwarded, or copied

message_sender_id :: Maybe MessageSender

Identifier of a user or chat that is selected to send messages in the chat; may be null if the user can't change message sender

positions :: Maybe [ChatPosition]

Positions of the chat in chat lists

last_message :: Maybe Message

Last message in the chat; may be null

permissions :: Maybe ChatPermissions

Actions that non-administrator chat members are allowed to take in the chat

photo :: Maybe ChatPhotoInfo

Chat photo; may be null

title :: Maybe String

Chat title

_type :: Maybe ChatType

Type of the chat

_id :: Maybe Int

Chat unique identifier