TD.Data.ChatAction

Description

Documentation

data ChatAction

Describes the different types of activity in a chat

Constructors

ChatActionTyping
The user is typing a message

ChatActionRecordingVideo
The user is recording a video

ChatActionUploadingVideo
The user is uploading a video @progress Upload progress, as a percentage

Fields

progress :: Maybe Int

ChatActionRecordingVoiceNote
The user is recording a voice note

ChatActionUploadingVoiceNote
The user is uploading a voice note @progress Upload progress, as a percentage

Fields

progress :: Maybe Int

ChatActionUploadingPhoto
The user is uploading a photo @progress Upload progress, as a percentage

Fields

progress :: Maybe Int

ChatActionUploadingDocument
The user is uploading a document @progress Upload progress, as a percentage

Fields

progress :: Maybe Int

ChatActionChoosingSticker
The user is picking a sticker to send

ChatActionChoosingLocation
The user is picking a location or venue to send

ChatActionChoosingContact
The user is picking a contact to send

ChatActionStartPlayingGame
The user has started to play a game

ChatActionRecordingVideoNote
The user is recording a video note

ChatActionUploadingVideoNote
The user is uploading a video note @progress Upload progress, as a percentage

Fields

progress :: Maybe Int

ChatActionWatchingAnimations
The user is watching animations sent by the other party by clicking on an animated emoji @emoji The animated emoji

Fields

emoji :: Maybe String

ChatActionCancel
The user has canceled the previous action