TD.Data.ChatActionBar

Description

Documentation

data ChatActionBar

Describes actions which must be possible to do through a chat action bar

Constructors

ChatActionBarReportSpam
The chat can be reported as spam using the method reportChat with the reason chatReportReasonSpam. If the chat is a private chat with a user with an emoji status, then a notice about emoji status usage must be shown

Fields

can_unarchive :: Maybe Bool

If true, the chat was automatically archived and can be moved back to the main chat list using addChatToList simultaneously with setting chat notification settings to default using setChatNotificationSettings

ChatActionBarReportUnrelatedLocation
The chat is a location-based supergroup, which can be reported as having unrelated location using the method reportChat with the reason chatReportReasonUnrelatedLocation

ChatActionBarInviteMembers
The chat is a recently created group chat to which new members can be invited

ChatActionBarReportAddBlock
The chat is a private or secret chat, which can be reported using the method reportChat, or the other user can be blocked using the method toggleMessageSenderIsBlocked, or the other user can be added to the contact list using the method addContact. If the chat is a private chat with a user with an emoji status, then a notice about emoji status usage must be shown

Fields

distance :: Maybe Int

If non-negative, the current user was found by the peer through searchChatsNearby and this is the distance between the users

can_unarchive :: Maybe Bool

If true, the chat was automatically archived and can be moved back to the main chat list using addChatToList simultaneously with setting chat notification settings to default using setChatNotificationSettings

ChatActionBarAddContact
The chat is a private or secret chat and the other user can be added to the contact list using the method addContact

ChatActionBarSharePhoneNumber
The chat is a private or secret chat with a mutual contact and the user's phone number can be shared with the other user using the method sharePhoneNumber

ChatActionBarJoinRequest
The chat is a private chat with an administrator of a chat to which the user sent join request

Fields

request_date :: Maybe Int

Point in time (Unix timestamp) when the join request was sent

is_channel :: Maybe Bool

True, if the join request was sent to a channel chat

title :: Maybe String

Title of the chat to which the join request was sent