TD.Data.ChatActiveStories

Documentation

data ChatActiveStories

Constructors

ChatActiveStories
Describes active stories posted by a chat

Fields

chat_id :: Maybe Int

Identifier of the chat that posted the stories

list :: Maybe StoryList

Identifier of the story list in which the stories are shown; may be null if the stories aren't shown in a story list

order :: Maybe Int

A parameter used to determine order of the stories in the story list; 0 if the stories doesn't need to be shown in the story list. Stories must be sorted by the pair (order, story_sender_chat_id) in descending order

max_read_story_id :: Maybe Int

Identifier of the last read active story

stories :: Maybe [StoryInfo]

Basic information about the stories; use getStory to get full information about the stories. The stories are in a chronological order (i.e., in order of increasing story identifiers)