TD.Data.ChatAdministratorRights

Description

Documentation

data ChatAdministratorRights

Constructors

ChatAdministratorRights
Describes rights of the administrator

Fields

is_anonymous :: Maybe Bool

True, if the administrator isn't shown in the chat member list and sends messages anonymously; applicable to supergroups only

can_manage_video_chats :: Maybe Bool

True, if the administrator can manage video chats

can_promote_members :: Maybe Bool

True, if the administrator can add new administrators with a subset of their own privileges or demote administrators that were directly or indirectly promoted by them

can_manage_topics :: Maybe Bool

True, if the administrator can manage topics; applicable to forum supergroups only

can_pin_messages :: Maybe Bool

True, if the administrator can pin messages; applicable to basic groups and supergroups only

can_restrict_members :: Maybe Bool

True, if the administrator can restrict, ban, or unban chat members; always true for channels

can_invite_users :: Maybe Bool

True, if the administrator can invite new users to the chat

can_delete_messages :: Maybe Bool

True, if the administrator can delete messages of other users

can_edit_messages :: Maybe Bool

True, if the administrator can edit messages of other users and pin messages; applicable to channels only

can_post_messages :: Maybe Bool

True, if the administrator can create channel posts; applicable to channels only

can_change_info :: Maybe Bool

True, if the administrator can change the chat title, photo, and other settings

can_manage_chat :: Maybe Bool

True, if the administrator can get chat event log, get chat statistics, get message statistics in channels, get channel members, see anonymous administrators in supergroups and ignore slow mode. Implied by any other privilege; applicable to supergroups and channels only