TD.Data.ChatBoostLevelFeatures

Documentation

data ChatBoostLevelFeatures

Constructors

ChatBoostLevelFeatures
Contains a list of features available on a specific chat boost level

Fields

level :: Maybe Int

Target chat boost level

story_per_day_count :: Maybe Int

Number of stories that the chat can publish daily

custom_emoji_reaction_count :: Maybe Int

Number of custom emoji reactions that can be added to the list of available reactions

title_color_count :: Maybe Int

Number of custom colors for chat title

profile_accent_color_count :: Maybe Int

Number of custom colors for profile photo background

can_set_profile_background_custom_emoji :: Maybe Bool

True, if custom emoji for profile background can be set

accent_color_count :: Maybe Int

Number of custom colors for background of empty chat photo, replies to messages and link previews

can_set_background_custom_emoji :: Maybe Bool

True, if custom emoji for reply header and link preview background can be set

can_set_emoji_status :: Maybe Bool

True, if emoji status can be set

chat_theme_background_count :: Maybe Int

Number of chat theme backgrounds that can be set as chat background

can_set_custom_background :: Maybe Bool

True, if custom background can be set in the chat for all users

can_set_custom_emoji_sticker_set :: Maybe Bool

True, if custom emoji sticker set can be set for the chat

can_enable_automatic_translation :: Maybe Bool

True, if automatic translation of messages can be enabled in the chat

can_recognize_speech :: Maybe Bool

True, if speech recognition can be used for video note and voice note messages by all users

can_disable_sponsored_messages :: Maybe Bool

True, if sponsored messages can be disabled in the chat