TD.Data.ChatBoostSource

Documentation

data ChatBoostSource

Describes source of a chat boost

Constructors

ChatBoostSourceGiftCode
The chat created a Telegram Premium gift code for a user

Fields

user_id :: Maybe Int

Identifier of a user, for which the gift code was created

gift_code :: Maybe Text

The created Telegram Premium gift code, which is known only if this is a gift code for the current user, or it has already been claimed

ChatBoostSourceGiveaway
The chat created a giveaway

Fields

user_id :: Maybe Int

Identifier of a user, for which the gift code was created

gift_code :: Maybe Text

The created Telegram Premium gift code, which is known only if this is a gift code for the current user, or it has already been claimed

star_count :: Maybe Int

Number of Telegram Stars distributed among winners of the giveaway

giveaway_message_id :: Maybe Int

Identifier of the corresponding giveaway message; can be an identifier of a deleted message

is_unclaimed :: Maybe Bool

True, if the winner for the corresponding giveaway prize wasn't chosen, because there were not enough participants

ChatBoostSourcePremium
A user with Telegram Premium subscription or gifted Telegram Premium boosted the chat

Fields

user_id :: Maybe Int

Identifier of a user, for which the gift code was created