TD.Data.ChatBoostStatus

Documentation

data ChatBoostStatus

Constructors

ChatBoostStatus
Describes current boost status of a chat

Fields

boost_url :: Maybe Text

An HTTP URL, which can be used to boost the chat

applied_slot_ids :: Maybe [Int]

Identifiers of boost slots of the current user applied to the chat

level :: Maybe Int

Current boost level of the chat

gift_code_boost_count :: Maybe Int

The number of boosts received by the chat from created Telegram Premium gift codes and giveaways; always 0 if the current user isn't an administrator in the chat

boost_count :: Maybe Int

The number of boosts received by the chat

current_level_boost_count :: Maybe Int

The number of boosts added to reach the current level

next_level_boost_count :: Maybe Int

The number of boosts needed to reach the next level; 0 if the next level isn't available

premium_member_count :: Maybe Int

Approximate number of Telegram Premium subscribers joined the chat; always 0 if the current user isn't an administrator in the chat

premium_member_percentage :: Maybe Double

A percentage of Telegram Premium subscribers joined the chat; always 0 if the current user isn't an administrator in the chat

prepaid_giveaways :: Maybe [PrepaidGiveaway]

The list of prepaid giveaways available for the chat; only for chat administrators