TD.Data.ChatEventAction

Documentation

data ChatEventAction

Represents a chat event

Constructors

ChatEventMessageEdited
A message was edited

Fields

old_message :: Maybe Message

The original message before the edit

new_message :: Maybe Message

The message after it was edited

ChatEventMessageDeleted
A message was deleted

Fields

message :: Maybe Message

Deleted message

can_report_anti_spam_false_positive :: Maybe Bool

True, if the message deletion can be reported via reportSupergroupAntiSpamFalsePositive

ChatEventMessagePinned
A message was pinned

Fields

message :: Maybe Message

Deleted message

ChatEventMessageUnpinned
A message was unpinned

Fields

message :: Maybe Message

Deleted message

ChatEventPollStopped
A poll in a message was stopped

Fields

message :: Maybe Message

Deleted message

ChatEventMemberJoined
A new member joined the chat

ChatEventMemberJoinedByInviteLink
A new member joined the chat via an invite link

Fields

invite_link :: Maybe ChatInviteLink

Invite link used to join the chat

via_chat_folder_invite_link :: Maybe Bool

True, if the user has joined the chat using an invite link for a chat folder

ChatEventMemberJoinedByRequest
A new member was accepted to the chat by an administrator

Fields

approver_user_id :: Maybe Int

User identifier of the chat administrator, approved user join request

invite_link :: Maybe ChatInviteLink

Invite link used to join the chat

ChatEventMemberInvited
A new chat member was invited

Fields

user_id :: Maybe Int

New member user identifier

status :: Maybe ChatMemberStatus

New member status

ChatEventMemberLeft
A member left the chat

ChatEventMemberPromoted
A chat member has gained/lost administrator status, or the list of their administrator privileges has changed

Fields

user_id :: Maybe Int

New member user identifier

old_status :: Maybe ChatMemberStatus

Previous status of the chat member

new_status :: Maybe ChatMemberStatus

New status of the chat member

ChatEventMemberRestricted
A chat member was restrictedunrestricted or bannedunbanned, or the list of their restrictions has changed

Fields

member_id :: Maybe MessageSender

Affected chat member identifier

old_status :: Maybe ChatMemberStatus

Previous status of the chat member

new_status :: Maybe ChatMemberStatus

New status of the chat member

ChatEventAvailableReactionsChanged
The chat available reactions were changed

Fields

old_available_reactions :: Maybe ChatAvailableReactions

Previous chat available reactions

new_available_reactions :: Maybe ChatAvailableReactions

New chat available reactions

ChatEventDescriptionChanged
The chat description was changed

Fields

old_description :: Maybe Text

Previous chat description

new_description :: Maybe Text

New chat description

ChatEventLinkedChatChanged
The linked chat of a supergroup was changed

Fields

old_linked_chat_id :: Maybe Int

Previous supergroup linked chat identifier

new_linked_chat_id :: Maybe Int

New supergroup linked chat identifier

ChatEventLocationChanged
The supergroup location was changed

Fields

old_location :: Maybe ChatLocation

Previous location; may be null

new_location :: Maybe ChatLocation

New location; may be null

ChatEventMessageAutoDeleteTimeChanged
The message auto-delete timer was changed

Fields

old_message_auto_delete_time :: Maybe Int

Previous value of message_auto_delete_time

new_message_auto_delete_time :: Maybe Int

New value of message_auto_delete_time

ChatEventPermissionsChanged
The chat permissions was changed

Fields

old_permissions :: Maybe ChatPermissions

Previous chat permissions

new_permissions :: Maybe ChatPermissions

New chat permissions

ChatEventPhotoChanged
The chat photo was changed

Fields

old_photo :: Maybe ChatPhoto

Previous chat photo value; may be null

new_photo :: Maybe ChatPhoto

New chat photo value; may be null

ChatEventSlowModeDelayChanged
The slow_mode_delay setting of a supergroup was changed

Fields

old_slow_mode_delay :: Maybe Int

Previous value of slow_mode_delay, in seconds

new_slow_mode_delay :: Maybe Int

New value of slow_mode_delay, in seconds

ChatEventStickerSetChanged
The supergroup sticker set was changed

Fields

old_sticker_set_id :: Maybe Int

Previous identifier of the chat sticker set; 0 if none

new_sticker_set_id :: Maybe Int

New identifier of the chat sticker set; 0 if none

ChatEventTitleChanged
The chat title was changed

Fields

old_title :: Maybe Text

Previous chat title

new_title :: Maybe Text

New chat title

ChatEventUsernameChanged
The chat editable username was changed

Fields

old_username :: Maybe Text

Previous chat username

new_username :: Maybe Text

New chat username

ChatEventActiveUsernamesChanged
The chat active usernames were changed

Fields

old_usernames :: Maybe [Text]

Previous list of active usernames

new_usernames :: Maybe [Text]

New list of active usernames

ChatEventHasProtectedContentToggled
The has_protected_content setting of a channel was toggled

Fields

has_protected_content :: Maybe Bool

New value of has_protected_content

ChatEventInvitesToggled
The can_invite_users permission of a supergroup chat was toggled

Fields

can_invite_users :: Maybe Bool

New value of can_invite_users permission

ChatEventIsAllHistoryAvailableToggled
The is_all_history_available setting of a supergroup was toggled

Fields

is_all_history_available :: Maybe Bool

New value of is_all_history_available

ChatEventHasAggressiveAntiSpamEnabledToggled
The has_aggressive_anti_spam_enabled setting of a supergroup was toggled

Fields

has_aggressive_anti_spam_enabled :: Maybe Bool

New value of has_aggressive_anti_spam_enabled

ChatEventSignMessagesToggled
The sign_messages setting of a channel was toggled

Fields

sign_messages :: Maybe Bool

New value of sign_messages

ChatEventInviteLinkEdited
A chat invite link was edited

Fields

old_invite_link :: Maybe ChatInviteLink

Previous information about the invite link

new_invite_link :: Maybe ChatInviteLink

New information about the invite link

ChatEventInviteLinkRevoked
A chat invite link was revoked

Fields

invite_link :: Maybe ChatInviteLink

Invite link used to join the chat

ChatEventInviteLinkDeleted
A revoked chat invite link was deleted

Fields

invite_link :: Maybe ChatInviteLink

Invite link used to join the chat

ChatEventVideoChatCreated
A video chat was created

Fields

group_call_id :: Maybe Int

Identifier of the video chat. The video chat can be received through the method getGroupCall

ChatEventVideoChatEnded
A video chat was ended

Fields

group_call_id :: Maybe Int

Identifier of the video chat. The video chat can be received through the method getGroupCall

ChatEventVideoChatMuteNewParticipantsToggled
The mute_new_participants setting of a video chat was toggled

Fields

mute_new_participants :: Maybe Bool

New value of the mute_new_participants setting

ChatEventVideoChatParticipantIsMutedToggled
A video chat participant was muted or unmuted

Fields

participant_id :: Maybe MessageSender

Identifier of the affected group call participant

is_muted :: Maybe Bool

New value of is_muted

ChatEventVideoChatParticipantVolumeLevelChanged
A video chat participant volume level was changed

Fields

participant_id :: Maybe MessageSender

Identifier of the affected group call participant

volume_level :: Maybe Int

New value of volume_level; 1-20000 in hundreds of percents

ChatEventIsForumToggled
The is_forum setting of a channel was toggled

Fields

is_forum :: Maybe Bool

New value of is_forum

ChatEventForumTopicCreated
A new forum topic was created

Fields

topic_info :: Maybe ForumTopicInfo

Information about the topic

ChatEventForumTopicEdited
A forum topic was edited

Fields

old_topic_info :: Maybe ForumTopicInfo

Old information about the topic

new_topic_info :: Maybe ForumTopicInfo

New information about the topic

ChatEventForumTopicToggleIsClosed
A forum topic was closed or reopened

Fields

topic_info :: Maybe ForumTopicInfo

Information about the topic

ChatEventForumTopicToggleIsHidden
The General forum topic was hidden or unhidden

Fields

topic_info :: Maybe ForumTopicInfo

Information about the topic

ChatEventForumTopicDeleted
A forum topic was deleted

Fields

topic_info :: Maybe ForumTopicInfo

Information about the topic

ChatEventForumTopicPinned
A pinned forum topic was changed

Fields

old_topic_info :: Maybe ForumTopicInfo

Old information about the topic

new_topic_info :: Maybe ForumTopicInfo

New information about the topic

defaultChatEventMessageEdited :: ChatEventAction

defaultChatEventMessageDeleted :: ChatEventAction

defaultChatEventMessagePinned :: ChatEventAction

defaultChatEventMessageUnpinned :: ChatEventAction

defaultChatEventPollStopped :: ChatEventAction

defaultChatEventMemberJoinedByInviteLink :: ChatEventAction

defaultChatEventMemberJoinedByRequest :: ChatEventAction

defaultChatEventMemberInvited :: ChatEventAction

defaultChatEventMemberPromoted :: ChatEventAction

defaultChatEventMemberRestricted :: ChatEventAction

defaultChatEventAvailableReactionsChanged :: ChatEventAction

defaultChatEventDescriptionChanged :: ChatEventAction

defaultChatEventLinkedChatChanged :: ChatEventAction

defaultChatEventLocationChanged :: ChatEventAction

defaultChatEventMessageAutoDeleteTimeChanged :: ChatEventAction

defaultChatEventPermissionsChanged :: ChatEventAction

defaultChatEventPhotoChanged :: ChatEventAction

defaultChatEventSlowModeDelayChanged :: ChatEventAction

defaultChatEventStickerSetChanged :: ChatEventAction

defaultChatEventTitleChanged :: ChatEventAction

defaultChatEventUsernameChanged :: ChatEventAction

defaultChatEventActiveUsernamesChanged :: ChatEventAction

defaultChatEventHasProtectedContentToggled :: ChatEventAction

defaultChatEventInvitesToggled :: ChatEventAction

defaultChatEventIsAllHistoryAvailableToggled :: ChatEventAction

defaultChatEventHasAggressiveAntiSpamEnabledToggled :: ChatEventAction

defaultChatEventSignMessagesToggled :: ChatEventAction

defaultChatEventInviteLinkEdited :: ChatEventAction

defaultChatEventInviteLinkRevoked :: ChatEventAction

defaultChatEventInviteLinkDeleted :: ChatEventAction

defaultChatEventVideoChatCreated :: ChatEventAction

defaultChatEventVideoChatEnded :: ChatEventAction

defaultChatEventVideoChatMuteNewParticipantsToggled :: ChatEventAction

defaultChatEventVideoChatParticipantIsMutedToggled :: ChatEventAction

defaultChatEventVideoChatParticipantVolumeLevelChanged :: ChatEventAction

defaultChatEventIsForumToggled :: ChatEventAction

defaultChatEventForumTopicCreated :: ChatEventAction

defaultChatEventForumTopicEdited :: ChatEventAction

defaultChatEventForumTopicToggleIsClosed :: ChatEventAction

defaultChatEventForumTopicToggleIsHidden :: ChatEventAction

defaultChatEventForumTopicDeleted :: ChatEventAction

defaultChatEventForumTopicPinned :: ChatEventAction