TD.Data.ChatFilterInfo

Documentation

data ChatFilterInfo

Constructors

ChatFilterInfo
Contains basic information about a chat filter

Fields

icon_name :: Maybe String

The chosen or default icon name for short filter representation. One of All, Unread, Unmuted, Bots, Channels, Groups, Private, Custom, Setup, Cat, Crown, Favorite, Flower, Game, Home, Love, Mask, Party, Sport, Study, Trade, Travel, Work

title :: Maybe String

The title of the filter; 1-12 characters without line feeds

_id :: Maybe Int

Unique chat filter identifier