TD.Data.ChatFolder

Documentation

data ChatFolder

Constructors

ChatFolder
Represents a folder for user chats

Fields

title :: Maybe Text

The title of the folder; 1-12 characters without line feeds

icon :: Maybe ChatFolderIcon

The chosen icon for the chat folder; may be null. If null, use getChatFolderDefaultIconName to get default icon name for the folder

color_id :: Maybe Int

The identifier of the chosen color for the chat folder icon; from -1 to 6. If -1, then color is disabled. Can't be changed if folder tags are disabled or the current user doesn't have Telegram Premium subscription

is_shareable :: Maybe Bool

True, if at least one link has been created for the folder

pinned_chat_ids :: Maybe [Int]

The chat identifiers of pinned chats in the folder. There can be up to getOption("chat_folder_chosen_chat_count_max") pinned and always included non-secret chats and the same number of secret chats, but the limit can be increased with Telegram Premium

included_chat_ids :: Maybe [Int]

The chat identifiers of always included chats in the folder. There can be up to getOption("chat_folder_chosen_chat_count_max") pinned and always included non-secret chats and the same number of secret chats, but the limit can be increased with Telegram Premium

excluded_chat_ids :: Maybe [Int]

The chat identifiers of always excluded chats in the folder. There can be up to getOption("chat_folder_chosen_chat_count_max") always excluded non-secret chats and the same number of secret chats, but the limit can be increased with Telegram Premium

exclude_muted :: Maybe Bool

True, if muted chats need to be excluded

exclude_read :: Maybe Bool

True, if read chats need to be excluded

exclude_archived :: Maybe Bool

True, if archived chats need to be excluded

include_contacts :: Maybe Bool

True, if contacts need to be included

include_non_contacts :: Maybe Bool

True, if non-contact users need to be included

include_bots :: Maybe Bool

True, if bots need to be included

include_groups :: Maybe Bool

True, if basic groups and supergroups need to be included

include_channels :: Maybe Bool

True, if channels need to be included

defaultChatFolder :: ChatFolder