TD.Data.ChatInviteLink

Description

Documentation

data ChatInviteLink

Constructors

ChatInviteLink
Contains a chat invite link

Fields

is_revoked :: Maybe Bool

True, if the link was revoked

is_primary :: Maybe Bool

True, if the link is primary. Primary invite link can't have name, expiration date, or usage limit. There is exactly one primary invite link for each administrator with can_invite_users right at a given time

creates_join_request :: Maybe Bool

True, if the link only creates join request. If true, total number of joining members will be unlimited

pending_join_request_count :: Maybe Int

Number of pending join requests created using this link

member_count :: Maybe Int

Number of chat members, which joined the chat using the link

member_limit :: Maybe Int

The maximum number of members, which can join the chat using the link simultaneously; 0 if not limited. Always 0 if the link requires approval

expiration_date :: Maybe Int

Point in time (Unix timestamp) when the link will expire; 0 if never

edit_date :: Maybe Int

Point in time (Unix timestamp) when the link was last edited; 0 if never or unknown

date :: Maybe Int

Point in time (Unix timestamp) when the link was created

creator_user_id :: Maybe Int

User identifier of an administrator created the link

name :: Maybe String

Name of the link

invite_link :: Maybe String

Chat invite link