TD.Data.ChatInviteLinkInfo

Documentation

data ChatInviteLinkInfo

Constructors

ChatInviteLinkInfo
Contains information about a chat invite link

Fields

chat_id :: Maybe Int

Chat identifier of the invite link; 0 if the user has no access to the chat before joining

accessible_for :: Maybe Int

If non-zero, the amount of time for which read access to the chat will remain available, in seconds

_type :: Maybe InviteLinkChatType

Type of the chat

title :: Maybe Text

Title of the chat

photo :: Maybe ChatPhotoInfo

Chat photo; may be null

accent_color_id :: Maybe Int

Identifier of the accent color for chat title and background of chat photo

description :: Maybe Text

Chat description

member_count :: Maybe Int

Number of members in the chat

member_user_ids :: Maybe [Int]

User identifiers of some chat members that may be known to the current user

subscription_info :: Maybe ChatInviteLinkSubscriptionInfo

Information about subscription plan that must be paid by the user to use the link; may be null if the link doesn't require subscription

creates_join_request :: Maybe Bool

True, if the link only creates join request

is_public :: Maybe Bool

True, if the chat is a public supergroup or channel, i.e. it has a username or it is a location-based supergroup

verification_status :: Maybe VerificationStatus

Information about verification status of the chat; may be null if none