TD.Data.ChatJoinResult

Documentation

data ChatJoinResult

Describes result of join of a chat by the current user

Constructors

ChatJoinResultSuccess
The chat was joined successfully

Fields

chat_id :: Maybe Int

Identifier of the chat

ChatJoinResultRequestSent
The join request was sent and have to be approved by administrators of the chat

ChatJoinResultGuardBotApprovalRequired
An approval from a guard bot through a Web App is required to join the chat

Fields

bot_user_id :: Maybe Int

Identifier of the guard bot

url :: Maybe WebAppUrl

The URL of the Web App to open

query_id :: Maybe Int

Unique identifier of the join request, which will be used in updateChatJoinResult

ChatJoinResultDeclined
The join was declined by the guard bot