TD.Data.ChatMember

Description

Documentation

data ChatMember

Constructors

ChatMember
Describes a user or a chat as a member of another chat

Fields

status :: Maybe ChatMemberStatus

Status of the member in the chat

joined_chat_date :: Maybe Int

Point in time (Unix timestamp) when the user joined the chat

inviter_user_id :: Maybe Int

Identifier of a user that invitedpromotedbanned this member in the chat; 0 if unknown

member_id :: Maybe MessageSender

Identifier of the chat member. Currently, other chats can be only Left or Banned. Only supergroups and channels can have other chats as Left or Banned members and these chats must be supergroups or channels