TD.Data.ChatMemberStatus

Description

Documentation

data ChatMemberStatus

Provides information about the status of a member in a chat

Constructors

ChatMemberStatusCreator
The user is the owner of the chat and has all the administrator privileges

Fields

is_member :: Maybe Bool

True, if the user is a member of the chat

is_anonymous :: Maybe Bool

True, if the creator isn't shown in the chat member list and sends messages anonymously; applicable to supergroups only

custom_title :: Maybe String

A custom title of the owner; 0-16 characters without emojis; applicable to supergroups only

ChatMemberStatusAdministrator
The user is a member of the chat and has some additional privileges. In basic groups, administrators can edit and delete messages sent by others, add new members, ban unprivileged members, and manage video chats. In supergroups and channels, there are more detailed options for administrator privileges

Fields

rights :: Maybe ChatAdministratorRights

Rights of the administrator

can_be_edited :: Maybe Bool

True, if the current user can edit the administrator privileges for the called user

custom_title :: Maybe String

A custom title of the owner; 0-16 characters without emojis; applicable to supergroups only

ChatMemberStatusMember
The user is a member of the chat, without any additional privileges or restrictions

ChatMemberStatusRestricted
The user is under certain restrictions in the chat. Not supported in basic groups and channels

Fields

permissions :: Maybe ChatPermissions

User permissions in the chat

restricted_until_date :: Maybe Int

Point in time (Unix timestamp) when restrictions will be lifted from the user; 0 if never. If the user is restricted for more than 366 days or for less than 30 seconds from the current time, the user is considered to be restricted forever

is_member :: Maybe Bool

True, if the user is a member of the chat

ChatMemberStatusLeft
The user or the chat is not a chat member

ChatMemberStatusBanned
The user or the chat was banned (and hence is not a member of the chat). Implies the user can't return to the chat, view messages, or be used as a participant identifier to join a video chat of the chat

Fields

banned_until_date :: Maybe Int

Point in time (Unix timestamp) when the user will be unbanned; 0 if never. If the user is banned for more than 366 days or for less than 30 seconds from the current time, the user is considered to be banned forever. Always 0 in basic groups