TD.Data.ChatMessageSender

Description

Documentation

data ChatMessageSender

Constructors

ChatMessageSender
Represents a message sender, which can be used to send messages in a chat sender Available message senders needs_premium True, if Telegram Premium is needed to use the message sender

Fields

needs_premium :: Maybe Bool

sender :: Maybe MessageSender