TD.Data.ChatNotificationSettings

Description

Documentation

data ChatNotificationSettings

Constructors

ChatNotificationSettings
Contains information about notification settings for a chat or a forum topic

Fields

disable_mention_notifications :: Maybe Bool

If true, notifications for messages with mentions will be created as for an ordinary unread message

use_default_disable_mention_notifications :: Maybe Bool

If true, disable_mention_notifications is ignored and the value for the relevant type of chat or the forum chat is used instead

disable_pinned_message_notifications :: Maybe Bool

If true, notifications for incoming pinned messages will be created as for an ordinary unread message

use_default_disable_pinned_message_notifications :: Maybe Bool

If true, disable_pinned_message_notifications is ignored and the value for the relevant type of chat or the forum chat is used instead

show_preview :: Maybe Bool

True, if message content must be displayed in notifications

use_default_show_preview :: Maybe Bool

If true, show_preview is ignored and the value for the relevant type of chat or the forum chat is used instead

sound_id :: Maybe Int

Identifier of the notification sound to be played; 0 if sound is disabled

use_default_sound :: Maybe Bool

If true, the value for the relevant type of chat or the forum chat is used instead of sound_id

mute_for :: Maybe Int

Time left before notifications will be unmuted, in seconds

use_default_mute_for :: Maybe Bool

If true, mute_for is ignored and the value for the relevant type of chat or the forum chat is used instead