TD.Data.ChatPermissions

Documentation

data ChatPermissions

Constructors

ChatPermissions
Describes actions that a user is allowed to take in a chat

Fields

can_send_basic_messages :: Maybe Bool

True, if the user can send text messages, contacts, invoices, locations, and venues

can_send_audios :: Maybe Bool

True, if the user can send music files

can_send_documents :: Maybe Bool

True, if the user can send documents

can_send_photos :: Maybe Bool

True, if the user can send photos

can_send_videos :: Maybe Bool

True, if the user can send videos

can_send_video_notes :: Maybe Bool

True, if the user can send video notes

can_send_voice_notes :: Maybe Bool

True, if the user can send voice notes

can_send_polls :: Maybe Bool

True, if the user can send polls

can_send_other_messages :: Maybe Bool

True, if the user can send animations, games, stickers, and dice and use inline bots

can_add_web_page_previews :: Maybe Bool

True, if the user may add a web page preview to their messages

can_change_info :: Maybe Bool

True, if the user can change the chat title, photo, and other settings

can_invite_users :: Maybe Bool

True, if the user can invite new users to the chat

can_pin_messages :: Maybe Bool

True, if the user can pin messages

can_manage_topics :: Maybe Bool

True, if the user can manage topics

defaultChatPermissions :: ChatPermissions