TD.Data.ChatPhotoStickerType

Description

Documentation

data ChatPhotoStickerType

Describes type of a sticker, which was used to create a chat photo

Constructors

ChatPhotoStickerTypeRegularOrMask
Information about the sticker, which was used to create the chat photo

Fields

sticker_id :: Maybe Int

Identifier of the sticker in the set

sticker_set_id :: Maybe Int

Sticker set identifier

ChatPhotoStickerTypeCustomEmoji
Information about the custom emoji, which was used to create the chat photo

Fields

custom_emoji_id :: Maybe Int

Identifier of the custom emoji