TD.Data.ChatRevenueTransactionType

Documentation

data ChatRevenueTransactionType

Describes type of transaction for revenue earned from sponsored messages in a chat

Constructors

ChatRevenueTransactionTypeEarnings
Describes earnings from sponsored messages in a chat in some time frame

Fields

start_date :: Maybe Int

Point in time (Unix timestamp) when the earnings started

end_date :: Maybe Int

Point in time (Unix timestamp) when the earnings ended

ChatRevenueTransactionTypeWithdrawal
Describes a withdrawal of earnings

Fields

withdrawal_date :: Maybe Int

Point in time (Unix timestamp) when the earnings withdrawal started

provider :: Maybe Text

Name of the payment provider

state :: Maybe ChatRevenueWithdrawalState

State of the withdrawal

ChatRevenueTransactionTypeRefund
Describes a refund for failed withdrawal of earnings

Fields

refund_date :: Maybe Int

Point in time (Unix timestamp) when the transaction was refunded

provider :: Maybe Text

Name of the payment provider