TD.Data.ChatStatistics

Description

Documentation

data ChatStatistics

Contains a detailed statistics about a chat

Constructors

ChatStatisticsSupergroup
A detailed statistics about a supergroup chat

Fields

top_inviters :: Maybe [ChatStatisticsInviterInfo]

List of most active inviters of new members in the last week

top_administrators :: Maybe [ChatStatisticsAdministratorActionsInfo]

List of most active administrators in the last week

top_senders :: Maybe [ChatStatisticsMessageSenderInfo]

List of users sent most messages in the last week

week_graph :: Maybe StatisticalGraph

A graph containing distribution of message views per day of week

day_graph :: Maybe StatisticalGraph

A graph containing distribution of message views per hour

action_graph :: Maybe StatisticalGraph

A graph containing number of different actions in the chat

message_content_graph :: Maybe StatisticalGraph

A graph containing distribution of sent messages by content type

language_graph :: Maybe StatisticalGraph

A graph containing distribution of active users per language

join_by_source_graph :: Maybe StatisticalGraph

A graph containing number of new member joins per source

join_graph :: Maybe StatisticalGraph

A graph containing number of members joined and left the chat

member_count_graph :: Maybe StatisticalGraph

A graph containing number of members in the chat

sender_count :: Maybe StatisticalValue

Number of users who sent messages to the chat

viewer_count :: Maybe StatisticalValue

Number of users who viewed messages in the chat

message_count :: Maybe StatisticalValue

Number of messages sent to the chat

member_count :: Maybe StatisticalValue

Number of members in the chat

period :: Maybe DateRange

A period to which the statistics applies

ChatStatisticsChannel
A detailed statistics about a channel chat

Fields

recent_message_interactions :: Maybe [ChatStatisticsMessageInteractionInfo]

Detailed statistics about number of views and shares of recently sent messages

instant_view_interaction_graph :: Maybe StatisticalGraph

A graph containing number of views of associated with the chat instant views

message_interaction_graph :: Maybe StatisticalGraph

A graph containing number of chat message views and shares

language_graph :: Maybe StatisticalGraph

A graph containing distribution of active users per language

join_by_source_graph :: Maybe StatisticalGraph

A graph containing number of new member joins per source

view_count_by_source_graph :: Maybe StatisticalGraph

A graph containing number of message views per source

view_count_by_hour_graph :: Maybe StatisticalGraph

A graph containing number of message views in a given hour in the last two weeks

mute_graph :: Maybe StatisticalGraph

A graph containing number of members muted and unmuted the chat

join_graph :: Maybe StatisticalGraph

A graph containing number of members joined and left the chat

member_count_graph :: Maybe StatisticalGraph

A graph containing number of members in the chat

enabled_notifications_percentage :: Maybe Float

A percentage of users with enabled notifications for the chat

mean_share_count :: Maybe StatisticalValue

Mean number of times the recently sent messages was shared

mean_view_count :: Maybe StatisticalValue

Mean number of times the recently sent messages was viewed

member_count :: Maybe StatisticalValue

Number of members in the chat

period :: Maybe DateRange

A period to which the statistics applies