TD.Data.ChatTransactionPurpose

Documentation

data ChatTransactionPurpose

Describes purpose of a transaction with a supergroup or a channel

Constructors

ChatTransactionPurposePaidMedia
Paid media were bought

Fields

message_id :: Maybe Int

Identifier of the corresponding message with paid media; can be 0 or an identifier of a deleted message

media :: Maybe [PaidMedia]

The bought media if the transaction wasn't refunded

ChatTransactionPurposeJoin
User joined the channel and subscribed to regular payments in Telegram Stars

Fields

period :: Maybe Int

The number of seconds between consecutive Telegram Star debiting

ChatTransactionPurposeReaction
User paid for a reaction

Fields

message_id :: Maybe Int

Identifier of the corresponding message with paid media; can be 0 or an identifier of a deleted message

ChatTransactionPurposeGiveaway
User received Telegram Stars from a giveaway

Fields

giveaway_message_id :: Maybe Int

Identifier of the message with giveaway; can be 0 or an identifier of a deleted message