TD.Data.CommunityAdministratorRights

Documentation

data CommunityAdministratorRights

Constructors

CommunityAdministratorRights
Describes rights of the administrator in a community

Fields

can_manage_community :: Maybe Bool

True, if the user is an administrator. Implied by any other privilege

can_change_info :: Maybe Bool

True, if the administrator can change the community name, photo, and other settings

can_edit_chat_list :: Maybe Bool

True, if the user can change the chats added to the community

can_promote_members :: Maybe Bool

True, if the administrator can add new administrators with a subset of their own privileges or demote administrators that were directly or indirectly promoted by them

can_ban_members :: Maybe Bool

True, if the administrator can ban, or unban community members