TD.Data.CommunityMemberStatus

Documentation

data CommunityMemberStatus

Provides information about the status of a member in a community

Constructors

CommunityMemberStatusCreator
The user is the owner of the community and has all the administrator privileges

CommunityMemberStatusAdministrator
The user is a member of the community and has some additional privileges

Fields

can_be_edited :: Maybe Bool

True, if the current user can edit the administrator privileges for the called user

rights :: Maybe CommunityAdministratorRights

Rights of the administrator

CommunityMemberStatusMember
The user is a member of the community, without any additional privileges or restrictions

CommunityMemberStatusLeft
The user or the chat is not a community member

CommunityMemberStatusBanned
The user or the chat was banned in the community; implies ban in all chats in the community