TD.Data.ConnectedAffiliateProgram

Documentation

data ConnectedAffiliateProgram

Constructors

ConnectedAffiliateProgram
Describes an affiliate program that was connected to an affiliate

Fields

url :: Maybe Text

The link that can be used to refer users if the program is still active

bot_user_id :: Maybe Int

User identifier of the bot created the program

parameters :: Maybe AffiliateProgramParameters

The parameters of the affiliate program

connection_date :: Maybe Int

Point in time (Unix timestamp) when the affiliate program was connected

is_disconnected :: Maybe Bool

True, if the program was canceled by the bot, or disconnected by the chat owner and isn't available anymore

user_count :: Maybe Int

The number of users that used the affiliate program

revenue_star_count :: Maybe Int

The number of Telegram Stars that were earned by the affiliate program