TD.Data.ConnectedWebsite

Documentation

data ConnectedWebsite

Constructors

ConnectedWebsite
Contains information about one website the current user is logged in with Telegram

Fields

_id :: Maybe Int

Website identifier

domain_name :: Maybe Text

The domain name of the website

bot_user_id :: Maybe Int

User identifier of a bot linked with the website

browser :: Maybe Text

The version of a browser used to log in

platform :: Maybe Text

Operating system the browser is running on

log_in_date :: Maybe Int

Point in time (Unix timestamp) when the user was logged in

last_active_date :: Maybe Int

Point in time (Unix timestamp) when obtained authorization was last used

ip :: Maybe Text

IP address from which the user was logged in, in human-readable format

location :: Maybe Text

Human-readable description of a country and a region from which the user was logged in, based on the IP address