TD.Data.DeviceToken

Description

Documentation

data DeviceToken

Represents a data needed to subscribe for push notifications through registerDevice method. To use specific push notification service, the correct application platform must be specified and a valid server authentication data must be uploaded at https://my.telegram.org

Constructors

DeviceTokenFirebaseCloudMessaging
A token for Firebase Cloud Messaging token Device registration token; may be empty to deregister a device encrypt True, if push notifications must be additionally encrypted

Fields

encrypt :: Maybe Bool

token :: Maybe String

DeviceTokenApplePush
A token for Apple Push Notification service device_token Device token; may be empty to deregister a device is_app_sandbox True, if App Sandbox is enabled

Fields

is_app_sandbox :: Maybe Bool

device_token :: Maybe String

DeviceTokenApplePushVoIP
A token for Apple Push Notification service VoIP notifications device_token Device token; may be empty to deregister a device is_app_sandbox True, if App Sandbox is enabled @encrypt True, if push notifications must be additionally encrypted

Fields

encrypt :: Maybe Bool

is_app_sandbox :: Maybe Bool

device_token :: Maybe String

DeviceTokenWindowsPush
A token for Windows Push Notification Services @access_token The access token that will be used to send notifications; may be empty to deregister a device

Fields

access_token :: Maybe String

DeviceTokenMicrosoftPush
A token for Microsoft Push Notification Service @channel_uri Push notification channel URI; may be empty to deregister a device

Fields

channel_uri :: Maybe String

DeviceTokenMicrosoftPushVoIP
A token for Microsoft Push Notification Service VoIP channel @channel_uri Push notification channel URI; may be empty to deregister a device

Fields

channel_uri :: Maybe String

DeviceTokenWebPush
A token for web Push API @endpoint Absolute URL exposed by the push service where the application server can send push messages; may be empty to deregister a device

Fields

auth_base64url :: Maybe String

p256dh_base64url :: Maybe String

Base64url-encoded P-256 elliptic curve Diffie-Hellman public key @auth_base64url Base64url-encoded authentication secret

endpoint :: Maybe String

DeviceTokenSimplePush
A token for Simple Push API for Firefox OS @endpoint Absolute URL exposed by the push service where the application server can send push messages; may be empty to deregister a device

Fields

endpoint :: Maybe String

DeviceTokenUbuntuPush
A token for Ubuntu Push Client service @token Token; may be empty to deregister a device

Fields

token :: Maybe String

DeviceTokenBlackBerryPush
A token for BlackBerry Push Service @token Token; may be empty to deregister a device

Fields

token :: Maybe String

DeviceTokenTizenPush
A token for Tizen Push Service @reg_id Push service registration identifier; may be empty to deregister a device

Fields

reg_id :: Maybe String