TD.Data.EmailAddressResetState

Description

Documentation

data EmailAddressResetState

Describes reset state of a email address

Constructors

EmailAddressResetStateAvailable
Email address can be reset after the given period. Call resetAuthenticationEmailAddress to reset it and allow the user to authorize with a code sent to the user's phone number

Fields

wait_period :: Maybe Int

Time required to wait before the email address can be reset; 0 if the user is subscribed to Telegram Premium

EmailAddressResetStatePending
Email address reset has already been requested. Call resetAuthenticationEmailAddress to check whether immediate reset is possible

Fields

reset_in :: Maybe Int

Left time before the email address will be reset, in seconds. updateAuthorizationState is not sent when this field changes