TD.Data.EmojiStatusType

Documentation

data EmojiStatusType

Describes type of emoji status

Constructors

EmojiStatusTypeCustomEmoji
A custom emoji set as emoji status

Fields

custom_emoji_id :: Maybe Int

Identifier of the custom emoji in stickerFormatTgs format

EmojiStatusTypeUpgradedGift
An upgraded gift set as emoji status

Fields

upgraded_gift_id :: Maybe Int

Identifier of the upgraded gift

gift_title :: Maybe Text

The title of the upgraded gift

gift_name :: Maybe Text

Unique name of the upgraded gift that can be used with internalLinkTypeUpgradedGift

model_custom_emoji_id :: Maybe Int

Custom emoji identifier of the model of the upgraded gift

symbol_custom_emoji_id :: Maybe Int

Custom emoji identifier of the symbol of the upgraded gift

backdrop_colors :: Maybe UpgradedGiftBackdropColors

Colors of the backdrop of the upgraded gift