TD.Data.FileType

Description

Documentation

data FileType

Represents the type of a file

Constructors

FileTypeNone
The data is not a file

FileTypeAnimation
The file is an animation

FileTypeAudio
The file is an audio file

FileTypeDocument
The file is a document

FileTypeNotificationSound
The file is a notification sound

FileTypePhoto
The file is a photo

FileTypeProfilePhoto
The file is a profile photo

FileTypeSecret
The file was sent to a secret chat (the file type is not known to the server)

FileTypeSecretThumbnail
The file is a thumbnail of a file from a secret chat

FileTypeSecure
The file is a file from Secure storage used for storing Telegram Passport files

FileTypeSticker
The file is a sticker

FileTypeThumbnail
The file is a thumbnail of another file

FileTypeUnknown
The file type is not yet known

FileTypeVideo
The file is a video

FileTypeVideoNote
The file is a video note

FileTypeVoiceNote
The file is a voice note

FileTypeWallpaper
The file is a wallpaper or a background pattern