TD.Data.ForumTopicInfo

Documentation

data ForumTopicInfo

Constructors

ForumTopicInfo
Contains basic information about a forum topic

Fields

message_thread_id :: Maybe Int

Message thread identifier of the topic

name :: Maybe Text

Name of the topic

icon :: Maybe ForumTopicIcon

Icon of the topic

creation_date :: Maybe Int

Point in time (Unix timestamp) when the topic was created

creator_id :: Maybe MessageSender

Identifier of the creator of the topic

is_general :: Maybe Bool

True, if the topic is the General topic list

is_outgoing :: Maybe Bool

True, if the topic was created by the current user

is_closed :: Maybe Bool

True, if the topic is closed

is_hidden :: Maybe Bool

True, if the topic is hidden above the topic list and closed; for General topic only