TD.Data.FoundWebApp

Documentation

data FoundWebApp

Constructors

FoundWebApp
Contains information about a Web App found by its short name

Fields

web_app :: Maybe WebApp

The Web App

request_write_access :: Maybe Bool

True, if the user must be asked for the permission to the bot to send them messages

skip_confirmation :: Maybe Bool

True, if there is no need to show an ordinary open URL confirmation before opening the Web App. The field must be ignored and confirmation must be shown anyway if the Web App link was hidden