TD.Data.GiftAuctionAcquiredGift

Documentation

data GiftAuctionAcquiredGift

Constructors

GiftAuctionAcquiredGift
Represents a gift that was acquired by the current user on an auction

Fields

receiver_id :: Maybe MessageSender

Receiver of the gift

date :: Maybe Int

Point in time (Unix timestamp) when the gift was acquired

star_count :: Maybe Int

The number of Telegram Stars that were paid for the gift

auction_round_number :: Maybe Int

Identifier of the auction round in which the gift was acquired

auction_round_position :: Maybe Int

Position of the user in the round among all auction participants

unique_gift_number :: Maybe Int

Unique number of the gift among gifts upgraded from the same gift after upgrade; 0 if yet unassigned

text :: Maybe FormattedText

Message added to the gift

is_private :: Maybe Bool

True, if the sender and gift text are shown only to the gift receiver; otherwise, everyone will be able to see them