TD.Data.GiftForResale

Documentation

data GiftForResale

Constructors

GiftForResale
Describes a gift available for resale

Fields

gift :: Maybe UpgradedGift

The gift

received_gift_id :: Maybe Text

Unique identifier of the received gift for the current user; only for the gifts owned by the current user