TD.Data.GiveawayInfo

Documentation

data GiveawayInfo

Contains information about a giveaway

Constructors

GiveawayInfoOngoing
Describes an ongoing giveaway

Fields

creation_date :: Maybe Int

Point in time (Unix timestamp) when the giveaway was created

status :: Maybe GiveawayParticipantStatus

Status of the current user in the giveaway

is_ended :: Maybe Bool

True, if the giveaway has ended and results are being prepared

GiveawayInfoCompleted
Describes a completed giveaway

Fields

creation_date :: Maybe Int

Point in time (Unix timestamp) when the giveaway was created

actual_winners_selection_date :: Maybe Int

Point in time (Unix timestamp) when the winners were selected. May be bigger than winners selection date specified in parameters of the giveaway

was_refunded :: Maybe Bool

True, if the giveaway was canceled and was fully refunded

is_winner :: Maybe Bool

True, if the current user is a winner of the giveaway

winner_count :: Maybe Int

Number of winners in the giveaway

activation_count :: Maybe Int

Number of winners, which activated their gift codes; for Telegram Premium giveaways only

gift_code :: Maybe Text

Telegram Premium gift code that was received by the current user; empty if the user isn't a winner in the giveaway or the giveaway isn't a Telegram Premium giveaway

won_star_count :: Maybe Int

The amount of Telegram Stars won by the current user; 0 if the user isn't a winner in the giveaway or the giveaway isn't a Telegram Star giveaway