TD.Data.GroupCall

Description

Documentation

data GroupCall

Constructors

GroupCall
Describes a group call

Fields

duration :: Maybe Int

Call duration, in seconds; for ended calls only

is_video_recorded :: Maybe Bool

True, if a video file is being recorded for the call

record_duration :: Maybe Int

Duration of the ongoing group call recording, in seconds; 0 if none. An updateGroupCall update is not triggered when value of this field changes, but the same recording goes on

can_toggle_mute_new_participants :: Maybe Bool

True, if the current user can enable or disable mute_new_participants setting

mute_new_participants :: Maybe Bool

True, if only group call administrators can unmute new participants

can_enable_video :: Maybe Bool

True, if the current user can broadcast video or share screen

is_my_video_paused :: Maybe Bool

True, if the current user's video is paused

is_my_video_enabled :: Maybe Bool

True, if the current user's video is enabled

recent_speakers :: Maybe [GroupCallRecentSpeaker]

At most 3 recently speaking users in the group call

loaded_all_participants :: Maybe Bool

True, if all group call participants are loaded

has_hidden_listeners :: Maybe Bool

True, if group call participants, which are muted, aren't returned in participant list

participant_count :: Maybe Int

Number of participants in the group call

can_be_managed :: Maybe Bool

True, if the current user can manage the group call

need_rejoin :: Maybe Bool

True, if user was kicked from the call because of network loss and the call needs to be rejoined

is_joined :: Maybe Bool

True, if the call is joined

is_rtmp_stream :: Maybe Bool

True, if the chat is an RTMP stream instead of an ordinary video chat

is_active :: Maybe Bool

True, if the call is active

enabled_start_notification :: Maybe Bool

True, if the group call is scheduled and the current user will receive a notification when the group call will start

scheduled_start_date :: Maybe Int

Point in time (Unix timestamp) when the group call is supposed to be started by an administrator; 0 if it is already active or was ended

title :: Maybe String

Group call title

_id :: Maybe Int

Group call identifier