TD.Data.GroupCallMessageLevel

Documentation

data GroupCallMessageLevel

Constructors

GroupCallMessageLevel
Represents a level of features for a message sent in a live story group call

Fields

min_star_count :: Maybe Int

The minimum number of Telegram Stars required to get features of the level

pin_duration :: Maybe Int

The amount of time the message of this level will be pinned, in seconds

max_text_length :: Maybe Int

The maximum allowed length of the message text

max_custom_emoji_count :: Maybe Int

The maximum allowed number of custom emoji in the message text

first_color :: Maybe Int

The first color used to show the message text in the RGB format

second_color :: Maybe Int

The second color used to show the message text in the RGB format

background_color :: Maybe Int

Background color for the message the RGB format