TD.Data.GroupCallParticipant

Description

Documentation

data GroupCallParticipant

Constructors

GroupCallParticipant
Represents a group call participant

Fields

order :: Maybe String

User's order in the group call participant list. Orders must be compared lexicographically. The bigger is order, the higher is user in the list. If order is empty, the user must be removed from the participant list

volume_level :: Maybe Int

Participant's volume level; 1-20000 in hundreds of percents

can_unmute_self :: Maybe Bool

True, if the participant is muted for all users, but can unmute themselves

is_muted_for_current_user :: Maybe Bool

True, if the participant is muted for the current user

is_muted_for_all_users :: Maybe Bool

True, if the participant is muted for all users

can_be_unmuted_for_current_user :: Maybe Bool

True, if the current user can unmute the participant for self

can_be_muted_for_current_user :: Maybe Bool

True, if the current user can mute the participant only for self

can_be_unmuted_for_all_users :: Maybe Bool

True, if the current user can allow the participant to unmute themselves or unmute the participant (if the participant is the current user)

can_be_muted_for_all_users :: Maybe Bool

True, if the current user can mute the participant for all other group call participants

is_hand_raised :: Maybe Bool

True, if the participant hand is raised

is_speaking :: Maybe Bool

True, if the participant is speaking as set by setGroupCallParticipantIsSpeaking

is_current_user :: Maybe Bool

True, if the participant is the current user

bio :: Maybe String

The participant user's bio or the participant chat's description

screen_sharing_video_info :: Maybe GroupCallParticipantVideoInfo

Information about user's screen sharing video channel; may be null if there is no active screen sharing video

video_info :: Maybe GroupCallParticipantVideoInfo

Information about user's video channel; may be null if there is no active video

screen_sharing_audio_source_id :: Maybe Int

User's screen sharing audio channel synchronization source identifier

audio_source_id :: Maybe Int

User's audio channel synchronization source identifier

participant_id :: Maybe MessageSender

Identifier of the group call participant