TD.Data.GroupCallParticipantVideoInfo

Description

Documentation

data GroupCallParticipantVideoInfo

Constructors

GroupCallParticipantVideoInfo
Contains information about a group call participant's video channel source_groups List of synchronization source groups of the video endpoint_id Video channel endpoint identifier

Fields

is_paused :: Maybe Bool

True if the video is paused. This flag needs to be ignored, if new video frames are received

endpoint_id :: Maybe String

source_groups :: Maybe [GroupCallVideoSourceGroup]