TD.Data.ImportedContacts

Documentation

data ImportedContacts

Constructors

ImportedContacts
Represents the result of an importContacts request

Fields

user_ids :: Maybe [Int]

User identifiers of the imported contacts in the same order as they were specified in the request; 0 if the contact is not yet a registered user

importer_count :: Maybe [Int]

The number of users that imported the corresponding contact; 0 for already registered users or if unavailable