TD.Data.InlineKeyboardButtonType

Description

Documentation

data InlineKeyboardButtonType

Describes the type of an inline keyboard button

Constructors

InlineKeyboardButtonTypeUrl
A button that opens a specified URL @url HTTP or tg:// URL to open

Fields

url :: Maybe String

InlineKeyboardButtonTypeLoginUrl
A button that opens a specified URL and automatically authorize the current user by calling getLoginUrlInfo url An HTTP URL to pass to getLoginUrlInfo id Unique button identifier @forward_text If non-empty, new text of the button in forwarded messages

Fields

forward_text :: Maybe String

_id :: Maybe Int

url :: Maybe String

InlineKeyboardButtonTypeWebApp
A button that opens a Web App by calling openWebApp @url An HTTP URL to pass to openWebApp

Fields

url :: Maybe String

InlineKeyboardButtonTypeCallback
A button that sends a callback query to a bot @data Data to be sent to the bot via a callback query

Fields

_data :: Maybe String

InlineKeyboardButtonTypeCallbackWithPassword
A button that asks for the 2-step verification password of the current user and then sends a callback query to a bot @data Data to be sent to the bot via a callback query

Fields

_data :: Maybe String

InlineKeyboardButtonTypeCallbackGame
A button with a game that sends a callback query to a bot. This button must be in the first column and row of the keyboard and can be attached only to a message with content of the type messageGame

InlineKeyboardButtonTypeSwitchInline
A button that forces an inline query to the bot to be inserted in the input field query Inline query to be sent to the bot in_current_chat True, if the inline query must be sent from the current chat

Fields

in_current_chat :: Maybe Bool

query :: Maybe String

InlineKeyboardButtonTypeBuy
A button to buy something. This button must be in the first column and row of the keyboard and can be attached only to a message with content of the type messageInvoice

InlineKeyboardButtonTypeUser
A button with a user reference to be handled in the same way as textEntityTypeMentionName entities @user_id User identifier

Fields

user_id :: Maybe Int