TD.Data.InlineQueryResult

Description

Documentation

data InlineQueryResult

Represents a single result of an inline query

Constructors

InlineQueryResultArticle
Represents a link to an article or web page id Unique identifier of the query result url URL of the result, if it exists hide_url True, if the URL must be not shown title Title of the result

Fields

thumbnail :: Maybe Thumbnail

description :: Maybe String

title :: Maybe String

hide_url :: Maybe Bool

url :: Maybe String

_id :: Maybe String

InlineQueryResultContact
Represents a user contact id Unique identifier of the query result contact A user contact @thumbnail Result thumbnail in JPEG format; may be null

Fields

thumbnail :: Maybe Thumbnail

contact :: Maybe Contact

_id :: Maybe String

InlineQueryResultLocation
Represents a point on the map id Unique identifier of the query result location Location result title Title of the result thumbnail Result thumbnail in JPEG format; may be null

Fields

thumbnail :: Maybe Thumbnail

title :: Maybe String

location :: Maybe Location

_id :: Maybe String

InlineQueryResultVenue
Represents information about a venue id Unique identifier of the query result venue Venue result @thumbnail Result thumbnail in JPEG format; may be null

Fields

thumbnail :: Maybe Thumbnail

venue :: Maybe Venue

_id :: Maybe String

InlineQueryResultGame
Represents information about a game id Unique identifier of the query result game Game result

Fields

game :: Maybe Game

_id :: Maybe String

InlineQueryResultAnimation
Represents an animation file id Unique identifier of the query result animation Animation file @title Animation title

Fields

title :: Maybe String

animation :: Maybe Animation

_id :: Maybe String

InlineQueryResultAudio
Represents an audio file id Unique identifier of the query result audio Audio file

Fields

audio :: Maybe Audio

_id :: Maybe String

InlineQueryResultDocument
Represents a document id Unique identifier of the query result document Document title Document title param_description Document description

Fields

description :: Maybe String

title :: Maybe String

document :: Maybe Document

_id :: Maybe String

InlineQueryResultPhoto
Represents a photo id Unique identifier of the query result photo Photo title Title of the result, if known param_description A short description of the result, if known

Fields

description :: Maybe String

title :: Maybe String

photo :: Maybe Photo

_id :: Maybe String

InlineQueryResultSticker
Represents a sticker id Unique identifier of the query result sticker Sticker

Fields

sticker :: Maybe Sticker

_id :: Maybe String

InlineQueryResultVideo
Represents a video id Unique identifier of the query result video Video title Title of the video param_description Description of the video

Fields

description :: Maybe String

title :: Maybe String

video :: Maybe Video

_id :: Maybe String

InlineQueryResultVoiceNote
Represents a voice note id Unique identifier of the query result voice_note Voice note @title Title of the voice note

Fields

title :: Maybe String

voice_note :: Maybe VoiceNote

_id :: Maybe String