TD.Data.InputFile

Description

Documentation

data InputFile

Points to a file

Constructors

InputFileId
A file defined by its unique ID @id Unique file identifier

Fields

_id :: Maybe Int

InputFileRemote
A file defined by its remote ID. The remote ID is guaranteed to be usable only if the corresponding file is still accessible to the user and known to TDLib. For example, if the file is from a message, then the message must be not deleted and accessible to the user. If the file database is disabled, then the corresponding object with the file must be preloaded by the application

Fields

__id :: Maybe String

Remote file identifier

InputFileLocal
A file defined by a local path @path Local path to the file

Fields

path :: Maybe String

InputFileGenerated
A file generated by the application

Fields

expected_size :: Maybe Int

Expected size of the generated file, in bytes; 0 if unknown

conversion :: Maybe String

String specifying the conversion applied to the original file; must be persistent across application restarts. Conversions beginning with # are reserved for internal TDLib usage

original_path :: Maybe String

Local path to a file from which the file is generated; may be empty if there is no such file