TD.Data.InputIdentityDocument

Description

Documentation

data InputIdentityDocument

Constructors

InputIdentityDocument
An identity document to be saved to Telegram Passport number Document number; 1-24 characters expiry_date Document expiry date; pass null if not applicable @front_side Front side of the document

Fields

translation :: Maybe [InputFile]

selfie :: Maybe InputFile

reverse_side :: Maybe InputFile

Reverse side of the document; only for driver license and identity card; pass null otherwise selfie Selfie with the document; pass null if unavailable translation List of files containing a certified English translation of the document

front_side :: Maybe InputFile

expiry_date :: Maybe Date

number :: Maybe String