TD.Data.InputInlineQueryResult

Description

Documentation

data InputInlineQueryResult

Represents a single result of an inline query; for bots only

Constructors

InputInlineQueryResultAnimation
Represents a link to an animated GIF or an animated (i.e., without sound) H.264/MPEG-4 AVC video

Fields

input_message_content :: Maybe InputMessageContent

The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageAnimation, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

reply_markup :: Maybe ReplyMarkup

The message reply markup; pass null if none. Must be of type replyMarkupInlineKeyboard or null

video_height :: Maybe Int

video_width :: Maybe Int

video_duration :: Maybe Int

Duration of the video, in seconds video_width Width of the video video_height Height of the video

video_mime_type :: Maybe String

video_url :: Maybe String

The URL of the video file (file size must not exceed 1MB) @video_mime_type MIME type of the video file. Must be one of "imagegif" and "videomp4"

thumbnail_mime_type :: Maybe String

thumbnail_url :: Maybe String

URL of the result thumbnail (JPEG, GIF, or MPEG4), if it exists @thumbnail_mime_type MIME type of the video thumbnail. If non-empty, must be one of "imagejpeg", "imagegif" and "video/mp4"

title :: Maybe String

_id :: Maybe String

Unique identifier of the query result @title Title of the query result

InputInlineQueryResultArticle
Represents a link to an article or web page id Unique identifier of the query result url URL of the result, if it exists hide_url True, if the URL must be not shown title Title of the result

Fields

input_message_content :: Maybe InputMessageContent

The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageAnimation, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

reply_markup :: Maybe ReplyMarkup

The message reply markup; pass null if none. Must be of type replyMarkupInlineKeyboard or null

thumbnail_height :: Maybe Int

thumbnail_width :: Maybe Int

thumbnail_url :: Maybe String

URL of the result thumbnail (JPEG, GIF, or MPEG4), if it exists @thumbnail_mime_type MIME type of the video thumbnail. If non-empty, must be one of "imagejpeg", "imagegif" and "video/mp4"

description :: Maybe String

title :: Maybe String

hide_url :: Maybe Bool

url :: Maybe String

_id :: Maybe String

Unique identifier of the query result @title Title of the query result

InputInlineQueryResultAudio
Represents a link to an MP3 audio file id Unique identifier of the query result title Title of the audio file @performer Performer of the audio file

Fields

input_message_content :: Maybe InputMessageContent

The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageAnimation, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

reply_markup :: Maybe ReplyMarkup

The message reply markup; pass null if none. Must be of type replyMarkupInlineKeyboard or null

audio_duration :: Maybe Int

audio_url :: Maybe String

The URL of the audio file @audio_duration Audio file duration, in seconds

performer :: Maybe String

title :: Maybe String

_id :: Maybe String

Unique identifier of the query result @title Title of the query result

InputInlineQueryResultContact
Represents a user contact id Unique identifier of the query result contact User contact thumbnail_url URL of the result thumbnail, if it exists thumbnail_width Thumbnail width, if known @thumbnail_height Thumbnail height, if known

Fields

input_message_content :: Maybe InputMessageContent

The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageAnimation, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

reply_markup :: Maybe ReplyMarkup

The message reply markup; pass null if none. Must be of type replyMarkupInlineKeyboard or null

thumbnail_height :: Maybe Int

thumbnail_width :: Maybe Int

thumbnail_url :: Maybe String

URL of the result thumbnail (JPEG, GIF, or MPEG4), if it exists @thumbnail_mime_type MIME type of the video thumbnail. If non-empty, must be one of "imagejpeg", "imagegif" and "video/mp4"

contact :: Maybe Contact

_id :: Maybe String

Unique identifier of the query result @title Title of the query result

InputInlineQueryResultDocument
Represents a link to a file id Unique identifier of the query result title Title of the resulting file param_description Short description of the result, if known document_url URL of the file @mime_type MIME type of the file content; only "applicationpdf" and "applicationzip" are currently allowed

Fields

input_message_content :: Maybe InputMessageContent

The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageAnimation, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

reply_markup :: Maybe ReplyMarkup

The message reply markup; pass null if none. Must be of type replyMarkupInlineKeyboard or null

thumbnail_height :: Maybe Int

thumbnail_width :: Maybe Int

thumbnail_url :: Maybe String

URL of the result thumbnail (JPEG, GIF, or MPEG4), if it exists @thumbnail_mime_type MIME type of the video thumbnail. If non-empty, must be one of "imagejpeg", "imagegif" and "video/mp4"

mime_type :: Maybe String

document_url :: Maybe String

description :: Maybe String

title :: Maybe String

_id :: Maybe String

Unique identifier of the query result @title Title of the query result

InputInlineQueryResultGame
Represents a game id Unique identifier of the query result game_short_name Short name of the game @reply_markup The message reply markup; pass null if none. Must be of type replyMarkupInlineKeyboard or null

Fields

reply_markup :: Maybe ReplyMarkup

The message reply markup; pass null if none. Must be of type replyMarkupInlineKeyboard or null

game_short_name :: Maybe String

_id :: Maybe String

Unique identifier of the query result @title Title of the query result

InputInlineQueryResultLocation
Represents a point on the map id Unique identifier of the query result location Location result

Fields

input_message_content :: Maybe InputMessageContent

The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageAnimation, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

reply_markup :: Maybe ReplyMarkup

The message reply markup; pass null if none. Must be of type replyMarkupInlineKeyboard or null

thumbnail_height :: Maybe Int

thumbnail_width :: Maybe Int

thumbnail_url :: Maybe String

URL of the result thumbnail (JPEG, GIF, or MPEG4), if it exists @thumbnail_mime_type MIME type of the video thumbnail. If non-empty, must be one of "imagejpeg", "imagegif" and "video/mp4"

title :: Maybe String

live_period :: Maybe Int

Amount of time relative to the message sent time until the location can be updated, in seconds

location :: Maybe Location

_id :: Maybe String

Unique identifier of the query result @title Title of the query result

InputInlineQueryResultPhoto
Represents link to a JPEG image id Unique identifier of the query result title Title of the result, if known param_description A short description of the result, if known thumbnail_url URL of the photo thumbnail, if it exists

Fields

input_message_content :: Maybe InputMessageContent

The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageAnimation, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

reply_markup :: Maybe ReplyMarkup

The message reply markup; pass null if none. Must be of type replyMarkupInlineKeyboard or null

photo_height :: Maybe Int

photo_width :: Maybe Int

photo_url :: Maybe String

The URL of the JPEG photo (photo size must not exceed 5MB) photo_width Width of the photo photo_height Height of the photo

thumbnail_url :: Maybe String

URL of the result thumbnail (JPEG, GIF, or MPEG4), if it exists @thumbnail_mime_type MIME type of the video thumbnail. If non-empty, must be one of "imagejpeg", "imagegif" and "video/mp4"

description :: Maybe String

title :: Maybe String

_id :: Maybe String

Unique identifier of the query result @title Title of the query result

InputInlineQueryResultSticker
Represents a link to a WEBP, TGS, or WEBM sticker id Unique identifier of the query result thumbnail_url URL of the sticker thumbnail, if it exists

Fields

input_message_content :: Maybe InputMessageContent

The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageAnimation, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

reply_markup :: Maybe ReplyMarkup

The message reply markup; pass null if none. Must be of type replyMarkupInlineKeyboard or null

sticker_height :: Maybe Int

sticker_width :: Maybe Int

sticker_url :: Maybe String

The URL of the WEBP, TGS, or WEBM sticker (sticker file size must not exceed 5MB) sticker_width Width of the sticker sticker_height Height of the sticker

thumbnail_url :: Maybe String

URL of the result thumbnail (JPEG, GIF, or MPEG4), if it exists @thumbnail_mime_type MIME type of the video thumbnail. If non-empty, must be one of "imagejpeg", "imagegif" and "video/mp4"

_id :: Maybe String

Unique identifier of the query result @title Title of the query result

InputInlineQueryResultVenue
Represents information about a venue id Unique identifier of the query result venue Venue result thumbnail_url URL of the result thumbnail, if it exists thumbnail_width Thumbnail width, if known @thumbnail_height Thumbnail height, if known

Fields

input_message_content :: Maybe InputMessageContent

The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageAnimation, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

reply_markup :: Maybe ReplyMarkup

The message reply markup; pass null if none. Must be of type replyMarkupInlineKeyboard or null

thumbnail_height :: Maybe Int

thumbnail_width :: Maybe Int

thumbnail_url :: Maybe String

URL of the result thumbnail (JPEG, GIF, or MPEG4), if it exists @thumbnail_mime_type MIME type of the video thumbnail. If non-empty, must be one of "imagejpeg", "imagegif" and "video/mp4"

venue :: Maybe Venue

_id :: Maybe String

Unique identifier of the query result @title Title of the query result

InputInlineQueryResultVideo
Represents a link to a page containing an embedded video player or a video file id Unique identifier of the query result title Title of the result @param_description A short description of the result, if known

Fields

input_message_content :: Maybe InputMessageContent

The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageAnimation, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

reply_markup :: Maybe ReplyMarkup

The message reply markup; pass null if none. Must be of type replyMarkupInlineKeyboard or null

video_duration :: Maybe Int

Duration of the video, in seconds video_width Width of the video video_height Height of the video

video_height :: Maybe Int

video_width :: Maybe Int

mime_type :: Maybe String

video_url :: Maybe String

The URL of the video file (file size must not exceed 1MB) @video_mime_type MIME type of the video file. Must be one of "imagegif" and "videomp4"

thumbnail_url :: Maybe String

URL of the result thumbnail (JPEG, GIF, or MPEG4), if it exists @thumbnail_mime_type MIME type of the video thumbnail. If non-empty, must be one of "imagejpeg", "imagegif" and "video/mp4"

description :: Maybe String

title :: Maybe String

_id :: Maybe String

Unique identifier of the query result @title Title of the query result

InputInlineQueryResultVoiceNote
Represents a link to an opus-encoded audio file within an OGG container, single channel audio id Unique identifier of the query result title Title of the voice note

Fields

input_message_content :: Maybe InputMessageContent

The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageAnimation, inputMessageInvoice, inputMessageLocation, inputMessageVenue or inputMessageContact

reply_markup :: Maybe ReplyMarkup

The message reply markup; pass null if none. Must be of type replyMarkupInlineKeyboard or null

voice_note_duration :: Maybe Int

voice_note_url :: Maybe String

The URL of the voice note file @voice_note_duration Duration of the voice note, in seconds

title :: Maybe String

_id :: Maybe String

Unique identifier of the query result @title Title of the query result