TD.Data.InputInvoice

Documentation

data InputInvoice

Describes an invoice to process

Constructors

InputInvoiceMessage
An invoice from a message of the type messageInvoice

Fields

chat_id :: Maybe Int

Chat identifier of the message

message_id :: Maybe Int

Message identifier

InputInvoiceName
An invoice from a link of the type internalLinkTypeInvoice

Fields

name :: Maybe Text

Name of the invoice

InputInvoiceTelegram
An invoice for a payment toward Telegram; must not be used in the in-store apps

Fields

purpose :: Maybe TelegramPaymentPurpose

Transaction purpose