TD.Data.InputMessageContent

Description

Documentation

data InputMessageContent

The content of a message to send

Constructors

InputMessageText
A text message

Fields

clear_draft :: Maybe Bool

True, if a chat message draft must be deleted

disable_web_page_preview :: Maybe Bool

True, if rich web page previews for URLs in the message text must be disabled

text :: Maybe FormattedText

Formatted text to be sent; 1-getOption("message_text_length_max") characters. Only Bold, Italic, Underline, Strikethrough, Spoiler, CustomEmoji, Code, Pre, PreCode, TextUrl and MentionName entities are allowed to be specified manually

InputMessageAnimation
An animation message (GIF-style).

Fields

has_spoiler :: Maybe Bool

True, if the animation preview must be covered by a spoiler animation; not supported in secret chats

caption :: Maybe FormattedText

Animation caption; pass null to use an empty caption; 0-getOption("message_caption_length_max") characters

height :: Maybe Int

Height of the animation; may be replaced by the server

width :: Maybe Int

Width of the animation; may be replaced by the server

duration :: Maybe Int

Duration of the animation, in seconds

added_sticker_file_ids :: Maybe [Int]

File identifiers of the stickers added to the animation, if applicable

thumbnail :: Maybe InputThumbnail

Animation thumbnail; pass null to skip thumbnail uploading

animation :: Maybe InputFile

Animation file to be sent

InputMessageAudio
An audio message

Fields

caption :: Maybe FormattedText

Animation caption; pass null to use an empty caption; 0-getOption("message_caption_length_max") characters

performer :: Maybe String

Performer of the audio; 0-64 characters, may be replaced by the server

title :: Maybe String

Title of the audio; 0-64 characters; may be replaced by the server

duration :: Maybe Int

Duration of the animation, in seconds

album_cover_thumbnail :: Maybe InputThumbnail

Thumbnail of the cover for the album; pass null to skip thumbnail uploading

audio :: Maybe InputFile

Audio file to be sent

InputMessageDocument
A document message (general file)

Fields

caption :: Maybe FormattedText

Animation caption; pass null to use an empty caption; 0-getOption("message_caption_length_max") characters

disable_content_type_detection :: Maybe Bool

If true, automatic file type detection will be disabled and the document will always be sent as file. Always true for files sent to secret chats

thumbnail :: Maybe InputThumbnail

Animation thumbnail; pass null to skip thumbnail uploading

document :: Maybe InputFile

Document to be sent

InputMessagePhoto
A photo message

Fields

has_spoiler :: Maybe Bool

True, if the animation preview must be covered by a spoiler animation; not supported in secret chats

self_destruct_time :: Maybe Int

Photo self-destruct time, in seconds (0-60). A non-zero self-destruct time can be specified only in private chats

caption :: Maybe FormattedText

Animation caption; pass null to use an empty caption; 0-getOption("message_caption_length_max") characters

height :: Maybe Int

Height of the animation; may be replaced by the server

width :: Maybe Int

Width of the animation; may be replaced by the server

added_sticker_file_ids :: Maybe [Int]

File identifiers of the stickers added to the animation, if applicable

thumbnail :: Maybe InputThumbnail

Animation thumbnail; pass null to skip thumbnail uploading

photo :: Maybe InputFile

Photo to send. The photo must be at most 10 MB in size. The photo's width and height must not exceed 10000 in total. Width and height ratio must be at most 20

InputMessageSticker
A sticker message

Fields

emoji :: Maybe String

Emoji used to choose the sticker

height :: Maybe Int

Height of the animation; may be replaced by the server

width :: Maybe Int

Width of the animation; may be replaced by the server

thumbnail :: Maybe InputThumbnail

Animation thumbnail; pass null to skip thumbnail uploading

sticker :: Maybe InputFile

Sticker to be sent

InputMessageVideo
A video message

Fields

has_spoiler :: Maybe Bool

True, if the animation preview must be covered by a spoiler animation; not supported in secret chats

self_destruct_time :: Maybe Int

Photo self-destruct time, in seconds (0-60). A non-zero self-destruct time can be specified only in private chats

caption :: Maybe FormattedText

Animation caption; pass null to use an empty caption; 0-getOption("message_caption_length_max") characters

supports_streaming :: Maybe Bool

True, if the video is supposed to be streamed

height :: Maybe Int

Height of the animation; may be replaced by the server

width :: Maybe Int

Width of the animation; may be replaced by the server

duration :: Maybe Int

Duration of the animation, in seconds

added_sticker_file_ids :: Maybe [Int]

File identifiers of the stickers added to the animation, if applicable

thumbnail :: Maybe InputThumbnail

Animation thumbnail; pass null to skip thumbnail uploading

video :: Maybe InputFile

Video to be sent

InputMessageVideoNote
A video note message

Fields

_length :: Maybe Int

Video width and height; must be positive and not greater than 640

duration :: Maybe Int

Duration of the animation, in seconds

thumbnail :: Maybe InputThumbnail

Animation thumbnail; pass null to skip thumbnail uploading

video_note :: Maybe InputFile

Video note to be sent

InputMessageVoiceNote
A voice note message

Fields

caption :: Maybe FormattedText

Animation caption; pass null to use an empty caption; 0-getOption("message_caption_length_max") characters

waveform :: Maybe String

Waveform representation of the voice note in 5-bit format

duration :: Maybe Int

Duration of the animation, in seconds

voice_note :: Maybe InputFile

Voice note to be sent

InputMessageLocation
A message with a location

Fields

proximity_alert_radius :: Maybe Int

For live locations, a maximum distance to another chat member for proximity alerts, in meters (0-100000). Pass 0 if the notification is disabled. Can't be enabled in channels and Saved Messages

heading :: Maybe Int

For live locations, a direction in which the location moves, in degrees; 1-360. Pass 0 if unknown

live_period :: Maybe Int

Period for which the location can be updated, in seconds; must be between 60 and 86400 for a live location and 0 otherwise

location :: Maybe Location

Location to be sent

InputMessageVenue
A message with information about a venue @venue Venue to send

Fields

venue :: Maybe Venue

InputMessageContact
A message containing a user contact @contact Contact to send

Fields

contact :: Maybe Contact

InputMessageDice
A dice message emoji Emoji on which the dice throw animation is based clear_draft True, if the chat message draft must be deleted

Fields

clear_draft :: Maybe Bool

True, if a chat message draft must be deleted

emoji :: Maybe String

Emoji used to choose the sticker

InputMessageGame
A message with a game; not supported for channels or secret chats bot_user_id User identifier of the bot that owns the game game_short_name Short name of the game

Fields

game_short_name :: Maybe String

bot_user_id :: Maybe Int

InputMessageInvoice
A message with an invoice; can be used only by bots

Fields

extended_media_content :: Maybe InputMessageContent

The content of extended media attached to the invoice. The content of the message to be sent. Must be one of the following types: inputMessagePhoto, inputMessageVideo

start_parameter :: Maybe String

Unique invoice bot deep link parameter for the generation of this invoice. If empty, it would be possible to pay directly from forwards of the invoice message

provider_data :: Maybe String

JSON-encoded data about the invoice, which will be shared with the payment provider

provider_token :: Maybe String

Payment provider token

payload :: Maybe String

The invoice payload

photo_height :: Maybe Int

Product photo height

photo_width :: Maybe Int

Product photo width

photo_size :: Maybe Int

Product photo size

photo_url :: Maybe String

Product photo URL; optional

description :: Maybe String

title :: Maybe String

Title of the audio; 0-64 characters; may be replaced by the server

invoice :: Maybe Invoice

Invoice

InputMessagePoll
A message with a poll. Polls can't be sent to secret chats. Polls can be sent only to a private chat with a bot

Fields

is_closed :: Maybe Bool

True, if the poll needs to be sent already closed; for bots only

close_date :: Maybe Int

Point in time (Unix timestamp) when the poll will automatically be closed; for bots only

open_period :: Maybe Int

Amount of time the poll will be active after creation, in seconds; for bots only

_type :: Maybe PollType

Type of the poll

is_anonymous :: Maybe Bool

True, if the poll voters are anonymous. Non-anonymous polls can't be sent or forwarded to channels

options :: Maybe [String]

List of poll answer options, 2-10 strings 1-100 characters each

question :: Maybe String

Poll question; 1-255 characters (up to 300 characters for bots)

InputMessageForwarded
A forwarded message

Fields

copy_options :: Maybe MessageCopyOptions

Options to be used to copy content of the message without reference to the original sender; pass null to forward the message as usual

in_game_share :: Maybe Bool

True, if a game message is being shared from a launched game; applies only to game messages

message_id :: Maybe Int

Identifier of the message to forward

from_chat_id :: Maybe Int

Identifier for the chat this forwarded message came from