TD.Data.InputMessageReplyTo

Documentation

data InputMessageReplyTo

Contains information about the message or the story to be replied

Constructors

InputMessageReplyToMessage
Describes a message to be replied

Fields

chat_id :: Maybe Int

The identifier of the chat to which the message to be replied belongs; pass 0 if the message to be replied is in the same chat. Must always be 0 for replies in secret chats. A message can be replied in another chat or topic only if message.can_be_replied_in_another_chat

message_id :: Maybe Int

The identifier of the message to be replied in the same or the specified chat

quote :: Maybe InputTextQuote

Quote from the message to be replied; pass null if none. Must always be null for replies in secret chats

InputMessageReplyToStory
Describes a story to be replied

Fields

story_sender_chat_id :: Maybe Int

The identifier of the sender of the story. Currently, stories can be replied only in the sender's chat

story_id :: Maybe Int

The identifier of the story