TD.Data.InputPageBlock

Documentation

data InputPageBlock

Describes a block of a rich message to send

Constructors

InputPageBlockSectionHeading
A section heading

Fields

text :: Maybe RichText

Text of the section heading

size :: Maybe Int

Relative size of the text font; 1-6, 1 is the largest, 6 is the smallest

InputPageBlockParagraph
A text paragraph

Fields

text :: Maybe RichText

Text of the section heading

InputPageBlockPreformatted
A preformatted text paragraph

Fields

text :: Maybe RichText

Text of the section heading

language :: Maybe Text

Programming language for which the text needs to be formatted

InputPageBlockFooter
The footer of the page

Fields

footer :: Maybe RichText

Footer

InputPageBlockThinking
A "Thinking..." placeholder; for pending rich messages only; for bots only

Fields

text :: Maybe RichText

Text of the section heading

InputPageBlockDivider
An empty block separating the page

InputPageBlockMathematicalExpression
A mathematical expression

Fields

expression :: Maybe Text

The expression in LaTeX format

InputPageBlockAnchor
An invisible anchor

Fields

name :: Maybe Text

Name of the anchor

InputPageBlockList
A list of data blocks

Fields

items :: Maybe [InputPageBlockListItem]

The items of the list

InputPageBlockBlockQuote
A block quote

Fields

blocks :: Maybe [InputPageBlock]

Quote blocks

credit :: Maybe RichText

Quote credit; pass null if none

InputPageBlockPullQuote
A pull quote

Fields

text :: Maybe RichText

Text of the section heading

credit :: Maybe RichText

Quote credit; pass null if none

InputPageBlockAnimation
An animation

Fields

animation :: Maybe InputAnimation

The animation to be sent

caption :: Maybe PageBlockCaption

Animation caption; pass null if none

has_spoiler :: Maybe Bool

True, if the animation preview must be covered by a spoiler animation

InputPageBlockAudio
An audio file

Fields

audio :: Maybe InputAudio

The audio to be sent

caption :: Maybe PageBlockCaption

Animation caption; pass null if none

InputPageBlockPhoto
A photo

Fields

photo :: Maybe InputPhoto

The photo to be sent

caption :: Maybe PageBlockCaption

Animation caption; pass null if none

has_spoiler :: Maybe Bool

True, if the animation preview must be covered by a spoiler animation

InputPageBlockVideo
A video

Fields

video :: Maybe InputVideo

The video to be sent

caption :: Maybe PageBlockCaption

Animation caption; pass null if none

has_spoiler :: Maybe Bool

True, if the animation preview must be covered by a spoiler animation

InputPageBlockVoiceNote
A voice note

Fields

voice_note :: Maybe InputVoiceNote

The voice note to be sent

caption :: Maybe PageBlockCaption

Animation caption; pass null if none

InputPageBlockCollage
A collage

Fields

blocks :: Maybe [InputPageBlock]

Quote blocks

caption :: Maybe PageBlockCaption

Animation caption; pass null if none

InputPageBlockSlideshow
A slideshow

Fields

blocks :: Maybe [InputPageBlock]

Quote blocks

caption :: Maybe PageBlockCaption

Animation caption; pass null if none

InputPageBlockTable
A table

Fields

_caption :: Maybe RichText

Table caption

cells :: Maybe [[PageBlockTableCell]]

Table cells

is_bordered :: Maybe Bool

True, if the table is bordered

is_striped :: Maybe Bool

True, if the table is striped

InputPageBlockDetails
A collapsible block

Fields

header :: Maybe RichText

Always visible heading for the block

blocks :: Maybe [InputPageBlock]

Quote blocks

is_open :: Maybe Bool

True, if the block is open by default

InputPageBlockMap
A map. The map's width and height must not exceed 10000 in total. Width and height ratio must be at most 20

Fields

location :: Maybe Location

Location of the map center

zoom :: Maybe Int

Map zoom level; 0-24

width :: Maybe Int

Map width; 0-10000

height :: Maybe Int

Map height; 0-10000

caption :: Maybe PageBlockCaption

Animation caption; pass null if none