TD.Data.InputPageBlockListItem

Documentation

data InputPageBlockListItem

Constructors

InputPageBlockListItem
Describes an item of a list page block to be sent

Fields

blocks :: Maybe [InputPageBlock]

Item blocks

has_checkbox :: Maybe Bool

True, if the item has a checkbox

is_checked :: Maybe Bool

True, if the item is checked

value :: Maybe Int

Value of the item; pass 0 for unordered lists

_type :: Maybe Text

Type of the item numbering type; must be one of "a" for a lowercase letter, A for an uppercase letter, "i" for lowercase Roman numerals, I for uppercase Roman numerals, "1" for decimal numbers, or empty for unordered lists

defaultInputPageBlockListItem :: InputPageBlockListItem