TD.Data.InputSticker

Description

Documentation

data InputSticker

Constructors

InputSticker
A sticker to be added to a sticker set

Fields

_type :: Maybe StickerType

Sticker type

emojis :: Maybe String

Emojis corresponding to the sticker

sticker :: Maybe InputFile

File with the sticker; must fit in a 512x512 square. For WEBP stickers and masks the file must be in PNG format, which will be converted to WEBP server-side. Otherwise, the file must be local or uploaded within a week. See https://core.telegram.org/animated_stickers#technical-requirements for technical requirements