TD.Data.InputVoiceNote

Documentation

data InputVoiceNote

Constructors

InputVoiceNote
A video note to be sent

Fields

voice_note :: Maybe InputFile

Voice note file to be sent. The voice note must be encoded with the Opus codec and stored inside an OGG container with a single audio channel, or be in MP3 or M4A format as regular audio

duration :: Maybe Int

Duration of the voice note, in seconds

waveform :: Maybe ByteString

Waveform representation of the voice note in 5-bit format

defaultInputVoiceNote :: InputVoiceNote