TD.Data.InternalLinkType

Description

Documentation

data InternalLinkType

Describes an internal https://t.me or tg: link, which must be processed by the application in a special way

Constructors

InternalLinkTypeActiveSessions
The link is a link to the active sessions section of the application. Use getActiveSessions to handle the link

InternalLinkTypeAttachmentMenuBot
The link is a link to an attachment menu bot to be opened in the specified or a chosen chat. Process given target_chat to open the chat. Then, call searchPublicChat with the given bot username, check that the user is a bot and can be added to attachment menu. Then, use getAttachmentMenuBot to receive information about the bot. If the bot isn't added to attachment menu, then user needs to confirm adding the bot to attachment menu. If user confirms adding, then use toggleBotIsAddedToAttachmentMenu to add it. If the attachment menu bot can't be used in the opened chat, show an error to the user. If the bot is added to attachment menu and can be used in the chat, then use openWebApp with the given URL

Fields

url :: Maybe String

URL to be passed to openWebApp

bot_username :: Maybe String

Username of the bot

target_chat :: Maybe TargetChat

Target chat to be opened

InternalLinkTypeAuthenticationCode
The link contains an authentication code. Call checkAuthenticationCode with the code if the current authorization state is authorizationStateWaitCode @code The authentication code

Fields

code :: Maybe String

InternalLinkTypeBackground
The link is a link to a background. Call searchBackground with the given background name to process the link @background_name Name of the background

Fields

background_name :: Maybe String

InternalLinkTypeBotAddToChannel
The link is a link to a Telegram bot, which is supposed to be added to a channel chat as an administrator. Call searchPublicChat with the given bot username and check that the user is a bot, ask the current user to select a channel chat to add the bot to as an administrator. Then, call getChatMember to receive the current bot rights in the chat and if the bot already is an administrator, check that the current user can edit its administrator rights and combine received rights with the requested administrator rights. Then, show confirmation box to the user, and call setChatMemberStatus with the chosen chat and confirmed rights

Fields

administrator_rights :: Maybe ChatAdministratorRights

Expected administrator rights for the bot

bot_username :: Maybe String

Username of the bot

InternalLinkTypeBotStart
The link is a link to a chat with a Telegram bot. Call searchPublicChat with the given bot username, check that the user is a bot, show START button in the chat with the bot, and then call sendBotStartMessage with the given start parameter after the button is pressed

Fields

autostart :: Maybe Bool

True, if sendBotStartMessage must be called automatically without showing the START button

start_parameter :: Maybe String

The parameter to be passed to sendBotStartMessage

bot_username :: Maybe String

Username of the bot

InternalLinkTypeBotStartInGroup
The link is a link to a Telegram bot, which is supposed to be added to a group chat. Call searchPublicChat with the given bot username, check that the user is a bot and can be added to groups, ask the current user to select a basic group or a supergroup chat to add the bot to, taking into account that bots can be added to a public supergroup only by administrators of the supergroup. If administrator rights are provided by the link, call getChatMember to receive the current bot rights in the chat and if the bot already is an administrator, check that the current user can edit its administrator rights, combine received rights with the requested administrator rights, show confirmation box to the user, and call setChatMemberStatus with the chosen chat and confirmed administrator rights. Before call to setChatMemberStatus it may be required to upgrade the chosen basic group chat to a supergroup chat. Then, if start_parameter isn't empty, call sendBotStartMessage with the given start parameter and the chosen chat; otherwise, just send /start message with bot's username added to the chat.

Fields

administrator_rights :: Maybe ChatAdministratorRights

Expected administrator rights for the bot

start_parameter :: Maybe String

The parameter to be passed to sendBotStartMessage

bot_username :: Maybe String

Username of the bot

InternalLinkTypeChangePhoneNumber
The link is a link to the change phone number section of the app

InternalLinkTypeChatInvite
The link is a chat invite link. Call checkChatInviteLink with the given invite link to process the link @invite_link Internal representation of the invite link

Fields

invite_link :: Maybe String

InternalLinkTypeDefaultMessageAutoDeleteTimerSettings
The link is a link to the default message auto-delete timer settings section of the app settings

InternalLinkTypeEditProfileSettings
The link is a link to the edit profile section of the app settings

InternalLinkTypeFilterSettings
The link is a link to the filter section of the app settings

InternalLinkTypeGame
The link is a link to a game. Call searchPublicChat with the given bot username, check that the user is a bot, ask the current user to select a chat to send the game, and then call sendMessage with inputMessageGame

Fields

game_short_name :: Maybe String

Short name of the game

bot_username :: Maybe String

Username of the bot

InternalLinkTypeInstantView
The link must be opened in an Instant View. Call getWebPageInstantView with the given URL to process the link url URL to be passed to getWebPageInstantView fallback_url An URL to open if getWebPageInstantView fails

Fields

fallback_url :: Maybe String

url :: Maybe String

URL to be passed to openWebApp

InternalLinkTypeInvoice
The link is a link to an invoice. Call getPaymentForm with the given invoice name to process the link @invoice_name Name of the invoice

Fields

invoice_name :: Maybe String

InternalLinkTypeLanguagePack
The link is a link to a language pack. Call getLanguagePackInfo with the given language pack identifier to process the link @language_pack_id Language pack identifier

Fields

language_pack_id :: Maybe String

InternalLinkTypeLanguageSettings
The link is a link to the language section of the app settings

InternalLinkTypeMessage
The link is a link to a Telegram message or a forum topic. Call getMessageLinkInfo with the given URL to process the link @url URL to be passed to getMessageLinkInfo

Fields

url :: Maybe String

URL to be passed to openWebApp

InternalLinkTypeMessageDraft
The link contains a message draft text. A share screen needs to be shown to the user, then the chosen chat must be opened and the text is added to the input field

Fields

contains_link :: Maybe Bool

True, if the first line of the text contains a link. If true, the input field needs to be focused and the text after the link must be selected

text :: Maybe FormattedText

Message draft text

InternalLinkTypePassportDataRequest
The link contains a request of Telegram passport data. Call getPassportAuthorizationForm with the given parameters to process the link if the link was received from outside of the application; otherwise, ignore it

Fields

callback_url :: Maybe String

An HTTP URL to open once the request is finished, canceled, or failed with the parameters tg_passport=success, tg_passport=cancel, or tg_passport=error&error=... respectively. If empty, then onActivityResult method must be used to return response on Android, or the link tgbot{bot_user_id}:/passportsuccess or tgbot{bot_user_id}:/passportcancel must be opened otherwise

nonce :: Maybe String

Unique request identifier provided by the service

public_key :: Maybe String

Service's public key

scope :: Maybe String

Telegram Passport element types requested by the service

bot_user_id :: Maybe Int

User identifier of the service's bot

InternalLinkTypePhoneNumberConfirmation
The link can be used to confirm ownership of a phone number to prevent account deletion. Call sendPhoneNumberConfirmationCode with the given hash and phone number to process the link

Fields

phone_number :: Maybe String

Phone number value from the link

hash :: Maybe String

Hash value from the link

InternalLinkTypePremiumFeatures
The link is a link to the Premium features screen of the application from which the user can subscribe to Telegram Premium. Call getPremiumFeatures with the given referrer to process the link @referrer Referrer specified in the link

Fields

referrer :: Maybe String

InternalLinkTypePrivacyAndSecuritySettings
The link is a link to the privacy and security section of the app settings

InternalLinkTypeProxy
The link is a link to a proxy. Call addProxy with the given parameters to process the link and add the proxy

Fields

_type :: Maybe ProxyType

Type of the proxy

port :: Maybe Int

Proxy server port

server :: Maybe String

Proxy server IP address

InternalLinkTypePublicChat
The link is a link to a chat by its username. Call searchPublicChat with the given chat username to process the link @chat_username Username of the chat

Fields

chat_username :: Maybe String

InternalLinkTypeQrCodeAuthentication
The link can be used to login the current user on another device, but it must be scanned from QR-code using in-app camera. An alert similar to "This code can be used to allow someone to log in to your Telegram account. To confirm Telegram login, please go to Settings > Devices > Scan QR and scan the code" needs to be shown

InternalLinkTypeRestorePurchases
The link forces restore of App Store purchases when opened. For official iOS application only

InternalLinkTypeSettings
The link is a link to application settings

InternalLinkTypeStickerSet
The link is a link to a sticker set. Call searchStickerSet with the given sticker set name to process the link and show the sticker set

Fields

expect_custom_emoji :: Maybe Bool

True, if the sticker set is expected to contain custom emoji

sticker_set_name :: Maybe String

Name of the sticker set

InternalLinkTypeTheme
The link is a link to a theme. TDLib has no theme support yet @theme_name Name of the theme

Fields

theme_name :: Maybe String

InternalLinkTypeThemeSettings
The link is a link to the theme section of the app settings

InternalLinkTypeUnknownDeepLink
The link is an unknown tg: link. Call getDeepLinkInfo to process the link @link Link to be passed to getDeepLinkInfo

Fields

link :: Maybe String

InternalLinkTypeUnsupportedProxy
The link is a link to an unsupported proxy. An alert can be shown to the user

InternalLinkTypeUserPhoneNumber
The link is a link to a user by its phone number. Call searchUserByPhoneNumber with the given phone number to process the link @phone_number Phone number of the user

Fields

phone_number :: Maybe String

Phone number value from the link

InternalLinkTypeUserToken
The link is a link to a user by a temporary token. Call searchUserByToken with the given token to process the link @token The token

Fields

token :: Maybe String

InternalLinkTypeVideoChat
The link is a link to a video chat. Call searchPublicChat with the given chat username, and then joinGroupCall with the given invite hash to process the link

Fields

is_live_stream :: Maybe Bool

True, if the video chat is expected to be a live stream in a channel or a broadcast group

invite_hash :: Maybe String

If non-empty, invite hash to be used to join the video chat without being muted by administrators

chat_username :: Maybe String

InternalLinkTypeWebApp
The link is a link to a Web App. Call searchPublicChat with the given bot username, check that the user is a bot, then call searchWebApp with the received bot and the given web_app_short_name. Process received foundWebApp by showing a confirmation dialog if needed, then calling getWebAppLinkUrl and opening the returned URL

Fields

start_parameter :: Maybe String

The parameter to be passed to sendBotStartMessage

web_app_short_name :: Maybe String

Short name of the Web App

bot_username :: Maybe String

Username of the bot