TD.Data.InviteGroupCallParticipantResult

Documentation

data InviteGroupCallParticipantResult

Describes result of group call participant invitation

Constructors

InviteGroupCallParticipantResultUserPrivacyRestricted
The user can't be invited due to their privacy settings

InviteGroupCallParticipantResultUserAlreadyParticipant
The user can't be invited because they are already a participant of the call

InviteGroupCallParticipantResultUserWasBanned
The user can't be invited because they were banned by the owner of the call and can be invited back only by the owner of the group call

InviteGroupCallParticipantResultSuccess
The user was invited and a service message of the type messageGroupCall was sent which can be used in declineGroupCallInvitation to cancel the invitation

Fields

chat_id :: Maybe Int

Identifier of the chat with the invitation message

message_id :: Maybe Int

Identifier of the message