TD.Data.KeyboardButtonType

Description

Documentation

data KeyboardButtonType

Describes a keyboard button type

Constructors

KeyboardButtonTypeText
A simple button, with text that must be sent when the button is pressed

KeyboardButtonTypeRequestPhoneNumber
A button that sends the user's phone number when pressed; available only in private chats

KeyboardButtonTypeRequestLocation
A button that sends the user's location when pressed; available only in private chats

KeyboardButtonTypeRequestPoll
A button that allows the user to create and send a poll when pressed; available only in private chats force_regular If true, only regular polls must be allowed to create force_quiz If true, only polls in quiz mode must be allowed to create

Fields

force_quiz :: Maybe Bool

force_regular :: Maybe Bool

KeyboardButtonTypeRequestUser
A button that requests a user to be shared by the current user; available only in private chats. Use the method shareUserWithBot to complete the request

Fields

user_is_premium :: Maybe Bool

True, if the shared user must be a Telegram Premium user; otherwise, the shared user must no be a Telegram Premium user. Ignored if restrict_user_is_premium is false

restrict_user_is_premium :: Maybe Bool

True, if the shared user must or must not be a Telegram Premium user

user_is_bot :: Maybe Bool

True, if the shared user must be a bot; otherwise, the shared user must no be a bot. Ignored if restrict_user_is_bot is false

restrict_user_is_bot :: Maybe Bool

True, if the shared user must or must not be a bot

_id :: Maybe Int

Unique button identifier

KeyboardButtonTypeRequestChat
A button that requests a chat to be shared by the current user; available only in private chats. Use the method shareChatWithBot to complete the request

Fields

bot_is_member :: Maybe Bool

True, if the bot must be a member of the chat; for basic group and supergroup chats only

bot_administrator_rights :: Maybe ChatAdministratorRights

Expected bot administrator rights in the chat; may be null if they aren't restricted

user_administrator_rights :: Maybe ChatAdministratorRights

Expected user administrator rights in the chat; may be null if they aren't restricted

chat_is_created :: Maybe Bool

True, if the chat must be created by the current user

chat_has_username :: Maybe Bool

True, if the chat must have a username; otherwise, the chat must not have a username. Ignored if restrict_chat_has_username is false

restrict_chat_has_username :: Maybe Bool

True, if the chat must or must not have a username

chat_is_forum :: Maybe Bool

True, if the chat must be a forum supergroup; otherwise, the chat must not be a forum supergroup. Ignored if restrict_chat_is_forum is false

restrict_chat_is_forum :: Maybe Bool

True, if the chat must or must not be a forum supergroup

chat_is_channel :: Maybe Bool

True, if the chat must be a channel; otherwise, a basic group or a supergroup chat is shared

_id :: Maybe Int

Unique button identifier

KeyboardButtonTypeWebApp
A button that opens a Web App by calling getWebAppUrl @url An HTTP URL to pass to getWebAppUrl

Fields

url :: Maybe String