TD.Data.LanguagePackInfo

Description

Documentation

data LanguagePackInfo

Constructors

LanguagePackInfo
Contains information about a language pack @id Unique language pack identifier

Fields

translation_url :: Maybe String

local_string_count :: Maybe Int

Total number of non-deleted strings from the language pack available locally @translation_url Link to language translation interface; empty for custom local language packs

translated_string_count :: Maybe Int

total_string_count :: Maybe Int

Total number of non-deleted strings from the language pack @translated_string_count Total number of translated strings from the language pack

is_installed :: Maybe Bool

True, if the language pack is installed by the current user

is_beta :: Maybe Bool

is_rtl :: Maybe Bool

is_official :: Maybe Bool

True, if the language pack is official is_rtl True, if the language pack strings are RTL is_beta True, if the language pack is a beta language pack

plural_code :: Maybe String

A language code to be used to apply plural forms. See https://www.unicode.org/cldr/charts/latest/supplemental/language_plural_rules.html for more information

native_name :: Maybe String

name :: Maybe String

Language name @native_name Name of the language in that language

base_language_pack_id :: Maybe String

Identifier of a base language pack; may be empty. If a string is missed in the language pack, then it must be fetched from base language pack. Unsupported in custom language packs

_id :: Maybe String