TD.Data.LinkPreviewType

Documentation

data LinkPreviewType

Describes type of link preview

Constructors

LinkPreviewTypeAlbum
The link is a link to a media album consisting of photos and videos

Fields

media :: Maybe [LinkPreviewAlbumMedia]

The list of album media

caption :: Maybe Text

Album caption

LinkPreviewTypeAnimation
The link is a link to an animation

Fields

animation :: Maybe Animation

The animation

LinkPreviewTypeApp
The link is a link to an app at App Store or Google Play

Fields

photo :: Maybe Photo

Photo for the app

LinkPreviewTypeArticle
The link is a link to a web site

Fields

photo :: Maybe Photo

Photo for the app

LinkPreviewTypeAudio
The link is a link to an audio

Fields

audio :: Maybe Audio

The audio description

LinkPreviewTypeBackground
The link is a link to a background. Link preview title and description are available only for filled backgrounds

Fields

document :: Maybe Document

Document with the background; may be null for filled backgrounds

background_type :: Maybe BackgroundType

Type of the background; may be null if unknown

LinkPreviewTypeChannelBoost
The link is a link to boost a channel chat

Fields

_photo :: Maybe ChatPhoto

Photo of the chat; may be null

LinkPreviewTypeChat
The link is a link to a chat

Fields

_type :: Maybe InviteLinkChatType

Type of the chat

_photo :: Maybe ChatPhoto

Photo of the chat; may be null

creates_join_request :: Maybe Bool

True, if the link only creates join request

LinkPreviewTypeDirectMessagesChat
The link is a link to a direct messages chat of a channel

Fields

_photo :: Maybe ChatPhoto

Photo of the chat; may be null

LinkPreviewTypeDocument
The link is a link to a general file

Fields

document :: Maybe Document

Document with the background; may be null for filled backgrounds

LinkPreviewTypeEmbeddedAnimationPlayer
The link is a link to an animation player

Fields

url :: Maybe Text

URL of the external animation player

thumbnail :: Maybe Photo

Thumbnail of the animation; may be null if unknown

duration :: Maybe Int

Duration of the animation, in seconds

width :: Maybe Int

Expected width of the embedded player

height :: Maybe Int

Expected height of the embedded player

LinkPreviewTypeEmbeddedAudioPlayer
The link is a link to an audio player

Fields

url :: Maybe Text

URL of the external animation player

thumbnail :: Maybe Photo

Thumbnail of the animation; may be null if unknown

duration :: Maybe Int

Duration of the animation, in seconds

width :: Maybe Int

Expected width of the embedded player

height :: Maybe Int

Expected height of the embedded player

LinkPreviewTypeEmbeddedVideoPlayer
The link is a link to a video player

Fields

url :: Maybe Text

URL of the external animation player

thumbnail :: Maybe Photo

Thumbnail of the animation; may be null if unknown

duration :: Maybe Int

Duration of the animation, in seconds

width :: Maybe Int

Expected width of the embedded player

height :: Maybe Int

Expected height of the embedded player

LinkPreviewTypeExternalAudio
The link is a link to an audio file

Fields

url :: Maybe Text

URL of the external animation player

mime_type :: Maybe Text

MIME type of the audio file

duration :: Maybe Int

Duration of the animation, in seconds

LinkPreviewTypeExternalVideo
The link is a link to a video file

Fields

url :: Maybe Text

URL of the external animation player

mime_type :: Maybe Text

MIME type of the audio file

width :: Maybe Int

Expected width of the embedded player

height :: Maybe Int

Expected height of the embedded player

duration :: Maybe Int

Duration of the animation, in seconds

LinkPreviewTypeGiftAuction
The link is a link to a gift auction

Fields

gift :: Maybe Gift

The gift

auction_end_date :: Maybe Int

Point in time (Unix timestamp) when the auction will be ended

LinkPreviewTypeGiftCollection
The link is a link to a gift collection

Fields

icons :: Maybe [Sticker]

Icons for some gifts from the collection; may be empty

LinkPreviewTypeGroupCall
The link is a link to a group call that isn't bound to a chat

LinkPreviewTypeInvoice
The link is a link to an invoice

LinkPreviewTypeLiveStory
The link is a link to a live story group call

Fields

story_poster_chat_id :: Maybe Int

The identifier of the chat that posted the story

story_id :: Maybe Int

Story identifier

LinkPreviewTypeMessage
The link is a link to a text or a poll Telegram message

LinkPreviewTypePhoto
The link is a link to a photo

Fields

photo :: Maybe Photo

Photo for the app

LinkPreviewTypePremiumGiftCode
The link is a link to a Telegram Premium gift code

LinkPreviewTypeShareableChatFolder
The link is a link to a shareable chat folder

LinkPreviewTypeSticker
The link is a link to a sticker

Fields

sticker :: Maybe Sticker

The sticker. It can be an arbitrary WEBP image and can have dimensions bigger than 512

LinkPreviewTypeStickerSet
The link is a link to a sticker set

Fields

stickers :: Maybe [Sticker]

Up to 4 stickers from the sticker set

LinkPreviewTypeStory
The link is a link to a story. Link preview description is unavailable

Fields

story_poster_chat_id :: Maybe Int

The identifier of the chat that posted the story

story_id :: Maybe Int

Story identifier

LinkPreviewTypeStoryAlbum
The link is a link to an album of stories

Fields

photo_icon :: Maybe Photo

Icon of the album; may be null if none

video_icon :: Maybe Video

Video icon of the album; may be null if none

LinkPreviewTypeSupergroupBoost
The link is a link to boost a supergroup chat

Fields

_photo :: Maybe ChatPhoto

Photo of the chat; may be null

LinkPreviewTypeTheme
The link is a link to a cloud theme. TDLib has no theme support yet

Fields

documents :: Maybe [Document]

The list of files with theme description

settings :: Maybe ThemeSettings

Settings for the cloud theme; may be null if unknown

LinkPreviewTypeUnsupported
The link preview type is unsupported yet

LinkPreviewTypeUpgradedGift
The link is a link to an upgraded gift

Fields

_gift :: Maybe UpgradedGift

The gift

LinkPreviewTypeUser
The link is a link to a user

Fields

_photo :: Maybe ChatPhoto

Photo of the chat; may be null

is_bot :: Maybe Bool

True, if the user is a bot

LinkPreviewTypeVideo
The link is a link to a video

Fields

video :: Maybe Video

The video description

cover :: Maybe Photo

Cover of the video; may be null if none

start_timestamp :: Maybe Int

Timestamp from which the video playing must start, in seconds

LinkPreviewTypeVideoChat
The link is a link to a video chat

Fields

_photo :: Maybe ChatPhoto

Photo of the chat; may be null

is_live_stream :: Maybe Bool

True, if the video chat is expected to be a live stream in a channel or a broadcast group

joins_as_speaker :: Maybe Bool

True, if the user can use the link to join the video chat without being muted by administrators

LinkPreviewTypeVideoNote
The link is a link to a video note message

Fields

video_note :: Maybe VideoNote

The video note

LinkPreviewTypeVoiceNote
The link is a link to a voice note message

Fields

voice_note :: Maybe VoiceNote

The voice note

LinkPreviewTypeWebApp
The link is a link to a Web App

Fields

photo :: Maybe Photo

Photo for the app