TD.Data.LocalFile

Description

Documentation

data LocalFile

Constructors

LocalFile
Represents a local file

Fields

downloaded_size :: Maybe Int

Total downloaded file size, in bytes. Can be used only for calculating download progress. The actual file size may be bigger, and some parts of it may contain garbage

downloaded_prefix_size :: Maybe Int

If is_downloading_completed is false, then only some prefix of the file starting from download_offset is ready to be read. downloaded_prefix_size is the size of that prefix in bytes

download_offset :: Maybe Int

Download will be started from this offset. downloaded_prefix_size is calculated from this offset

is_downloading_completed :: Maybe Bool

True, if the local copy is fully available

is_downloading_active :: Maybe Bool

True, if the file is currently being downloaded (or a local copy is being generated by some other means)

can_be_deleted :: Maybe Bool

True, if the file can be deleted

can_be_downloaded :: Maybe Bool

True, if it is possible to download or generate the file

path :: Maybe String

Local path to the locally available file part; may be empty