TD.Data.LoginUrlInfo

Description

Documentation

data LoginUrlInfo

Contains information about an inline button of type inlineKeyboardButtonTypeLoginUrl

Constructors

LoginUrlInfoOpen
An HTTP url needs to be open url The URL to open skip_confirm True, if there is no need to show an ordinary open URL confirm

Fields

skip_confirm :: Maybe Bool

url :: Maybe String

LoginUrlInfoRequestConfirmation
An authorization confirmation dialog needs to be shown to the user url An HTTP URL to be opened domain A domain of the URL

Fields

request_write_access :: Maybe Bool

bot_user_id :: Maybe Int

User identifier of a bot linked with the website @request_write_access True, if the user needs to be requested to give the permission to the bot to send them messages

domain :: Maybe String

url :: Maybe String