TD.Data.Message

Documentation

data Message

Constructors

Message
Describes a message

Fields

_id :: Maybe Int

Message identifier; unique for the chat to which the message belongs

sender_id :: Maybe MessageSender

Identifier of the sender of the message

chat_id :: Maybe Int

Chat identifier

sending_state :: Maybe MessageSendingState

The sending state of the message; may be null if the message isn't being sent and didn't fail to be sent

scheduling_state :: Maybe MessageSchedulingState

The scheduling state of the message; may be null if the message isn't scheduled

is_outgoing :: Maybe Bool

True, if the message is outgoing

is_pinned :: Maybe Bool

True, if the message is pinned

can_be_edited :: Maybe Bool

True, if the message can be edited. For live location and poll messages this fields shows whether editMessageLiveLocation or stopPoll can be used with this message by the application

can_be_forwarded :: Maybe Bool

True, if the message can be forwarded

can_be_replied_in_another_chat :: Maybe Bool

True, if the message can be replied in another chat or topic

can_be_saved :: Maybe Bool

True, if content of the message can be saved locally or copied

can_be_deleted_only_for_self :: Maybe Bool

True, if the message can be deleted only for the current user while other users will continue to see it

can_be_deleted_for_all_users :: Maybe Bool

True, if the message can be deleted for all users

can_get_added_reactions :: Maybe Bool

True, if the list of added reactions is available through getMessageAddedReactions

can_get_statistics :: Maybe Bool

True, if the message statistics are available through getMessageStatistics

can_get_message_thread :: Maybe Bool

True, if information about the message thread is available through getMessageThread and getMessageThreadHistory

can_get_viewers :: Maybe Bool

True, if chat members already viewed the message can be received through getMessageViewers

can_get_media_timestamp_links :: Maybe Bool

True, if media timestamp links can be generated for media timestamp entities in the message text, caption or web page description through getMessageLink

can_report_reactions :: Maybe Bool

True, if reactions on the message can be reported through reportMessageReactions

has_timestamped_media :: Maybe Bool

True, if media timestamp entities refers to a media in this message as opposed to a media in the replied message

is_channel_post :: Maybe Bool

True, if the message is a channel post. All messages to channels are channel posts, all other messages are not channel posts

is_topic_message :: Maybe Bool

True, if the message is a forum topic message

contains_unread_mention :: Maybe Bool

True, if the message contains an unread mention for the current user

date :: Maybe Int

Point in time (Unix timestamp) when the message was sent

edit_date :: Maybe Int

Point in time (Unix timestamp) when the message was last edited

forward_info :: Maybe MessageForwardInfo

Information about the initial message sender; may be null if none or unknown

import_info :: Maybe MessageImportInfo

Information about the initial message for messages created with importMessages; may be null if the message isn't imported

interaction_info :: Maybe MessageInteractionInfo

Information about interactions with the message; may be null if none

unread_reactions :: Maybe [UnreadReaction]

Information about unread reactions added to the message

reply_to :: Maybe MessageReplyTo

Information about the message or the story this message is replying to; may be null if none

message_thread_id :: Maybe Int

If non-zero, the identifier of the message thread the message belongs to; unique within the chat to which the message belongs

self_destruct_type :: Maybe MessageSelfDestructType

The message's self-destruct type; may be null if none

self_destruct_in :: Maybe Double

Time left before the message self-destruct timer expires, in seconds; 0 if self-destruction isn't scheduled yet

auto_delete_in :: Maybe Double

Time left before the message will be automatically deleted by message_auto_delete_time setting of the chat, in seconds; 0 if never

via_bot_user_id :: Maybe Int

If non-zero, the user identifier of the bot through which this message was sent

author_signature :: Maybe Text

For channel posts and anonymous group messages, optional author signature

media_album_id :: Maybe Int

Unique identifier of an album this message belongs to. Only audios, documents, photos and videos can be grouped together in albums

restriction_reason :: Maybe Text

If non-empty, contains a human-readable description of the reason why access to this message must be restricted

content :: Maybe MessageContent

Content of the message

reply_markup :: Maybe ReplyMarkup

Reply markup for the message; may be null if none