TD.Data.MessageContent

Documentation

data MessageContent

Contains the content of a message

Constructors

MessageText
A text message

Fields

text :: Maybe FormattedText

Text of the message

link_preview :: Maybe LinkPreview

A link preview attached to the message; may be null

link_preview_options :: Maybe LinkPreviewOptions

Options which were used for generation of the link preview; may be null if default options were used

MessageAnimation
An animation message (GIF-style).

Fields

animation :: Maybe Animation

The animation description

caption :: Maybe FormattedText

Animation caption

show_caption_above_media :: Maybe Bool

True, if the caption must be shown above the animation; otherwise, the caption must be shown below the animation

has_spoiler :: Maybe Bool

True, if the animation preview must be covered by a spoiler animation

is_secret :: Maybe Bool

True, if the animation thumbnail must be blurred and the animation must be shown only while tapped

MessageAudio
An audio message

Fields

audio :: Maybe Audio

The audio description

caption :: Maybe FormattedText

Animation caption

MessageDocument
A document message (general file)

Fields

document :: Maybe Document

The document description

caption :: Maybe FormattedText

Animation caption

MessagePaidMedia
A message with paid media

Fields

star_count :: Maybe Int

Number of Telegram Stars needed to buy access to the media in the message

media :: Maybe [PaidMedia]

Information about the media

caption :: Maybe FormattedText

Animation caption

show_caption_above_media :: Maybe Bool

True, if the caption must be shown above the animation; otherwise, the caption must be shown below the animation

MessagePhoto
A photo message

Fields

photo :: Maybe Photo

The photo

caption :: Maybe FormattedText

Animation caption

show_caption_above_media :: Maybe Bool

True, if the caption must be shown above the animation; otherwise, the caption must be shown below the animation

has_spoiler :: Maybe Bool

True, if the animation preview must be covered by a spoiler animation

is_secret :: Maybe Bool

True, if the animation thumbnail must be blurred and the animation must be shown only while tapped

MessageSticker
A sticker message

Fields

sticker :: Maybe Sticker

The sticker description

is_premium :: Maybe Bool

True, if premium animation of the sticker must be played

MessageVideo
A video message

Fields

video :: Maybe Video

The video description

alternative_videos :: Maybe [AlternativeVideo]

Alternative qualities of the video

cover :: Maybe Photo

Cover of the video; may be null if none

start_timestamp :: Maybe Int

Timestamp from which the video playing must start, in seconds

caption :: Maybe FormattedText

Animation caption

show_caption_above_media :: Maybe Bool

True, if the caption must be shown above the animation; otherwise, the caption must be shown below the animation

has_spoiler :: Maybe Bool

True, if the animation preview must be covered by a spoiler animation

is_secret :: Maybe Bool

True, if the animation thumbnail must be blurred and the animation must be shown only while tapped

MessageVideoNote
A video note message

Fields

video_note :: Maybe VideoNote

The video note description

is_viewed :: Maybe Bool

True, if at least one of the recipients has viewed the video note

is_secret :: Maybe Bool

True, if the animation thumbnail must be blurred and the animation must be shown only while tapped

MessageVoiceNote
A voice note message

Fields

voice_note :: Maybe VoiceNote

The voice note description

caption :: Maybe FormattedText

Animation caption

is_listened :: Maybe Bool

True, if at least one of the recipients has listened to the voice note

MessageExpiredPhoto
A self-destructed photo message

MessageExpiredVideo
A self-destructed video message

MessageExpiredVideoNote
A self-destructed video note message

MessageExpiredVoiceNote
A self-destructed voice note message

MessageLocation
A message with a location

Fields

location :: Maybe Location

The location description

live_period :: Maybe Int

Time relative to the message send date, for which the location can be updated, in seconds; if 0x7FFFFFFF, then location can be updated forever

expires_in :: Maybe Int

Left time for which the location can be updated, in seconds. If 0, then the location can't be updated anymore. The update updateMessageContent is not sent when this field changes

heading :: Maybe Int

For live locations, a direction in which the location moves, in degrees; 1-360. If 0 the direction is unknown

proximity_alert_radius :: Maybe Int

For live locations, a maximum distance to another chat member for proximity alerts, in meters (0-100000). 0 if the notification is disabled. Available only to the message sender

MessageVenue
A message with information about a venue

Fields

venue :: Maybe Venue

The venue description

MessageContact
A message with a user contact

Fields

contact :: Maybe Contact

The contact description

MessageAnimatedEmoji
A message with an animated emoji

Fields

animated_emoji :: Maybe AnimatedEmoji

The animated emoji

emoji :: Maybe Text

The corresponding emoji

MessageDice
A dice message. The dice value is randomly generated by the server

Fields

initial_state :: Maybe DiceStickers

The animated stickers with the initial dice animation; may be null if unknown. The update updateMessageContent will be sent when the sticker became known

final_state :: Maybe DiceStickers

The animated stickers with the final dice animation; may be null if unknown. The update updateMessageContent will be sent when the sticker became known

emoji :: Maybe Text

The corresponding emoji

value :: Maybe Int

The dice value. If the value is 0, the dice don't have final state yet

success_animation_frame_number :: Maybe Int

Number of frame after which a success animation like a shower of confetti needs to be shown on updateMessageSendSucceeded

MessageGame
A message with a game

Fields

game :: Maybe Game

The game description

MessagePoll
A message with a poll

Fields

poll :: Maybe Poll

The poll description

MessageStory
A message with a forwarded story

Fields

story_poster_chat_id :: Maybe Int

Identifier of the chat that posted the story

story_id :: Maybe Int

Story identifier

via_mention :: Maybe Bool

True, if the story was automatically forwarded because of a mention of the user

MessageInvoice
A message with an invoice from a bot. Use getInternalLink with internalLinkTypeBotStart to share the invoice

Fields

product_info :: Maybe ProductInfo

Information about the product

currency :: Maybe Text

Currency for the product price

total_amount :: Maybe Int

Product total price in the smallest units of the currency

start_parameter :: Maybe Text

Unique invoice bot start_parameter to be passed to getInternalLink

is_test :: Maybe Bool

True, if the invoice is a test invoice

need_shipping_address :: Maybe Bool

True, if the shipping address must be specified

receipt_message_id :: Maybe Int

The identifier of the message with the receipt, after the product has been purchased

paid_media :: Maybe PaidMedia

Extended media attached to the invoice; may be null if none

paid_media_caption :: Maybe FormattedText

Extended media caption; may be null if none

MessageCall
A message with information about an ended call

Fields

is_video :: Maybe Bool

True, if the call was a video call

discard_reason :: Maybe CallDiscardReason

Reason why the call was discarded

duration :: Maybe Int

Call duration, in seconds

MessageGroupCall
A message with information about a group call not bound to a chat. If the message is incoming, the call isn't active, isn't missed, and has no duration, and getOption("can_accept_calls") is true, then incoming call screen must be shown to the user. Use joinGroupCall to accept the call or declineGroupCallInvitation to decline it. If the call become active or missed, then the call screen must be hidden

Fields

is_active :: Maybe Bool

True, if the call is active, i.e. the called user joined the call

was_missed :: Maybe Bool

True, if the called user missed or declined the call

is_video :: Maybe Bool

True, if the call was a video call

duration :: Maybe Int

Call duration, in seconds

other_participant_ids :: Maybe [MessageSender]

Identifiers of some other call participants

MessageVideoChatScheduled
A new video chat was scheduled

Fields

group_call_id :: Maybe Int

Identifier of the video chat. The video chat can be received through the method getGroupCall

start_date :: Maybe Int

Point in time (Unix timestamp) when the group call is expected to be started by an administrator

MessageVideoChatStarted
A newly created video chat

Fields

group_call_id :: Maybe Int

Identifier of the video chat. The video chat can be received through the method getGroupCall

MessageVideoChatEnded
A message with information about an ended video chat

Fields

duration :: Maybe Int

Call duration, in seconds

MessageInviteVideoChatParticipants
A message with information about an invitation to a video chat

Fields

group_call_id :: Maybe Int

Identifier of the video chat. The video chat can be received through the method getGroupCall

user_ids :: Maybe [Int]

Invited user identifiers

MessageBasicGroupChatCreate
A newly created basic group

Fields

title :: Maybe Text

Title of the basic group

member_user_ids :: Maybe [Int]

User identifiers of members in the basic group

MessageSupergroupChatCreate
A newly created supergroup or channel

Fields

title :: Maybe Text

Title of the basic group

MessageChatChangeTitle
An updated chat title

Fields

title :: Maybe Text

Title of the basic group

MessageChatChangePhoto
An updated chat photo

Fields

_photo :: Maybe ChatPhoto

New chat photo

MessageChatDeletePhoto
A deleted chat photo

MessageChatAddMembers
New chat members were added

Fields

member_user_ids :: Maybe [Int]

User identifiers of members in the basic group

MessageChatJoinByLink
A new member joined the chat via an invite link

MessageChatJoinByRequest
A new member was accepted to the chat by an administrator

MessageChatDeleteMember
A chat member was deleted

Fields

user_id :: Maybe Int

User identifier of the deleted chat member

MessageChatUpgradeTo
A basic group was upgraded to a supergroup and was deactivated as the result

Fields

supergroup_id :: Maybe Int

Identifier of the supergroup to which the basic group was upgraded

MessageChatUpgradeFrom
A supergroup has been created from a basic group

Fields

title :: Maybe Text

Title of the basic group

basic_group_id :: Maybe Int

The identifier of the original basic group

MessagePinMessage
A message has been pinned

Fields

message_id :: Maybe Int

Identifier of the pinned message, can be an identifier of a deleted message or 0

MessageScreenshotTaken
A screenshot of a message in the chat has been taken

MessageChatSetBackground
A new background was set in the chat

Fields

old_background_message_id :: Maybe Int

Identifier of the message with a previously set same background; 0 if none. Can be an identifier of a deleted message

background :: Maybe ChatBackground

The new background

only_for_self :: Maybe Bool

True, if the background was set only for self

MessageChatSetTheme
A theme in the chat has been changed

Fields

theme_name :: Maybe Text

If non-empty, name of a new theme, set for the chat. Otherwise, chat theme was reset to the default one

MessageChatSetMessageAutoDeleteTime
The auto-delete or self-destruct timer for messages in the chat has been changed

Fields

message_auto_delete_time :: Maybe Int

New value auto-delete or self-destruct time, in seconds; 0 if disabled

from_user_id :: Maybe Int

If not 0, a user identifier, which default setting was automatically applied

MessageChatBoost
The chat was boosted by the sender of the message

Fields

boost_count :: Maybe Int

Number of times the chat was boosted

MessageForumTopicCreated
A forum topic has been created

Fields

name :: Maybe Text

Name of the topic

icon :: Maybe ForumTopicIcon

Icon of the topic

MessageForumTopicEdited
A forum topic has been edited

Fields

name :: Maybe Text

Name of the topic

edit_icon_custom_emoji_id :: Maybe Bool

True, if icon's custom_emoji_id is changed

icon_custom_emoji_id :: Maybe Int

New unique identifier of the custom emoji shown on the topic icon; 0 if none. Must be ignored if edit_icon_custom_emoji_id is false

MessageForumTopicIsClosedToggled
A forum topic has been closed or opened

Fields

is_closed :: Maybe Bool

True, if the topic was closed; otherwise, the topic was reopened

MessageForumTopicIsHiddenToggled
A General forum topic has been hidden or unhidden

Fields

is_hidden :: Maybe Bool

True, if the topic was hidden; otherwise, the topic was unhidden

MessageSuggestProfilePhoto
A profile photo was suggested to a user in a private chat

Fields

_photo :: Maybe ChatPhoto

New chat photo

MessageCustomServiceAction
A non-standard action has happened in the chat

Fields

_text :: Maybe Text

Message text to be shown in the chat

MessageGameScore
A new high score was achieved in a game

Fields

game_message_id :: Maybe Int

Identifier of the message with the game, can be an identifier of a deleted message

game_id :: Maybe Int

Identifier of the game; may be different from the games presented in the message with the game

score :: Maybe Int

New score

MessagePaymentSuccessful
A payment has been sent to a bot or a business account

Fields

invoice_chat_id :: Maybe Int

Identifier of the chat, containing the corresponding invoice message

invoice_message_id :: Maybe Int

Identifier of the message with the corresponding invoice; can be 0 or an identifier of a deleted message

currency :: Maybe Text

Currency for the product price

total_amount :: Maybe Int

Product total price in the smallest units of the currency

subscription_until_date :: Maybe Int

Point in time (Unix timestamp) when the subscription will expire; 0 if unknown or the payment isn't recurring

is_recurring :: Maybe Bool

True, if this is a recurring payment

is_first_recurring :: Maybe Bool

True, if this is the first recurring payment

invoice_name :: Maybe Text

Name of the invoice; may be empty if unknown

MessagePaymentSuccessfulBot
A payment has been received by the bot or the business account

Fields

currency :: Maybe Text

Currency for the product price

total_amount :: Maybe Int

Product total price in the smallest units of the currency

subscription_until_date :: Maybe Int

Point in time (Unix timestamp) when the subscription will expire; 0 if unknown or the payment isn't recurring

is_recurring :: Maybe Bool

True, if this is a recurring payment

is_first_recurring :: Maybe Bool

True, if this is the first recurring payment

invoice_payload :: Maybe ByteString

Invoice payload

shipping_option_id :: Maybe Text

Identifier of the shipping option chosen by the user; may be empty if not applicable; for bots only

order_info :: Maybe OrderInfo

Information about the order; may be null; for bots only

telegram_payment_charge_id :: Maybe Text

Telegram payment identifier

provider_payment_charge_id :: Maybe Text

Provider payment identifier

MessagePaymentRefunded
A payment has been refunded

Fields

owner_id :: Maybe MessageSender

Identifier of the previous owner of the Telegram Stars that refunds them

currency :: Maybe Text

Currency for the product price

total_amount :: Maybe Int

Product total price in the smallest units of the currency

invoice_payload :: Maybe ByteString

Invoice payload

telegram_payment_charge_id :: Maybe Text

Telegram payment identifier

provider_payment_charge_id :: Maybe Text

Provider payment identifier

MessageGiftedPremium
Telegram Premium was gifted to a user

Fields

gifter_user_id :: Maybe Int

The identifier of a user that gifted Telegram Premium; 0 if the gift was anonymous or is outgoing

receiver_user_id :: Maybe Int

The identifier of a user that received Telegram Premium; 0 if the gift is incoming

text :: Maybe FormattedText

Text of the message

currency :: Maybe Text

Currency for the product price

amount :: Maybe Int

The paid amount, in the smallest units of the currency

cryptocurrency :: Maybe Text

Cryptocurrency used to pay for the gift; may be empty if none

cryptocurrency_amount :: Maybe Int

The paid amount, in the smallest units of the cryptocurrency; 0 if none

month_count :: Maybe Int

Number of months the Telegram Premium subscription will be active

sticker :: Maybe Sticker

The sticker description

MessagePremiumGiftCode
A Telegram Premium gift code was created for the user

Fields

creator_id :: Maybe MessageSender

Identifier of a chat or a user that created the gift code; may be null if unknown

text :: Maybe FormattedText

Text of the message

is_from_giveaway :: Maybe Bool

True, if the gift code was created for a giveaway

is_unclaimed :: Maybe Bool

True, if the winner for the corresponding Telegram Premium subscription wasn't chosen

currency :: Maybe Text

Currency for the product price

amount :: Maybe Int

The paid amount, in the smallest units of the currency

cryptocurrency :: Maybe Text

Cryptocurrency used to pay for the gift; may be empty if none

cryptocurrency_amount :: Maybe Int

The paid amount, in the smallest units of the cryptocurrency; 0 if none

month_count :: Maybe Int

Number of months the Telegram Premium subscription will be active

sticker :: Maybe Sticker

The sticker description

code :: Maybe Text

The gift code

MessageGiveawayCreated
A giveaway was created for the chat. Use telegramPaymentPurposePremiumGiveaway, storePaymentPurposePremiumGiveaway, telegramPaymentPurposeStarGiveaway, or storePaymentPurposeStarGiveaway to create a giveaway

Fields

star_count :: Maybe Int

Number of Telegram Stars needed to buy access to the media in the message

MessageGiveaway
A giveaway

Fields

parameters :: Maybe GiveawayParameters

Giveaway parameters

winner_count :: Maybe Int

Number of users which will receive Telegram Premium subscription gift codes

prize :: Maybe GiveawayPrize

Prize of the giveaway

sticker :: Maybe Sticker

The sticker description

MessageGiveawayCompleted
A giveaway without public winners has been completed for the chat

Fields

giveaway_message_id :: Maybe Int

Identifier of the message with the giveaway; can be 0 if the message was deleted

winner_count :: Maybe Int

Number of users which will receive Telegram Premium subscription gift codes

is_star_giveaway :: Maybe Bool

True, if the giveaway is a Telegram Star giveaway

unclaimed_prize_count :: Maybe Int

Number of undistributed prizes; for Telegram Premium giveaways only

MessageGiveawayWinners
A giveaway with public winners has been completed for the chat

Fields

boosted_chat_id :: Maybe Int

Identifier of the supergroup or channel chat, which was automatically boosted by the winners of the giveaway

giveaway_message_id :: Maybe Int

Identifier of the message with the giveaway; can be 0 if the message was deleted

additional_chat_count :: Maybe Int

Number of other chats that participated in the giveaway

actual_winners_selection_date :: Maybe Int

Point in time (Unix timestamp) when the winners were selected. May be bigger than winners selection date specified in parameters of the giveaway

only_new_members :: Maybe Bool

True, if only new members of the chats were eligible for the giveaway

was_refunded :: Maybe Bool

True, if the giveaway was canceled and was fully refunded

prize :: Maybe GiveawayPrize

Prize of the giveaway

prize_description :: Maybe Text

Additional description of the giveaway prize

winner_count :: Maybe Int

Number of users which will receive Telegram Premium subscription gift codes

winner_user_ids :: Maybe [Int]

Up to 100 user identifiers of the winners of the giveaway

unclaimed_prize_count :: Maybe Int

Number of undistributed prizes; for Telegram Premium giveaways only

MessageGiftedStars
Telegram Stars were gifted to a user

Fields

gifter_user_id :: Maybe Int

The identifier of a user that gifted Telegram Premium; 0 if the gift was anonymous or is outgoing

receiver_user_id :: Maybe Int

The identifier of a user that received Telegram Premium; 0 if the gift is incoming

currency :: Maybe Text

Currency for the product price

amount :: Maybe Int

The paid amount, in the smallest units of the currency

cryptocurrency :: Maybe Text

Cryptocurrency used to pay for the gift; may be empty if none

cryptocurrency_amount :: Maybe Int

The paid amount, in the smallest units of the cryptocurrency; 0 if none

star_count :: Maybe Int

Number of Telegram Stars needed to buy access to the media in the message

transaction_id :: Maybe Text

Identifier of the transaction for Telegram Stars purchase; for receiver only

sticker :: Maybe Sticker

The sticker description

MessageGiveawayPrizeStars
A Telegram Stars were received by the current user from a giveaway

Fields

star_count :: Maybe Int

Number of Telegram Stars needed to buy access to the media in the message

transaction_id :: Maybe Text

Identifier of the transaction for Telegram Stars purchase; for receiver only

boosted_chat_id :: Maybe Int

Identifier of the supergroup or channel chat, which was automatically boosted by the winners of the giveaway

giveaway_message_id :: Maybe Int

Identifier of the message with the giveaway; can be 0 if the message was deleted

is_unclaimed :: Maybe Bool

True, if the winner for the corresponding Telegram Premium subscription wasn't chosen

sticker :: Maybe Sticker

The sticker description

MessageGift
A regular gift was received or sent by the current user, or the current user was notified about a channel gift

Fields

gift :: Maybe Gift

The gift

sender_id :: Maybe MessageSender

Sender of the gift

received_gift_id :: Maybe Text

Unique identifier of the received gift for the current user; only for the receiver of the gift

text :: Maybe FormattedText

Text of the message

sell_star_count :: Maybe Int

Number of Telegram Stars that can be claimed by the receiver instead of the regular gift; 0 if the gift can't be sold by the receiver

prepaid_upgrade_star_count :: Maybe Int

Number of Telegram Stars that were paid by the sender for the ability to upgrade the gift

is_private :: Maybe Bool

True, if the sender and gift text are shown only to the gift receiver; otherwise, everyone will be able to see them

is_saved :: Maybe Bool

True, if the gift is displayed on the user's or the channel's profile page; only for the receiver of the gift

can_be_upgraded :: Maybe Bool

True, if the gift can be upgraded to a unique gift; only for the receiver of the gift

was_converted :: Maybe Bool

True, if the gift was converted to Telegram Stars; only for the receiver of the gift

was_upgraded :: Maybe Bool

True, if the gift was upgraded to a unique gift

was_refunded :: Maybe Bool

True, if the giveaway was canceled and was fully refunded

upgraded_received_gift_id :: Maybe Text

Identifier of the corresponding upgraded gift; may be empty if unknown. Use getReceivedGift to get information about the gift

MessageUpgradedGift
An upgraded gift was received or sent by the current user, or the current user was notified about a channel gift

Fields

_gift :: Maybe UpgradedGift

The gift

sender_id :: Maybe MessageSender

Sender of the gift

received_gift_id :: Maybe Text

Unique identifier of the received gift for the current user; only for the receiver of the gift

is_upgrade :: Maybe Bool

True, if the gift was obtained by upgrading of a previously received gift; otherwise, this is a transferred gift

is_saved :: Maybe Bool

True, if the gift is displayed on the user's or the channel's profile page; only for the receiver of the gift

can_be_transferred :: Maybe Bool

True, if the gift can be transferred to another owner; only for the receiver of the gift

was_transferred :: Maybe Bool

True, if the gift was transferred to another owner; only for the receiver of the gift

transfer_star_count :: Maybe Int

Number of Telegram Stars that must be paid to transfer the upgraded gift; only for the receiver of the gift

export_date :: Maybe Int

Point in time (Unix timestamp) when the gift can be transferred to the TON blockchain as an NFT; 0 if NFT export isn't possible; only for the receiver of the gift

MessageRefundedUpgradedGift
A gift which purchase, upgrade or transfer were refunded

Fields

gift :: Maybe Gift

The gift

sender_id :: Maybe MessageSender

Sender of the gift

is_upgrade :: Maybe Bool

True, if the gift was obtained by upgrading of a previously received gift; otherwise, this is a transferred gift

MessagePaidMessagesRefunded
Paid messages were refunded

Fields

message_count :: Maybe Int

The number of refunded messages

star_count :: Maybe Int

Number of Telegram Stars needed to buy access to the media in the message

MessagePaidMessagePriceChanged
A price for paid messages was changed in the supergroup chat

Fields

paid_message_star_count :: Maybe Int

The new number of Telegram Stars that must be paid by non-administrator users of the supergroup chat for each sent message

MessageContactRegistered
A contact has registered with Telegram

MessageUsersShared
The current user shared users, which were requested by the bot

Fields

users :: Maybe [SharedUser]

The shared users

button_id :: Maybe Int

Identifier of the keyboard button with the request

MessageChatShared
The current user shared a chat, which was requested by the bot

Fields

chat :: Maybe SharedChat

The shared chat

button_id :: Maybe Int

Identifier of the keyboard button with the request

MessageBotWriteAccessAllowed
The user allowed the bot to send messages

Fields

reason :: Maybe BotWriteAccessAllowReason

The reason why the bot was allowed to write messages

MessageWebAppDataSent
Data from a Web App has been sent to a bot

Fields

button_text :: Maybe Text

Text of the keyboardButtonTypeWebApp button, which opened the Web App

MessageWebAppDataReceived
Data from a Web App has been received; for bots only

Fields

button_text :: Maybe Text

Text of the keyboardButtonTypeWebApp button, which opened the Web App

_data :: Maybe Text

The data

MessagePassportDataSent
Telegram Passport data has been sent to a bot

Fields

types :: Maybe [PassportElementType]

List of Telegram Passport element types sent

MessagePassportDataReceived
Telegram Passport data has been received; for bots only

Fields

elements :: Maybe [EncryptedPassportElement]

List of received Telegram Passport elements

credentials :: Maybe EncryptedCredentials

Encrypted data credentials

MessageProximityAlertTriggered
A user in the chat came within proximity alert range

Fields

traveler_id :: Maybe MessageSender

The identifier of a user or chat that triggered the proximity alert

watcher_id :: Maybe MessageSender

The identifier of a user or chat that subscribed for the proximity alert

distance :: Maybe Int

The distance between the users

MessageUnsupported
A message content that is not supported in the current TDLib version