TD.Data.MessageEffect

Documentation

data MessageEffect

Constructors

MessageEffect
Contains information about an effect added to a message

Fields

_id :: Maybe Int

Unique identifier of the effect

static_icon :: Maybe Sticker

Static icon for the effect in WEBP format; may be null if none

emoji :: Maybe Text

Emoji corresponding to the effect that can be used if static icon isn't available

is_premium :: Maybe Bool

True, if Telegram Premium subscription is required to use the effect

_type :: Maybe MessageEffectType

Type of the effect