TD.Data.MessageExtendedMedia

Documentation

data MessageExtendedMedia

Describes a media, which is attached to an invoice

Constructors

MessageExtendedMediaPreview
The media is hidden until the invoice is paid

Fields

width :: Maybe Int

Media width; 0 if unknown

height :: Maybe Int

Media height; 0 if unknown

duration :: Maybe Int

Media duration; 0 if unknown

minithumbnail :: Maybe Minithumbnail

Media minithumbnail; may be null

caption :: Maybe FormattedText

Media caption

MessageExtendedMediaPhoto
The media is a photo

Fields

photo :: Maybe Photo

The photo

caption :: Maybe FormattedText

Media caption

MessageExtendedMediaVideo
The media is a video

Fields

video :: Maybe Video

The video

caption :: Maybe FormattedText

Media caption

MessageExtendedMediaUnsupported
The media is unsupported

Fields

caption :: Maybe FormattedText

Media caption