TD.Data.MessageForwardOrigin

Description

Documentation

data MessageForwardOrigin

Contains information about the origin of a forwarded message

Constructors

MessageForwardOriginUser
The message was originally sent by a known user @sender_user_id Identifier of the user that originally sent the message

Fields

sender_user_id :: Maybe Int

MessageForwardOriginChat
The message was originally sent on behalf of a chat

Fields

author_signature :: Maybe String

For messages originally sent by an anonymous chat administrator, original message author signature

sender_chat_id :: Maybe Int

Identifier of the chat that originally sent the message

MessageForwardOriginHiddenUser
The message was originally sent by a user, which is hidden by their privacy settings @sender_name Name of the sender

Fields

sender_name :: Maybe String

MessageForwardOriginChannel
The message was originally a post in a channel

Fields

author_signature :: Maybe String

For messages originally sent by an anonymous chat administrator, original message author signature

message_id :: Maybe Int

Message identifier of the original message

chat_id :: Maybe Int

Identifier of the chat from which the message was originally forwarded

MessageForwardOriginMessageImport
The message was imported from an exported message history @sender_name Name of the sender

Fields

sender_name :: Maybe String