TD.Data.MessageOrigin

Documentation

data MessageOrigin

Contains information about the origin of a message

Constructors

MessageOriginUser
The message was originally sent by a known user

Fields

sender_user_id :: Maybe Int

Identifier of the user that originally sent the message

MessageOriginHiddenUser
The message was originally sent by a user, which is hidden by their privacy settings

Fields

sender_name :: Maybe Text

Name of the sender

MessageOriginChat
The message was originally sent on behalf of a chat

Fields

sender_chat_id :: Maybe Int

Identifier of the chat that originally sent the message

author_signature :: Maybe Text

For messages originally sent by an anonymous chat administrator, original message author signature

MessageOriginChannel
The message was originally a post in a channel

Fields

chat_id :: Maybe Int

Identifier of the channel chat to which the message was originally sent

message_id :: Maybe Int

Message identifier of the original message

author_signature :: Maybe Text

For messages originally sent by an anonymous chat administrator, original message author signature