TD.Data.MessageReadDate

Documentation

data MessageReadDate

Describes read date of a recent outgoing message in a private chat

Constructors

MessageReadDateRead
Contains read date of the message

Fields

read_date :: Maybe Int

Point in time (Unix timestamp) when the message was read by the other user

MessageReadDateUnread
The message is unread yet

MessageReadDateTooOld
The message is too old to get read date

MessageReadDateUserPrivacyRestricted
The read date is unknown due to privacy settings of the other user

MessageReadDateMyPrivacyRestricted
The read date is unknown due to privacy settings of the current user, but will be known if the user subscribes to Telegram Premium