TD.Data.MessageReplyInfo

Description

Documentation

data MessageReplyInfo

Constructors

MessageReplyInfo
Contains information about replies to a message

Fields

last_message_id :: Maybe Int

Identifier of the last reply to the message

last_read_outbox_message_id :: Maybe Int

Identifier of the last read outgoing reply to the message

last_read_inbox_message_id :: Maybe Int

Identifier of the last read incoming reply to the message

recent_replier_ids :: Maybe [MessageSender]

Identifiers of at most 3 recent repliers to the message; available in channels with a discussion supergroup. The users and chats are expected to be inaccessible: only their photo and name will be available

reply_count :: Maybe Int

Number of times the message was directly or indirectly replied