TD.Data.MessageReplyTo

Documentation

data MessageReplyTo

Contains information about the message or the story a message is replying to

Constructors

MessageReplyToMessage
Describes a message replied by a given message

Fields

chat_id :: Maybe Int

The identifier of the chat to which the message belongs; may be 0 if the replied message is in unknown chat

message_id :: Maybe Int

The identifier of the message; may be 0 if the replied message is in unknown chat

quote :: Maybe TextQuote

Chosen quote from the replied message; may be null if none

origin :: Maybe MessageOrigin

Information about origin of the message if the message was from another chat or topic; may be null for messages from the same chat

origin_send_date :: Maybe Int

Point in time (Unix timestamp) when the message was sent if the message was from another chat or topic; 0 for messages from the same chat

content :: Maybe MessageContent

Media content of the message if the message was from another chat or topic; may be null for messages from the same chat and messages without media. Can be only one of the following types: messageAnimation, messageAudio, messageContact, messageDice, messageDocument, messageGame, messageGiveaway, messageGiveawayWinners, messageInvoice, messageLocation, messagePaidMedia, messagePhoto, messagePoll, messageSticker, messageStory, messageText (for link preview), messageVenue, messageVideo, messageVideoNote, or messageVoiceNote

MessageReplyToStory
Describes a story replied by a given message

Fields

story_sender_chat_id :: Maybe Int

The identifier of the sender of the story

story_id :: Maybe Int

The identifier of the story