TD.Data.MessageSendOptions

Description

Documentation

data MessageSendOptions

Constructors

MessageSendOptions
Options to be used when a message is sent

Fields

sending_id :: Maybe Int

Non-persistent identifier, which will be returned back in messageSendingStatePending object and can be used to match sent messages and corresponding updateNewMessage updates

scheduling_state :: Maybe MessageSchedulingState

Message scheduling state; pass null to send message immediately. Messages sent to a secret chat, live location messages and self-destructing messages can't be scheduled

update_order_of_installed_sticker_sets :: Maybe Bool

Pass true if the user explicitly chosen a sticker or a custom emoji from an installed sticker set; applicable only to sendMessage and sendMessageAlbum

protect_content :: Maybe Bool

Pass true if the content of the message must be protected from forwarding and saving; for bots only

from_background :: Maybe Bool

Pass true if the message is sent from the background

disable_notification :: Maybe Bool

Pass true to disable notification for the message