TD.Data.MessageSendingState

Documentation

data MessageSendingState

Contains information about the sending state of the message

Constructors

MessageSendingStatePending
The message is being sent now, but has not yet been delivered to the server

Fields

sending_id :: Maybe Int

Non-persistent message sending identifier, specified by the application

MessageSendingStateFailed
The message failed to be sent

Fields

_error :: Maybe Error

The cause of the message sending failure

can_retry :: Maybe Bool

True, if the message can be re-sent

need_another_sender :: Maybe Bool

True, if the message can be re-sent only on behalf of a different sender

need_another_reply_quote :: Maybe Bool

True, if the message can be re-sent only if another quote is chosen in the message that is replied by the given message

need_drop_reply :: Maybe Bool

True, if the message can be re-sent only if the message to be replied is removed. This will be done automatically by resendMessages

retry_after :: Maybe Double

Time left before the message can be re-sent, in seconds. No update is sent when this field changes