TD.Data.MessageSponsorType

Documentation

data MessageSponsorType

Describes type of a message sponsor

Constructors

MessageSponsorTypeBot
The sponsor is a bot

Fields

bot_user_id :: Maybe Int

User identifier of the bot

link :: Maybe InternalLinkType

An internal link to be opened when the sponsored message is clicked

MessageSponsorTypeWebApp
The sponsor is a web app

Fields

web_app_title :: Maybe Text

Web App title

link :: Maybe InternalLinkType

An internal link to be opened when the sponsored message is clicked

MessageSponsorTypePublicChannel
The sponsor is a public channel chat

Fields

chat_id :: Maybe Int

Sponsor chat identifier

link :: Maybe InternalLinkType

An internal link to be opened when the sponsored message is clicked

MessageSponsorTypePrivateChannel
The sponsor is a private channel chat

Fields

title :: Maybe Text

Title of the chat

invite_link :: Maybe Text

Invite link for the channel

MessageSponsorTypeWebsite
The sponsor is a website

Fields

url :: Maybe Text

URL of the website

name :: Maybe Text

Name of the website