TD.Data.MessageThreadInfo

Description

Documentation

data MessageThreadInfo

Constructors

MessageThreadInfo
Contains information about a message thread

Fields

draft_message :: Maybe DraftMessage

A draft of a message in the message thread; may be null

messages :: Maybe [Message]

The messages from which the thread starts. The messages are returned in a reverse chronological order (i.e., in order of decreasing message_id)

unread_message_count :: Maybe Int

Approximate number of unread messages in the message thread

reply_info :: Maybe MessageReplyInfo

Information about the message thread; may be null for forum topic threads

message_thread_id :: Maybe Int

Message thread identifier, unique within the chat

chat_id :: Maybe Int

Identifier of the chat to which the message thread belongs