TD.Data.NetworkStatisticsEntry

Description

Documentation

data NetworkStatisticsEntry

Contains statistics about network usage

Constructors

NetworkStatisticsEntryFile
Contains information about the total amount of data that was used to send and receive files

Fields

received_bytes :: Maybe Int

sent_bytes :: Maybe Int

Total number of bytes sent @received_bytes Total number of bytes received

network_type :: Maybe NetworkType

Type of the network the data was sent through. Call setNetworkType to maintain the actual network type

file_type :: Maybe FileType

Type of the file the data is part of; pass null if the data isn't related to files

NetworkStatisticsEntryCall
Contains information about the total amount of data that was used for calls

Fields

duration :: Maybe Float

Total call duration, in seconds

received_bytes :: Maybe Int

sent_bytes :: Maybe Int

Total number of bytes sent @received_bytes Total number of bytes received

network_type :: Maybe NetworkType

Type of the network the data was sent through. Call setNetworkType to maintain the actual network type