TD.Data.NotificationGroupType

Description

Documentation

data NotificationGroupType

Describes the type of notifications in a notification group

Constructors

NotificationGroupTypeMessages
A group containing notifications of type notificationTypeNewMessage and notificationTypeNewPushMessage with ordinary unread messages

NotificationGroupTypeMentions
A group containing notifications of type notificationTypeNewMessage and notificationTypeNewPushMessage with unread mentions of the current user, replies to their messages, or a pinned message

NotificationGroupTypeSecretChat
A group containing a notification of type notificationTypeNewSecretChat

NotificationGroupTypeCalls
A group containing notifications of type notificationTypeNewCall