TD.Data.NotificationType

Documentation

data NotificationType

Contains detailed information about a notification

Constructors

NotificationTypeNewMessage
New message was received

Fields

message :: Maybe Message

The message

show_preview :: Maybe Bool

True, if message content must be displayed in notifications

NotificationTypeNewSecretChat
New secret chat was created

NotificationTypeNewCall
New call was received

Fields

call_id :: Maybe Int

Call identifier

NotificationTypeNewPushMessage
New message was received through a push notification

Fields

message_id :: Maybe Int

The message identifier. The message will not be available in the chat history, but the identifier can be used in viewMessages, or as a message to be replied in the same chat

sender_id :: Maybe MessageSender

Identifier of the sender of the message. Corresponding user or chat may be inaccessible

sender_name :: Maybe Text

Name of the sender

is_outgoing :: Maybe Bool

True, if the message is outgoing

content :: Maybe PushMessageContent

Push message content