TD.Data.OauthLinkInfo

Documentation

data OauthLinkInfo

Constructors

OauthLinkInfo
Information about the OAuth authorization

Fields

user_id :: Maybe Int

Identifier of the user for which the link was generated; may be 0 if unknown. The corresponding user may be unknown. If the user is logged in the app, then they must be chosen for authorization by default

url :: Maybe Text

An HTTP URL where the user authorizes

domain :: Maybe Text

A domain of the URL

bot_user_id :: Maybe Int

User identifier of a bot linked with the website

request_write_access :: Maybe Bool

True, if the user must be asked for the permission to the bot to send them messages

request_phone_number_access :: Maybe Bool

True, if the user must be asked for the permission to share their phone number

browser :: Maybe Text

The version of a browser used for the authorization

platform :: Maybe Text

Operating system the browser is running on

ip_address :: Maybe Text

IP address from which the authorization is performed, in human-readable format

location :: Maybe Text

Human-readable description of a country and a region from which the authorization is performed, based on the IP address

match_code_first :: Maybe Bool

True, if code matching dialog must be shown first and checkOauthRequestMatchCode must be called before acceptOauthRequest. Otherwise, checkOauthRequestMatchCode must not be called

match_codes :: Maybe [Text]

The list of codes to match; may be empty if irrelevant