TD.Data.PageBlock

Documentation

data PageBlock

Describes a block of an instant view web page

Constructors

PageBlockTitle
The title of a page

Fields

title :: Maybe RichText

Title

PageBlockSubtitle
The subtitle of a page

Fields

subtitle :: Maybe RichText

Subtitle

PageBlockAuthorDate
The author and publishing date of a page

Fields

author :: Maybe RichText

Author

publish_date :: Maybe Int

Point in time (Unix timestamp) when the article was published; 0 if unknown

PageBlockHeader
A header

Fields

header :: Maybe RichText

Header

PageBlockSubheader
A subheader

Fields

subheader :: Maybe RichText

Subheader

PageBlockKicker
A kicker

Fields

kicker :: Maybe RichText

Kicker

PageBlockParagraph
A text paragraph

Fields

text :: Maybe RichText

Paragraph text

PageBlockPreformatted
A preformatted text paragraph

Fields

text :: Maybe RichText

Paragraph text

language :: Maybe Text

Programming language for which the text needs to be formatted

PageBlockFooter
The footer of a page

Fields

footer :: Maybe RichText

Footer

PageBlockDivider
An empty block separating a page

PageBlockAnchor
An invisible anchor on a page, which can be used in a URL to open the page from the specified anchor

Fields

name :: Maybe Text

Name of the anchor

PageBlockList
A list of data blocks

Fields

items :: Maybe [PageBlockListItem]

The items of the list

PageBlockBlockQuote
A block quote

Fields

text :: Maybe RichText

Paragraph text

credit :: Maybe RichText

Quote credit

PageBlockPullQuote
A pull quote

Fields

text :: Maybe RichText

Paragraph text

credit :: Maybe RichText

Quote credit

PageBlockAnimation
An animation

Fields

animation :: Maybe Animation

Animation file; may be null

caption :: Maybe PageBlockCaption

Animation caption

need_autoplay :: Maybe Bool

True, if the animation must be played automatically

PageBlockAudio
An audio file

Fields

audio :: Maybe Audio

Audio file; may be null

caption :: Maybe PageBlockCaption

Animation caption

PageBlockPhoto
A photo

Fields

photo :: Maybe Photo

Photo file; may be null

caption :: Maybe PageBlockCaption

Animation caption

url :: Maybe Text

URL that needs to be opened when the photo is clicked

PageBlockVideo
A video

Fields

video :: Maybe Video

Video file; may be null

caption :: Maybe PageBlockCaption

Animation caption

need_autoplay :: Maybe Bool

True, if the animation must be played automatically

is_looped :: Maybe Bool

True, if the video must be looped

PageBlockVoiceNote
A voice note

Fields

voice_note :: Maybe VoiceNote

Voice note; may be null

caption :: Maybe PageBlockCaption

Animation caption

PageBlockCover
A page cover

Fields

cover :: Maybe PageBlock

Cover

PageBlockEmbedded
An embedded web page

Fields

url :: Maybe Text

URL that needs to be opened when the photo is clicked

html :: Maybe Text

HTML-markup of the embedded page

poster_photo :: Maybe Photo

Poster photo, if available; may be null

width :: Maybe Int

Block width; 0 if unknown

height :: Maybe Int

Block height; 0 if unknown

caption :: Maybe PageBlockCaption

Animation caption

is_full_width :: Maybe Bool

True, if the block must be full width

allow_scrolling :: Maybe Bool

True, if scrolling needs to be allowed

PageBlockEmbeddedPost
An embedded post

Fields

url :: Maybe Text

URL that needs to be opened when the photo is clicked

_author :: Maybe Text

Post author

author_photo :: Maybe Photo

Post author photo; may be null

date :: Maybe Int

Point in time (Unix timestamp) when the post was created; 0 if unknown

page_blocks :: Maybe [PageBlock]

Post content

caption :: Maybe PageBlockCaption

Animation caption

PageBlockCollage
A collage

Fields

page_blocks :: Maybe [PageBlock]

Post content

caption :: Maybe PageBlockCaption

Animation caption

PageBlockSlideshow
A slideshow

Fields

page_blocks :: Maybe [PageBlock]

Post content

caption :: Maybe PageBlockCaption

Animation caption

PageBlockChatLink
A link to a chat

Fields

_title :: Maybe Text

Chat title

_photo :: Maybe ChatPhotoInfo

Chat photo; may be null

username :: Maybe Text

Chat username by which all other information about the chat can be resolved

PageBlockTable
A table

Fields

_caption :: Maybe RichText

Table caption

cells :: Maybe [[PageBlockTableCell]]

Table cells

is_bordered :: Maybe Bool

True, if the table is bordered

is_striped :: Maybe Bool

True, if the table is striped

PageBlockDetails
A collapsible block

Fields

header :: Maybe RichText

Header

page_blocks :: Maybe [PageBlock]

Post content

is_open :: Maybe Bool

True, if the block is open by default

PageBlockRelatedArticles
Related articles

Fields

header :: Maybe RichText

Header

articles :: Maybe [PageBlockRelatedArticle]

List of related articles

PageBlockMap
A map

Fields

location :: Maybe Location

Location of the map center

zoom :: Maybe Int

Map zoom level

width :: Maybe Int

Block width; 0 if unknown

height :: Maybe Int

Block height; 0 if unknown

caption :: Maybe PageBlockCaption

Animation caption

defaultPageBlockTitle :: PageBlock

defaultPageBlockSubtitle :: PageBlock

defaultPageBlockAuthorDate :: PageBlock

defaultPageBlockHeader :: PageBlock

defaultPageBlockSubheader :: PageBlock

defaultPageBlockKicker :: PageBlock

defaultPageBlockParagraph :: PageBlock

defaultPageBlockPreformatted :: PageBlock

defaultPageBlockFooter :: PageBlock

defaultPageBlockAnchor :: PageBlock

defaultPageBlockList :: PageBlock

defaultPageBlockBlockQuote :: PageBlock

defaultPageBlockPullQuote :: PageBlock

defaultPageBlockAnimation :: PageBlock

defaultPageBlockAudio :: PageBlock

defaultPageBlockPhoto :: PageBlock

defaultPageBlockVideo :: PageBlock

defaultPageBlockVoiceNote :: PageBlock

defaultPageBlockCover :: PageBlock

defaultPageBlockEmbedded :: PageBlock

defaultPageBlockEmbeddedPost :: PageBlock

defaultPageBlockCollage :: PageBlock

defaultPageBlockSlideshow :: PageBlock

defaultPageBlockChatLink :: PageBlock

defaultPageBlockTable :: PageBlock

defaultPageBlockDetails :: PageBlock

defaultPageBlockRelatedArticles :: PageBlock

defaultPageBlockMap :: PageBlock