TD.Data.PassportSuitableElement

Documentation

data PassportSuitableElement

Constructors

PassportSuitableElement
Contains information about a Telegram Passport element that was requested by a service

Fields

_type :: Maybe PassportElementType

Type of the element

is_selfie_required :: Maybe Bool

True, if a selfie is required with the identity document

is_translation_required :: Maybe Bool

True, if a certified English translation is required with the document

is_native_name_required :: Maybe Bool

True, if personal details must include the user's name in the language of their country of residence