TD.Data.PasswordState

Documentation

data PasswordState

Constructors

PasswordState
Represents the current state of 2-step verification

Fields

has_password :: Maybe Bool

True, if a 2-step verification password is set

password_hint :: Maybe Text

Hint for the password; may be empty

has_recovery_email_address :: Maybe Bool

True, if a recovery email is set

has_passport_data :: Maybe Bool

True, if some Telegram Passport elements were saved

recovery_email_address_code_info :: Maybe EmailAddressAuthenticationCodeInfo

Information about the recovery email address to which the confirmation email was sent; may be null

login_email_address_pattern :: Maybe Text

Pattern of the email address set up for logging in

pending_reset_date :: Maybe Int

If not 0, point in time (Unix timestamp) after which the 2-step verification password can be reset immediately using resetPassword