TD.Data.PaymentForm

Description

Documentation

data PaymentForm

Constructors

PaymentForm
Contains information about an invoice payment form

Fields

product_photo :: Maybe Photo

Product photo; may be null

product_description :: Maybe FormattedText

Product description

product_title :: Maybe String

Product title

need_password :: Maybe Bool

True, if the user will be able to save credentials, if sets up a 2-step verification password

can_save_credentials :: Maybe Bool

True, if the user can choose to save credentials

saved_credentials :: Maybe [SavedCredentials]

The list of saved payment credentials

saved_order_info :: Maybe OrderInfo

Saved server-side order information; may be null

additional_payment_options :: Maybe [PaymentOption]

The list of additional payment options

payment_provider :: Maybe PaymentProvider

Information about the payment provider

payment_provider_user_id :: Maybe Int

User identifier of the payment provider bot

seller_bot_user_id :: Maybe Int

User identifier of the seller bot

invoice :: Maybe Invoice

Full information about the invoice

_id :: Maybe Int

The payment form identifier