TD.Data.PaymentFormType

Documentation

data PaymentFormType

Describes type of payment form

Constructors

PaymentFormTypeRegular
The payment form is for a regular payment

Fields

invoice :: Maybe Invoice

Full information about the invoice

payment_provider_user_id :: Maybe Int

User identifier of the payment provider bot

payment_provider :: Maybe PaymentProvider

Information about the payment provider

additional_payment_options :: Maybe [PaymentOption]

The list of additional payment options

saved_order_info :: Maybe OrderInfo

Saved server-side order information; may be null

saved_credentials :: Maybe [SavedCredentials]

The list of saved payment credentials

can_save_credentials :: Maybe Bool

True, if the user can choose to save credentials

need_password :: Maybe Bool

True, if the user will be able to save credentials, if sets up a 2-step verification password

PaymentFormTypeStars
The payment form is for a payment in Telegram Stars

Fields

star_count :: Maybe Int

Number of Telegram Stars that will be paid

PaymentFormTypeStarSubscription
The payment form is for a payment in Telegram Stars for subscription

Fields

pricing :: Maybe StarSubscriptionPricing

Information about subscription plan