TD.Data.PaymentReceipt

Description

Documentation

data PaymentReceipt

Constructors

PaymentReceipt
Contains information about a successful payment

Fields

tip_amount :: Maybe Int

The amount of tip chosen by the buyer in the smallest units of the currency

credentials_title :: Maybe String

Title of the saved credentials chosen by the buyer

shipping_option :: Maybe ShippingOption

Chosen shipping option; may be null

order_info :: Maybe OrderInfo

Order information; may be null

invoice :: Maybe Invoice

Information about the invoice

payment_provider_user_id :: Maybe Int

User identifier of the payment provider bot

seller_bot_user_id :: Maybe Int

User identifier of the seller bot

date :: Maybe Int

Point in time (Unix timestamp) when the payment was made

photo :: Maybe Photo

Product photo; may be null

description :: Maybe FormattedText

title :: Maybe String

Product title