TD.Data.PaymentReceiptType

Documentation

data PaymentReceiptType

Describes type of successful payment

Constructors

PaymentReceiptTypeRegular
The payment was done using a third-party payment provider

Fields

payment_provider_user_id :: Maybe Int

User identifier of the payment provider bot

invoice :: Maybe Invoice

Information about the invoice

order_info :: Maybe OrderInfo

Order information; may be null

shipping_option :: Maybe ShippingOption

Chosen shipping option; may be null

credentials_title :: Maybe Text

Title of the saved credentials chosen by the buyer

tip_amount :: Maybe Int

The amount of tip chosen by the buyer in the smallest units of the currency

PaymentReceiptTypeStars
The payment was done using Telegram Stars

Fields

star_count :: Maybe Int

Number of Telegram Stars that were paid

transaction_id :: Maybe Text

Unique identifier of the transaction that can be used to dispute it