TD.Data.PhoneNumberInfo

Description

Documentation

data PhoneNumberInfo

Constructors

PhoneNumberInfo
Contains information about a phone number

Fields

is_anonymous :: Maybe Bool

True, if the phone number was bought on Fragment and isn't tied to a SIM card

formatted_phone_number :: Maybe String

The phone number without country calling code formatted accordingly to local rules. Expected digits are returned as -, but even more digits might be entered by the user

country_calling_code :: Maybe String

The part of the phone number denoting country calling code or its part

country :: Maybe CountryInfo

Information about the country to which the phone number belongs; may be null