TD.Data.Poll

Documentation

data Poll

Constructors

Poll
Describes a poll id Unique poll identifier question Poll question; 1-300 characters @options List of poll answer options

Fields

is_closed :: Maybe Bool

close_date :: Maybe Int

open_period :: Maybe Int

Amount of time the poll will be active after creation, in seconds close_date Point in time (Unix timestamp) when the poll will automatically be closed is_closed True, if the poll is closed

_type :: Maybe PollType

is_anonymous :: Maybe Bool

True, if the poll is anonymous @type Type of the poll

recent_voter_user_ids :: Maybe [Int]

total_voter_count :: Maybe Int

Total number of voters, participating in the poll @recent_voter_user_ids User identifiers of recent voters, if the poll is non-anonymous

options :: Maybe [PollOption]

question :: Maybe String

_id :: Maybe Int