TD.Data.PollType

Documentation

data PollType

Describes the type of poll

Constructors

PollTypeRegular
A regular poll

PollTypeQuiz
A poll in quiz mode, which has predefined correct answers

Fields

correct_option_ids :: Maybe [Int]

Increasing list of 0-based identifiers of the correct answer options; empty for a yet unanswered poll

explanation :: Maybe FormattedText

Text that is shown when the user chooses an incorrect answer or taps on the lamp icon; empty for a yet unanswered poll

explanation_media :: Maybe MessageContent

Media that is shown when the user chooses an incorrect answer or taps on the lamp icon; may be null if none or the poll is unanswered yet. Currently, can be only of the types messageAnimation, messageAudio, messageDocument, messageLocation, messagePhoto, messageVenue, or messageVideo without caption