TD.Data.PremiumFeature

Description

Documentation

data PremiumFeature

Describes a feature available to Premium users

Constructors

PremiumFeatureIncreasedLimits
Increased limits

PremiumFeatureIncreasedUploadFileSize
Increased maximum upload file size

PremiumFeatureImprovedDownloadSpeed
Improved download speed

PremiumFeatureVoiceRecognition
The ability to convert voice notes to text

PremiumFeatureDisabledAds
Disabled ads

PremiumFeatureUniqueReactions
Allowed to use more reactions

PremiumFeatureUniqueStickers
Allowed to use premium stickers with unique effects

PremiumFeatureCustomEmoji
Allowed to use custom emoji stickers in message texts and captions

PremiumFeatureAdvancedChatManagement
Ability to change position of the main chat list, archive and mute all new chats from non-contacts, and completely disable notifications about the user's contacts joined Telegram

PremiumFeatureProfileBadge
A badge in the user's profile

PremiumFeatureEmojiStatus
A emoji status shown along with the user's name

PremiumFeatureAnimatedProfilePhoto
Profile photo animation on message and chat screens

PremiumFeatureForumTopicIcon
The ability to set a custom emoji as a forum topic icon

PremiumFeatureAppIcons
Allowed to set a premium appllication icons

PremiumFeatureRealTimeChatTranslation
Allowed to translate chat messages real-time