TD.Data.PremiumFeatures

Description

Documentation

data PremiumFeatures

Constructors

PremiumFeatures
Contains information about features, available to Premium users features The list of available features limits The list of limits, increased for Premium users

Fields

payment_link :: Maybe InternalLinkType

An internal link to be opened to pay for Telegram Premium if store payment isn't possible; may be null if direct payment isn't available

limits :: Maybe [PremiumLimit]

features :: Maybe [PremiumFeature]