TD.Data.PremiumGiveawayParameters

Documentation

data PremiumGiveawayParameters

Constructors

PremiumGiveawayParameters
Describes parameters of a Telegram Premium giveaway

Fields

boosted_chat_id :: Maybe Int

Identifier of the supergroup or channel chat, which will be automatically boosted by the winners of the giveaway for duration of the Premium subscription. If the chat is a channel, then can_post_messages right is required in the channel, otherwise, the user must be an administrator in the supergroup

additional_chat_ids :: Maybe [Int]

Identifiers of other supergroup or channel chats that must be subscribed by the users to be eligible for the giveaway. There can be up to getOption("giveaway_additional_chat_count_max") additional chats

winners_selection_date :: Maybe Int

Point in time (Unix timestamp) when the giveaway is expected to be performed; must be 60-getOption("giveaway_duration_max") seconds in the future in scheduled giveaways

only_new_members :: Maybe Bool

True, if only new members of the chats will be eligible for the giveaway

has_public_winners :: Maybe Bool

True, if the list of winners of the giveaway will be available to everyone

country_codes :: Maybe [Text]

The list of two-letter ISO 3166-1 alpha-2 codes of countries, users from which will be eligible for the giveaway. If empty, then all users can participate in the giveaway. There can be up to getOption("giveaway_country_count_max") chosen countries. Users with phone number that was bought at https://fragment.com can participate in any giveaway and the country code FT must not be specified in the list

prize_description :: Maybe Text

Additional description of the giveaway prize; 0-128 characters

defaultPremiumGiveawayParameters :: PremiumGiveawayParameters