TD.Data.PremiumGiveawayParticipantStatus

Documentation

data PremiumGiveawayParticipantStatus

Contains information about status of a user in a Telegram Premium giveaway

Constructors

PremiumGiveawayParticipantStatusEligible
The user is eligible for the giveaway

PremiumGiveawayParticipantStatusParticipating
The user participates in the giveaway

PremiumGiveawayParticipantStatusAlreadyWasMember
The user can't participate in the giveaway, because they have already been member of the chat

Fields

joined_chat_date :: Maybe Int

Point in time (Unix timestamp) when the user joined the chat

PremiumGiveawayParticipantStatusAdministrator
The user can't participate in the giveaway, because they are an administrator in one of the chats that created the giveaway

Fields

chat_id :: Maybe Int

Identifier of the chat administered by the user

PremiumGiveawayParticipantStatusDisallowedCountry
The user can't participate in the giveaway, because they phone number is from a disallowed country

Fields

user_country_code :: Maybe Text

A two-letter ISO 3166-1 alpha-2 country code of the user's country