TD.Data.PremiumLimitType

Description

Documentation

data PremiumLimitType

Describes type of a limit, increased for Premium users

Constructors

PremiumLimitTypeSupergroupCount
The maximum number of joined supergroups and channels

PremiumLimitTypePinnedChatCount
The maximum number of pinned chats in the main chat list

PremiumLimitTypeCreatedPublicChatCount
The maximum number of created public chats

PremiumLimitTypeSavedAnimationCount
The maximum number of saved animations

PremiumLimitTypeFavoriteStickerCount
The maximum number of favorite stickers

PremiumLimitTypeChatFolderCount
The maximum number of chat folders

PremiumLimitTypeChatFolderChosenChatCount
The maximum number of pinned and always included, or always excluded chats in a chat folder

PremiumLimitTypePinnedArchivedChatCount
The maximum number of pinned chats in the archive chat list

PremiumLimitTypeCaptionLength
The maximum length of sent media caption

PremiumLimitTypeBioLength
The maximum length of the user's bio

PremiumLimitTypeChatFolderInviteLinkCount
The maximum number of invite links for a chat folder

PremiumLimitTypeShareableChatFolderCount
The maximum number of added shareable chat folders