TD.Data.PremiumPaymentOption

Documentation

data PremiumPaymentOption

Constructors

PremiumPaymentOption
Describes an option for buying Telegram Premium to a user

Fields

currency :: Maybe Text

ISO 4217 currency code for Telegram Premium subscription payment

amount :: Maybe Int

The amount to pay, in the smallest units of the currency

discount_percentage :: Maybe Int

The discount associated with this option, as a percentage

month_count :: Maybe Int

Number of month the Telegram Premium subscription will be active

store_product_id :: Maybe Text

Identifier of the store product associated with the option

payment_link :: Maybe InternalLinkType

An internal link to be opened for buying Telegram Premium to the user if store payment isn't possible; may be null if direct payment isn't available