TD.Data.PremiumSource

Documentation

data PremiumSource

Describes a source from which the Premium features screen is opened

Constructors

PremiumSourceLimitExceeded
A limit was exceeded

Fields

limit_type :: Maybe PremiumLimitType

Type of the exceeded limit

PremiumSourceFeature
A user tried to use a Premium feature

Fields

feature :: Maybe PremiumFeature

The used feature

PremiumSourceBusinessFeature
A user tried to use a Business feature

Fields

_feature :: Maybe BusinessFeature

The used feature; pass null if none specific feature was used

PremiumSourceStoryFeature
A user tried to use a Premium story feature

Fields

__feature :: Maybe PremiumStoryFeature

The used feature

PremiumSourceLink
A user opened an internal link of the type internalLinkTypePremiumFeatures

Fields

referrer :: Maybe Text

The referrer from the link

PremiumSourceSettings
A user opened the Premium features screen from settings