TD.Data.PremiumState

Documentation

data PremiumState

Constructors

PremiumState
Contains state of Telegram Premium subscription and promotion videos for Premium features

Fields

state :: Maybe FormattedText

Text description of the state of the current Premium subscription; may be empty if the current user has no Telegram Premium subscription

payment_options :: Maybe [PremiumStatePaymentOption]

The list of available options for buying Telegram Premium

animations :: Maybe [PremiumFeaturePromotionAnimation]

The list of available promotion animations for Premium features