TD.Data.PremiumStatePaymentOption

Description

Documentation

data PremiumStatePaymentOption

Constructors

PremiumStatePaymentOption
Describes an option for buying or upgrading Telegram Premium for self

Fields

last_transaction_id :: Maybe String

Identifier of the last in-store transaction for the currently used option

is_upgrade :: Maybe Bool

True, if the payment option can be used to upgrade the existing Telegram Premium subscription

is_current :: Maybe Bool

True, if this is the currently used Telegram Premium subscription option

payment_option :: Maybe PremiumPaymentOption

Information about the payment option