TD.Data.ProfileAccentColor

Documentation

data ProfileAccentColor

Constructors

ProfileAccentColor
Contains information about supported accent color for user profile photo background

Fields

_id :: Maybe Int

Profile accent color identifier

light_theme_colors :: Maybe ProfileAccentColors

Accent colors expected to be used in light themes

dark_theme_colors :: Maybe ProfileAccentColors

Accent colors expected to be used in dark themes

min_supergroup_chat_boost_level :: Maybe Int

The minimum chat boost level required to use the color in a supergroup chat

min_channel_chat_boost_level :: Maybe Int

The minimum chat boost level required to use the color in a channel chat