TD.Data.ProfilePhoto

Description

Documentation

data ProfilePhoto

Constructors

ProfilePhoto
Describes a user profile photo @id Photo identifier; 0 for an empty photo. Can be used to find a photo in a list of user profile photos

Fields

has_animation :: Maybe Bool

True, if the photo has animated variant

minithumbnail :: Maybe Minithumbnail

User profile photo minithumbnail; may be null

big :: Maybe File

A big (640x640) user profile photo. The file can be downloaded only before the photo is changed

small :: Maybe File

A small (160x160) user profile photo. The file can be downloaded only before the photo is changed

_id :: Maybe Int