TD.Data.ProfileTab

Documentation

data ProfileTab

Describes a tab shown in a user or a chat profile

Constructors

ProfileTabPosts
A tab with stories posted by the user or the channel chat and saved to profile

ProfileTabGifts
A tab with gifts received by the user or the channel chat

ProfileTabMedia
A tab with photos and videos posted by the channel

ProfileTabFiles
A tab with documents posted by the channel

ProfileTabLinks
A tab with messages posted by the channel and containing links

ProfileTabMusic
A tab with audio messages posted by the channel

ProfileTabVoice
A tab with voice notes posted by the channel

ProfileTabGifs
A tab with animations posted by the channel