TD.Data.PublicChatType

Description

Documentation

data PublicChatType

Describes a type of public chats

Constructors

PublicChatTypeHasUsername
The chat is public, because it has an active username

PublicChatTypeIsLocationBased
The chat is public, because it is a location-based supergroup