TD.Data.PushMessageContent

Documentation

data PushMessageContent

Contains content of a push message notification

Constructors

PushMessageContentHidden
A general message with hidden content

Fields

is_pinned :: Maybe Bool

True, if the message is a pinned message with the specified content

PushMessageContentAnimation
An animation message (GIF-style).

Fields

animation :: Maybe Animation

Message content; may be null

caption :: Maybe Text

Animation caption

is_pinned :: Maybe Bool

True, if the message is a pinned message with the specified content

PushMessageContentAudio
An audio message

Fields

audio :: Maybe Audio

Message content; may be null

is_pinned :: Maybe Bool

True, if the message is a pinned message with the specified content

PushMessageContentContact
A message with a user contact

Fields

name :: Maybe Text

Contact's name

is_pinned :: Maybe Bool

True, if the message is a pinned message with the specified content

PushMessageContentContactRegistered
A contact has registered with Telegram

PushMessageContentDocument
A document message (a general file)

Fields

document :: Maybe Document

Message content; may be null

is_pinned :: Maybe Bool

True, if the message is a pinned message with the specified content

PushMessageContentGame
A message with a game

Fields

title :: Maybe Text

Game title, empty for pinned game message

is_pinned :: Maybe Bool

True, if the message is a pinned message with the specified content

PushMessageContentGameScore
A new high score was achieved in a game

Fields

title :: Maybe Text

Game title, empty for pinned game message

score :: Maybe Int

New score, 0 for pinned message

is_pinned :: Maybe Bool

True, if the message is a pinned message with the specified content

PushMessageContentInvoice
A message with an invoice from a bot

Fields

price :: Maybe Text

Product price

is_pinned :: Maybe Bool

True, if the message is a pinned message with the specified content

PushMessageContentLocation
A message with a location

Fields

is_live :: Maybe Bool

True, if the location is live

is_pinned :: Maybe Bool

True, if the message is a pinned message with the specified content

PushMessageContentPhoto
A photo message

Fields

photo :: Maybe Photo

Message content; may be null

caption :: Maybe Text

Animation caption

is_secret :: Maybe Bool

True, if the photo is secret

is_pinned :: Maybe Bool

True, if the message is a pinned message with the specified content

PushMessageContentPoll
A message with a poll

Fields

question :: Maybe Text

Poll question

is_regular :: Maybe Bool

True, if the poll is regular and not in quiz mode

is_pinned :: Maybe Bool

True, if the message is a pinned message with the specified content

PushMessageContentScreenshotTaken
A screenshot of a message in the chat has been taken

PushMessageContentSticker
A message with a sticker

Fields

sticker :: Maybe Sticker

Message content; may be null

emoji :: Maybe Text

Emoji corresponding to the sticker; may be empty

is_pinned :: Maybe Bool

True, if the message is a pinned message with the specified content

PushMessageContentStory
A message with a story

Fields

is_pinned :: Maybe Bool

True, if the message is a pinned message with the specified content

PushMessageContentText
A text message

Fields

text :: Maybe Text

Message text

is_pinned :: Maybe Bool

True, if the message is a pinned message with the specified content

PushMessageContentVideo
A video message

Fields

video :: Maybe Video

Message content; may be null

caption :: Maybe Text

Animation caption

is_secret :: Maybe Bool

True, if the photo is secret

is_pinned :: Maybe Bool

True, if the message is a pinned message with the specified content

PushMessageContentVideoNote
A video note message

Fields

video_note :: Maybe VideoNote

Message content; may be null

is_pinned :: Maybe Bool

True, if the message is a pinned message with the specified content

PushMessageContentVoiceNote
A voice note message

Fields

voice_note :: Maybe VoiceNote

Message content; may be null

is_pinned :: Maybe Bool

True, if the message is a pinned message with the specified content

PushMessageContentBasicGroupChatCreate
A newly created basic group

PushMessageContentChatAddMembers
New chat members were invited to a group

Fields

member_name :: Maybe Text

Name of the added member

is_current_user :: Maybe Bool

True, if the current user was added to the group

is_returned :: Maybe Bool

True, if the user has returned to the group themselves

PushMessageContentChatChangePhoto
A chat photo was edited

PushMessageContentChatChangeTitle
A chat title was edited

Fields

title :: Maybe Text

Game title, empty for pinned game message

PushMessageContentChatSetBackground
A chat background was edited

Fields

is_same :: Maybe Bool

True, if the set background is the same as the background of the current user

PushMessageContentChatSetTheme
A chat theme was edited

Fields

theme_name :: Maybe Text

If non-empty, name of a new theme, set for the chat. Otherwise, the chat theme was reset to the default one

PushMessageContentChatDeleteMember
A chat member was deleted

Fields

member_name :: Maybe Text

Name of the added member

is_current_user :: Maybe Bool

True, if the current user was added to the group

is_left :: Maybe Bool

True, if the user has left the group themselves

PushMessageContentChatJoinByLink
A new member joined the chat via an invite link

PushMessageContentChatJoinByRequest
A new member was accepted to the chat by an administrator

PushMessageContentRecurringPayment
A new recurring payment was made by the current user

Fields

amount :: Maybe Text

The paid amount

PushMessageContentSuggestProfilePhoto
A profile photo was suggested to the user

PushMessageContentMessageForwards
A forwarded messages

Fields

total_count :: Maybe Int

Number of forwarded messages

PushMessageContentMediaAlbum
A media album

Fields

total_count :: Maybe Int

Number of forwarded messages

has_photos :: Maybe Bool

True, if the album has at least one photo

has_videos :: Maybe Bool

True, if the album has at least one video file

has_audios :: Maybe Bool

True, if the album has at least one audio file

has_documents :: Maybe Bool

True, if the album has at least one document

defaultPushMessageContentHidden :: PushMessageContent

defaultPushMessageContentAnimation :: PushMessageContent

defaultPushMessageContentAudio :: PushMessageContent

defaultPushMessageContentContact :: PushMessageContent

defaultPushMessageContentDocument :: PushMessageContent

defaultPushMessageContentGame :: PushMessageContent

defaultPushMessageContentGameScore :: PushMessageContent

defaultPushMessageContentInvoice :: PushMessageContent

defaultPushMessageContentLocation :: PushMessageContent

defaultPushMessageContentPhoto :: PushMessageContent

defaultPushMessageContentPoll :: PushMessageContent

defaultPushMessageContentSticker :: PushMessageContent

defaultPushMessageContentStory :: PushMessageContent

defaultPushMessageContentText :: PushMessageContent

defaultPushMessageContentVideo :: PushMessageContent

defaultPushMessageContentVideoNote :: PushMessageContent

defaultPushMessageContentVoiceNote :: PushMessageContent

defaultPushMessageContentChatAddMembers :: PushMessageContent

defaultPushMessageContentChatChangeTitle :: PushMessageContent

defaultPushMessageContentChatSetBackground :: PushMessageContent

defaultPushMessageContentChatSetTheme :: PushMessageContent

defaultPushMessageContentChatDeleteMember :: PushMessageContent

defaultPushMessageContentRecurringPayment :: PushMessageContent

defaultPushMessageContentMessageForwards :: PushMessageContent

defaultPushMessageContentMediaAlbum :: PushMessageContent