TD.Data.RemoteFile

Description

Documentation

data RemoteFile

Constructors

RemoteFile
Represents a remote file

Fields

uploaded_size :: Maybe Int

Size of the remote available part of the file, in bytes; 0 if unknown

is_uploading_completed :: Maybe Bool

True, if a remote copy is fully available

is_uploading_active :: Maybe Bool

True, if the file is currently being uploaded (or a remote copy is being generated by some other means)

unique_id :: Maybe String

Unique file identifier; may be empty if unknown. The unique file identifier which is the same for the same file even for different users and is persistent over time

_id :: Maybe String

Remote file identifier; may be empty. Can be used by the current user across application restarts or even from other devices. Uniquely identifies a file, but a file can have a lot of different valid identifiers. If the ID starts with "http://" or "https://", it represents the HTTP URL of the file. TDLib is currently unable to download files if only their URL is known. If downloadFile/addFileToDownloads is called on such a file or if it is sent to a secret chat, TDLib starts a file generation process by sending updateFileGenerationStart to the application with the HTTP URL in the original_path and "" as the conversion string. Application must generate the file by downloading it to the specified location