TD.Data.RichText

Documentation

data RichText

Describes a formatted text object

Constructors

RichTextPlain
A plain text

Fields

text :: Maybe Text

Text

RichTextBold
A bold rich text

Fields

_text :: Maybe RichText

Text

RichTextItalic
An italicized rich text

Fields

_text :: Maybe RichText

Text

RichTextUnderline
An underlined rich text

Fields

_text :: Maybe RichText

Text

RichTextStrikethrough
A strikethrough rich text

Fields

_text :: Maybe RichText

Text

RichTextSpoiler
A spoilered rich text

Fields

_text :: Maybe RichText

Text

RichTextDateTime
A date and time

Fields

_text :: Maybe RichText

Text

unix_time :: Maybe Int

Point in time (Unix timestamp) representing the date and time

formatting_type :: Maybe DateTimeFormattingType

Date and time formatting type; may be null if none and the original text must not be changed

RichTextMention
A mention of a Telegram user or chat by a username

Fields

_text :: Maybe RichText

Text

username :: Maybe Text

The username

RichTextHashtag
A hashtag

Fields

_text :: Maybe RichText

Text

hashtag :: Maybe Text

The hashtag

RichTextCashtag
A cashtag

Fields

_text :: Maybe RichText

Text

cashtag :: Maybe Text

The cashtag

RichTextBotCommand
A bot command

Fields

_text :: Maybe RichText

Text

bot_command :: Maybe Text

The bot command

RichTextFixed
A fixed-width rich text

Fields

_text :: Maybe RichText

Text

RichTextMentionName
A rich text that serves as a mention of a user

Fields

_text :: Maybe RichText

Text

user_id :: Maybe Int

Identifier of the mentioned user

RichTextUrl
A rich text URL link

Fields

_text :: Maybe RichText

Text

url :: Maybe Text

URL

is_cached :: Maybe Bool

True, if the URL has cached instant view server-side; instant view only

RichTextEmailAddress
A rich text email address

Fields

_text :: Maybe RichText

Text

email_address :: Maybe Text

Email address

RichTextBankCardNumber
A bank card number

Fields

_text :: Maybe RichText

Text

bank_card_number :: Maybe Text

The number of the bank card

RichTextSubscript
A subscript rich text

Fields

_text :: Maybe RichText

Text

RichTextSuperscript
A superscript rich text

Fields

_text :: Maybe RichText

Text

RichTextMarked
A marked rich text

Fields

_text :: Maybe RichText

Text

RichTextPhoneNumber
A rich text phone number

Fields

_text :: Maybe RichText

Text

phone_number :: Maybe Text

Phone number

RichTextCustomEmoji
A custom emoji

Fields

custom_emoji_id :: Maybe Int

Unique identifier of the custom emoji

alternative_text :: Maybe Text

Alternative text for the custom emoji

RichTextIcon
A small image inside the text; instant view only

Fields

document :: Maybe Document

The image represented as a document. The image can be in GIF, JPEG or PNG format

width :: Maybe Int

Width of a bounding box in which the image must be shown; 0 if unknown

height :: Maybe Int

Height of a bounding box in which the image must be shown; 0 if unknown

RichTextMathematicalExpression
A mathematical expression

Fields

expression :: Maybe Text

The expression in LaTeX format

RichTextReference
A reference

Fields

name :: Maybe Text

Reference name

_text :: Maybe RichText

Text

RichTextReferenceLink
A link to a reference on the same page

Fields

_text :: Maybe RichText

Text

reference_name :: Maybe Text

The reference name

url :: Maybe Text

URL

RichTextAnchor
An anchor

Fields

name :: Maybe Text

Reference name

RichTextAnchorLink
A link to an anchor on the same page

Fields

_text :: Maybe RichText

Text

anchor_name :: Maybe Text

The anchor name. If the name is empty, the link must bring back to top

url :: Maybe Text

URL

RichTexts
A concatenation of rich texts

Fields

texts :: Maybe [RichText]

Texts