TD.Data.SavedMessagesTopic

Documentation

data SavedMessagesTopic

Constructors

SavedMessagesTopic
Contains information about a Saved Messages topic

Fields

_id :: Maybe Int

Unique topic identifier

_type :: Maybe SavedMessagesTopicType

Type of the topic

is_pinned :: Maybe Bool

True, if the topic is pinned

order :: Maybe Int

A parameter used to determine order of the topic in the topic list. Topics must be sorted by the order in descending order

last_message :: Maybe Message

Last message in the topic; may be null if none or unknown

draft_message :: Maybe DraftMessage

A draft of a message in the topic; may be null if none