TD.Data.SavedMessagesTopicType

Documentation

data SavedMessagesTopicType

Describes type of Saved Messages topic

Constructors

SavedMessagesTopicTypeMyNotes
Topic containing messages sent by the current user of forwarded from an unknown chat

SavedMessagesTopicTypeAuthorHidden
Topic containing messages forwarded from a user with hidden privacy

SavedMessagesTopicTypeSavedFromChat
Topic containing messages forwarded from a specific chat

Fields

chat_id :: Maybe Int

Identifier of the chat