TD.Data.SearchMessagesFilter

Description

Documentation

data SearchMessagesFilter

Represents a filter for message search results

Constructors

SearchMessagesFilterEmpty
Returns all found messages, no filter is applied

SearchMessagesFilterAnimation
Returns only animation messages

SearchMessagesFilterAudio
Returns only audio messages

SearchMessagesFilterDocument
Returns only document messages

SearchMessagesFilterPhoto
Returns only photo messages

SearchMessagesFilterVideo
Returns only video messages

SearchMessagesFilterVoiceNote
Returns only voice note messages

SearchMessagesFilterPhotoAndVideo
Returns only photo and video messages

SearchMessagesFilterUrl
Returns only messages containing URLs

SearchMessagesFilterChatPhoto
Returns only messages containing chat photos

SearchMessagesFilterVideoNote
Returns only video note messages

SearchMessagesFilterVoiceAndVideoNote
Returns only voice and video note messages

SearchMessagesFilterMention
Returns only messages with mentions of the current user, or messages that are replies to their messages

SearchMessagesFilterUnreadMention
Returns only messages with unread mentions of the current user, or messages that are replies to their messages. When using this filter the results can't be additionally filtered by a query, a message thread or by the sending user

SearchMessagesFilterUnreadReaction
Returns only messages with unread reactions for the current user. When using this filter the results can't be additionally filtered by a query, a message thread or by the sending user

SearchMessagesFilterFailedToSend
Returns only failed to send messages. This filter can be used only if the message database is used

SearchMessagesFilterPinned
Returns only pinned messages