TD.Data.Session

Documentation

data Session

Constructors

Session
Contains information about one session in a Telegram application used by the current user. Sessions must be shown to the user in the returned order

Fields

_id :: Maybe Int

Session identifier

is_current :: Maybe Bool

True, if this session is the current session

is_password_pending :: Maybe Bool

True, if a 2-step verification password is needed to complete authorization of the session

is_unconfirmed :: Maybe Bool

True, if the session wasn't confirmed from another session

can_accept_secret_chats :: Maybe Bool

True, if incoming secret chats can be accepted by the session

can_accept_calls :: Maybe Bool

True, if incoming calls can be accepted by the session

_type :: Maybe SessionType

Session type based on the system and application version, which can be used to display a corresponding icon

api_id :: Maybe Int

Telegram API identifier, as provided by the application

application_name :: Maybe Text

Name of the application, as provided by the application

application_version :: Maybe Text

The version of the application, as provided by the application

is_official_application :: Maybe Bool

True, if the application is an official application or uses the api_id of an official application

device_model :: Maybe Text

Model of the device the application has been run or is running on, as provided by the application

platform :: Maybe Text

Operating system the application has been run or is running on, as provided by the application

system_version :: Maybe Text

Version of the operating system the application has been run or is running on, as provided by the application

log_in_date :: Maybe Int

Point in time (Unix timestamp) when the user has logged in

last_active_date :: Maybe Int

Point in time (Unix timestamp) when the session was last used

ip_address :: Maybe Text

IP address from which the session was created, in human-readable format

location :: Maybe Text

A human-readable description of the location from which the session was created, based on the IP address