TD.Data.SessionDeviceType

Documentation

data SessionDeviceType

Represents the type of device from which session was created

Constructors

SessionDeviceTypeAndroid
The session is running on an Android device

SessionDeviceTypeApple
The session is running on a generic Apple device

SessionDeviceTypeBrave
The session is running on the Brave browser

SessionDeviceTypeChrome
The session is running on the Chrome browser

SessionDeviceTypeEdge
The session is running on the Edge browser

SessionDeviceTypeFirefox
The session is running on the Firefox browser

SessionDeviceTypeIpad
The session is running on an iPad device

SessionDeviceTypeIphone
The session is running on an iPhone device

SessionDeviceTypeLinux
The session is running on a Linux device

SessionDeviceTypeMac
The session is running on a Mac device

SessionDeviceTypeOpera
The session is running on the Opera browser

SessionDeviceTypeSafari
The session is running on the Safari browser

SessionDeviceTypeUbuntu
The session is running on an Ubuntu device

SessionDeviceTypeUnknown
The session is running on an unknown type of device

SessionDeviceTypeVivaldi
The session is running on the Vivaldi browser

SessionDeviceTypeWindows
The session is running on a Windows device

SessionDeviceTypeXbox
The session is running on an Xbox console