TD.Data.SessionType

Description

Documentation

data SessionType

Represents the type of a session

Constructors

SessionTypeAndroid
The session is running on an Android device

SessionTypeApple
The session is running on a generic Apple device

SessionTypeBrave
The session is running on the Brave browser

SessionTypeChrome
The session is running on the Chrome browser

SessionTypeEdge
The session is running on the Edge browser

SessionTypeFirefox
The session is running on the Firefox browser

SessionTypeIpad
The session is running on an iPad device

SessionTypeIphone
The session is running on an iPhone device

SessionTypeLinux
The session is running on a Linux device

SessionTypeMac
The session is running on a Mac device

SessionTypeOpera
The session is running on the Opera browser

SessionTypeSafari
The session is running on the Safari browser

SessionTypeUbuntu
The session is running on an Ubuntu device

SessionTypeUnknown
The session is running on an unknown type of device

SessionTypeVivaldi
The session is running on the Vivaldi browser

SessionTypeWindows
The session is running on a Windows device

SessionTypeXbox
The session is running on an Xbox console