TD.Data.SponsoredMessage

Description

Documentation

data SponsoredMessage

Constructors

SponsoredMessage
Describes a sponsored message

Fields

additional_info :: Maybe String

If non-empty, additional information about the sponsored message to be shown along with the message

sponsor_info :: Maybe String

If non-empty, information about the sponsor to be shown along with the message

content :: Maybe MessageContent

Content of the message. Currently, can be only of the type messageText

link :: Maybe InternalLinkType

An internal link to be opened when the sponsored message is clicked; may be null if the sponsor chat needs to be opened instead

show_chat_photo :: Maybe Bool

True, if the sponsor's chat photo must be shown

sponsor_chat_info :: Maybe ChatInviteLinkInfo

Information about the sponsor chat; may be null unless sponsor_chat_id == 0

sponsor_chat_id :: Maybe Int

Sponsor chat identifier; 0 if the sponsor chat is accessible through an invite link

is_recommended :: Maybe Bool

True, if the message needs to be labeled as "recommended" instead of "sponsored"

message_id :: Maybe Int

Message identifier; unique for the chat to which the sponsored message belongs among both ordinary and sponsored messages