TD.Data.StarGiveawayPaymentOption

Documentation

data StarGiveawayPaymentOption

Constructors

StarGiveawayPaymentOption
Describes an option for creating of Telegram Star giveaway. Use telegramPaymentPurposeStarGiveaway for out-of-store payments

Fields

currency :: Maybe Text

ISO 4217 currency code for the payment

amount :: Maybe Int

The amount to pay, in the smallest units of the currency

star_count :: Maybe Int

Number of Telegram Stars that will be distributed among winners

store_product_id :: Maybe Text

Identifier of the store product associated with the option; may be empty if none

yearly_boost_count :: Maybe Int

Number of times the chat will be boosted for one year if the option is chosen

winner_options :: Maybe [StarGiveawayWinnerOption]

Allowed options for the number of giveaway winners

is_default :: Maybe Bool

True, if the option must be chosen by default

is_additional :: Maybe Bool

True, if the option must be shown only in the full list of payment options