TD.Data.StarRevenueStatistics

Documentation

data StarRevenueStatistics

Constructors

StarRevenueStatistics
A detailed statistics about Telegram Stars earned by a bot or a chat

Fields

revenue_by_day_graph :: Maybe StatisticalGraph

A graph containing amount of revenue in a given day

status :: Maybe StarRevenueStatus

Telegram Star revenue status

usd_rate :: Maybe Double

Current conversion rate of a Telegram Star to USD