TD.Data.StarRevenueStatus

Documentation

data StarRevenueStatus

Constructors

StarRevenueStatus
Contains information about Telegram Stars earned by a bot or a chat

Fields

total_count :: Maybe Int

Total number of Telegram Stars earned

current_count :: Maybe Int

The number of Telegram Stars that aren't withdrawn yet

available_count :: Maybe Int

The number of Telegram Stars that are available for withdrawal

withdrawal_enabled :: Maybe Bool

True, if Telegram Stars can be withdrawn now or later

next_withdrawal_in :: Maybe Int

Time left before the next withdrawal can be started, in seconds; 0 if withdrawal can be started now