TD.Data.StarSubscriptionType

Documentation

data StarSubscriptionType

Describes type of subscription paid in Telegram Stars

Constructors

StarSubscriptionTypeChannel
Describes a subscription to a channel chat

Fields

can_reuse :: Maybe Bool

True, if the subscription is active and the user can use the method reuseStarSubscription to join the subscribed chat again

invite_link :: Maybe Text

The invite link that can be used to renew the subscription if it has been expired; may be empty, if the link isn't available anymore

StarSubscriptionTypeBot
Describes a subscription in a bot or a business account

Fields

is_canceled_by_bot :: Maybe Bool

True, if the subscription was canceled by the bot and can't be extended

title :: Maybe Text

Subscription invoice title

photo :: Maybe Photo

Subscription invoice photo

invoice_link :: Maybe Text

The link to the subscription invoice