TD.Data.StarTransactionPartner

Documentation

data StarTransactionPartner

Describes source or recipient of a transaction with Telegram Stars

Constructors

StarTransactionPartnerTelegram
The transaction is a transaction with Telegram through a bot

StarTransactionPartnerAppStore
The transaction is a transaction with App Store

StarTransactionPartnerGooglePlay
The transaction is a transaction with Google Play

StarTransactionPartnerFragment
The transaction is a transaction with Fragment

Fields

withdrawal_state :: Maybe RevenueWithdrawalState

State of the withdrawal; may be null for refunds from Fragment

StarTransactionPartnerTelegramAds
The transaction is a transaction with Telegram Ad platform

StarTransactionPartnerBot
The transaction is a transaction with a bot

Fields

user_id :: Maybe Int

Identifier of the bot

purpose :: Maybe BotTransactionPurpose

Purpose of the transaction

StarTransactionPartnerBusiness
The transaction is a transaction with a business account

Fields

user_id :: Maybe Int

Identifier of the bot

media :: Maybe [PaidMedia]

The bought media if the trancastion wasn't refunded

StarTransactionPartnerChat
The transaction is a transaction with a supergroup or a channel chat

Fields

chat_id :: Maybe Int

Identifier of the chat

_purpose :: Maybe ChatTransactionPurpose

Purpose of the transaction

StarTransactionPartnerUser
The transaction is a transcation with another user

Fields

user_id :: Maybe Int

Identifier of the bot

__purpose :: Maybe UserTransactionPurpose

Purpose of the transaction

StarTransactionPartnerUnsupported
The transaction is a transaction with unknown partner