TD.Data.StarTransactionSource

Documentation

data StarTransactionSource

Describes source or recipient of a transaction with Telegram stars

Constructors

StarTransactionSourceTelegram
The transaction is a transaction with Telegram through a bot

StarTransactionSourceAppStore
The transaction is a transaction with App Store

StarTransactionSourceGooglePlay
The transaction is a transaction with Google Play

StarTransactionSourceFragment
The transaction is a transaction with Fragment

StarTransactionSourceUser
The transaction is a transaction with another user

Fields

user_id :: Maybe Int

Identifier of the user

product_info :: Maybe ProductInfo

Information about the bought product; may be null if none

StarTransactionSourceUnsupported
The transaction is a transaction with unknown source