TD.Data.StarTransactionType

Documentation

data StarTransactionType

Describes type of transaction with Telegram Stars

Constructors

StarTransactionTypePremiumBotDeposit
The transaction is a deposit of Telegram Stars from the Premium bot; relevant for regular users only

StarTransactionTypeAppStoreDeposit
The transaction is a deposit of Telegram Stars from App Store; relevant for regular users only

StarTransactionTypeGooglePlayDeposit
The transaction is a deposit of Telegram Stars from Google Play; relevant for regular users only

StarTransactionTypeFragmentDeposit
The transaction is a deposit of Telegram Stars from Fragment; relevant for regular users and bots only

StarTransactionTypeUserDeposit
The transaction is a deposit of Telegram Stars by another user; relevant for regular users only

Fields

user_id :: Maybe Int

Identifier of the user that gifted Telegram Stars; 0 if the user was anonymous

sticker :: Maybe Sticker

The sticker to be shown in the transaction information; may be null if unknown

StarTransactionTypeGiveawayDeposit
The transaction is a deposit of Telegram Stars from a giveaway; relevant for regular users only

Fields

chat_id :: Maybe Int

Identifier of a supergroup or a channel chat that created the giveaway

giveaway_message_id :: Maybe Int

Identifier of the message with the giveaway; may be 0 or an identifier of a deleted message

StarTransactionTypeFragmentWithdrawal
The transaction is a withdrawal of earned Telegram Stars to Fragment; relevant for regular users, bots, supergroup and channel chats only

Fields

withdrawal_state :: Maybe RevenueWithdrawalState

State of the withdrawal; may be null for refunds from Fragment

StarTransactionTypeTelegramAdsWithdrawal
The transaction is a withdrawal of earned Telegram Stars to Telegram Ad platform; relevant for bots and channel chats only

StarTransactionTypeTelegramApiUsage
The transaction is a payment for Telegram API usage; relevant for bots only

Fields

request_count :: Maybe Int

The number of billed requests

StarTransactionTypeBotPaidMediaPurchase
The transaction is a purchase of paid media from a bot or a business account by the current user; relevant for regular users only

Fields

user_id :: Maybe Int

Identifier of the user that gifted Telegram Stars; 0 if the user was anonymous

media :: Maybe [PaidMedia]

The bought media if the transaction wasn't refunded

StarTransactionTypeBotPaidMediaSale
The transaction is a sale of paid media by the bot or a business account managed by the bot; relevant for bots only

Fields

user_id :: Maybe Int

Identifier of the user that gifted Telegram Stars; 0 if the user was anonymous

media :: Maybe [PaidMedia]

The bought media if the transaction wasn't refunded

payload :: Maybe Text

Bot-provided payload

affiliate :: Maybe AffiliateInfo

Information about the affiliate which received commission from the transaction; may be null if none

StarTransactionTypeChannelPaidMediaPurchase
The transaction is a purchase of paid media from a channel by the current user; relevant for regular users only

Fields

chat_id :: Maybe Int

Identifier of a supergroup or a channel chat that created the giveaway

message_id :: Maybe Int

Identifier of the corresponding message with paid media; may be 0 or an identifier of a deleted message

media :: Maybe [PaidMedia]

The bought media if the transaction wasn't refunded

StarTransactionTypeChannelPaidMediaSale
The transaction is a sale of paid media by the channel chat; relevant for channel chats only

Fields

user_id :: Maybe Int

Identifier of the user that gifted Telegram Stars; 0 if the user was anonymous

message_id :: Maybe Int

Identifier of the corresponding message with paid media; may be 0 or an identifier of a deleted message

media :: Maybe [PaidMedia]

The bought media if the transaction wasn't refunded

StarTransactionTypeBotInvoicePurchase
The transaction is a purchase of a product from a bot or a business account by the current user; relevant for regular users only

Fields

user_id :: Maybe Int

Identifier of the user that gifted Telegram Stars; 0 if the user was anonymous

product_info :: Maybe ProductInfo

Information about the bought product

StarTransactionTypeBotInvoiceSale
The transaction is a sale of a product by the bot; relevant for bots only

Fields

user_id :: Maybe Int

Identifier of the user that gifted Telegram Stars; 0 if the user was anonymous

product_info :: Maybe ProductInfo

Information about the bought product

invoice_payload :: Maybe ByteString

Invoice payload

affiliate :: Maybe AffiliateInfo

Information about the affiliate which received commission from the transaction; may be null if none

StarTransactionTypeBotSubscriptionPurchase
The transaction is a purchase of a subscription from a bot or a business account by the current user; relevant for regular users only

Fields

user_id :: Maybe Int

Identifier of the user that gifted Telegram Stars; 0 if the user was anonymous

subscription_period :: Maybe Int

The number of seconds between consecutive Telegram Star debitings

product_info :: Maybe ProductInfo

Information about the bought product

StarTransactionTypeBotSubscriptionSale
The transaction is a sale of a subscription by the bot; relevant for bots only

Fields

user_id :: Maybe Int

Identifier of the user that gifted Telegram Stars; 0 if the user was anonymous

subscription_period :: Maybe Int

The number of seconds between consecutive Telegram Star debitings

product_info :: Maybe ProductInfo

Information about the bought product

invoice_payload :: Maybe ByteString

Invoice payload

affiliate :: Maybe AffiliateInfo

Information about the affiliate which received commission from the transaction; may be null if none

StarTransactionTypeChannelSubscriptionPurchase
The transaction is a purchase of a subscription to a channel chat by the current user; relevant for regular users only

Fields

chat_id :: Maybe Int

Identifier of a supergroup or a channel chat that created the giveaway

subscription_period :: Maybe Int

The number of seconds between consecutive Telegram Star debitings

StarTransactionTypeChannelSubscriptionSale
The transaction is a sale of a subscription by the channel chat; relevant for channel chats only

Fields

user_id :: Maybe Int

Identifier of the user that gifted Telegram Stars; 0 if the user was anonymous

subscription_period :: Maybe Int

The number of seconds between consecutive Telegram Star debitings

StarTransactionTypeGiftAuctionBid
The transaction is a bid on a gift auction; relevant for regular users only

Fields

owner_id :: Maybe MessageSender

Identifier of the user that will receive the gift

gift :: Maybe Gift

The gift

StarTransactionTypeGiftPurchase
The transaction is a purchase of a regular gift; relevant for regular users and bots only

Fields

owner_id :: Maybe MessageSender

Identifier of the user that will receive the gift

gift :: Maybe Gift

The gift

StarTransactionTypeGiftPurchaseOffer
The transaction is an offer of gift purchase; relevant for regular users only

Fields

_gift :: Maybe UpgradedGift

The gift

StarTransactionTypeGiftTransfer
The transaction is a transfer of an upgraded gift; relevant for regular users only

Fields

owner_id :: Maybe MessageSender

Identifier of the user that will receive the gift

_gift :: Maybe UpgradedGift

The gift

StarTransactionTypeGiftOriginalDetailsDrop
The transaction is a drop of original details of an upgraded gift; relevant for regular users only

Fields

owner_id :: Maybe MessageSender

Identifier of the user that will receive the gift

_gift :: Maybe UpgradedGift

The gift

StarTransactionTypeGiftSale
The transaction is a sale of a received gift; relevant for regular users and channel chats only

Fields

user_id :: Maybe Int

Identifier of the user that gifted Telegram Stars; 0 if the user was anonymous

gift :: Maybe Gift

The gift

StarTransactionTypeGiftUpgrade
The transaction is an upgrade of a gift; relevant for regular users only

Fields

user_id :: Maybe Int

Identifier of the user that gifted Telegram Stars; 0 if the user was anonymous

_gift :: Maybe UpgradedGift

The gift

StarTransactionTypeGiftUpgradePurchase
The transaction is a purchase of an upgrade of a gift owned by another user or channel; relevant for regular users only

Fields

owner_id :: Maybe MessageSender

Identifier of the user that will receive the gift

gift :: Maybe Gift

The gift

StarTransactionTypeUpgradedGiftPurchase
The transaction is a purchase of an upgraded gift for some user or channel; relevant for regular users only

Fields

user_id :: Maybe Int

Identifier of the user that gifted Telegram Stars; 0 if the user was anonymous

_gift :: Maybe UpgradedGift

The gift

StarTransactionTypeUpgradedGiftSale
The transaction is a sale of an upgraded gift; relevant for regular users only

Fields

user_id :: Maybe Int

Identifier of the user that gifted Telegram Stars; 0 if the user was anonymous

_gift :: Maybe UpgradedGift

The gift

commission_per_mille :: Maybe Int

The number of Telegram Stars received by the Telegram for each 1000 Telegram Stars received by the seller of the gift

commission_star_amount :: Maybe StarAmount

The amount of Telegram Stars that were received by Telegram; can be negative for refunds

via_offer :: Maybe Bool

True, if the gift was sold through a purchase offer

StarTransactionTypeChannelPaidReactionSend
The transaction is a sending of a paid reaction to a message in a channel chat by the current user; relevant for regular users only

Fields

chat_id :: Maybe Int

Identifier of a supergroup or a channel chat that created the giveaway

message_id :: Maybe Int

Identifier of the corresponding message with paid media; may be 0 or an identifier of a deleted message

StarTransactionTypeChannelPaidReactionReceive
The transaction is a receiving of a paid reaction to a message by the channel chat; relevant for channel chats only

Fields

user_id :: Maybe Int

Identifier of the user that gifted Telegram Stars; 0 if the user was anonymous

message_id :: Maybe Int

Identifier of the corresponding message with paid media; may be 0 or an identifier of a deleted message

StarTransactionTypeAffiliateProgramCommission
The transaction is a receiving of a commission from an affiliate program; relevant for regular users, bots and channel chats only

Fields

chat_id :: Maybe Int

Identifier of a supergroup or a channel chat that created the giveaway

commission_per_mille :: Maybe Int

The number of Telegram Stars received by the Telegram for each 1000 Telegram Stars received by the seller of the gift

StarTransactionTypePaidMessageSend
The transaction is a sending of a paid message; relevant for regular users only

Fields

chat_id :: Maybe Int

Identifier of a supergroup or a channel chat that created the giveaway

message_count :: Maybe Int

Number of sent paid messages

StarTransactionTypePaidMessageReceive
The transaction is a receiving of a paid message; relevant for regular users, supergroup and channel chats only

Fields

sender_id :: Maybe MessageSender

Identifier of the sender of the message

message_count :: Maybe Int

Number of sent paid messages

commission_per_mille :: Maybe Int

The number of Telegram Stars received by the Telegram for each 1000 Telegram Stars received by the seller of the gift

commission_star_amount :: Maybe StarAmount

The amount of Telegram Stars that were received by Telegram; can be negative for refunds

StarTransactionTypePaidGroupCallMessageSend
The transaction is a sending of a paid group call message; relevant for regular users only

Fields

chat_id :: Maybe Int

Identifier of a supergroup or a channel chat that created the giveaway

StarTransactionTypePaidGroupCallMessageReceive
The transaction is a receiving of a paid group call message; relevant for regular users and channel chats only

Fields

sender_id :: Maybe MessageSender

Identifier of the sender of the message

commission_per_mille :: Maybe Int

The number of Telegram Stars received by the Telegram for each 1000 Telegram Stars received by the seller of the gift

commission_star_amount :: Maybe StarAmount

The amount of Telegram Stars that were received by Telegram; can be negative for refunds

StarTransactionTypePaidGroupCallReactionSend
The transaction is a sending of a paid group reaction; relevant for regular users only

Fields

chat_id :: Maybe Int

Identifier of a supergroup or a channel chat that created the giveaway

StarTransactionTypePaidGroupCallReactionReceive
The transaction is a receiving of a paid group call reaction; relevant for regular users and channel chats only

Fields

sender_id :: Maybe MessageSender

Identifier of the sender of the message

commission_per_mille :: Maybe Int

The number of Telegram Stars received by the Telegram for each 1000 Telegram Stars received by the seller of the gift

commission_star_amount :: Maybe StarAmount

The amount of Telegram Stars that were received by Telegram; can be negative for refunds

StarTransactionTypeSuggestedPostPaymentSend
The transaction is a payment for a suggested post; relevant for regular users only

Fields

chat_id :: Maybe Int

Identifier of a supergroup or a channel chat that created the giveaway

StarTransactionTypeSuggestedPostPaymentReceive
The transaction is a receiving of a payment for a suggested post by the channel chat; relevant for channel chats only

Fields

user_id :: Maybe Int

Identifier of the user that gifted Telegram Stars; 0 if the user was anonymous

StarTransactionTypePremiumPurchase
The transaction is a purchase of Telegram Premium subscription; relevant for regular users and bots only

Fields

user_id :: Maybe Int

Identifier of the user that gifted Telegram Stars; 0 if the user was anonymous

month_count :: Maybe Int

Number of months the Telegram Premium subscription will be active

sticker :: Maybe Sticker

The sticker to be shown in the transaction information; may be null if unknown

StarTransactionTypeBusinessBotTransferSend
The transaction is a transfer of Telegram Stars to a business bot; relevant for regular users only

Fields

user_id :: Maybe Int

Identifier of the user that gifted Telegram Stars; 0 if the user was anonymous

StarTransactionTypeBusinessBotTransferReceive
The transaction is a transfer of Telegram Stars from a business account; relevant for bots only

Fields

user_id :: Maybe Int

Identifier of the user that gifted Telegram Stars; 0 if the user was anonymous

StarTransactionTypePublicPostSearch
The transaction is a payment for search of posts in public Telegram channels; relevant for regular users only

StarTransactionTypeUnsupported
The transaction is a transaction of an unsupported type