TD.Data.Sticker

Documentation

data Sticker

Constructors

Sticker
Describes a sticker

Fields

_id :: Maybe Int

Unique sticker identifier within the set; 0 if none

set_id :: Maybe Int

Identifier of the sticker set to which the sticker belongs; 0 if none

width :: Maybe Int

Sticker width; as defined by the sender

height :: Maybe Int

Sticker height; as defined by the sender

emoji :: Maybe Text

Emoji corresponding to the sticker

format :: Maybe StickerFormat

Sticker format

full_type :: Maybe StickerFullType

Sticker's full type

outline :: Maybe [ClosedVectorPath]

Sticker's outline represented as a list of closed vector paths; may be empty. The coordinate system origin is in the upper-left corner

thumbnail :: Maybe Thumbnail

Sticker thumbnail in WEBP or JPEG format; may be null

sticker :: Maybe File

File containing the sticker