TD.Data.StickerSet

Documentation

data StickerSet

Constructors

StickerSet
Represents a sticker set

Fields

_id :: Maybe Int

Identifier of the sticker set

title :: Maybe Text

Title of the sticker set

name :: Maybe Text

Name of the sticker set

thumbnail :: Maybe Thumbnail

Sticker set thumbnail in WEBP, TGS, or WEBM format with width and height 100; may be null. The file can be downloaded only before the thumbnail is changed

thumbnail_outline :: Maybe [ClosedVectorPath]

Sticker set thumbnail's outline represented as a list of closed vector paths; may be empty. The coordinate system origin is in the upper-left corner

is_owned :: Maybe Bool

True, if the sticker set is owned by the current user

is_installed :: Maybe Bool

True, if the sticker set has been installed by the current user

is_archived :: Maybe Bool

True, if the sticker set has been archived. A sticker set can't be installed and archived simultaneously

is_official :: Maybe Bool

True, if the sticker set is official

sticker_type :: Maybe StickerType

Type of the stickers in the set

needs_repainting :: Maybe Bool

True, if stickers in the sticker set are custom emoji that must be repainted; for custom emoji sticker sets only

is_allowed_as_chat_emoji_status :: Maybe Bool

True, if stickers in the sticker set are custom emoji that can be used as chat emoji status; for custom emoji sticker sets only

is_viewed :: Maybe Bool

True for already viewed trending sticker sets

stickers :: Maybe [Sticker]

List of stickers in this set

emojis :: Maybe [Emojis]

A list of emojis corresponding to the stickers in the same order. The list is only for informational purposes, because a sticker is always sent with a fixed emoji from the corresponding Sticker object