TD.Data.StorageStatisticsByChat

Description

Documentation

data StorageStatisticsByChat

Constructors

StorageStatisticsByChat
Contains the storage usage statistics for a specific chat

Fields

by_file_type :: Maybe [StorageStatisticsByFileType]

Statistics split by file types

count :: Maybe Int

Total number of files in the chat

size :: Maybe Int

Total size of the files in the chat, in bytes

chat_id :: Maybe Int

Chat identifier; 0 if none