TD.Data.StorePaymentPurpose

Documentation

data StorePaymentPurpose

Describes a purpose of an in-store payment

Constructors

StorePaymentPurposePremiumSubscription
The user subscribing to Telegram Premium

Fields

is_restore :: Maybe Bool

Pass true if this is a restore of a Telegram Premium purchase; only for App Store

is_upgrade :: Maybe Bool

Pass true if this is an upgrade from a monthly subscription to early subscription; only for App Store

StorePaymentPurposeGiftedPremium
The user gifting Telegram Premium to another user

Fields

user_id :: Maybe Int

Identifier of the user to which Premium was gifted

currency :: Maybe Text

ISO 4217 currency code of the payment currency

amount :: Maybe Int

Paid amount, in the smallest units of the currency

StorePaymentPurposePremiumGiftCodes
The user creating Telegram Premium gift codes for other users

Fields

boosted_chat_id :: Maybe Int

Identifier of the supergroup or channel chat, which will be automatically boosted by the users for duration of the Premium subscription and which is administered by the user; 0 if none

currency :: Maybe Text

ISO 4217 currency code of the payment currency

amount :: Maybe Int

Paid amount, in the smallest units of the currency

user_ids :: Maybe [Int]

Identifiers of the users which can activate the gift codes

StorePaymentPurposePremiumGiveaway
The user creating a Telegram Premium giveaway

Fields

parameters :: Maybe PremiumGiveawayParameters

Giveaway parameters

currency :: Maybe Text

ISO 4217 currency code of the payment currency

amount :: Maybe Int

Paid amount, in the smallest units of the currency

StorePaymentPurposeStars
The user buying Telegram stars

Fields

currency :: Maybe Text

ISO 4217 currency code of the payment currency

amount :: Maybe Int

Paid amount, in the smallest units of the currency

star_count :: Maybe Int

Number of bought stars