TD.Data.Story

Documentation

data Story

Constructors

Story
Represents a story

Fields

_id :: Maybe Int

Unique story identifier among stories of the given sender

sender_chat_id :: Maybe Int

Identifier of the chat that posted the story

sender_id :: Maybe MessageSender

Identifier of the sender of the story; may be null if the story is posted on behalf of the sender_chat_id

date :: Maybe Int

Point in time (Unix timestamp) when the story was published

is_being_sent :: Maybe Bool

True, if the story is being sent by the current user

is_being_edited :: Maybe Bool

True, if the story is being edited by the current user

is_edited :: Maybe Bool

True, if the story was edited

is_posted_to_chat_page :: Maybe Bool

True, if the story is saved in the sender's profile and will be available there after expiration

is_visible_only_for_self :: Maybe Bool

True, if the story is visible only for the current user

can_be_deleted :: Maybe Bool

True, if the story can be deleted

can_be_edited :: Maybe Bool

True, if the story can be edited

can_be_forwarded :: Maybe Bool

True, if the story can be forwarded as a message. Otherwise, screenshots and saving of the story content must be also forbidden

can_be_replied :: Maybe Bool

True, if the story can be replied in the chat with the story sender

can_toggle_is_posted_to_chat_page :: Maybe Bool

True, if the story's is_posted_to_chat_page value can be changed

can_get_statistics :: Maybe Bool

True, if the story statistics are available through getStoryStatistics

can_get_interactions :: Maybe Bool

True, if interactions with the story can be received through getStoryInteractions

has_expired_viewers :: Maybe Bool

True, if users viewed the story can't be received, because the story has expired more than getOption("story_viewers_expiration_delay") seconds ago

repost_info :: Maybe StoryRepostInfo

Information about the original story; may be null if the story wasn't reposted

interaction_info :: Maybe StoryInteractionInfo

Information about interactions with the story; may be null if the story isn't owned or there were no interactions

chosen_reaction_type :: Maybe ReactionType

Type of the chosen reaction; may be null if none

privacy_settings :: Maybe StoryPrivacySettings

Privacy rules affecting story visibility; may be approximate for non-owned stories

content :: Maybe StoryContent

Content of the story

areas :: Maybe [StoryArea]

Clickable areas to be shown on the story content

caption :: Maybe FormattedText

Caption of the story