TD.Data.StoryAreaPosition

Documentation

data StoryAreaPosition

Constructors

StoryAreaPosition
Describes position of a clickable rectangle area on a story media

Fields

x_percentage :: Maybe Double

The abscissa of the rectangle's center, as a percentage of the media width

y_percentage :: Maybe Double

The ordinate of the rectangle's center, as a percentage of the media height

width_percentage :: Maybe Double

The width of the rectangle, as a percentage of the media width

height_percentage :: Maybe Double

The height of the rectangle, as a percentage of the media height

rotation_angle :: Maybe Double

Clockwise rotation angle of the rectangle, in degrees; 0-360

corner_radius_percentage :: Maybe Double

The radius of the rectangle corner rounding, as a percentage of the media width

defaultStoryAreaPosition :: StoryAreaPosition