TD.Data.StoryOrigin

Documentation

data StoryOrigin

Contains information about the origin of a story that was reposted

Constructors

StoryOriginPublicStory
The original story was a public story that was posted by a known chat

Fields

chat_id :: Maybe Int

Identifier of the chat that posted original story

story_id :: Maybe Int

Story identifier of the original story

StoryOriginHiddenUser
The original story was posted by an unknown user

Fields

poster_name :: Maybe Text

Name of the user or the chat that posted the story