TD.Data.StoryPrivacySettings

Description

Documentation

data StoryPrivacySettings

Describes privacy settings of a story

Constructors

StoryPrivacySettingsEveryone
The story can be viewed by everyone @except_user_ids Identifiers of the users that can't see the story; always unknown and empty for non-owned stories

Fields

except_user_ids :: Maybe [Int]

StoryPrivacySettingsContacts
The story can be viewed by all contacts except chosen users @except_user_ids User identifiers of the contacts that can't see the story; always unknown and empty for non-owned stories

Fields

except_user_ids :: Maybe [Int]

StoryPrivacySettingsCloseFriends
The story can be viewed by all close friends

StoryPrivacySettingsSelectedUsers
The story can be viewed by certain specified users @user_ids Identifiers of the users; always unknown and empty for non-owned stories

Fields

user_ids :: Maybe [Int]