TD.Data.StoryPublicForward

Documentation

data StoryPublicForward

Describes a public forward or repost of a story

Constructors

StoryPublicForwardMessage
Contains a public forward of a story as a message

Fields

message :: Maybe Message

Information about the message with the story

StoryPublicForwardStory
Contains a public repost of a story as a story

Fields

story :: Maybe Story

Information about the reposted story