TD.Data.StoryVideo

Documentation

data StoryVideo

Constructors

StoryVideo
Describes a video file posted as a story

Fields

duration :: Maybe Double

Duration of the video, in seconds

width :: Maybe Int

Video width

height :: Maybe Int

Video height

has_stickers :: Maybe Bool

True, if stickers were added to the video. The list of corresponding sticker sets can be received using getAttachedStickerSets

is_animation :: Maybe Bool

True, if the video has no sound

minithumbnail :: Maybe Minithumbnail

Video minithumbnail; may be null

thumbnail :: Maybe Thumbnail

Video thumbnail in JPEG or MPEG4 format; may be null

preload_prefix_size :: Maybe Int

Size of file prefix, which is expected to be preloaded, in bytes

cover_frame_timestamp :: Maybe Double

Timestamp of the frame used as video thumbnail

video :: Maybe File

File containing the video