TD.Data.SuggestedPostInfo

Documentation

data SuggestedPostInfo

Constructors

SuggestedPostInfo
Contains information about a suggested post. If the post can be approved or declined, then changes to the post can be also suggested. Use sendMessage with reply to the message and suggested post information to suggest message changes. Use addOffer to suggest price or time changes

Fields

price :: Maybe SuggestedPostPrice

Price of the suggested post; may be null if the post is non-paid

send_date :: Maybe Int

Point in time (Unix timestamp) when the post is expected to be published; 0 if the specific date isn't set yet

state :: Maybe SuggestedPostState

State of the post

can_be_approved :: Maybe Bool

True, if the suggested post can be approved by the current user using approveSuggestedPost; updates aren't sent when value of this field changes

can_be_declined :: Maybe Bool

True, if the suggested post can be declined by the current user using declineSuggestedPost; updates aren't sent when value of this field changes