TD.Data.Supergroup

Documentation

data Supergroup

Constructors

Supergroup
Represents a supergroup or channel with zero or more members (subscribers in the case of channels). From the point of view of the system, a channel is a special kind of a supergroup: only administrators can post and see the list of members, and posts from all administrators use the name and photo of the channel instead of individual names and profile photos. Unlike supergroups, channels can have an unlimited number of subscribers

Fields

_id :: Maybe Int

Supergroup or channel identifier

usernames :: Maybe Usernames

Usernames of the supergroup or channel; may be null

date :: Maybe Int

Point in time (Unix timestamp) when the current user joined, or the point in time when the supergroup or channel was created, in case the user is not a member

status :: Maybe ChatMemberStatus

Status of the current user in the supergroup or channel; custom title will always be empty

member_count :: Maybe Int

Number of members in the supergroup or channel; 0 if unknown. Currently, it is guaranteed to be known only if the supergroup or channel was received through getChatSimilarChats, getChatsToPostStories, getCreatedPublicChats, getGroupsInCommon, getInactiveSupergroupChats, getRecommendedChats, getSuitableDiscussionChats, getUserPrivacySettingRules, getVideoChatAvailableParticipants, searchPublicChats, or in chatFolderInviteLinkInfo.missing_chat_ids, or in userFullInfo.personal_chat_id, or for chats with messages or stories from publicForwards and foundStories

boost_level :: Maybe Int

Approximate boost level for the chat

has_automatic_translation :: Maybe Bool

True, if automatic translation of messages is enabled in the channel

has_linked_chat :: Maybe Bool

True, if the channel has a discussion group, or the supergroup is the designated discussion group for a channel

has_location :: Maybe Bool

True, if the supergroup is connected to a location, i.e. the supergroup is a location-based supergroup

sign_messages :: Maybe Bool

True, if messages sent to the channel contains name of the sender. This field is only applicable to channels

show_message_sender :: Maybe Bool

True, if messages sent to the channel have information about the sender user. This field is only applicable to channels

join_to_send_messages :: Maybe Bool

True, if users need to join the supergroup before they can send messages. May be false only for discussion supergroups and channel direct messages groups

join_by_request :: Maybe Bool

True, if all users directly joining the supergroup need to be approved by supergroup administrators. May be true only for non-broadcast supergroups with username, location, or a linked chat

is_slow_mode_enabled :: Maybe Bool

True, if the slow mode is enabled in the supergroup

is_channel :: Maybe Bool

True, if the supergroup is a channel

is_broadcast_group :: Maybe Bool

True, if the supergroup is a broadcast group, i.e. only administrators can send messages and there is no limit on the number of members

is_forum :: Maybe Bool

True, if the supergroup is a forum with topics

is_direct_messages_group :: Maybe Bool

True, if the supergroup is a direct message group for a channel chat

is_administered_direct_messages_group :: Maybe Bool

True, if the supergroup is a direct messages group for a channel chat that is administered by the current user

verification_status :: Maybe VerificationStatus

Information about verification status of the supergroup or channel; may be null if none

has_direct_messages_group :: Maybe Bool

True, if the channel has direct messages group

has_forum_tabs :: Maybe Bool

True, if the supergroup is a forum, which topics are shown in the same way as in channel direct messages groups

restriction_info :: Maybe RestrictionInfo

Information about the restrictions that must be applied to the corresponding supergroup or channel chat; may be null if none

paid_message_star_count :: Maybe Int

Number of Telegram Stars that must be paid by non-administrator users of the supergroup chat for each sent message

active_story_state :: Maybe ActiveStoryState

State of active stories of the supergroup or channel; may be null if there are no active stories