TD.Data.SupergroupFullInfo

Description

Documentation

data SupergroupFullInfo

Constructors

SupergroupFullInfo
Contains full information about a supergroup or channel

Fields

upgraded_from_max_message_id :: Maybe Int

Identifier of the last message in the basic group from which supergroup was upgraded; 0 if none

upgraded_from_basic_group_id :: Maybe Int

Identifier of the basic group from which supergroup was upgraded; 0 if none

bot_commands :: Maybe [BotCommands]

List of commands of bots in the group

invite_link :: Maybe ChatInviteLink

Primary invite link for the chat; may be null. For chat administrators with can_invite_users right only

location :: Maybe ChatLocation

Location to which the supergroup is connected; may be null

sticker_set_id :: Maybe Int

Identifier of the supergroup sticker set; 0 if none

has_aggressive_anti_spam_enabled :: Maybe Bool

True, if aggressive anti-spam checks are enabled in the supergroup. The value of this field is only available to chat administrators

is_all_history_available :: Maybe Bool

True, if new chat members will have access to old messages. In public, discussion, of forum groups and all channels, old messages are always available, so this option affects only private non-forum supergroups without a linked chat. The value of this field is only available to chat administrators

can_toggle_aggressive_anti_spam :: Maybe Bool

True, if aggressive anti-spam checks can be enabled or disabled in the supergroup

can_get_statistics :: Maybe Bool

True, if the supergroup or channel statistics are available

can_set_location :: Maybe Bool

True, if the supergroup location can be changed

can_set_sticker_set :: Maybe Bool

True, if the supergroup sticker set can be changed

can_set_username :: Maybe Bool

True, if the chat username can be changed

can_hide_members :: Maybe Bool

True, if non-administrators and non-bots can be hidden in responses to getSupergroupMembers and searchChatMembers for non-administrators

has_hidden_members :: Maybe Bool

True, if non-administrators can receive only administrators and bots using getSupergroupMembers or searchChatMembers

can_get_members :: Maybe Bool

True, if members of the chat can be retrieved via getSupergroupMembers or searchChatMembers

slow_mode_delay_expires_in :: Maybe Float

Time left before next message can be sent in the supergroup, in seconds. An updateSupergroupFullInfo update is not triggered when value of this field changes, but both new and old values are non-zero

slow_mode_delay :: Maybe Int

Delay between consecutive sent messages for non-administrator supergroup members, in seconds

linked_chat_id :: Maybe Int

Chat identifier of a discussion group for the channel, or a channel, for which the supergroup is the designated discussion group; 0 if none or unknown

banned_count :: Maybe Int

Number of users banned from chat; 0 if unknown

restricted_count :: Maybe Int

Number of restricted users in the supergroup; 0 if unknown

administrator_count :: Maybe Int

Number of privileged users in the supergroup or channel; 0 if unknown

member_count :: Maybe Int

Number of members in the supergroup or channel; 0 if unknown

description :: Maybe String

photo :: Maybe ChatPhoto

Chat photo; may be null if empty or unknown. If non-null, then it is the same photo as in chat.photo