TD.Data.SupergroupMembersFilter

Description

Documentation

data SupergroupMembersFilter

Specifies the kind of chat members to return in getSupergroupMembers

Constructors

SupergroupMembersFilterRecent
Returns recently active users in reverse chronological order

SupergroupMembersFilterContacts
Returns contacts of the user, which are members of the supergroup or channel @query Query to search for

Fields

query :: Maybe String

SupergroupMembersFilterAdministrators
Returns the owner and administrators

SupergroupMembersFilterSearch
Used to search for supergroup or channel members via a (string) query @query Query to search for

Fields

query :: Maybe String

SupergroupMembersFilterRestricted
Returns restricted supergroup members; can be used only by administrators @query Query to search for

Fields

query :: Maybe String

SupergroupMembersFilterBanned
Returns users banned from the supergroup or channel; can be used only by administrators @query Query to search for

Fields

query :: Maybe String

SupergroupMembersFilterMention
Returns users which can be mentioned in the supergroup query Query to search for message_thread_id If non-zero, the identifier of the current message thread

Fields

message_thread_id :: Maybe Int

query :: Maybe String

SupergroupMembersFilterBots
Returns bot members of the supergroup or channel