TD.Data.TargetChat

Description

Documentation

data TargetChat

Describes the target chat to be opened

Constructors

TargetChatCurrent
The currently opened chat needs to be kept

TargetChatChosen
The chat needs to be chosen by the user among chats of the specified types

Fields

allow_channel_chats :: Maybe Bool

True, if channel chats are allowed

allow_group_chats :: Maybe Bool

True, if basic group and supergroup chats are allowed

allow_bot_chats :: Maybe Bool

True, if private chats with other bots are allowed

allow_user_chats :: Maybe Bool

True, if private chats with ordinary users are allowed

TargetChatInternalLink
The chat needs to be open with the provided internal link @link An internal link pointing to the chat

Fields

link :: Maybe InternalLinkType