TD.Data.TdlibParameters

Description

Documentation

data TdlibParameters

Constructors

TdlibParameters
Contains parameters for TDLib initialization

Fields

ignore_file_names :: Maybe Bool

If set to true, original file names will be ignored. Otherwise, downloaded files will be saved under names as close as possible to the original name

enable_storage_optimizer :: Maybe Bool

If set to true, old files will automatically be deleted

application_version :: Maybe String

Application version; must be non-empty

system_version :: Maybe String

Version of the operating system the application is being run on. If empty, the version is automatically detected by TDLib

device_model :: Maybe String

Model of the device the application is being run on; must be non-empty

system_language_code :: Maybe String

IETF language tag of the user's operating system language; must be non-empty

api_hash :: Maybe String

Application identifier hash for Telegram API access, which can be obtained at https://my.telegram.org

api_id :: Maybe Int

Application identifier for Telegram API access, which can be obtained at https://my.telegram.org

use_secret_chats :: Maybe Bool

If set to true, support for secret chats will be enabled

use_message_database :: Maybe Bool

If set to true, the library will maintain a cache of chats and messages. Implies use_chat_info_database

use_chat_info_database :: Maybe Bool

If set to true, the library will maintain a cache of users, basic groups, supergroups, channels and secret chats. Implies use_file_database

use_file_database :: Maybe Bool

If set to true, information about downloaded and uploaded files will be saved between application restarts

files_directory :: Maybe String

The path to the directory for storing files; if empty, database_directory will be used

database_directory :: Maybe String

The path to the directory for the persistent database; if empty, the current working directory will be used

use_test_dc :: Maybe Bool

If set to true, the Telegram test environment will be used instead of the production environment