TD.Data.TextEntityType

Description

Documentation

data TextEntityType

Represents a part of the text which must be formatted differently

Constructors

TextEntityTypeMention
A mention of a user, a supergroup, or a channel by their username

TextEntityTypeHashtag
A hashtag text, beginning with "#"

TextEntityTypeCashtag
A cashtag text, beginning with "$" and consisting of capital English letters (e.g., "$USD")

TextEntityTypeBotCommand
A bot command, beginning with "/"

TextEntityTypeUrl
An HTTP URL

TextEntityTypeEmailAddress
An email address

TextEntityTypePhoneNumber
A phone number

TextEntityTypeBankCardNumber
A bank card number. The getBankCardInfo method can be used to get information about the bank card

TextEntityTypeBold
A bold text

TextEntityTypeItalic
An italic text

TextEntityTypeUnderline
An underlined text

TextEntityTypeStrikethrough
A strikethrough text

TextEntityTypeSpoiler
A spoiler text

TextEntityTypeCode
Text that must be formatted as if inside a code HTML tag

TextEntityTypePre
Text that must be formatted as if inside a pre HTML tag

TextEntityTypePreCode
Text that must be formatted as if inside pre, and code HTML tags @language Programming language of the code; as defined by the sender

Fields

language :: Maybe String

TextEntityTypeTextUrl
A text description shown instead of a raw URL @url HTTP or tg:// URL to be opened when the link is clicked

Fields

url :: Maybe String

TextEntityTypeMentionName
A text shows instead of a raw mention of the user (e.g., when the user has no username) @user_id Identifier of the mentioned user

Fields

user_id :: Maybe Int

TextEntityTypeCustomEmoji
A custom emoji. The text behind a custom emoji must be an emoji. Only premium users can use premium custom emoji @custom_emoji_id Unique identifier of the custom emoji

Fields

custom_emoji_id :: Maybe Int

TextEntityTypeMediaTimestamp
A media timestamp @media_timestamp Timestamp from which a videoaudiovideo note/voice note playing must start, in seconds. The media can be in the content or the web page preview of the current message, or in the same places in the replied message

Fields

media_timestamp :: Maybe Int