TD.Data.ThemeSettings

Description

Documentation

data ThemeSettings

Constructors

ThemeSettings
Describes theme settings

Fields

outgoing_message_accent_color :: Maybe Int

Accent color of outgoing messages in ARGB format

animate_outgoing_message_fill :: Maybe Bool

If true, the freeform gradient fill needs to be animated on every sent message

outgoing_message_fill :: Maybe BackgroundFill

The fill to be used as a background for outgoing messages

background :: Maybe Background

The background to be used in chats; may be null

accent_color :: Maybe Int

Theme accent color in ARGB format