TD.Data.ThumbnailFormat

Documentation

data ThumbnailFormat

Describes format of a thumbnail

Constructors

ThumbnailFormatJpeg
The thumbnail is in JPEG format

ThumbnailFormatGif
The thumbnail is in static GIF format. It will be used only for some bot inline results

ThumbnailFormatMpeg4
The thumbnail is in MPEG4 format. It will be used only for some animations and videos

ThumbnailFormatPng
The thumbnail is in PNG format. It will be used only for background patterns

ThumbnailFormatTgs
The thumbnail is in TGS format. It will be used only for TGS sticker sets

ThumbnailFormatWebm
The thumbnail is in WEBM format. It will be used only for WEBM sticker sets

ThumbnailFormatWebp
The thumbnail is in WEBP format. It will be used only for some stickers