TD.Data.TonTransactionType

Documentation

data TonTransactionType

Describes type of transaction with Toncoins

Constructors

TonTransactionTypeFragmentDeposit
The transaction is a deposit of Toncoins from Fragment

Fields

is_gift :: Maybe Bool

True, if the transaction is a gift from another user

sticker :: Maybe Sticker

The sticker to be shown in the transaction information; may be null if unknown

TonTransactionTypeFragmentWithdrawal
The transaction is a withdrawal of earned Toncoins to Fragment

Fields

withdrawal_state :: Maybe RevenueWithdrawalState

State of the withdrawal; may be null for refunds from Fragment

TonTransactionTypeSuggestedPostPayment
The transaction is a payment for a suggested post

Fields

chat_id :: Maybe Int

Identifier of the channel chat that posted the post

TonTransactionTypeGiftPurchaseOffer
The transaction is an offer of gift purchase

Fields

gift :: Maybe UpgradedGift

The gift

TonTransactionTypeUpgradedGiftPurchase
The transaction is a purchase of an upgraded gift for some user or channel

Fields

user_id :: Maybe Int

Identifier of the user that sold the gift

gift :: Maybe UpgradedGift

The gift

TonTransactionTypeUpgradedGiftSale
The transaction is a sale of an upgraded gift

Fields

user_id :: Maybe Int

Identifier of the user that sold the gift

gift :: Maybe UpgradedGift

The gift

commission_per_mille :: Maybe Int

The number of Toncoins received by the Telegram for each 1000 Toncoins received by the seller of the gift

commission_toncoin_amount :: Maybe Int

The amount of Toncoins that were received by the Telegram; in the smallest units of the currency

via_offer :: Maybe Bool

True, if the gift was sold through a purchase offer

TonTransactionTypeUnsupported
The transaction is a transaction of an unsupported type