TD.Data.TopChatCategory

Description

Documentation

data TopChatCategory

Represents the categories of chats for which a list of frequently used chats can be retrieved

Constructors

TopChatCategoryUsers
A category containing frequently used private chats with non-bot users

TopChatCategoryBots
A category containing frequently used private chats with bot users

TopChatCategoryGroups
A category containing frequently used basic groups and supergroups

TopChatCategoryChannels
A category containing frequently used channels

TopChatCategoryInlineBots
A category containing frequently used chats with inline bots sorted by their usage in inline mode

TopChatCategoryCalls
A category containing frequently used chats used for calls

TopChatCategoryForwardChats
A category containing frequently used chats used to forward messages