TD.Data.UnconfirmedSession

Documentation

data UnconfirmedSession

Constructors

UnconfirmedSession
Contains information about an unconfirmed session

Fields

_id :: Maybe Int

Session identifier

log_in_date :: Maybe Int

Point in time (Unix timestamp) when the user has logged in

device_model :: Maybe Text

Model of the device that was used for the session creation, as provided by the application

location :: Maybe Text

A human-readable description of the location from which the session was created, based on the IP address