TD.Data.UnreadReaction

Documentation

data UnreadReaction

Constructors

UnreadReaction
Contains information about an unread reaction to a message

Fields

_type :: Maybe ReactionType

Type of the reaction

sender_id :: Maybe MessageSender

Identifier of the sender, added the reaction

is_big :: Maybe Bool

True, if the reaction was added with a big animation