TD.Data.Update

Documentation

data Update

Contains notifications about data changes

Constructors

UpdateAuthorizationState
The user authorization state has changed

Fields

authorization_state :: Maybe AuthorizationState

New authorization state

UpdateNewMessage
A new message was received; can also be an outgoing message

Fields

message :: Maybe Message

The new message

UpdateMessageSendAcknowledged
A request to send a message has reached the Telegram server. This doesn't mean that the message will be sent successfully. This update is sent only if the option "use_quick_ack" is set to true. This update may be sent multiple times for the same message

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

message_id :: Maybe Int

A temporary message identifier

UpdateMessageSendSucceeded
A message has been successfully sent

Fields

message :: Maybe Message

The new message

old_message_id :: Maybe Int

The previous temporary message identifier

UpdateMessageSendFailed
A message failed to send. Be aware that some messages being sent can be irrecoverably deleted, in which case updateDeleteMessages will be received instead of this update

Fields

message :: Maybe Message

The new message

old_message_id :: Maybe Int

The previous temporary message identifier

_error :: Maybe Error

The cause of the message sending failure

UpdateMessageContent
The message content has changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

message_id :: Maybe Int

A temporary message identifier

new_content :: Maybe MessageContent

New message content

UpdateMessageEdited
A message was edited. Changes in the message content will come in a separate updateMessageContent

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

message_id :: Maybe Int

A temporary message identifier

edit_date :: Maybe Int

Point in time (Unix timestamp) when the message was edited

reply_markup :: Maybe ReplyMarkup

New message reply markup; may be null

UpdateMessageIsPinned
The message pinned state was changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

message_id :: Maybe Int

A temporary message identifier

is_pinned :: Maybe Bool

True, if the message is pinned

UpdateMessageInteractionInfo
The information about interactions with a message has changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

message_id :: Maybe Int

A temporary message identifier

interaction_info :: Maybe MessageInteractionInfo

New information about interactions with the message; may be null

UpdateMessageContentOpened
The message content was opened. Updates voice note messages to "listened", video note messages to "viewed" and starts the self-destruct timer

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

message_id :: Maybe Int

A temporary message identifier

UpdateMessageMentionRead
A message with an unread mention was read

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

message_id :: Maybe Int

A temporary message identifier

unread_mention_count :: Maybe Int

The new number of unread mention messages left in the chat

UpdateMessageUnreadReactions
The list of unread reactions added to a message was changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

message_id :: Maybe Int

A temporary message identifier

unread_reactions :: Maybe [UnreadReaction]

The new list of unread reactions

unread_reaction_count :: Maybe Int

The new number of messages with unread reactions left in the chat

UpdateMessageFactCheck
A fact-check added to a message was changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

message_id :: Maybe Int

A temporary message identifier

fact_check :: Maybe FactCheck

The new fact-check

UpdateMessageSuggestedPostInfo
Information about suggested post of a message was changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

message_id :: Maybe Int

A temporary message identifier

suggested_post_info :: Maybe SuggestedPostInfo

The new information about the suggested post

UpdateMessageLiveLocationViewed
A message with a live location was viewed. When the update is received, the application is expected to update the live location

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

message_id :: Maybe Int

A temporary message identifier

UpdateVideoPublished
An automatically scheduled message with video has been successfully sent after conversion

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

message_id :: Maybe Int

A temporary message identifier

UpdateNewChat
A new chat has been loaded/created. This update is guaranteed to come before the chat identifier is returned to the application. The chat field changes will be reported through separate updates

Fields

chat :: Maybe Chat

The chat

UpdateChatTitle
The title of a chat was changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

title :: Maybe Text

The new chat title

UpdateChatPhoto
A chat photo was changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

photo :: Maybe ChatPhotoInfo

The new chat photo; may be null

UpdateChatAccentColors
Chat accent colors have changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

accent_color_id :: Maybe Int

The new chat accent color identifier

background_custom_emoji_id :: Maybe Int

The new identifier of a custom emoji to be shown on the reply header and link preview background; 0 if none

upgraded_gift_colors :: Maybe UpgradedGiftColors

Color scheme based on an upgraded gift to be used for the chat instead of accent_color_id and background_custom_emoji_id; may be null if none

profile_accent_color_id :: Maybe Int

The new chat profile accent color identifier; -1 if none

profile_background_custom_emoji_id :: Maybe Int

The new identifier of a custom emoji to be shown on the profile background; 0 if none

UpdateChatPermissions
Chat permissions were changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

permissions :: Maybe ChatPermissions

The new chat permissions

UpdateChatLastMessage
The last message of a chat was changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

last_message :: Maybe Message

The new last message in the chat; may be null if the last message became unknown. While the last message is unknown, new messages can be added to the chat without corresponding updateNewMessage update

positions :: Maybe [ChatPosition]

The new chat positions in the chat lists

UpdateChatPosition
The position of a chat in a chat list has changed. An updateChatLastMessage or updateChatDraftMessage update might be sent instead of the update

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

position :: Maybe ChatPosition

New chat position. If new order is 0, then the chat needs to be removed from the list

UpdateChatAddedToList
A chat was added to a chat list

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

chat_list :: Maybe ChatList

The chat list to which the chat was added

UpdateChatRemovedFromList
A chat was removed from a chat list

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

chat_list :: Maybe ChatList

The chat list to which the chat was added

UpdateChatReadInbox
Incoming messages were read or the number of unread messages has been changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

last_read_inbox_message_id :: Maybe Int

Identifier of the last read incoming message

unread_count :: Maybe Int

The number of unread messages left in the chat

UpdateChatReadOutbox
Outgoing messages were read

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

last_read_outbox_message_id :: Maybe Int

Identifier of last read outgoing message

UpdateChatActionBar
The chat action bar was changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

action_bar :: Maybe ChatActionBar

The new value of the action bar; may be null

UpdateChatBusinessBotManageBar
The bar for managing business bot was changed in a chat

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

business_bot_manage_bar :: Maybe BusinessBotManageBar

The new value of the business bot manage bar; may be null

UpdateChatAvailableReactions
The chat available reactions were changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

available_reactions :: Maybe ChatAvailableReactions

The new reactions, available in the chat

UpdateChatDraftMessage
A chat draft has changed. Be aware that the update may come in the currently opened chat but with old content of the draft. If the user has changed the content of the draft, this update mustn't be applied

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

draft_message :: Maybe DraftMessage

The new draft message; may be null if none

positions :: Maybe [ChatPosition]

The new chat positions in the chat lists

UpdateChatEmojiStatus
Chat emoji status has changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

emoji_status :: Maybe EmojiStatus

The new chat emoji status; may be null

UpdateChatMessageSender
The message sender that is selected to send messages in a chat has changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

message_sender_id :: Maybe MessageSender

New value of message_sender_id; may be null if the user can't change message sender

UpdateChatMessageAutoDeleteTime
The message auto-delete or self-destruct timer setting for a chat was changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

message_auto_delete_time :: Maybe Int

New value of message_auto_delete_time

UpdateChatNotificationSettings
Notification settings for a chat were changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

notification_settings :: Maybe ChatNotificationSettings

The new notification settings

UpdateChatPendingJoinRequests
The chat pending join requests were changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

pending_join_requests :: Maybe ChatJoinRequestsInfo

The new data about pending join requests; may be null

UpdateChatReplyMarkup
The default chat reply markup was changed. Can occur because new messages with reply markup were received or because an old reply markup was hidden by the user

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

reply_markup_message_id :: Maybe Int

Identifier of the message from which reply markup needs to be used; 0 if there is no default custom reply markup in the chat

UpdateChatBackground
The chat background was changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

background :: Maybe ChatBackground

The new chat background; may be null if background was reset to default

UpdateChatTheme
The chat theme was changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

theme :: Maybe ChatTheme

The new theme of the chat; may be null if theme was reset to default

UpdateChatUnreadMentionCount
The chat unread_mention_count has changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

unread_mention_count :: Maybe Int

The new number of unread mention messages left in the chat

UpdateChatUnreadReactionCount
The chat unread_reaction_count has changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

unread_reaction_count :: Maybe Int

The new number of messages with unread reactions left in the chat

UpdateChatVideoChat
A chat video chat state has changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

video_chat :: Maybe VideoChat

New value of video_chat

UpdateChatDefaultDisableNotification
The value of the default disable_notification parameter, used when a message is sent to the chat, was changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

default_disable_notification :: Maybe Bool

The new default_disable_notification value

UpdateChatHasProtectedContent
A chat content was allowed or restricted for saving

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

has_protected_content :: Maybe Bool

New value of has_protected_content

UpdateChatIsTranslatable
Translation of chat messages was enabled or disabled

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

is_translatable :: Maybe Bool

New value of is_translatable

UpdateChatIsMarkedAsUnread
A chat was marked as unread or was read

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

is_marked_as_unread :: Maybe Bool

New value of is_marked_as_unread

UpdateChatViewAsTopics
A chat default appearance has changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

view_as_topics :: Maybe Bool

New value of view_as_topics

UpdateChatBlockList
A chat was blocked or unblocked

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

block_list :: Maybe BlockList

Block list to which the chat is added; may be null if none

UpdateChatHasScheduledMessages
A chat's has_scheduled_messages field has changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

has_scheduled_messages :: Maybe Bool

New value of has_scheduled_messages

UpdateChatFolders
The list of chat folders or a chat folder has changed

Fields

chat_folders :: Maybe [ChatFolderInfo]

The new list of chat folders

main_chat_list_position :: Maybe Int

Position of the main chat list among chat folders, 0-based

are_tags_enabled :: Maybe Bool

True, if folder tags are enabled

UpdateChatOnlineMemberCount
The number of online group members has changed. This update with non-zero number of online group members is sent only for currently opened chats. There is no guarantee that it is sent just after the number of online users has changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

online_member_count :: Maybe Int

New number of online members in the chat, or 0 if unknown

UpdateSavedMessagesTopic
Basic information about a Saved Messages topic has changed. This update is guaranteed to come before the topic identifier is returned to the application

Fields

topic :: Maybe SavedMessagesTopic

New data about the topic

UpdateSavedMessagesTopicCount
Number of Saved Messages topics has changed

Fields

topic_count :: Maybe Int

Approximate total number of Saved Messages topics

UpdateDirectMessagesChatTopic
Basic information about a topic in a channel direct messages chat administered by the current user has changed. This update is guaranteed to come before the topic identifier is returned to the application

Fields

_topic :: Maybe DirectMessagesChatTopic

New data about the topic

UpdateTopicMessageCount
Number of messages in a topic has changed; for Saved Messages and channel direct messages chat topics only

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

topic_id :: Maybe MessageTopic

Identifier of the topic

message_count :: Maybe Int

Approximate number of messages in the topic

UpdateQuickReplyShortcut
Basic information about a quick reply shortcut has changed. This update is guaranteed to come before the quick shortcut name is returned to the application

Fields

shortcut :: Maybe QuickReplyShortcut

New data about the shortcut

UpdateQuickReplyShortcutDeleted
A quick reply shortcut and all its messages were deleted

Fields

shortcut_id :: Maybe Int

The identifier of the deleted shortcut

UpdateQuickReplyShortcuts
The list of quick reply shortcuts has changed

Fields

shortcut_ids :: Maybe [Int]

The new list of identifiers of quick reply shortcuts

UpdateQuickReplyShortcutMessages
The list of quick reply shortcut messages has changed

Fields

shortcut_id :: Maybe Int

The identifier of the deleted shortcut

messages :: Maybe [QuickReplyMessage]

The new list of quick reply messages for the shortcut in order from the first to the last sent

UpdateForumTopicInfo
Basic information about a topic in a forum chat was changed

Fields

info :: Maybe ForumTopicInfo

New information about the topic

UpdateForumTopic
Information about a topic in a forum chat was changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

forum_topic_id :: Maybe Int

Forum topic identifier of the topic

is_pinned :: Maybe Bool

True, if the message is pinned

last_read_inbox_message_id :: Maybe Int

Identifier of the last read incoming message

last_read_outbox_message_id :: Maybe Int

Identifier of last read outgoing message

unread_mention_count :: Maybe Int

The new number of unread mention messages left in the chat

unread_reaction_count :: Maybe Int

The new number of messages with unread reactions left in the chat

notification_settings :: Maybe ChatNotificationSettings

The new notification settings

draft_message :: Maybe DraftMessage

The new draft message; may be null if none

UpdateScopeNotificationSettings
Notification settings for some type of chats were updated

Fields

scope :: Maybe NotificationSettingsScope

Types of chats for which notification settings were updated

_notification_settings :: Maybe ScopeNotificationSettings

The new notification settings

UpdateReactionNotificationSettings
Notification settings for reactions were updated

Fields

__notification_settings :: Maybe ReactionNotificationSettings

The new notification settings

UpdateNotification
A notification was changed

Fields

notification_group_id :: Maybe Int

Unique notification group identifier

notification :: Maybe Notification

Changed notification

UpdateNotificationGroup
A list of active notifications in a notification group has changed

Fields

notification_group_id :: Maybe Int

Unique notification group identifier

_type :: Maybe NotificationGroupType

New type of the notification group

chat_id :: Maybe Int

The chat identifier of the sent message

notification_settings_chat_id :: Maybe Int

Chat identifier, which notification settings must be applied to the added notifications

notification_sound_id :: Maybe Int

Identifier of the notification sound to be played; 0 if sound is disabled

total_count :: Maybe Int

Total number of unread notifications in the group, can be bigger than number of active notifications

added_notifications :: Maybe [Notification]

List of added group notifications, sorted by notification identifier

removed_notification_ids :: Maybe [Int]

Identifiers of removed group notifications, sorted by notification identifier

UpdateActiveNotifications
Contains active notifications that were shown on previous application launches. This update is sent only if the message database is used. In that case it comes once before any updateNotification and updateNotificationGroup update

Fields

groups :: Maybe [NotificationGroup]

Lists of active notification groups

UpdateHavePendingNotifications
Describes whether there are some pending notification updates. Can be used to prevent application from killing, while there are some pending notifications

Fields

have_delayed_notifications :: Maybe Bool

True, if there are some delayed notification updates, which will be sent soon

have_unreceived_notifications :: Maybe Bool

True, if there can be some yet unreceived notifications, which are being fetched from the server

UpdateDeleteMessages
Some messages were deleted

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

message_ids :: Maybe [Int]

Identifiers of the deleted messages

is_permanent :: Maybe Bool

True, if the messages are permanently deleted by a user (as opposed to just becoming inaccessible)

from_cache :: Maybe Bool

True, if the messages are deleted only from the cache and can possibly be retrieved again in the future

UpdateChatAction
A message sender activity in the chat has changed

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

topic_id :: Maybe MessageTopic

Identifier of the topic

sender_id :: Maybe MessageSender

Identifier of a message sender performing the action

action :: Maybe ChatAction

The action

UpdatePendingTextMessage
A new pending text message was received in a chat with a bot. The message must be shown in the chat for at most getOption("pending_text_message_period") seconds, replace any other pending message with the same draft_id, and be deleted whenever any incoming message from the bot in the message thread is received

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

forum_topic_id :: Maybe Int

Forum topic identifier of the topic

draft_id :: Maybe Int

Unique identifier of the message draft within the message thread

text :: Maybe FormattedText

Text of the pending message

UpdateUserStatus
The user went online or offline

Fields

user_id :: Maybe Int

User identifier

status :: Maybe UserStatus

New status of the user

UpdateUser
Some data of a user has changed. This update is guaranteed to come before the user identifier is returned to the application

Fields

user :: Maybe User

New data about the user

UpdateBasicGroup
Some data of a basic group has changed. This update is guaranteed to come before the basic group identifier is returned to the application

Fields

basic_group :: Maybe BasicGroup

New data about the group

UpdateSupergroup
Some data of a supergroup or a channel has changed. This update is guaranteed to come before the supergroup identifier is returned to the application

Fields

supergroup :: Maybe Supergroup

New data about the supergroup

UpdateSecretChat
Some data of a secret chat has changed. This update is guaranteed to come before the secret chat identifier is returned to the application

Fields

secret_chat :: Maybe SecretChat

New data about the secret chat

UpdateUserFullInfo
Some data in userFullInfo has been changed

Fields

user_id :: Maybe Int

User identifier

user_full_info :: Maybe UserFullInfo

New full information about the user

UpdateBasicGroupFullInfo
Some data in basicGroupFullInfo has been changed

Fields

basic_group_id :: Maybe Int

Identifier of a basic group

basic_group_full_info :: Maybe BasicGroupFullInfo

New full information about the group

UpdateSupergroupFullInfo
Some data in supergroupFullInfo has been changed

Fields

supergroup_id :: Maybe Int

Identifier of the supergroup or channel

supergroup_full_info :: Maybe SupergroupFullInfo

New full information about the supergroup

UpdateServiceNotification
A service notification from the server was received. Upon receiving this the application must show a popup with the content of the notification

Fields

__type :: Maybe Text

Notification type. If type begins with AUTH_KEY_DROP_, then two buttons Cancel and "Log out" must be shown under notification; if user presses the second, all local data must be destroyed using Destroy method

content :: Maybe MessageContent

Notification content

UpdateFile
Information about a file was updated

Fields

file :: Maybe File

New data about the file

UpdateFileGenerationStart
The file generation process needs to be started by the application. Use setFileGenerationProgress and finishFileGeneration to generate the file

Fields

generation_id :: Maybe Int

Unique identifier for the generation process

original_path :: Maybe Text

The original path specified by the application in inputFileGenerated

destination_path :: Maybe Text

The path to a file that must be created and where the new file must be generated by the application. If the application has no access to the path, it can use writeGeneratedFilePart to generate the file

conversion :: Maybe Text

If the conversion is "" than original_path contains an HTTP/HTTPS URL of a file that must be downloaded by the application. Otherwise, this is the conversion specified by the application in inputFileGenerated

UpdateFileGenerationStop
File generation is no longer needed

Fields

generation_id :: Maybe Int

Unique identifier for the generation process

UpdateFileDownloads
The state of the file download list has changed

Fields

total_size :: Maybe Int

Total size of files in the file download list, in bytes

total_count :: Maybe Int

Total number of unread notifications in the group, can be bigger than number of active notifications

downloaded_size :: Maybe Int

Total downloaded size of files in the file download list, in bytes

UpdateFileAddedToDownloads
A file was added to the file download list. This update is sent only after file download list is loaded for the first time

Fields

file_download :: Maybe FileDownload

The added file download

counts :: Maybe DownloadedFileCounts

New number of being downloaded and recently downloaded files found

UpdateFileDownload
A file download was changed. This update is sent only after file download list is loaded for the first time

Fields

file_id :: Maybe Int

File identifier

complete_date :: Maybe Int

Point in time (Unix timestamp) when the file downloading was completed; 0 if the file downloading isn't completed

is_paused :: Maybe Bool

True, if downloading of the file is paused

counts :: Maybe DownloadedFileCounts

New number of being downloaded and recently downloaded files found

UpdateFileRemovedFromDownloads
A file was removed from the file download list. This update is sent only after file download list is loaded for the first time

Fields

file_id :: Maybe Int

File identifier

counts :: Maybe DownloadedFileCounts

New number of being downloaded and recently downloaded files found

UpdateApplicationVerificationRequired
A request can't be completed unless application verification is performed; for official mobile applications only. The method setApplicationVerificationToken must be called once the verification is completed or failed

Fields

verification_id :: Maybe Int

Unique identifier for the verification process

nonce :: Maybe Text

Unique base64url-encoded nonce for the classic Play Integrity verification (https:/developer.android.comgoogleplayintegrity/classic) for Android, or a unique string to compare with verify_nonce field from a push notification for iOS

cloud_project_number :: Maybe Int

Cloud project number to pass to the Play Integrity API on Android

UpdateApplicationRecaptchaVerificationRequired
A request can't be completed unless reCAPTCHA verification is performed; for official mobile applications only. The method setApplicationVerificationToken must be called once the verification is completed or failed

Fields

verification_id :: Maybe Int

Unique identifier for the verification process

_action :: Maybe Text

The action for the check

recaptcha_key_id :: Maybe Text

Identifier of the reCAPTCHA key

UpdateCall
New call was created or information about a call was updated

Fields

call :: Maybe Call

New data about a call

UpdateGroupCall
Information about a group call was updated

Fields

group_call :: Maybe GroupCall

New data about the group call

UpdateGroupCallParticipant
Information about a group call participant was changed. The updates are sent only after the group call is received through getGroupCall and only if the call is joined or being joined

Fields

group_call_id :: Maybe Int

Identifier of the group call

participant :: Maybe GroupCallParticipant

New data about the participant

UpdateGroupCallParticipants
The list of group call participants that can send and receive encrypted call data has changed; for group calls not bound to a chat only

Fields

group_call_id :: Maybe Int

Identifier of the group call

participant_user_ids :: Maybe [Int]

New list of group call participant user identifiers. The identifiers may be invalid or the corresponding users may be unknown. The participants must be shown in the list of group call participants even if there is no information about them

UpdateGroupCallVerificationState
The verification state of an encrypted group call has changed; for group calls not bound to a chat only

Fields

group_call_id :: Maybe Int

Identifier of the group call

generation :: Maybe Int

The call state generation to which the emoji corresponds. If generation is different for two users, then their emoji may be also different

emojis :: Maybe [Text]

Group call state fingerprint represented as 4 emoji; may be empty if the state isn't verified yet

UpdateNewGroupCallMessage
A new message was received in a group call

Fields

group_call_id :: Maybe Int

Identifier of the group call

_message :: Maybe GroupCallMessage

The message

UpdateNewGroupCallPaidReaction
A new paid reaction was received in a live story group call

Fields

group_call_id :: Maybe Int

Identifier of the group call

sender_id :: Maybe MessageSender

Identifier of a message sender performing the action

star_count :: Maybe Int

The number of Telegram Stars that were paid to send the reaction

UpdateGroupCallMessageSendFailed
A group call message failed to send

Fields

group_call_id :: Maybe Int

Identifier of the group call

_message_id :: Maybe Int

Message identifier

_error :: Maybe Error

The cause of the message sending failure

UpdateGroupCallMessagesDeleted
Some group call messages were deleted

Fields

group_call_id :: Maybe Int

Identifier of the group call

_message_ids :: Maybe [Int]

Identifiers of the deleted messages

UpdateLiveStoryTopDonors
The list of top donors in live story group call has changed

Fields

group_call_id :: Maybe Int

Identifier of the group call

donors :: Maybe LiveStoryDonors

New list of live story donors

UpdateNewCallSignalingData
New call signaling data arrived

Fields

call_id :: Maybe Int

The call identifier

_data :: Maybe ByteString

The data

UpdateGiftAuctionState
State of a gift auction was updated

Fields

state :: Maybe GiftAuctionState

New state of the auction

UpdateActiveGiftAuctions
The list of auctions in which participate the current user has changed

Fields

states :: Maybe [GiftAuctionState]

New states of the auctions

UpdateUserPrivacySettingRules
Some privacy setting rules have been changed

Fields

setting :: Maybe UserPrivacySetting

The privacy setting

rules :: Maybe UserPrivacySettingRules

New privacy rules

UpdateUnreadMessageCount
Number of unread messages in a chat list has changed. This update is sent only if the message database is used

Fields

chat_list :: Maybe ChatList

The chat list to which the chat was added

unread_count :: Maybe Int

The number of unread messages left in the chat

unread_unmuted_count :: Maybe Int

Total number of unread messages in unmuted chats

UpdateUnreadChatCount
Number of unread chats, i.e. with unread messages or marked as unread, has changed. This update is sent only if the message database is used

Fields

chat_list :: Maybe ChatList

The chat list to which the chat was added

total_count :: Maybe Int

Total number of unread notifications in the group, can be bigger than number of active notifications

unread_count :: Maybe Int

The number of unread messages left in the chat

unread_unmuted_count :: Maybe Int

Total number of unread messages in unmuted chats

marked_as_unread_count :: Maybe Int

Total number of chats marked as unread

marked_as_unread_unmuted_count :: Maybe Int

Total number of unmuted chats marked as unread

UpdateStory
A story was changed

Fields

story :: Maybe Story

The new information about the story

UpdateStoryDeleted
A story became inaccessible

Fields

story_poster_chat_id :: Maybe Int

Identifier of the chat that posted the story

story_id :: Maybe Int

Story identifier

UpdateStoryPostSucceeded
A story has been successfully posted

Fields

story :: Maybe Story

The new information about the story

old_story_id :: Maybe Int

The previous temporary story identifier

UpdateStoryPostFailed
A story failed to post. If the story posting is canceled, then updateStoryDeleted will be received instead of this update

Fields

story :: Maybe Story

The new information about the story

_error :: Maybe Error

The cause of the message sending failure

error_type :: Maybe CanPostStoryResult

Type of the error; may be null if unknown

UpdateChatActiveStories
The list of active stories posted by a specific chat has changed

Fields

active_stories :: Maybe ChatActiveStories

The new list of active stories

UpdateStoryListChatCount
Number of chats in a story list has changed

Fields

story_list :: Maybe StoryList

The story list

chat_count :: Maybe Int

Approximate total number of chats with active stories in the list

UpdateStoryStealthMode
Story stealth mode settings have changed

Fields

active_until_date :: Maybe Int

Point in time (Unix timestamp) until stealth mode is active; 0 if it is disabled

cooldown_until_date :: Maybe Int

Point in time (Unix timestamp) when stealth mode can be enabled again; 0 if there is no active cooldown

UpdateTrustedMiniAppBots
Lists of bots which Mini Apps must be allowed to read text from clipboard and must be opened without a warning

Fields

bot_user_ids :: Maybe [Int]

List of user identifiers of the bots; the corresponding users may not be sent using updateUser updates and may not be accessible

UpdateOption
An option changed its value

Fields

name :: Maybe Text

The option name

value :: Maybe OptionValue

The new option value

UpdateStickerSet
A sticker set has changed

Fields

sticker_set :: Maybe StickerSet

The sticker set

UpdateInstalledStickerSets
The list of installed sticker sets was updated

Fields

sticker_type :: Maybe StickerType

Type of the affected stickers

sticker_set_ids :: Maybe [Int]

The new list of installed ordinary sticker sets

UpdateTrendingStickerSets
The list of trending sticker sets was updated or some of them were viewed

Fields

sticker_type :: Maybe StickerType

Type of the affected stickers

sticker_sets :: Maybe TrendingStickerSets

The prefix of the list of trending sticker sets with the newest trending sticker sets

UpdateRecentStickers
The list of recently used stickers was updated

Fields

is_attached :: Maybe Bool

True, if the list of stickers attached to photo or video files was updated; otherwise, the list of sent stickers is updated

sticker_ids :: Maybe [Int]

The new list of file identifiers of recently used stickers

UpdateFavoriteStickers
The list of favorite stickers was updated

Fields

sticker_ids :: Maybe [Int]

The new list of file identifiers of recently used stickers

UpdateSavedAnimations
The list of saved animations was updated

Fields

animation_ids :: Maybe [Int]

The new list of file identifiers of saved animations

UpdateSavedNotificationSounds
The list of saved notification sounds was updated. This update may not be sent until information about a notification sound was requested for the first time

Fields

notification_sound_ids :: Maybe [Int]

The new list of identifiers of saved notification sounds

UpdateDefaultBackground
The default background has changed

Fields

for_dark_theme :: Maybe Bool

True, if default background for dark theme has changed

_background :: Maybe Background

The new default background; may be null

UpdateEmojiChatThemes
The list of available emoji chat themes has changed

Fields

chat_themes :: Maybe [EmojiChatTheme]

The new list of emoji chat themes

UpdateAccentColors
The list of supported accent colors has changed

Fields

colors :: Maybe [AccentColor]

Information about supported colors; colors with identifiers 0 (red), 1 (orange), 2 (purple/violet), 3 (green), 4 (cyan), 5 (blue), 6 (pink) must always be supported and aren't included in the list. The exact colors for the accent colors with identifiers 0-6 must be taken from the app theme

available_accent_color_ids :: Maybe [Int]

The list of accent color identifiers, which can be set through setAccentColor and setChatAccentColor. The colors must be shown in the specified order

UpdateProfileAccentColors
The list of supported accent colors for user profiles has changed

Fields

_colors :: Maybe [ProfileAccentColor]

Information about supported colors

available_accent_color_ids :: Maybe [Int]

The list of accent color identifiers, which can be set through setAccentColor and setChatAccentColor. The colors must be shown in the specified order

UpdateLanguagePackStrings
Some language pack strings have been updated

Fields

localization_target :: Maybe Text

Localization target to which the language pack belongs

language_pack_id :: Maybe Text

Identifier of the updated language pack

strings :: Maybe [LanguagePackString]

List of changed language pack strings; empty if all strings have changed

UpdateConnectionState
The connection state has changed. This update must be used only to show a human-readable description of the connection state

Fields

_state :: Maybe ConnectionState

The new connection state

UpdateFreezeState
The freeze state of the current user's account has changed

Fields

is_frozen :: Maybe Bool

True, if the account is frozen

freezing_date :: Maybe Int

Point in time (Unix timestamp) when the account was frozen; 0 if the account isn't frozen

deletion_date :: Maybe Int

Point in time (Unix timestamp) when the account will be deleted and can't be unfrozen; 0 if the account isn't frozen

appeal_link :: Maybe Text

The link to open to send an appeal to unfreeze the account

UpdateAgeVerificationParameters
The parameters for age verification of the current user's account has changed

Fields

parameters :: Maybe AgeVerificationParameters

Parameters for the age verification; may be null if age verification isn't needed

UpdateTermsOfService
New terms of service must be accepted by the user. If the terms of service are declined, then the deleteAccount method must be called with the reason "Decline ToS update"

Fields

terms_of_service_id :: Maybe Text

Identifier of the terms of service

terms_of_service :: Maybe TermsOfService

The new terms of service

UpdateUnconfirmedSession
The first unconfirmed session has changed

Fields

session :: Maybe UnconfirmedSession

The unconfirmed session; may be null if none

UpdateAttachmentMenuBots
The list of bots added to attachment or side menu has changed

Fields

bots :: Maybe [AttachmentMenuBot]

The new list of bots. The bots must not be shown on scheduled messages screen

UpdateWebAppMessageSent
A message was sent by an opened Web App, so the Web App needs to be closed

Fields

web_app_launch_id :: Maybe Int

Identifier of Web App launch

UpdateActiveEmojiReactions
The list of active emoji reactions has changed

Fields

emojis :: Maybe [Text]

Group call state fingerprint represented as 4 emoji; may be empty if the state isn't verified yet

UpdateAvailableMessageEffects
The list of available message effects has changed

Fields

reaction_effect_ids :: Maybe [Int]

The new list of available message effects from emoji reactions

sticker_effect_ids :: Maybe [Int]

The new list of available message effects from Premium stickers

UpdateDefaultReactionType
The type of default reaction has changed

Fields

reaction_type :: Maybe ReactionType

The new type of the default reaction

UpdateDefaultPaidReactionType
The type of default paid reaction has changed

Fields

___type :: Maybe PaidReactionType

The new type of the default paid reaction

UpdateSavedMessagesTags
Tags used in Saved Messages or a Saved Messages topic have changed

Fields

saved_messages_topic_id :: Maybe Int

Identifier of Saved Messages topic which tags were changed; 0 if tags for the whole chat has changed

tags :: Maybe SavedMessagesTags

The new tags

UpdateActiveLiveLocationMessages
The list of messages with active live location that need to be updated by the application has changed. The list is persistent across application restarts only if the message database is used

Fields

_messages :: Maybe [Message]

The list of messages with active live locations

UpdateOwnedStarCount
The number of Telegram Stars owned by the current user has changed

Fields

star_amount :: Maybe StarAmount

The new amount of owned Telegram Stars

UpdateOwnedTonCount
The number of Toncoins owned by the current user has changed

Fields

ton_amount :: Maybe Int

The new amount of owned Toncoins; in the smallest units of the cryptocurrency

UpdateChatRevenueAmount
The revenue earned from sponsored messages in a chat has changed. If chat revenue screen is opened, then getChatRevenueTransactions may be called to fetch new transactions

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

revenue_amount :: Maybe ChatRevenueAmount

New amount of earned revenue

UpdateStarRevenueStatus
The Telegram Star revenue earned by a user or a chat has changed. If Telegram Star transaction screen of the chat is opened, then getStarTransactions may be called to fetch new transactions

Fields

owner_id :: Maybe MessageSender

Identifier of the owner of the Telegram Stars

_status :: Maybe StarRevenueStatus

New Telegram Star revenue status

UpdateTonRevenueStatus
The Toncoin revenue earned by the current user has changed. If Toncoin transaction screen of the chat is opened, then getTonTransactions may be called to fetch new transactions

Fields

__status :: Maybe TonRevenueStatus

New Toncoin revenue status

UpdateSpeechRecognitionTrial
The parameters of speech recognition without Telegram Premium subscription has changed

Fields

max_media_duration :: Maybe Int

The maximum allowed duration of media for speech recognition without Telegram Premium subscription, in seconds

weekly_count :: Maybe Int

The total number of allowed speech recognitions per week; 0 if none

left_count :: Maybe Int

Number of left speech recognition attempts this week

next_reset_date :: Maybe Int

Point in time (Unix timestamp) when the weekly number of tries will reset; 0 if unknown

UpdateGroupCallMessageLevels
The levels of live story group call messages have changed

Fields

levels :: Maybe [GroupCallMessageLevel]

New description of the levels in decreasing order of groupCallMessageLevel.min_star_count

UpdateDiceEmojis
The list of supported dice emojis has changed

Fields

emojis :: Maybe [Text]

Group call state fingerprint represented as 4 emoji; may be empty if the state isn't verified yet

UpdateAnimatedEmojiMessageClicked
Some animated emoji message was clicked and a big animated sticker must be played if the message is visible on the screen. chatActionWatchingAnimations with the text of the message needs to be sent if the sticker is played

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

message_id :: Maybe Int

A temporary message identifier

sticker :: Maybe Sticker

The animated sticker to be played

UpdateAnimationSearchParameters
The parameters of animation search through getOption("animation_search_bot_username") bot has changed

Fields

provider :: Maybe Text

Name of the animation search provider

emojis :: Maybe [Text]

Group call state fingerprint represented as 4 emoji; may be empty if the state isn't verified yet

UpdateSuggestedActions
The list of suggested to the user actions has changed

Fields

added_actions :: Maybe [SuggestedAction]

Added suggested actions

removed_actions :: Maybe [SuggestedAction]

Removed suggested actions

UpdateSpeedLimitNotification
Download or upload file speed for the user was limited, but it can be restored by subscription to Telegram Premium. The notification can be postponed until a being downloaded or uploaded file is visible to the user. Use getOption("premium_download_speedup") or getOption("premium_upload_speedup") to get expected speedup after subscription to Telegram Premium

Fields

is_upload :: Maybe Bool

True, if upload speed was limited; false, if download speed was limited

UpdateContactCloseBirthdays
The list of contacts that had birthdays recently or will have birthday soon has changed

Fields

close_birthday_users :: Maybe [CloseBirthdayUser]

List of contact users with close birthday

UpdateAutosaveSettings
Autosave settings for some type of chats were updated

Fields

_scope :: Maybe AutosaveSettingsScope

Type of chats for which autosave settings were updated

settings :: Maybe ScopeAutosaveSettings

The new autosave settings; may be null if the settings are reset to default

UpdateBusinessConnection
A business connection has changed; for bots only

Fields

connection :: Maybe BusinessConnection

New data about the connection

UpdateNewBusinessMessage
A new message was added to a business account; for bots only

Fields

connection_id :: Maybe Text

Unique identifier of the business connection

__message :: Maybe BusinessMessage

The new message

UpdateBusinessMessageEdited
A message in a business account was edited; for bots only

Fields

connection_id :: Maybe Text

Unique identifier of the business connection

__message :: Maybe BusinessMessage

The new message

UpdateBusinessMessagesDeleted
Messages in a business account were deleted; for bots only

Fields

connection_id :: Maybe Text

Unique identifier of the business connection

chat_id :: Maybe Int

The chat identifier of the sent message

message_ids :: Maybe [Int]

Identifiers of the deleted messages

UpdateNewInlineQuery
A new incoming inline query; for bots only

Fields

_id :: Maybe Int

Unique query identifier

sender_user_id :: Maybe Int

Identifier of the user who sent the query

user_location :: Maybe Location

User location; may be null

chat_type :: Maybe ChatType

The type of the chat from which the query originated; may be null if unknown

query :: Maybe Text

Text of the query

offset :: Maybe Text

Offset of the first entry to return

UpdateNewChosenInlineResult
The user has chosen a result of an inline query; for bots only

Fields

sender_user_id :: Maybe Int

Identifier of the user who sent the query

user_location :: Maybe Location

User location; may be null

query :: Maybe Text

Text of the query

result_id :: Maybe Text

Identifier of the chosen result

inline_message_id :: Maybe Text

Identifier of the sent inline message, if known

UpdateNewCallbackQuery
A new incoming callback query; for bots only

Fields

_id :: Maybe Int

Unique query identifier

sender_user_id :: Maybe Int

Identifier of the user who sent the query

chat_id :: Maybe Int

The chat identifier of the sent message

message_id :: Maybe Int

A temporary message identifier

chat_instance :: Maybe Int

Identifier that uniquely corresponds to the chat to which the message was sent

payload :: Maybe CallbackQueryPayload

Query payload

UpdateNewInlineCallbackQuery
A new incoming callback query from a message sent via a bot; for bots only

Fields

_id :: Maybe Int

Unique query identifier

sender_user_id :: Maybe Int

Identifier of the user who sent the query

inline_message_id :: Maybe Text

Identifier of the sent inline message, if known

chat_instance :: Maybe Int

Identifier that uniquely corresponds to the chat to which the message was sent

payload :: Maybe CallbackQueryPayload

Query payload

UpdateNewBusinessCallbackQuery
A new incoming callback query from a business message; for bots only

Fields

_id :: Maybe Int

Unique query identifier

sender_user_id :: Maybe Int

Identifier of the user who sent the query

connection_id :: Maybe Text

Unique identifier of the business connection

__message :: Maybe BusinessMessage

The new message

chat_instance :: Maybe Int

Identifier that uniquely corresponds to the chat to which the message was sent

payload :: Maybe CallbackQueryPayload

Query payload

UpdateNewShippingQuery
A new incoming shipping query; for bots only. Only for invoices with flexible price

Fields

_id :: Maybe Int

Unique query identifier

sender_user_id :: Maybe Int

Identifier of the user who sent the query

invoice_payload :: Maybe Text

Invoice payload

shipping_address :: Maybe Address

User shipping address

UpdateNewPreCheckoutQuery
A new incoming pre-checkout query; for bots only. Contains full information about a checkout

Fields

_id :: Maybe Int

Unique query identifier

sender_user_id :: Maybe Int

Identifier of the user who sent the query

currency :: Maybe Text

Currency for the product price

total_amount :: Maybe Int

Total price for the product, in the smallest units of the currency

_invoice_payload :: Maybe ByteString

Invoice payload

shipping_option_id :: Maybe Text

Identifier of a shipping option chosen by the user; may be empty if not applicable

order_info :: Maybe OrderInfo

Information about the order; may be null

UpdateNewCustomEvent
A new incoming event; for bots only

Fields

event :: Maybe Text

A JSON-serialized event

UpdateNewCustomQuery
A new incoming query; for bots only

Fields

_id :: Maybe Int

Unique query identifier

__data :: Maybe Text

JSON-serialized query data

timeout :: Maybe Int

Query timeout

UpdatePoll
A poll was updated; for bots only

Fields

poll :: Maybe Poll

New data about the poll

UpdatePollAnswer
A user changed the answer to a poll; for bots only

Fields

poll_id :: Maybe Int

Unique poll identifier

voter_id :: Maybe MessageSender

Identifier of the message sender that changed the answer to the poll

option_ids :: Maybe [Int]

0-based identifiers of answer options, chosen by the user

UpdateChatMember
User rights changed in a chat; for bots only

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

actor_user_id :: Maybe Int

Identifier of the user, changing the rights

date :: Maybe Int

Point in time (Unix timestamp) when the user rights were changed

invite_link :: Maybe ChatInviteLink

If user has joined the chat using an invite link, the invite link; may be null

via_join_request :: Maybe Bool

True, if the user has joined the chat after sending a join request and being approved by an administrator

via_chat_folder_invite_link :: Maybe Bool

True, if the user has joined the chat using an invite link for a chat folder

old_chat_member :: Maybe ChatMember

Previous chat member

new_chat_member :: Maybe ChatMember

New chat member

UpdateNewChatJoinRequest
A user sent a join request to a chat; for bots only

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

request :: Maybe ChatJoinRequest

Join request

user_chat_id :: Maybe Int

Chat identifier of the private chat with the user

invite_link :: Maybe ChatInviteLink

If user has joined the chat using an invite link, the invite link; may be null

UpdateChatBoost
A chat boost has changed; for bots only

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

boost :: Maybe ChatBoost

New information about the boost

UpdateMessageReaction
User changed its reactions on a message with public reactions; for bots only

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

message_id :: Maybe Int

A temporary message identifier

actor_id :: Maybe MessageSender

Identifier of the user or chat that changed reactions

date :: Maybe Int

Point in time (Unix timestamp) when the user rights were changed

old_reaction_types :: Maybe [ReactionType]

Old list of chosen reactions

new_reaction_types :: Maybe [ReactionType]

New list of chosen reactions

UpdateMessageReactions
Reactions added to a message with anonymous reactions have changed; for bots only

Fields

chat_id :: Maybe Int

The chat identifier of the sent message

message_id :: Maybe Int

A temporary message identifier

date :: Maybe Int

Point in time (Unix timestamp) when the user rights were changed

reactions :: Maybe [MessageReaction]

The list of reactions added to the message

UpdatePaidMediaPurchased
Paid media were purchased by a user; for bots only

Fields

user_id :: Maybe Int

User identifier

_payload :: Maybe Text

Bot-specified payload for the paid media