TD.Data.UpgradedGift

Documentation

data UpgradedGift

Constructors

UpgradedGift
Describes an upgraded gift that can be transferred to another owner or transferred to the TON blockchain as an NFT

Fields

_id :: Maybe Int

Unique identifier of the gift

title :: Maybe Text

The title of the upgraded gift

name :: Maybe Text

Unique name of the upgraded gift that can be used with internalLinkTypeUpgradedGift

number :: Maybe Int

Unique number of the upgraded gift among gifts upgraded from the same gift

total_upgraded_count :: Maybe Int

Total number of gifts that were upgraded from the same gift

max_upgraded_count :: Maybe Int

The maximum number of gifts that can be upgraded from the same gift

owner_id :: Maybe MessageSender

Identifier of the user or the chat that owns the upgraded gift; may be null if none or unknown

owner_address :: Maybe Text

Address of the gift NFT owner in TON blockchain; may be empty if none

owner_name :: Maybe Text

Name of the owner for the case when owner identifier and address aren't known

model :: Maybe UpgradedGiftModel

Model of the upgraded gift

symbol :: Maybe UpgradedGiftSymbol

Symbol of the upgraded gift

backdrop :: Maybe UpgradedGiftBackdrop

Backdrop of the upgraded gift

original_details :: Maybe UpgradedGiftOriginalDetails

Information about the originally sent gift; may be null if unknown