TD.Data.UpgradedGiftOrigin

Documentation

data UpgradedGiftOrigin

Describes origin from which the upgraded gift was obtained

Constructors

UpgradedGiftOriginUpgrade
The gift was obtained by upgrading of a previously received gift

Fields

gift_message_id :: Maybe Int

Identifier of the message with the regular gift that was upgraded; can be 0 or an identifier of a deleted message

UpgradedGiftOriginTransfer
The gift was transferred from another owner

UpgradedGiftOriginResale
The gift was bought from another user

Fields

star_count :: Maybe Int

Number of Telegram Stars that were paid by the sender for the gift