TD.Data.UpgradedGiftValueInfo

Documentation

data UpgradedGiftValueInfo

Constructors

UpgradedGiftValueInfo
Contains information about value of an upgraded gift

Fields

currency :: Maybe Text

ISO 4217 currency code of the currency in which the prices are represented

value :: Maybe Int

Estimated value of the gift; in the smallest units of the currency

is_value_average :: Maybe Bool

True, if the value is calculated as average value of similar sold gifts. Otherwise, it is based on the sale price of the gift

initial_sale_date :: Maybe Int

Point in time (Unix timestamp) when the corresponding regular gift was originally purchased

initial_sale_star_count :: Maybe Int

Amount of Telegram Stars that were paid for the gift

initial_sale_price :: Maybe Int

Initial price of the gift; in the smallest units of the currency

last_sale_date :: Maybe Int

Point in time (Unix timestamp) when the upgraded gift was purchased last time; 0 if never

last_sale_price :: Maybe Int

Last purchase price of the gift; in the smallest units of the currency; 0 if the gift has never been resold

is_last_sale_on_fragment :: Maybe Bool

True, if the last sale was completed on Fragment

minimum_price :: Maybe Int

The current minimum price of gifts upgraded from the same gift; in the smallest units of the currency; 0 if there are no such gifts

average_sale_price :: Maybe Int

The average sale price in the last month of gifts upgraded from the same gift; in the smallest units of the currency; 0 if there were no such sales

telegram_listed_gift_count :: Maybe Int

Number of gifts upgraded from the same gift being resold on Telegram

fragment_listed_gift_count :: Maybe Int

Number of gifts upgraded from the same gift being resold on Fragment

fragment_url :: Maybe Text

The HTTPS link to the Fragment for the gift; may be empty if there are no such gifts being sold on Fragment