TD.Data.User

Documentation

data User

Constructors

User
Represents a user

Fields

_id :: Maybe Int

User identifier

first_name :: Maybe Text

First name of the user

last_name :: Maybe Text

Last name of the user

usernames :: Maybe Usernames

Usernames of the user; may be null

phone_number :: Maybe Text

Phone number of the user

status :: Maybe UserStatus

Current online status of the user

profile_photo :: Maybe ProfilePhoto

Profile photo of the user; may be null

accent_color_id :: Maybe Int

Identifier of the accent color for name, and backgrounds of profile photo, reply header, and link preview

background_custom_emoji_id :: Maybe Int

Identifier of a custom emoji to be shown on the reply header and link preview background; 0 if none

profile_accent_color_id :: Maybe Int

Identifier of the accent color for the user's profile; -1 if none

profile_background_custom_emoji_id :: Maybe Int

Identifier of a custom emoji to be shown on the background of the user's profile; 0 if none

emoji_status :: Maybe EmojiStatus

Emoji status to be shown instead of the default Telegram Premium badge; may be null

is_contact :: Maybe Bool

The user is a contact of the current user

is_mutual_contact :: Maybe Bool

The user is a contact of the current user and the current user is a contact of the user

is_close_friend :: Maybe Bool

The user is a close friend of the current user; implies that the user is a contact

is_verified :: Maybe Bool

True, if the user is verified

is_premium :: Maybe Bool

True, if the user is a Telegram Premium user

is_support :: Maybe Bool

True, if the user is Telegram support account

restriction_reason :: Maybe Text

If non-empty, it contains a human-readable description of the reason why access to this user must be restricted

is_scam :: Maybe Bool

True, if many users reported this user as a scam

is_fake :: Maybe Bool

True, if many users reported this user as a fake account

has_active_stories :: Maybe Bool

True, if the user has non-expired stories available to the current user

has_unread_active_stories :: Maybe Bool

True, if the user has unread non-expired stories available to the current user

restricts_new_chats :: Maybe Bool

True, if the user may restrict new chats with non-contacts. Use canSendMessageToUser to check whether the current user can message the user or try to create a chat with them

have_access :: Maybe Bool

If false, the user is inaccessible, and the only information known about the user is inside this class. Identifier of the user can't be passed to any method

_type :: Maybe UserType

Type of the user

language_code :: Maybe Text

IETF language tag of the user's language; only available to bots

added_to_attachment_menu :: Maybe Bool

True, if the user added the current bot to attachment menu; only available to bots