TD.Data.UserFullInfo

Documentation

data UserFullInfo

Constructors

UserFullInfo
Contains full information about a user

Fields

personal_photo :: Maybe ChatPhoto

User profile photo set by the current user for the contact; may be null. If null and user.profile_photo is null, then the photo is empty; otherwise, it is unknown. If non-null, then it is the same photo as in user.profile_photo and chat.photo. This photo isn't returned in the list of user photos

photo :: Maybe ChatPhoto

User profile photo; may be null. If null and user.profile_photo is null, then the photo is empty; otherwise, it is unknown. If non-null and personal_photo is null, then it is the same photo as in user.profile_photo and chat.photo

public_photo :: Maybe ChatPhoto

User profile photo visible if the main photo is hidden by privacy settings; may be null. If null and user.profile_photo is null, then the photo is empty; otherwise, it is unknown. If non-null and both photo and personal_photo are null, then it is the same photo as in user.profile_photo and chat.photo. This photo isn't returned in the list of user photos

block_list :: Maybe BlockList

Block list to which the user is added; may be null if none

can_be_called :: Maybe Bool

True, if the user can be called

supports_video_calls :: Maybe Bool

True, if a video call can be created with the user

has_private_calls :: Maybe Bool

True, if the user can't be called due to their privacy settings

has_private_forwards :: Maybe Bool

True, if the user can't be linked in forwarded messages due to their privacy settings

has_restricted_voice_and_video_note_messages :: Maybe Bool

True, if voice and video notes can't be sent or forwarded to the user

has_posted_to_profile_stories :: Maybe Bool

True, if the user has posted to profile stories

has_sponsored_messages_enabled :: Maybe Bool

True, if the user always enabled sponsored messages; known only for the current user

need_phone_number_privacy_exception :: Maybe Bool

True, if the current user needs to explicitly allow to share their phone number with the user when the method addContact is used

set_chat_background :: Maybe Bool

True, if the user set chat background for both chat users and it wasn't reverted yet

bio :: Maybe FormattedText

A short user bio; may be null for bots

birthdate :: Maybe Birthdate

Birthdate of the user; may be null if unknown

personal_chat_id :: Maybe Int

Identifier of the personal chat of the user; 0 if none

gift_count :: Maybe Int

Number of saved to profile gifts for other users or the total number of received gifts for the current user

group_in_common_count :: Maybe Int

Number of group chats where both the other user and the current user are a member; 0 for the current user

incoming_paid_message_star_count :: Maybe Int

Number of Telegram Stars that must be paid by the user for each sent message to the current user

outgoing_paid_message_star_count :: Maybe Int

Number of Telegram Stars that must be paid by the current user for each sent message to the user

gift_settings :: Maybe GiftSettings

Settings for gift receiving for the user

bot_verification :: Maybe BotVerification

Information about verification status of the user provided by a bot; may be null if none or unknown

business_info :: Maybe BusinessInfo

Information about business settings for Telegram Business accounts; may be null if none

bot_info :: Maybe BotInfo

For bots, information about the bot; may be null if the user isn't a bot