TD.Data.UserGift

Documentation

data UserGift

Constructors

UserGift
Represents a gift received by a user

Fields

sender_user_id :: Maybe Int

Identifier of the user that sent the gift; 0 if unknown

text :: Maybe FormattedText

Message added to the gift

is_private :: Maybe Bool

True, if the sender and gift text are shown only to the gift receiver; otherwise, everyone are able to see them

is_saved :: Maybe Bool

True, if the gift is displayed on the user's profile page; may be false only for the receiver of the gift

date :: Maybe Int

Point in time (Unix timestamp) when the gift was sent

gift :: Maybe Gift

The gift

message_id :: Maybe Int

Identifier of the message with the gift in the chat with the sender of the gift; can be 0 or an identifier of a deleted message; only for the gift receiver

sell_star_count :: Maybe Int

Number of Telegram Stars that can be claimed by the receiver instead of the gift; only for the gift receiver