TD.Data.UserPrivacySetting

Description

Documentation

data UserPrivacySetting

Describes available user privacy settings

Constructors

UserPrivacySettingShowStatus
A privacy setting for managing whether the user's online status is visible

UserPrivacySettingShowProfilePhoto
A privacy setting for managing whether the user's profile photo is visible

UserPrivacySettingShowLinkInForwardedMessages
A privacy setting for managing whether a link to the user's account is included in forwarded messages

UserPrivacySettingShowPhoneNumber
A privacy setting for managing whether the user's phone number is visible

UserPrivacySettingAllowChatInvites
A privacy setting for managing whether the user can be invited to chats

UserPrivacySettingAllowCalls
A privacy setting for managing whether the user can be called

UserPrivacySettingAllowPeerToPeerCalls
A privacy setting for managing whether peer-to-peer connections can be used for calls

UserPrivacySettingAllowFindingByPhoneNumber
A privacy setting for managing whether the user can be found by their phone number. Checked only if the phone number is not known to the other user. Can be set only to "Allow contacts" or "Allow all"

UserPrivacySettingAllowPrivateVoiceAndVideoNoteMessages
A privacy setting for managing whether the user can receive voice and video messages in private chats