TD.Data.UserType

Description

Documentation

data UserType

Represents the type of a user. The following types are possible: regular users, deleted users and bots

Constructors

UserTypeRegular
A regular user

UserTypeDeleted
A deleted user or deleted bot. No information on the user besides the user identifier is available. It is not possible to perform any active actions on this type of user

UserTypeBot
A bot (see https://core.telegram.org/bots)

Fields

can_be_added_to_attachment_menu :: Maybe Bool

True, if the bot can be added to attachment menu

need_location :: Maybe Bool

True, if the location of the user is expected to be sent with every inline query to this bot

inline_query_placeholder :: Maybe String

Placeholder for inline queries (displayed on the application input field)

is_inline :: Maybe Bool

True, if the bot supports inline queries

can_read_all_group_messages :: Maybe Bool

True, if the bot can read all messages in basic group or supergroup chats and not just those addressed to the bot. In private and channel chats a bot can always read all messages

can_join_groups :: Maybe Bool

True, if the bot can be invited to basic group and supergroup chats

UserTypeUnknown
No information on the user besides the user identifier is available, yet this user has not been deleted. This object is extremely rare and must be handled like a deleted user. It is not possible to perform any actions on users of this type