TD.Data.Video

Documentation

data Video

Constructors

Video
Describes a video file

Fields

duration :: Maybe Int

Duration of the video, in seconds; as defined by the sender

width :: Maybe Int

Video width; as defined by the sender

height :: Maybe Int

Video height; as defined by the sender

file_name :: Maybe Text

Original name of the file; as defined by the sender

mime_type :: Maybe Text

MIME type of the file; as defined by the sender

has_stickers :: Maybe Bool

True, if stickers were added to the video. The list of corresponding sticker sets can be received using getAttachedStickerSets

supports_streaming :: Maybe Bool

True, if the video is expected to be streamed

minithumbnail :: Maybe Minithumbnail

Video minithumbnail; may be null

thumbnail :: Maybe Thumbnail

Video thumbnail in JPEG or MPEG4 format; as defined by the sender; may be null

video :: Maybe File

File containing the video