TD.Data.VideoChat

Description

Documentation

data VideoChat

Constructors

VideoChat
Describes a video chat

Fields

default_participant_id :: Maybe MessageSender

Default group call participant identifier to join the video chat; may be null

has_participants :: Maybe Bool

True, if the video chat has participants

group_call_id :: Maybe Int

Group call identifier of an active video chat; 0 if none. Full information about the video chat can be received through the method getGroupCall