TD.Data.VideoNote

Documentation

data VideoNote

Constructors

VideoNote
Describes a video note. The video must be equal in width and height, cropped to a circle, and stored in MPEG4 format

Fields

duration :: Maybe Int

Duration of the video, in seconds; as defined by the sender

waveform :: Maybe ByteString

A waveform representation of the video note's audio in 5-bit format; may be empty if unknown

_length :: Maybe Int

Video width and height; as defined by the sender

minithumbnail :: Maybe Minithumbnail

Video minithumbnail; may be null

thumbnail :: Maybe Thumbnail

Video thumbnail in JPEG format; as defined by the sender; may be null

speech_recognition_result :: Maybe SpeechRecognitionResult

Result of speech recognition in the video note; may be null

video :: Maybe File

File containing the video