TD.Data.WebPage

Description

Documentation

data WebPage

Constructors

WebPage
Describes a web page preview

Fields

instant_view_version :: Maybe Int

Version of web page instant view (currently, can be 1 or 2); 0 if none

voice_note :: Maybe VoiceNote

Preview of the content as a voice note, if available; may be null

video_note :: Maybe VideoNote

Preview of the content as a video note, if available; may be null

video :: Maybe Video

Preview of the content as a video, if available; may be null

sticker :: Maybe Sticker

Preview of the content as a sticker for small WEBP files, if available; may be null

document :: Maybe Document

Preview of the content as a document, if available; may be null

audio :: Maybe Audio

Preview of the content as an audio file, if available; may be null

animation :: Maybe Animation

Preview of the content as an animation, if available; may be null

author :: Maybe String

Author of the content

duration :: Maybe Int

Duration of the content, in seconds

embed_height :: Maybe Int

Height of the embedded preview

embed_width :: Maybe Int

Width of the embedded preview

embed_type :: Maybe String

MIME type of the embedded preview, (e.g., texthtml or videomp4)

embed_url :: Maybe String

URL to show in the embedded preview

photo :: Maybe Photo

Image representing the content; may be null

description :: Maybe FormattedText

title :: Maybe String

Title of the content

site_name :: Maybe String

Short name of the site (e.g., Google Docs, App Store)

_type :: Maybe String

Type of the web page. Can be: article, photo, audio, video, document, profile, app, or something else

display_url :: Maybe String

URL to display

url :: Maybe String

Original URL of the link