TD.Lib

Documentation

create :: IO Client

send :: ToJSON a => Client -> a -> IO ()

sendWExtra :: ToJSON a => Client -> a -> IO Extra

receive :: Client -> IO (Maybe (GeneralResult, Maybe Extra))

destroy :: Client -> IO ()

type Client = Ptr ()

data Extra

class ShortShow a where

Show class alternative that hides unset and bytes values, multiline text values are shown as single line

Methods

shortShow :: a -> String

createExtra :: Hashable a => a -> IO Extra

sendWMyExtra :: ToJSON a => Client -> a -> Extra -> IO ()