TD.Query.AcceptOauthRequest

Documentation

data AcceptOauthRequest

Accepts an OAuth authorization request. Returns an HTTP URL to open after successful authorization. May return an empty link if just a toast about successful login has to be shown. Returns HttpUrl

Constructors

AcceptOauthRequest

Fields

url :: Maybe Text

URL of the OAuth deep link

match_code :: Maybe Text

The matching code chosen by the user

allow_write_access :: Maybe Bool

Pass true if the current user allowed the bot that was returned in getOauthLinkInfo, to send them messages

allow_phone_number_access :: Maybe Bool

Pass true if the current user allowed the bot that was returned in getOauthLinkInfo, to access their phone number

defaultAcceptOauthRequest :: AcceptOauthRequest