TD.Query.AddBotMediaPreview

Documentation

data AddBotMediaPreview

Adds a new media preview to the beginning of the list of media previews of a bot. Returns the added preview after addition is completed server-side. The total number of previews must not exceed getOption("bot_media_preview_count_max") for the given language. Returns BotMediaPreview

Constructors

AddBotMediaPreview

Fields

bot_user_id :: Maybe Int

Identifier of the target bot. The bot must be owned and must have the main Web App

language_code :: Maybe Text

A two-letter ISO 639-1 language code for which preview is added. If empty, then the preview will be shown to all users for whose languages there are no dedicated previews. If non-empty, then there must be an official language pack of the same name, which is returned by getLocalizationTargetInfo

content :: Maybe InputStoryContent

Content of the added preview

defaultAddBotMediaPreview :: AddBotMediaPreview