TD.Query.AddContact

Description

Documentation

data AddContact

Adds a user to the contact list or edits an existing contact by their user identifier @contact The contact to add or edit; phone number may be empty and needs to be specified only if known, vCard is ignored

Constructors

AddContact

Fields

share_phone_number :: Maybe Bool

Pass true to share the current user's phone number with the new contact. A corresponding rule to userPrivacySettingShowPhoneNumber will be added if needed. Use the field userFullInfo.need_phone_number_privacy_exception to check whether the current user needs to be asked to share their phone number

contact :: Maybe Contact