TD.Query.AddFileToDownloads

Description

Documentation

data AddFileToDownloads

Adds a file from a message to the list of file downloads. Download progress and completion of the download will be notified through updateFile updates. If message database is used, the list of file downloads is persistent across application restarts. The downloading is independent from download using downloadFile, i.e. it continues if downloadFile is canceled or is used to download a part of the file

Constructors

AddFileToDownloads

Fields

priority :: Maybe Int

Priority of the download (1-32). The higher the priority, the earlier the file will be downloaded. If the priorities of two files are equal, then the last one for which downloadFile/addFileToDownloads was called will be downloaded first

message_id :: Maybe Int

Message identifier

chat_id :: Maybe Int

Chat identifier of the message with the file

file_id :: Maybe Int

Identifier of the file to download