TD.Query.AddGiftCollectionGifts

Documentation

data AddGiftCollectionGifts

Adds gifts to the beginning of a previously created collection. If the collection is owned by a channel chat, then requires can_post_messages administrator right in the channel chat. Returns the changed collection. Returns GiftCollection

Constructors

AddGiftCollectionGifts

Fields

owner_id :: Maybe MessageSender

Identifier of the user or the channel chat that owns the collection

collection_id :: Maybe Int

Identifier of the gift collection

received_gift_ids :: Maybe [Text]

Identifier of the gifts to add to the collection; 1-getOption("gift_collection_gift_count_max") identifiers. If after addition the collection has more than getOption("gift_collection_gift_count_max") gifts, then the last one are removed from the collection

defaultAddGiftCollectionGifts :: AddGiftCollectionGifts