TD.Query.AddPendingLiveStoryReaction

Documentation

data AddPendingLiveStoryReaction

Adds pending paid reaction in a live story group call. Can't be used in live stories posted by the current user. Call commitPendingLiveStoryReactions or removePendingLiveStoryReactions to actually send all pending reactions when the undo timer is over or abort the sending. Returns Ok

Constructors

AddPendingLiveStoryReaction

Fields

group_call_id :: Maybe Int

Group call identifier

star_count :: Maybe Int

Number of Telegram Stars to be used for the reaction. The total number of pending paid reactions must not exceed getOption("paid_group_call_message_star_count_max")

defaultAddPendingLiveStoryReaction :: AddPendingLiveStoryReaction