TD.Query.AddQuickReplyShortcutInlineQueryResultMessage

Documentation

data AddQuickReplyShortcutInlineQueryResultMessage

Adds a message to a quick reply shortcut via inline bot. If shortcut doesn't exist and there are less than getOption("quick_reply_shortcut_count_max") shortcuts, then a new shortcut is created. The shortcut must not contain more than getOption("quick_reply_shortcut_message_count_max") messages after adding the new message. Returns the added message. Returns QuickReplyMessage

Constructors

AddQuickReplyShortcutInlineQueryResultMessage

Fields

shortcut_name :: Maybe Text

Name of the target shortcut

reply_to_message_id :: Maybe Int

Identifier of a quick reply message in the same shortcut to be replied; pass 0 if none

query_id :: Maybe Int

Identifier of the inline query

result_id :: Maybe Text

Identifier of the inline query result

hide_via_bot :: Maybe Bool

Pass true to hide the bot, via which the message is sent. Can be used only for bots getOption("animation_search_bot_username"), getOption("photo_search_bot_username"), and getOption("venue_search_bot_username")

defaultAddQuickReplyShortcutInlineQueryResultMessage :: AddQuickReplyShortcutInlineQueryResultMessage