TD.Query.AddQuickReplyShortcutMessageAlbum

Documentation

data AddQuickReplyShortcutMessageAlbum

Adds 2-10 messages grouped together into an album to a quick reply shortcut. Currently, only audio, document, photo and video messages can be grouped into an album. Documents and audio files can be only grouped in an album with messages of the same type. Returns sent messages. Returns QuickReplyMessages

Constructors

AddQuickReplyShortcutMessageAlbum

Fields

shortcut_name :: Maybe Text

Name of the target shortcut

reply_to_message_id :: Maybe Int

Identifier of a quick reply message in the same shortcut to be replied; pass 0 if none

input_message_contents :: Maybe [InputMessageContent]

Contents of messages to be sent. At most 10 messages can be added to an album. All messages must have the same value of show_caption_above_media

defaultAddQuickReplyShortcutMessageAlbum :: AddQuickReplyShortcutMessageAlbum