TD.Query.AddStoryAlbumStories

Documentation

data AddStoryAlbumStories

Adds stories to the beginning of a previously created story album. If the album is owned by a supergroup or a channel chat, then requires can_edit_stories administrator right in the chat. Returns the changed album. Returns StoryAlbum

Constructors

AddStoryAlbumStories

Fields

chat_id :: Maybe Int

Identifier of the chat that owns the stories

story_album_id :: Maybe Int

Identifier of the story album

story_ids :: Maybe [Int]

Identifier of the stories to add to the album; 1-getOption("story_album_story_count_max") identifiers. If after addition the album has more than getOption("story_album_story_count_max") stories, then the last one are removed from the album

defaultAddStoryAlbumStories :: AddStoryAlbumStories