TD.Query.AnswerCallbackQuery

Documentation

data AnswerCallbackQuery

Sets the result of a callback query; for bots only callback_query_id Identifier of the callback query text Text of the answer show_alert Pass true to show an alert to the user instead of a toast notification url URL to be opened @cache_time Time during which the result of the query can be cached, in seconds

Constructors

AnswerCallbackQuery

Fields

cache_time :: Maybe Int

url :: Maybe String

show_alert :: Maybe Bool

text :: Maybe String

callback_query_id :: Maybe Int