TD.Query.AnswerInlineQuery

Description

Documentation

data AnswerInlineQuery

Sets the result of an inline query; for bots only

Constructors

AnswerInlineQuery

Fields

next_offset :: Maybe String

Offset for the next inline query; pass an empty string if there are no more results

cache_time :: Maybe Int

Allowed time to cache the results of the query, in seconds

results :: Maybe [InputInlineQueryResult]

The results of the query

button :: Maybe InlineQueryResultsButton

Button to be shown above inline query results; pass null if none

is_personal :: Maybe Bool

Pass true if results may be cached and returned only for the user that sent the query. By default, results may be returned to any user who sends the same query

inline_query_id :: Maybe Int

Identifier of the inline query