TD.Query.BanGroupCallParticipants

Documentation

data BanGroupCallParticipants

Bans users from a group call not bound to a chat; requires groupCall.is_owned. Only the owner of the group call can invite the banned users back. Returns Ok

Constructors

BanGroupCallParticipants

Fields

group_call_id :: Maybe Int

Group call identifier

user_ids :: Maybe [Int]

Identifiers of group call participants to ban; identifiers of unknown users from the update updateGroupCallParticipants can be also passed to the method

defaultBanGroupCallParticipants :: BanGroupCallParticipants