TD.Query.CanSendMessageToUser

Documentation

data CanSendMessageToUser

Check whether the current user can message another user or try to create a chat with them. Returns CanSendMessageToUserResult

Constructors

CanSendMessageToUser

Fields

user_id :: Maybe Int

Identifier of the other user

only_local :: Maybe Bool

Pass true to get only locally available information without sending network requests

defaultCanSendMessageToUser :: CanSendMessageToUser