TD.Query.ChangePhoneNumber

Description

Documentation

data ChangePhoneNumber

Changes the phone number of the user and sends an authentication code to the user's new phone number. On success, returns information about the sent code

Constructors

ChangePhoneNumber

Fields

settings :: Maybe PhoneNumberAuthenticationSettings

Settings for the authentication of the user's phone number; pass null to use default settings

phone_number :: Maybe String

The new phone number of the user in international format