TD.Query.CheckAuthenticationPasskey

Documentation

data CheckAuthenticationPasskey

Checks a passkey to log in to the corresponding account. Call getAuthenticationPasskeyParameters to get parameters for the passkey. Works only when the current authorization state is authorizationStateWaitPhoneNumber or authorizationStateWaitOtherDeviceConfirmation, or if there is no pending authentication query and the current authorization state is authorizationStateWaitPremiumPurchase, authorizationStateWaitEmailAddress, authorizationStateWaitEmailCode, authorizationStateWaitCode, authorizationStateWaitRegistration, or authorizationStateWaitPassword. Returns Ok

Constructors

CheckAuthenticationPasskey

Fields

credential_id :: Maybe Text

Base64url-encoded identifier of the credential

client_data :: Maybe Text

JSON-encoded client data

authenticator_data :: Maybe ByteString

Authenticator data of the application that created the credential

signature :: Maybe ByteString

Cryptographic signature of the credential

user_handle :: Maybe ByteString

User handle of the passkey

defaultCheckAuthenticationPasskey :: CheckAuthenticationPasskey