TD.Query.CreateChatInviteLink

Description

Documentation

data CreateChatInviteLink

Creates a new invite link for a chat. Available for basic groups, supergroups, and channels. Requires administrator privileges and can_invite_users right in the chat

Constructors

CreateChatInviteLink

Fields

creates_join_request :: Maybe Bool

Pass true if users joining the chat via the link need to be approved by chat administrators. In this case, member_limit must be 0

member_limit :: Maybe Int

The maximum number of chat members that can join the chat via the link simultaneously; 0-99999; pass 0 if not limited

expiration_date :: Maybe Int

Point in time (Unix timestamp) when the link will expire; pass 0 if never

name :: Maybe String

Invite link name; 0-32 characters

chat_id :: Maybe Int

Chat identifier