TD.Query.CreateForumTopic

Description

Documentation

data CreateForumTopic

Creates a topic in a forum supergroup chat; requires can_manage_topics rights in the supergroup

Constructors

CreateForumTopic

Fields

icon :: Maybe ForumTopicIcon

Icon of the topic. Icon color must be one of 0x6FB9F0, 0xFFD67E, 0xCB86DB, 0x8EEE98, 0xFF93B2, or 0xFB6F5F. Telegram Premium users can use any custom emoji as topic icon, other users can use only a custom emoji returned by getForumTopicDefaultIcons

name :: Maybe String

Name of the topic; 1-128 characters

chat_id :: Maybe Int

Identifier of the chat