TD.Query.CreateNewStickerSet

Documentation

data CreateNewStickerSet

Creates a new sticker set. Returns the newly created sticker set. Returns StickerSet

Constructors

CreateNewStickerSet

Fields

user_id :: Maybe Int

Sticker set owner; ignored for regular users

title :: Maybe Text

Sticker set title; 1-64 characters

name :: Maybe Text

Sticker set name. Can contain only English letters, digits and underscores. Must end with *"_by_username"* (*bot_username* is case insensitive) for bots; 0-64 characters. If empty, then the name returned by getSuggestedStickerSetName will be used automatically

sticker_type :: Maybe StickerType

Type of the stickers in the set

needs_repainting :: Maybe Bool

Pass true if stickers in the sticker set must be repainted; for custom emoji sticker sets only

stickers :: Maybe [InputSticker]

List of stickers to be added to the set; 1-200 stickers for custom emoji sticker sets, and 1-120 stickers otherwise. For TGS stickers, uploadStickerFile must be used before the sticker is shown

source :: Maybe Text

Source of the sticker set; may be empty if unknown

defaultCreateNewStickerSet :: CreateNewStickerSet