TD.Query.CreateNewSupergroupChat

Description

Documentation

data CreateNewSupergroupChat

Creates a new supergroup or channel and sends a corresponding messageSupergroupChatCreate. Returns the newly created chat

Constructors

CreateNewSupergroupChat

Fields

for_import :: Maybe Bool

Pass true to create a supergroup for importing messages using importMessages

message_auto_delete_time :: Maybe Int

Message auto-delete time value, in seconds; must be from 0 up to 365 * 86400 and be divisible by 86400. If 0, then messages aren't deleted automatically

location :: Maybe ChatLocation

Chat location if a location-based supergroup is being created; pass null to create an ordinary supergroup chat

description :: Maybe String

is_channel :: Maybe Bool

Pass true to create a channel chat; ignored if a forum is created

is_forum :: Maybe Bool

Pass true to create a forum supergroup chat

title :: Maybe String

Title of the new chat; 1-128 characters