TD.Query.CreatePrivateChat

Description

Documentation

data CreatePrivateChat

Returns an existing chat corresponding to a given user user_id User identifier force Pass true to create the chat without a network request. In this case all information about the chat except its type, title and photo can be incorrect

Constructors

CreatePrivateChat

Fields

force :: Maybe Bool

user_id :: Maybe Int