TD.Query.CreateVideoChat

Description

Documentation

data CreateVideoChat

Creates a video chat (a group call bound to a chat). Available only for basic groups, supergroups and channels; requires can_manage_video_chats rights

Constructors

CreateVideoChat

Fields

is_rtmp_stream :: Maybe Bool

Pass true to create an RTMP stream instead of an ordinary video chat; requires creator privileges

start_date :: Maybe Int

Point in time (Unix timestamp) when the group call is supposed to be started by an administrator; 0 to start the video chat immediately. The date must be at least 10 seconds and at most 8 days in the future

title :: Maybe String

Group call title; if empty, chat title will be used

chat_id :: Maybe Int

Identifier of a chat in which the video chat will be created