TD.Query.DeleteSavedMessagesTopicMessagesByDate

Documentation

data DeleteSavedMessagesTopicMessagesByDate

Deletes all messages between the specified dates in a Saved Messages topic. Messages sent in the last 30 seconds will not be deleted. Returns Ok

Constructors

DeleteSavedMessagesTopicMessagesByDate

Fields

saved_messages_topic_id :: Maybe Int

Identifier of Saved Messages topic which messages will be deleted

min_date :: Maybe Int

The minimum date of the messages to delete

max_date :: Maybe Int

The maximum date of the messages to delete

defaultDeleteSavedMessagesTopicMessagesByDate :: DeleteSavedMessagesTopicMessagesByDate