TD.Query.EditBotMediaPreview

Documentation

data EditBotMediaPreview

Replaces media preview in the list of media previews of a bot. Returns the new preview after edit is completed server-side. Returns BotMediaPreview

Constructors

EditBotMediaPreview

Fields

bot_user_id :: Maybe Int

Identifier of the target bot. The bot must be owned and must have the main Web App

language_code :: Maybe Text

Language code of the media preview to edit

file_id :: Maybe Int

File identifier of the media to replace

content :: Maybe InputStoryContent

Content of the new preview

defaultEditBotMediaPreview :: EditBotMediaPreview