TD.Query.EditBusinessMessageLiveLocation

Documentation

data EditBusinessMessageLiveLocation

Edits the content of a live location in a message sent on behalf of a business account; for bots only. Returns BusinessMessage

Constructors

EditBusinessMessageLiveLocation

Fields

business_connection_id :: Maybe Text

Unique identifier of business connection on behalf of which the message was sent

chat_id :: Maybe Int

The chat the message belongs to

message_id :: Maybe Int

Identifier of the message

reply_markup :: Maybe ReplyMarkup

The new message reply markup; pass null if none

location :: Maybe Location

New location content of the message; pass null to stop sharing the live location

live_period :: Maybe Int

New time relative to the message send date, for which the location can be updated, in seconds. If 0x7FFFFFFF specified, then the location can be updated forever. Otherwise, must not exceed the current live_period by more than a day, and the live location expiration date must remain in the next 90 days. Pass 0 to keep the current live_period

heading :: Maybe Int

The new direction in which the location moves, in degrees; 1-360. Pass 0 if unknown

proximity_alert_radius :: Maybe Int

The new maximum distance for proximity alerts, in meters (0-100000). Pass 0 if the notification is disabled

defaultEditBusinessMessageLiveLocation :: EditBusinessMessageLiveLocation