TD.Query.EditBusinessStory

Documentation

data EditBusinessStory

Changes a story posted by the bot on behalf of a business account; for bots only. Returns Story

Constructors

EditBusinessStory

Fields

story_poster_chat_id :: Maybe Int

Identifier of the chat that posted the story

story_id :: Maybe Int

Identifier of the story to edit

content :: Maybe InputStoryContent

New content of the story

areas :: Maybe InputStoryAreas

New clickable rectangle areas to be shown on the story media

caption :: Maybe FormattedText

New story caption

privacy_settings :: Maybe StoryPrivacySettings

The new privacy settings for the story

defaultEditBusinessStory :: EditBusinessStory