TD.Query.EditChatInviteLink

Documentation

data EditChatInviteLink

Edits a non-primary invite link for a chat. Available for basic groups, supergroups, and channels. Requires administrator privileges and can_invite_users right in the chat for own links and owner privileges for other links. Returns ChatInviteLink

Constructors

EditChatInviteLink

Fields

chat_id :: Maybe Int

Chat identifier

invite_link :: Maybe Text

Invite link to be edited

name :: Maybe Text

Invite link name; 0-32 characters

expiration_date :: Maybe Int

Point in time (Unix timestamp) when the link will expire; pass 0 if never

member_limit :: Maybe Int

The maximum number of chat members that can join the chat via the link simultaneously; 0-99999; pass 0 if not limited

creates_join_request :: Maybe Bool

Pass true if users joining the chat via the link need to be approved by chat administrators. In this case, member_limit must be 0

defaultEditChatInviteLink :: EditChatInviteLink