TD.Query.EditForumTopic

Description

Documentation

data EditForumTopic

Edits title and icon of a topic in a forum supergroup chat; requires can_manage_topics administrator right in the supergroup unless the user is creator of the topic

Constructors

EditForumTopic

Fields

icon_custom_emoji_id :: Maybe Int

Identifier of the new custom emoji for topic icon; pass 0 to remove the custom emoji. Ignored if edit_icon_custom_emoji is false. Telegram Premium users can use any custom emoji, other users can use only a custom emoji returned by getForumTopicDefaultIcons

edit_icon_custom_emoji :: Maybe Bool

Pass true to edit the icon of the topic. Icon of the General topic can't be edited

name :: Maybe String

New name of the topic; 0-128 characters. If empty, the previous topic name is kept

message_thread_id :: Maybe Int

Message thread identifier of the forum topic

chat_id :: Maybe Int

Identifier of the chat