TD.Query.EditInlineMessageMedia

Description

Documentation

data EditInlineMessageMedia

Edits the content of a message with an animation, an audio, a document, a photo or a video in an inline message sent via a bot; for bots only

Constructors

EditInlineMessageMedia

Fields

input_message_content :: Maybe InputMessageContent

New content of the message. Must be one of the following types: inputMessageAnimation, inputMessageAudio, inputMessageDocument, inputMessagePhoto or inputMessageVideo

reply_markup :: Maybe ReplyMarkup

The new message reply markup; pass null if none; for bots only

inline_message_id :: Maybe String

Inline message identifier