TD.Query.EditMessageLiveLocation

Description

Documentation

data EditMessageLiveLocation

Edits the message content of a live location. Messages can be edited for a limited period of time specified in the live location. Returns the edited message after the edit is completed on the server side

Constructors

EditMessageLiveLocation

Fields

proximity_alert_radius :: Maybe Int

The new maximum distance for proximity alerts, in meters (0-100000). Pass 0 if the notification is disabled

heading :: Maybe Int

The new direction in which the location moves, in degrees; 1-360. Pass 0 if unknown

location :: Maybe Location

New location content of the message; pass null to stop sharing the live location

reply_markup :: Maybe ReplyMarkup

The new message reply markup; pass null if none; for bots only

message_id :: Maybe Int

Identifier of the message

chat_id :: Maybe Int

The chat the message belongs to