TD.Query.EditQuickReplyMessage

Documentation

data EditQuickReplyMessage

Asynchronously edits the text, media or caption of a quick reply message. Use quickReplyMessage.can_be_edited to check whether a message can be edited. Media message can be edited only to a media message. The type of message content in an album can't be changed with exception of replacing a photo with a video or vice versa. Returns Ok

Constructors

EditQuickReplyMessage

Fields

shortcut_id :: Maybe Int

Unique identifier of the quick reply shortcut with the message

message_id :: Maybe Int

Identifier of the message

input_message_content :: Maybe InputMessageContent

New content of the message. Must be one of the following types: inputMessageText, inputMessageAnimation, inputMessageAudio, inputMessageDocument, inputMessagePhoto or inputMessageVideo

defaultEditQuickReplyMessage :: EditQuickReplyMessage