TD.Query.EncryptGroupCallData

Documentation

data EncryptGroupCallData

Encrypts group call data before sending them over network using tgcalls. Returns Data

Constructors

EncryptGroupCallData

Fields

group_call_id :: Maybe Int

Group call identifier. The call must not be a video chat

data_channel :: Maybe GroupCallDataChannel

Data channel for which data is encrypted

_data :: Maybe ByteString

Data to encrypt

unencrypted_prefix_size :: Maybe Int

Size of data prefix that must be kept unencrypted

defaultEncryptGroupCallData :: EncryptGroupCallData