TD.Query.ForwardMessages

Description

Documentation

data ForwardMessages

Forwards previously sent messages. Returns the forwarded messages in the same order as the message identifiers passed in message_ids. If a message can't be forwarded, null will be returned instead of the message

Constructors

ForwardMessages

Fields

only_preview :: Maybe Bool

Pass true to get fake messages instead of actually forwarding them

remove_caption :: Maybe Bool

Pass true to remove media captions of message copies. Ignored if send_copy is false

send_copy :: Maybe Bool

Pass true to copy content of the messages without reference to the original sender. Always true if the messages are forwarded to a secret chat or are local

options :: Maybe MessageSendOptions

Options to be used to send the messages; pass null to use default options

message_ids :: Maybe [Int]

Identifiers of the messages to forward. Message identifiers must be in a strictly increasing order. At most 100 messages can be forwarded simultaneously

from_chat_id :: Maybe Int

Identifier of the chat from which to forward messages

chat_id :: Maybe Int

Identifier of the chat to which to forward messages