TD.Query.GetBlockedMessageSenders

Description

Documentation

data GetBlockedMessageSenders

Returns users and chats that were blocked by the current user offset Number of users and chats to skip in the result; must be non-negative limit The maximum number of users and chats to return; up to 100

Constructors

GetBlockedMessageSenders

Fields

limit :: Maybe Int

offset :: Maybe Int