TD.Query.GetBotMediaPreviewInfo

Documentation

data GetBotMediaPreviewInfo

Returns the list of media previews for the given language and the list of languages for which the bot has dedicated previews. Returns BotMediaPreviewInfo

Constructors

GetBotMediaPreviewInfo

Fields

bot_user_id :: Maybe Int

Identifier of the target bot. The bot must be owned and must have the main Web App

language_code :: Maybe Text

A two-letter ISO 639-1 language code for which to get previews. If empty, then default previews are returned

defaultGetBotMediaPreviewInfo :: GetBotMediaPreviewInfo