TD.Query.GetChatHistory

Description

Documentation

data GetChatHistory

Returns messages in a chat. The messages are returned in a reverse chronological order (i.e., in order of decreasing message_id). For optimal performance, the number of returned messages is chosen by TDLib. This is an offline request if only_local is true

Constructors

GetChatHistory

Fields

only_local :: Maybe Bool

Pass true to get only messages that are available without sending network requests

limit :: Maybe Int

The maximum number of messages to be returned; must be positive and can't be greater than 100. If the offset is negative, the limit must be greater than or equal to -offset. For optimal performance, the number of returned messages is chosen by TDLib and can be smaller than the specified limit

offset :: Maybe Int

Specify 0 to get results from exactly the from_message_id or a negative offset up to 99 to get additionally some newer messages

from_message_id :: Maybe Int

Identifier of the message starting from which history must be fetched; use 0 to get results from the last message

chat_id :: Maybe Int

Chat identifier