TD.Query.GetChatInviteLinkMembers

Description

Documentation

data GetChatInviteLinkMembers

Returns chat members joined a chat via an invite link. Requires administrator privileges and can_invite_users right in the chat for own links and owner privileges for other links chat_id Chat identifier invite_link Invite link for which to return chat members

Constructors

GetChatInviteLinkMembers

Fields

limit :: Maybe Int

offset_member :: Maybe ChatInviteLinkMember

A chat member from which to return next chat members; pass null to get results from the beginning @limit The maximum number of chat members to return; up to 100

invite_link :: Maybe String

chat_id :: Maybe Int