TD.Query.GetChatMessagePosition

Description

Documentation

data GetChatMessagePosition

Returns approximate 1-based position of a message among messages, which can be found by the specified filter in the chat. Cannot be used in secret chats

Constructors

GetChatMessagePosition

Fields

message_thread_id :: Maybe Int

If not 0, only messages in the specified thread will be considered; supergroups only

_filter :: Maybe SearchMessagesFilter

Filter for message content; searchMessagesFilterEmpty, searchMessagesFilterUnreadMention, searchMessagesFilterUnreadReaction, and searchMessagesFilterFailedToSend are unsupported in this function

message_id :: Maybe Int

Message identifier

chat_id :: Maybe Int

Identifier of the chat in which to find message position