TD.Query.GetForumTopics

Description

Documentation

data GetForumTopics

Returns found forum topics in a forum chat. This is a temporary method for getting information about topic list from the server

Constructors

GetForumTopics

Fields

limit :: Maybe Int

The maximum number of forum topics to be returned; up to 100. For optimal performance, the number of returned forum topics is chosen by TDLib and can be smaller than the specified limit

offset_message_thread_id :: Maybe Int

The message thread identifier of the last found topic, or 0 for the first request

offset_message_id :: Maybe Int

The message identifier of the last message in the last found topic, or 0 for the first request

offset_date :: Maybe Int

The date starting from which the results need to be fetched. Use 0 or any date in the future to get results from the last topic

query :: Maybe String

Query to search for in the forum topic's name

chat_id :: Maybe Int

Identifier of the forum chat