TD.Query.GetGuardBotWebAppUrl

Documentation

data GetGuardBotWebAppUrl

Returns an HTTPS URL of a Web App of a guard bot to open after receiving chatJoinResultGuardBotApprovalRequired. Returns WebAppUrl

Constructors

GetGuardBotWebAppUrl

Fields

query_id :: Maybe Int

Unique identifier of the join request as received in chatJoinResultGuardBotApprovalRequired

parameters :: Maybe WebAppOpenParameters

Parameters to use to open the Web App

defaultGetGuardBotWebAppUrl :: GetGuardBotWebAppUrl