TD.Query.GetLanguagePackString

Description

Documentation

data GetLanguagePackString

Returns a string stored in the local database from the specified localization target and language pack by its key. Returns a 404 error if the string is not found. Can be called synchronously

Constructors

GetLanguagePackString

Fields

key :: Maybe String

Language pack key of the string to be returned

language_pack_id :: Maybe String

Language pack identifier

localization_target :: Maybe String

Localization target to which the language pack belongs

language_pack_database_path :: Maybe String

Path to the language pack database in which strings are stored