TD.Query.GetLinkPreview

Documentation

data GetLinkPreview

Returns a link preview by the text of a message. Do not call this function too often. Returns a 404 error if the text has no link preview. Returns LinkPreview

Constructors

GetLinkPreview

Fields

text :: Maybe FormattedText

Message text with formatting

link_preview_options :: Maybe LinkPreviewOptions

Options to be used for generation of the link preview; pass null to use default link preview options

defaultGetLinkPreview :: GetLinkPreview