TD.Query.GetMainWebApp

Documentation

data GetMainWebApp

Returns information needed to open the main Web App of a bot. Returns MainWebApp

Constructors

GetMainWebApp

Fields

chat_id :: Maybe Int

Identifier of the chat in which the Web App is opened; pass 0 if none

bot_user_id :: Maybe Int

Identifier of the target bot

start_parameter :: Maybe Text

Start parameter from internalLinkTypeMainWebApp

theme :: Maybe ThemeParameters

Preferred Web App theme; pass null to use the default theme

application_name :: Maybe Text

Short name of the current application; 0-64 English letters, digits, and underscores

defaultGetMainWebApp :: GetMainWebApp