TD.Query.GetMapThumbnailFile

Documentation

data GetMapThumbnailFile

Returns information about a file with a map thumbnail in PNG format. Only map thumbnail files with size less than 1MB can be downloaded location Location of the map center zoom Map zoom level; 13-20 width Map width in pixels before applying scale; 16-1024 height Map height in pixels before applying scale; 16-1024 scale Map scale; 1-3 chat_id Identifier of a chat in which the thumbnail will be shown. Use 0 if unknown

Constructors

GetMapThumbnailFile

Fields

chat_id :: Maybe Int

scale :: Maybe Int

height :: Maybe Int

width :: Maybe Int

zoom :: Maybe Int

location :: Maybe Location