TD.Query.GetMessageLink

Description

Documentation

data GetMessageLink

Returns an HTTPS link to a message in a chat. Available only for already sent messages in supergroups and channels, or if message.can_get_media_timestamp_links and a media timestamp link is generated. This is an offline request

Constructors

GetMessageLink

Fields

for_comment :: Maybe Bool

Pass true to create a link to the message as a channel post comment, or from a message thread

for_album :: Maybe Bool

Pass true to create a link for the whole media album

media_timestamp :: Maybe Int

If not 0, timestamp from which the videoaudiovideo note/voice note playing must start, in seconds. The media can be in the message content or in its web page preview

message_id :: Maybe Int

Identifier of the message

chat_id :: Maybe Int

Identifier of the chat to which the message belongs