TD.Query.GetPhoneNumberInfoSync

Description

Documentation

data GetPhoneNumberInfoSync

Returns information about a phone number by its prefix synchronously. getCountries must be called at least once after changing localization to the specified language if properly localized country information is expected. Can be called synchronously

Constructors

GetPhoneNumberInfoSync

Fields

phone_number_prefix :: Maybe String

language_code :: Maybe String

A two-letter ISO 639-1 language code for country information localization @phone_number_prefix The phone number prefix