TD.Query.GetReceivedGifts

Documentation

data GetReceivedGifts

Returns gifts received by the given user or chat. Returns ReceivedGifts

Constructors

GetReceivedGifts

Fields

business_connection_id :: Maybe Text

Unique identifier of business connection on behalf of which to send the request; for bots only

owner_id :: Maybe MessageSender

Identifier of the gift receiver

exclude_unsaved :: Maybe Bool

Pass true to exclude gifts that aren't saved to the chat's profile page. Always true for gifts received by other users and channel chats without can_post_messages administrator right

exclude_saved :: Maybe Bool

Pass true to exclude gifts that are saved to the chat's profile page. Always false for gifts received by other users and channel chats without can_post_messages administrator right

exclude_unlimited :: Maybe Bool

Pass true to exclude gifts that can be purchased unlimited number of times

exclude_limited :: Maybe Bool

Pass true to exclude gifts that can be purchased limited number of times

exclude_upgraded :: Maybe Bool

Pass true to exclude upgraded gifts

sort_by_price :: Maybe Bool

Pass true to sort results by gift price instead of send date

offset :: Maybe Text

Offset of the first entry to return as received from the previous request; use empty string to get the first chunk of results

limit :: Maybe Int

The maximum number of gifts to be returned; must be positive and can't be greater than 100. For optimal performance, the number of returned objects is chosen by TDLib and can be smaller than the specified limit

defaultGetReceivedGifts :: GetReceivedGifts