TD.Query.GetStarRevenueStatistics

Documentation

data GetStarRevenueStatistics

Returns detailed Telegram Star revenue statistics. Returns StarRevenueStatistics

Constructors

GetStarRevenueStatistics

Fields

owner_id :: Maybe MessageSender

Identifier of the owner of the Telegram Stars; can be identifier of the current user, an owned bot, or a supergroup or a channel chat with supergroupFullInfo.can_get_star_revenue_statistics == true

is_dark :: Maybe Bool

Pass true if a dark theme is used by the application

defaultGetStarRevenueStatistics :: GetStarRevenueStatistics