TD.Query.GetStarTransactions

Documentation

data GetStarTransactions

Returns the list of Telegram Star transactions for the specified owner. Returns StarTransactions

Constructors

GetStarTransactions

Fields

owner_id :: Maybe MessageSender

Identifier of the owner of the Telegram Stars; can be the identifier of the current user, identifier of an owned bot, or identifier of a supergroup or a channel chat with supergroupFullInfo.can_get_star_revenue_statistics == true

subscription_id :: Maybe Text

If non-empty, only transactions related to the Star Subscription will be returned

direction :: Maybe TransactionDirection

Direction of the transactions to receive; pass null to get all transactions

offset :: Maybe Text

Offset of the first transaction to return as received from the previous request; use empty string to get the first chunk of results

limit :: Maybe Int

The maximum number of transactions to return

defaultGetStarTransactions :: GetStarTransactions