TD.Query.GetWebAppLinkUrl

Description

Documentation

data GetWebAppLinkUrl

Returns an HTTPS URL of a Web App to open after a link of the type internalLinkTypeWebApp is clicked

Constructors

GetWebAppLinkUrl

Fields

allow_write_access :: Maybe Bool

Pass true if the current user allowed the bot to send them messages

application_name :: Maybe String

Short name of the application; 0-64 English letters, digits, and underscores

theme :: Maybe ThemeParameters

Preferred Web App theme; pass null to use the default theme

start_parameter :: Maybe String

Start parameter from internalLinkTypeWebApp

web_app_short_name :: Maybe String

Short name of the Web App

bot_user_id :: Maybe Int

Identifier of the target bot

chat_id :: Maybe Int

Identifier of the chat in which the link was clicked; pass 0 if none