TD.Query.GiftPremiumWithStars

Documentation

data GiftPremiumWithStars

Allows to buy a Telegram Premium subscription for another user with payment in Telegram Stars; for bots only. Returns Ok

Constructors

GiftPremiumWithStars

Fields

user_id :: Maybe Int

Identifier of the user which will receive Telegram Premium

star_count :: Maybe Int

The number of Telegram Stars to pay for subscription

month_count :: Maybe Int

Number of months the Telegram Premium subscription will be active for the user

text :: Maybe FormattedText

Text to show to the user receiving Telegram Premium; 0-getOption("gift_text_length_max") characters. Only Bold, Italic, Underline, Strikethrough, Spoiler, and CustomEmoji entities are allowed

defaultGiftPremiumWithStars :: GiftPremiumWithStars