TD.Query.InviteGroupCallParticipant

Documentation

data InviteGroupCallParticipant

Invites a user to an active group call; for group calls not bound to a chat only. Sends a service message of the type messageGroupCall. The group call can have at most getOption("group_call_participant_count_max") participants. Returns InviteGroupCallParticipantResult

Constructors

InviteGroupCallParticipant

Fields

group_call_id :: Maybe Int

Group call identifier

user_id :: Maybe Int

User identifier

is_video :: Maybe Bool

Pass true if the group call is a video call

defaultInviteGroupCallParticipant :: InviteGroupCallParticipant