TD.Query.OpenWebApp

Description

Documentation

data OpenWebApp

Informs TDLib that a Web App is being opened from attachment menu, a botMenuButton button, an internalLinkTypeAttachmentMenuBot link, or an inlineKeyboardButtonTypeWebApp button. For each bot, a confirmation alert about data sent to the bot must be shown once

Constructors

OpenWebApp

Fields

reply_to_message_id :: Maybe Int

Identifier of the replied message for the message sent by the Web App; 0 if none

message_thread_id :: Maybe Int

If not 0, a message thread identifier in which the message will be sent

application_name :: Maybe String

Short name of the application; 0-64 English letters, digits, and underscores

theme :: Maybe ThemeParameters

Preferred Web App theme; pass null to use the default theme

url :: Maybe String

The URL from an inlineKeyboardButtonTypeWebApp button, a botMenuButton button, or an internalLinkTypeAttachmentMenuBot link, or an empty string otherwise

bot_user_id :: Maybe Int

Identifier of the bot, providing the Web App

chat_id :: Maybe Int

Identifier of the chat in which the Web App is opened