TD.Query.ProcessPushNotification

Description

Documentation

data ProcessPushNotification

Handles a push notification. Returns error with code 406 if the push notification is not supported and connection to the server is required to fetch new data. Can be called before authorization

Constructors

ProcessPushNotification

Fields

payload :: Maybe String

JSON-encoded push notification payload with all fields sent by the server, and "google.sent_time" and "google.notification.sound" fields added