TD.Query.ReadFilePart

Description

Documentation

data ReadFilePart

Reads a part of a file from the TDLib file cache and returns read bytes. This method is intended to be used only if the application has no direct access to TDLib's file system, because it is usually slower than a direct read from the file

Constructors

ReadFilePart

Fields

count :: Maybe Int

Number of bytes to read. An error will be returned if there are not enough bytes available in the file from the specified position. Pass 0 to read all available data from the specified position

offset :: Maybe Int

The offset from which to read the file

file_id :: Maybe Int

Identifier of the file. The file must be located in the TDLib file cache