TD.Query.ReaddQuickReplyShortcutMessages

Documentation

data ReaddQuickReplyShortcutMessages

Readds quick reply messages which failed to add. Can be called only for messages for which messageSendingStateFailed.can_retry is true and after specified in messageSendingStateFailed.retry_after time passed. If a message is readded, the corresponding failed to send message is deleted. Returns the sent messages in the same order as the message identifiers passed in message_ids. If a message can't be readded, null will be returned instead of the message. Returns QuickReplyMessages

Constructors

ReaddQuickReplyShortcutMessages

Fields

shortcut_name :: Maybe Text

Name of the target shortcut

message_ids :: Maybe [Int]

Identifiers of the quick reply messages to readd. Message identifiers must be in a strictly increasing order

defaultReaddQuickReplyShortcutMessages :: ReaddQuickReplyShortcutMessages