TD.Query.ReorderGiftCollections

Documentation

data ReorderGiftCollections

Changes order of gift collections. If the collections are owned by a channel chat, then requires can_post_messages administrator right in the channel chat. Returns Ok

Constructors

ReorderGiftCollections

Fields

owner_id :: Maybe MessageSender

Identifier of the user or the channel chat that owns the collection

collection_ids :: Maybe [Int]

New order of gift collections

defaultReorderGiftCollections :: ReorderGiftCollections