TD.Query.SearchAffiliatePrograms

Documentation

data SearchAffiliatePrograms

Searches affiliate programs that can be connected to the given affiliate. Returns FoundAffiliatePrograms

Constructors

SearchAffiliatePrograms

Fields

affiliate :: Maybe AffiliateType

The affiliate for which affiliate programs are searched for

sort_order :: Maybe AffiliateProgramSortOrder

Sort order for the results

offset :: Maybe Text

Offset of the first affiliate program to return as received from the previous request; use empty string to get the first chunk of results

limit :: Maybe Int

The maximum number of affiliate programs to return

defaultSearchAffiliatePrograms :: SearchAffiliatePrograms