TD.Query.SearchChatMessages

Description

Documentation

data SearchChatMessages

Searches for messages with given words in the chat. Returns the results in reverse chronological order, i.e. in order of decreasing message_id. Cannot be used in secret chats with a non-empty query (searchSecretMessages must be used instead), or without an enabled message database. For optimal performance, the number of returned messages is chosen by TDLib and can be smaller than the specified limit. A combination of query, sender_id, filter and message_thread_id search criteria is expected to be supported, only if it is required for Telegram official application implementation

Constructors

SearchChatMessages

Fields

message_thread_id :: Maybe Int

If not 0, only messages in the specified thread will be returned; supergroups only

_filter :: Maybe SearchMessagesFilter

Additional filter for messages to search; pass null to search for all messages

limit :: Maybe Int

The maximum number of messages to be returned; must be positive and can't be greater than 100. If the offset is negative, the limit must be greater than -offset. For optimal performance, the number of returned messages is chosen by TDLib and can be smaller than the specified limit

offset :: Maybe Int

Specify 0 to get results from exactly the from_message_id or a negative offset to get the specified message and some newer messages

from_message_id :: Maybe Int

Identifier of the message starting from which history must be fetched; use 0 to get results from the last message

sender_id :: Maybe MessageSender

Identifier of the sender of messages to search for; pass null to search for messages from any sender. Not supported in secret chats

query :: Maybe String

Query to search for

chat_id :: Maybe Int

Identifier of the chat in which to search messages