TD.Query.SearchChats

Description

Documentation

data SearchChats

Searches for the specified query in the title and username of already known chats, this is an offline request. Returns chats in the order seen in the main chat list

Constructors

SearchChats

Fields

limit :: Maybe Int

The maximum number of chats to be returned

query :: Maybe String

Query to search for. If the query is empty, returns up to 50 recently found chats