TD.Query.SearchSavedMessages

Documentation

data SearchSavedMessages

Searches for messages tagged by the given reaction and with the given words in the Saved Messages chat; for Telegram Premium users only. Returns the results in reverse chronological order, i.e. in order of decreasing message_id For optimal performance, the number of returned messages is chosen by TDLib and can be smaller than the specified limit. Returns FoundChatMessages

Constructors

SearchSavedMessages

Fields

saved_messages_topic_id :: Maybe Int

If not 0, only messages in the specified Saved Messages topic will be considered; pass 0 to consider all messages

tag :: Maybe ReactionType

Tag to search for; pass null to return all suitable messages

query :: Maybe Text

Query to search for

from_message_id :: Maybe Int

Identifier of the message starting from which messages must be fetched; use 0 to get results from the last message

offset :: Maybe Int

Specify 0 to get results from exactly the message from_message_id or a negative offset to get the specified message and some newer messages

limit :: Maybe Int

The maximum number of messages to be returned; must be positive and can't be greater than 100. If the offset is negative, the limit must be greater than -offset. For optimal performance, the number of returned messages is chosen by TDLib and can be smaller than the specified limit

defaultSearchSavedMessages :: SearchSavedMessages