TD.Query.SearchStringsByPrefix

Description

Documentation

data SearchStringsByPrefix

Searches specified query by word prefixes in the provided strings. Returns 0-based positions of strings that matched. Can be called synchronously

Constructors

SearchStringsByPrefix

Fields

return_none_for_empty_query :: Maybe Bool

Pass true to receive no results for an empty query

limit :: Maybe Int

The maximum number of objects to return

query :: Maybe String

Query to search for

strings :: Maybe [String]

The strings to search in for the query