TD.Query.SendBusinessMessage

Documentation

data SendBusinessMessage

Sends a message on behalf of a business account; for bots only. Returns the message after it was sent. Returns BusinessMessage

Constructors

SendBusinessMessage

Fields

business_connection_id :: Maybe Text

Unique identifier of business connection on behalf of which to send the request

chat_id :: Maybe Int

Target chat

reply_to :: Maybe InputMessageReplyTo

Information about the message to be replied; pass null if none

disable_notification :: Maybe Bool

Pass true to disable notification for the message

protect_content :: Maybe Bool

Pass true if the content of the message must be protected from forwarding and saving

effect_id :: Maybe Int

Identifier of the effect to apply to the message

reply_markup :: Maybe ReplyMarkup

Markup for replying to the message; pass null if none

input_message_content :: Maybe InputMessageContent

The content of the message to be sent

defaultSendBusinessMessage :: SendBusinessMessage