TD.Query.SendBusinessMessageAlbum

Documentation

data SendBusinessMessageAlbum

Sends 2-10 messages grouped together into an album on behalf of a business account; for bots only. Currently, only audio, document, photo and video messages can be grouped into an album. Documents and audio files can be only grouped in an album with messages of the same type. Returns sent messages. Returns BusinessMessages

Constructors

SendBusinessMessageAlbum

Fields

business_connection_id :: Maybe Text

Unique identifier of business connection on behalf of which to send the request

chat_id :: Maybe Int

Target chat

reply_to :: Maybe InputMessageReplyTo

Information about the message to be replied; pass null if none

disable_notification :: Maybe Bool

Pass true to disable notification for the message

protect_content :: Maybe Bool

Pass true if the content of the message must be protected from forwarding and saving

effect_id :: Maybe Int

Identifier of the effect to apply to the message

input_message_contents :: Maybe [InputMessageContent]

Contents of messages to be sent. At most 10 messages can be added to an album. All messages must have the same value of show_caption_above_media

defaultSendBusinessMessageAlbum :: SendBusinessMessageAlbum