TD.Query.SendEphemeralMessage

Documentation

data SendEphemeralMessage

Sends an ephemeral message which will be received only by one bot in a chat. Currently, only ephemeral bot commands and replies to bot ephemeral messages can be sent using the method. The message is persistent across application restarts only if the message database is used. Returns the sent message. Returns Message

Constructors

SendEphemeralMessage

Fields

chat_id :: Maybe Int

Target chat

topic_id :: Maybe MessageTopic

Topic in which the message will be sent; pass null if none

receiver_user_id :: Maybe Int

Identifier of the user who will receive the message

callback_query_id :: Maybe Int

Identifier of the callback query which triggered the message; for bots only

replace_callback_query_message :: Maybe Bool

Pass true if the ephemeral message must replace the message from which the callback query originated; for bots only

reply_to :: Maybe InputMessageReplyTo

Information about the message to be replied; pass null if none. The message can be an incoming ephemeral message

protect_content :: Maybe Bool

Pass true if the content of the message must be protected from forwarding and saving; for bots only

sending_id :: Maybe Int

Non-persistent identifier, which will be returned back in messageSendingStatePending object and can be used to match sent messages and corresponding updateNewMessage updates

only_preview :: Maybe Bool

Pass true to get a fake message instead of actually sending them

reply_markup :: Maybe ReplyMarkup

Markup for replying to the message; pass null if none; for bots only

input_message_content :: Maybe InputMessageContent

The content of the message to be sent. Must be one of the following types: inputMessageText, inputMessageAnimation, inputMessageAudio, inputMessageDocument, inputMessagePhoto, inputMessageRichMessage, inputMessageSticker, inputMessageVideo, inputMessageVideoNote, inputMessageVoiceNote, inputMessageLocation, inputMessageVenue, inputMessageContact

defaultSendEphemeralMessage :: SendEphemeralMessage