TD.Query.SendGiftPurchaseOffer

Documentation

data SendGiftPurchaseOffer

Sends an offer to purchase an upgraded gift. Returns Ok

Constructors

SendGiftPurchaseOffer

Fields

owner_id :: Maybe MessageSender

Identifier of the user or the channel chat that currently owns the gift and will receive the offer

gift_name :: Maybe Text

Name of the upgraded gift

price :: Maybe GiftResalePrice

The price that the user agreed to pay for the gift

duration :: Maybe Int

Duration of the offer, in seconds; must be one of 21600, 43200, 86400, 129600, 172800, or 259200. Can also be 120 if Telegram test environment is used

paid_message_star_count :: Maybe Int

The number of Telegram Stars the user agreed to pay additionally for sending of the offer message to the current gift owner; pass userFullInfo.outgoing_paid_message_star_count for users and 0 otherwise

defaultSendGiftPurchaseOffer :: SendGiftPurchaseOffer