TD.Query.SendGroupCallMessage

Documentation

data SendGroupCallMessage

Sends a message to other participants of a group call. Requires groupCall.can_send_messages right. Returns Ok

Constructors

SendGroupCallMessage

Fields

group_call_id :: Maybe Int

Group call identifier

text :: Maybe FormattedText

Text of the message to send; 1-getOption("group_call_message_text_length_max") characters for non-live-stories; see updateGroupCallMessageLevels for live story restrictions, which depends on paid_message_star_count. Can't contain line feeds for live stories

paid_message_star_count :: Maybe Int

The number of Telegram Stars the user agreed to pay to send the message; for live stories only; 0-getOption("paid_group_call_message_star_count_max"). Must be 0 for messages sent to live stories posted by the current user

defaultSendGroupCallMessage :: SendGroupCallMessage