TD.Query.SendInlineQueryResultMessage

Documentation

data SendInlineQueryResultMessage

Sends the result of an inline query as a message. Returns the sent message. Always clears a chat draft message. Returns Message

Constructors

SendInlineQueryResultMessage

Fields

chat_id :: Maybe Int

Target chat

topic_id :: Maybe MessageTopic

Topic in which the message will be sent; pass null if none

reply_to :: Maybe InputMessageReplyTo

Information about the message or story to be replied; pass null if none

options :: Maybe MessageSendOptions

Options to be used to send the message; pass null to use default options

query_id :: Maybe Int

Identifier of the inline query

result_id :: Maybe Text

Identifier of the inline query result

hide_via_bot :: Maybe Bool

Pass true to hide the bot, via which the message is sent. Can be used only for bots getOption("animation_search_bot_username"), getOption("photo_search_bot_username"), and getOption("venue_search_bot_username")

defaultSendInlineQueryResultMessage :: SendInlineQueryResultMessage