TD.Query.SendMessageAlbum

Documentation

data SendMessageAlbum

Sends 2-10 messages grouped together into an album. Currently, only audio, document, photo and video messages can be grouped into an album. Documents and audio files can be only grouped in an album with messages of the same type. Returns sent messages. Returns Messages

Constructors

SendMessageAlbum

Fields

chat_id :: Maybe Int

Target chat

message_thread_id :: Maybe Int

If not 0, the message thread identifier in which the messages will be sent

reply_to :: Maybe InputMessageReplyTo

Information about the message or story to be replied; pass null if none

options :: Maybe MessageSendOptions

Options to be used to send the messages; pass null to use default options

input_message_contents :: Maybe [InputMessageContent]

Contents of messages to be sent. At most 10 messages can be added to an album. All messages must have the same value of show_caption_above_media

defaultSendMessageAlbum :: SendMessageAlbum