TD.Query.SendMessageViewMetrics

Documentation

data SendMessageViewMetrics

Informs TDLib about details of a message view by the user from a chat, a message thread or a forum topic history. The method must be called if the message wasn't seen for more than 300 milliseconds, the viewport was destroyed, or the total view duration exceeded 5 minutes. Returns Ok

Constructors

SendMessageViewMetrics

Fields

chat_id :: Maybe Int

Chat identifier

message_id :: Maybe Int

The identifier of the message being viewed

time_in_view_ms :: Maybe Int

The amount of time the message was seen by at least 1 pixel; in milliseconds

active_time_in_view_ms :: Maybe Int

The amount of time the message was seen by at least 1 pixel within 15 seconds after any action from the user; in milliseconds

height_to_viewport_ratio_per_mille :: Maybe Int

The ratio of the post height to the viewport height in 1/1000 fractions

seen_range_ratio_per_mille :: Maybe Int

The ratio of the viewed post height to the full post height in 1/1000 fractions; 0-1000

defaultSendMessageViewMetrics :: SendMessageViewMetrics