TD.Query.SendPaymentForm

Documentation

data SendPaymentForm

Sends a filled-out payment form to the bot for final verification. Returns PaymentResult

Constructors

SendPaymentForm

Fields

input_invoice :: Maybe InputInvoice

The invoice

payment_form_id :: Maybe Int

Payment form identifier returned by getPaymentForm

order_info_id :: Maybe Text

Identifier returned by validateOrderInfo, or an empty string

shipping_option_id :: Maybe Text

Identifier of a chosen shipping option, if applicable

credentials :: Maybe InputCredentials

The credentials chosen by user for payment; pass null for a payment in Telegram Stars

tip_amount :: Maybe Int

Chosen by the user amount of tip in the smallest units of the currency

defaultSendPaymentForm :: SendPaymentForm