TD.Query.SendQuickReplyShortcutMessages

Documentation

data SendQuickReplyShortcutMessages

Sends messages from a quick reply shortcut. Requires Telegram Business subscription. Can't be used to send paid messages. Returns Messages

Constructors

SendQuickReplyShortcutMessages

Fields

chat_id :: Maybe Int

Identifier of the chat to which to send messages. The chat must be a private chat with a regular user

shortcut_id :: Maybe Int

Unique identifier of the quick reply shortcut

sending_id :: Maybe Int

Non-persistent identifier, which will be returned back in messageSendingStatePending object and can be used to match sent messages and corresponding updateNewMessage updates

defaultSendQuickReplyShortcutMessages :: SendQuickReplyShortcutMessages