TD.Query.SendResoldGift

Documentation

data SendResoldGift

Sends an upgraded gift that is available for resale to another user or channel chat; gifts already owned by the current user must be transferred using transferGift and can't be passed to the method. Returns GiftResaleResult

Constructors

SendResoldGift

Fields

gift_name :: Maybe Text

Name of the upgraded gift to send

owner_id :: Maybe MessageSender

Identifier of the user or the channel chat that will receive the gift

price :: Maybe GiftResalePrice

The price that the user agreed to pay for the gift

defaultSendResoldGift :: SendResoldGift