TD.Query.SendRichMessageDraft

Documentation

data SendRichMessageDraft

Sends a draft for a being generated rich message; for bots only. Returns Ok

Constructors

SendRichMessageDraft

Fields

chat_id :: Maybe Int

Chat identifier

forum_topic_id :: Maybe Int

The forum topic identifier in which the message will be sent; pass 0 if none

draft_id :: Maybe Int

Unique identifier of the draft

can_stop :: Maybe Bool

Pass true to show the user a button to stop further drafts

keep_on_stop :: Maybe Bool

Pass true to keep the current draft when the user stops further generation

message :: Maybe InputRichMessage

Draft of the message; file upload isn't supported

defaultSendRichMessageDraft :: SendRichMessageDraft