TD.Query.SendStory

Documentation

data SendStory

Sends a new story to a chat; requires can_post_stories right for supergroup and channel chats. Returns a temporary story. Returns Story

Constructors

SendStory

Fields

chat_id :: Maybe Int

Identifier of the chat that will post the story. Pass Saved Messages chat identifier when posting a story on behalf of the current user

content :: Maybe InputStoryContent

Content of the story

areas :: Maybe InputStoryAreas

Clickable rectangle areas to be shown on the story media; pass null if none

caption :: Maybe FormattedText

Story caption; pass null to use an empty caption; 0-getOption("story_caption_length_max") characters; can have entities only if getOption("can_use_text_entities_in_story_caption")

privacy_settings :: Maybe StoryPrivacySettings

The privacy settings for the story; ignored for stories sent to supergroup and channel chats

active_period :: Maybe Int

Period after which the story is moved to archive, in seconds; must be one of 6 * 3600, 12 * 3600, 86400, or 2 * 86400 for Telegram Premium users, and 86400 otherwise

from_story_full_id :: Maybe StoryFullId

Full identifier of the original story, which content was used to create the story; pass null if the story isn't repost of another story

is_posted_to_chat_page :: Maybe Bool

Pass true to keep the story accessible after expiration

protect_content :: Maybe Bool

Pass true if the content of the story must be protected from forwarding and screenshotting

defaultSendStory :: SendStory