TD.Query.SetApplicationVerificationToken

Documentation

data SetApplicationVerificationToken

Informs TDLib that application or reCAPTCHA verification has been completed. Can be called before authorization. Returns Ok

Constructors

SetApplicationVerificationToken

Fields

verification_id :: Maybe Int

Unique identifier for the verification process as received from updateApplicationVerificationRequired or updateApplicationRecaptchaVerificationRequired

token :: Maybe Text

Play Integrity API token for the Android application, or secret from push notification for the iOS application for application verification, or reCAPTCHA token for reCAPTCHA verifications; pass an empty string to abort verification and receive the error VERIFICATION_FAILED for the request

defaultSetApplicationVerificationToken :: SetApplicationVerificationToken