TD.Query.SetChatBackground

Documentation

data SetChatBackground

Sets the background in a specific chat. Supported only in private and secret chats with non-deleted users, and in chats with sufficient boost level and can_change_info administrator right. Returns Ok

Constructors

SetChatBackground

Fields

chat_id :: Maybe Int

Chat identifier

background :: Maybe InputBackground

The input background to use; pass null to create a new filled or chat theme background

_type :: Maybe BackgroundType

Background type; pass null to use default background type for the chosen background; backgroundTypeChatTheme isn't supported for private and secret chats. Use chatBoostLevelFeatures.chat_theme_background_count and chatBoostLevelFeatures.can_set_custom_background to check whether the background type can be set in the boosted chat

dark_theme_dimming :: Maybe Int

Dimming of the background in dark themes, as a percentage; 0-100. Applied only to Wallpaper and Fill types of background

only_for_self :: Maybe Bool

Pass true to set background only for self; pass false to set background for all chat users. Always false for backgrounds set in boosted chats. Background can be set for both users only by Telegram Premium users and if set background isn't of the type inputBackgroundPrevious

defaultSetChatBackground :: SetChatBackground