TD.Query.SetChatDirectMessagesGroup

Documentation

data SetChatDirectMessagesGroup

Changes direct messages group settings for a channel chat; requires owner privileges in the chat. Returns Ok

Constructors

SetChatDirectMessagesGroup

Fields

chat_id :: Maybe Int

Identifier of the channel chat

is_enabled :: Maybe Bool

Pass true if the direct messages group is enabled for the channel chat; pass false otherwise

paid_message_star_count :: Maybe Int

The new number of Telegram Stars that must be paid for each message that is sent to the direct messages chat unless the sender is an administrator of the channel chat; 0-getOption("paid_message_star_count_max"). The channel will receive getOption("paid_message_earnings_per_mille") Telegram Stars for each 1000 Telegram Stars paid for message sending. Requires supergroupFullInfo.can_enable_paid_messages for positive amounts

defaultSetChatDirectMessagesGroup :: SetChatDirectMessagesGroup