TD.Query.SetCommands

Description

Documentation

data SetCommands

Sets the list of commands supported by the bot for the given user scope and language; for bots only

Constructors

SetCommands

Fields

commands :: Maybe [BotCommand]

List of the bot's commands

language_code :: Maybe String

A two-letter ISO 639-1 language code. If empty, the commands will be applied to all users from the given scope, for which language there are no dedicated commands

scope :: Maybe BotCommandScope

The scope to which the commands are relevant; pass null to change commands in the default bot command scope