TD.Query.SetDefaultBackground

Documentation

data SetDefaultBackground

Sets default background for chats; adds the background to the list of installed backgrounds. Returns Background

Constructors

SetDefaultBackground

Fields

background :: Maybe InputBackground

The input background to use; pass null to create a new filled background

_type :: Maybe BackgroundType

Background type; pass null to use the default type of the remote background; backgroundTypeChatTheme isn't supported

for_dark_theme :: Maybe Bool

Pass true if the background is set for a dark theme

defaultSetDefaultBackground :: SetDefaultBackground