TD.Query.SetPinnedGifts

Documentation

data SetPinnedGifts

Changes the list of pinned gifts on the current user's or the channel's profile page; requires can_post_messages administrator right in the channel chat. Returns Ok

Constructors

SetPinnedGifts

Fields

owner_id :: Maybe MessageSender

Identifier of the user or the channel chat that received the gifts

received_gift_ids :: Maybe [Text]

New list of pinned gifts. All gifts must be upgraded and saved on the profile page first. There can be up to getOption("pinned_gift_count_max") pinned gifts

defaultSetPinnedGifts :: SetPinnedGifts