TD.Query.SetStoryReaction

Description

Documentation

data SetStoryReaction

Changes chosen reaction on a story

Constructors

SetStoryReaction

Fields

update_recent_reactions :: Maybe Bool

Pass true if the reaction needs to be added to recent reactions

reaction_type :: Maybe ReactionType

Type of the reaction to set; pass null to remove the reaction. reactionTypeCustomEmoji reactions can be used only by Telegram Premium users

story_id :: Maybe Int

The identifier of the story

story_sender_chat_id :: Maybe Int

The identifier of the sender of the story