TD.Query.SetTdlibParameters

Description

Documentation

data SetTdlibParameters

Sets the parameters for TDLib initialization. Works only when the current authorization state is authorizationStateWaitTdlibParameters

Constructors

SetTdlibParameters

Fields

ignore_file_names :: Maybe Bool

Pass true to ignore original file names for downloaded files. Otherwise, downloaded files are saved under names as close as possible to the original name

enable_storage_optimizer :: Maybe Bool

Pass true to automatically delete old files in background

application_version :: Maybe String

Application version; must be non-empty

system_version :: Maybe String

Version of the operating system the application is being run on. If empty, the version is automatically detected by TDLib

device_model :: Maybe String

Model of the device the application is being run on; must be non-empty

system_language_code :: Maybe String

IETF language tag of the user's operating system language; must be non-empty

api_hash :: Maybe String

Application identifier hash for Telegram API access, which can be obtained at https://my.telegram.org

api_id :: Maybe Int

Application identifier for Telegram API access, which can be obtained at https://my.telegram.org

use_secret_chats :: Maybe Bool

Pass true to enable support for secret chats

use_message_database :: Maybe Bool

Pass true to keep cache of chats and messages between restarts. Implies use_chat_info_database

use_chat_info_database :: Maybe Bool

Pass true to keep cache of users, basic groups, supergroups, channels and secret chats between restarts. Implies use_file_database

use_file_database :: Maybe Bool

Pass true to keep information about downloaded and uploaded files between application restarts

database_encryption_key :: Maybe String

Encryption key for the database. If the encryption key is invalid, then an error with code 401 will be returned

files_directory :: Maybe String

The path to the directory for storing files; if empty, database_directory will be used

database_directory :: Maybe String

The path to the directory for the persistent database; if empty, the current working directory will be used

use_test_dc :: Maybe Bool

Pass true to use Telegram test environment instead of the production environment