TD.Query.ShareChatWithBot

Description

Documentation

data ShareChatWithBot

Shares a chat after pressing a keyboardButtonTypeRequestChat button with the bot

Constructors

ShareChatWithBot

Fields

only_check :: Maybe Bool

Pass true to check that the chat can be shared by the button instead of actually sharing it. Doesn't check bot_is_member and bot_administrator_rights restrictions. If the bot must be a member, then all chats from getGroupsInCommon and all chats, where the user can add the bot, are suitable. In the latter case the bot will be automatically added to the chat. If the bot must be an administrator, then all chats, where the bot already has requested rights or can be added to administrators by the user, are suitable. In the latter case the bot will be automatically granted requested rights

shared_chat_id :: Maybe Int

Identifier of the shared chat

button_id :: Maybe Int

Identifier of the button

message_id :: Maybe Int

Identifier of the message with the button

chat_id :: Maybe Int

Identifier of the chat with the bot