TD.Query.StartLiveStory

Documentation

data StartLiveStory

Starts a new live story on behalf of a chat; requires can_post_stories administrator right for channel chats. Returns StartLiveStoryResult

Constructors

StartLiveStory

Fields

chat_id :: Maybe Int

Identifier of the chat that will start the live story. Pass Saved Messages chat identifier when starting a live story on behalf of the current user, or a channel chat identifier

privacy_settings :: Maybe StoryPrivacySettings

The privacy settings for the story; ignored for stories posted on behalf of channel chats

protect_content :: Maybe Bool

Pass true if the content of the story must be protected from screenshotting

is_rtmp_stream :: Maybe Bool

Pass true to create an RTMP stream instead of an ordinary group call

enable_messages :: Maybe Bool

Pass true to allow viewers of the story to send messages

paid_message_star_count :: Maybe Int

The minimum number of Telegram Stars that must be paid by viewers for each sent message to the call; 0-getOption("paid_group_call_message_star_count_max")

defaultStartLiveStory :: StartLiveStory