TD.Query.StopPoll

Description

Documentation

data StopPoll

Stops a poll. A poll in a message can be stopped when the message has can_be_edited flag set

Constructors

StopPoll

Fields

reply_markup :: Maybe ReplyMarkup

The new message reply markup; pass null if none; for bots only

message_id :: Maybe Int

Identifier of the message containing the poll

chat_id :: Maybe Int

Identifier of the chat to which the poll belongs