TD.Query.ToggleChatIsPinned

Description

Documentation

data ToggleChatIsPinned

Changes the pinned state of a chat. There can be up to GetOption("pinned_chat_count_max")GetOption("pinned_archived_chat_count_max") pinned non-secret chats and the same number of secret chats in the mainarchive chat list. The limit can be increased with Telegram Premium

Constructors

ToggleChatIsPinned

Fields

is_pinned :: Maybe Bool

chat_id :: Maybe Int

chat_list :: Maybe ChatList

Chat list in which to change the pinned state of the chat chat_id Chat identifier is_pinned Pass true to pin the chat; pass false to unpin it