TD.Query.ToggleGroupCallMuteNewParticipants

Description

Documentation

data ToggleGroupCallMuteNewParticipants

Toggles whether new participants of a group call can be unmuted only by administrators of the group call. Requires groupCall.can_toggle_mute_new_participants group call flag

Constructors

ToggleGroupCallMuteNewParticipants

Fields

mute_new_participants :: Maybe Bool

group_call_id :: Maybe Int

Group call identifier @mute_new_participants New value of the mute_new_participants setting