TD.Query.ToggleSupergroupUsernameIsActive

Documentation

data ToggleSupergroupUsernameIsActive

Changes active state for a username of a supergroup or channel, requires owner privileges in the supergroup or channel. The editable username can't be disabled. May return an error with a message USERNAMES_ACTIVE_TOO_MUCH if the maximum number of active usernames has been reached. Returns Ok

Constructors

ToggleSupergroupUsernameIsActive

Fields

supergroup_id :: Maybe Int

Identifier of the supergroup or channel

username :: Maybe Text

The username to change

is_active :: Maybe Bool

Pass true to activate the username; pass false to disable it

defaultToggleSupergroupUsernameIsActive :: ToggleSupergroupUsernameIsActive