TD.Query.TranslateText

Description

Documentation

data TranslateText

Translates a text to the given language. Returns a 404 error if the translation can't be performed

Constructors

TranslateText

Fields

to_language_code :: Maybe String

A two-letter ISO 639-1 language code of the language to which the message is translated

from_language_code :: Maybe String

A two-letter ISO 639-1 language code of the language from which the message is translated. If empty, the language will be detected automatically

text :: Maybe String

Text to translate