TD.Reply.InputChatPhoto

Documentation

data InputChatPhoto

Describes a photo to be set as a user profile or chat photo

Constructors

InputChatPhotoPrevious
A previously used profile photo of the current user @chat_photo_id Identifier of the current user's profile photo to reuse

Fields

chat_photo_id :: Maybe Int

InputChatPhotoStatic
A static photo in JPEG format @photo Photo to be set as profile photo. Only inputFileLocal and inputFileGenerated are allowed

Fields

photo :: Maybe InputFile

InputChatPhotoAnimation
An animation in MPEG4 format; must be square, at most 10 seconds long, have width between 160 and 800 and be at most 2MB in size

Fields

main_frame_timestamp :: Maybe Float

Timestamp of the frame, which will be used as static chat photo

animation :: Maybe InputFile

Animation to be set as profile photo. Only inputFileLocal and inputFileGenerated are allowed