API.InputSticker

Documentation

data InputSticker

Describes a sticker that needs to be added to a sticker set

Constructors

InputStickerStatic
A static sticker in PNG format, which will be converted to WEBP server-side

sticker PNG image with the sticker; must be up to 512 KB in size and fit in a 512x512 square

emojis Emojis corresponding to the sticker

mask_position For masks, position where the mask should be placed; may be null

Fields

mask_position :: Maybe MaskPosition

emojis :: Maybe String

sticker :: Maybe InputFile

InputStickerAnimated
An animated sticker in TGS format

sticker File with the animated sticker. Only local or uploaded within a week files are supported. See https://core.telegram.org/animated_stickers#technical-requirements for technical requirements

emojis Emojis corresponding to the sticker

Fields

emojis :: Maybe String

sticker :: Maybe InputFile