API.Message

Documentation

data Message

Describes a message

id Message identifier; unique for the chat to which the message belongs

sender_id Identifier of the sender of the message

chat_id Chat identifier

sending_state The sending state of the message; may be null

scheduling_state The scheduling state of the message; may be null

is_outgoing True, if the message is outgoing

is_pinned True, if the message is pinned

can_be_edited True, if the message can be edited. For live location and poll messages this fields shows whether editMessageLiveLocation or stopPoll can be used with this message by the application

can_be_forwarded True, if the message can be forwarded

can_be_saved True, if content of the message can be saved locally or copied

can_be_deleted_only_for_self True, if the message can be deleted only for the current user while other users will continue to see it

can_be_deleted_for_all_users True, if the message can be deleted for all users

can_get_added_reactions True, if the list of added reactions is available through getMessageAddedReactions

can_get_statistics True, if the message statistics are available through getMessageStatistics

can_get_message_thread True, if information about the message thread is available through getMessageThread

can_get_viewers True, if chat members already viewed the message can be received through getMessageViewers

can_get_media_timestamp_links True, if media timestamp links can be generated for media timestamp entities in the message text, caption or web page description through getMessageLink

has_timestamped_media True, if media timestamp entities refers to a media in this message as opposed to a media in the replied message

is_channel_post True, if the message is a channel post. All messages to channels are channel posts, all other messages are not channel posts

contains_unread_mention True, if the message contains an unread mention for the current user

date Point in time (Unix timestamp) when the message was sent

edit_date Point in time (Unix timestamp) when the message was last edited

forward_info Information about the initial message sender; may be null

interaction_info Information about interactions with the message; may be null

unread_reactions Information about unread reactions added to the message

reply_in_chat_id If non-zero, the identifier of the chat to which the replied message belongs; Currently, only messages in the Replies chat can have different reply_in_chat_id and chat_id

reply_to_message_id If non-zero, the identifier of the message this message is replying to; can be the identifier of a deleted message

message_thread_id If non-zero, the identifier of the message thread the message belongs to; unique within the chat to which the message belongs

ttl For self-destructing messages, the message's TTL (Time To Live), in seconds; 0 if none. TDLib will send updateDeleteMessages or updateMessageContent once the TTL expires

ttl_expires_in Time left before the message expires, in seconds. If the TTL timer isn't started yet, equals to the value of the ttl field

via_bot_user_id If non-zero, the user identifier of the bot through which this message was sent

author_signature For channel posts and anonymous group messages, optional author signature

media_album_id Unique identifier of an album this message belongs to. Only audios, documents, photos and videos can be grouped together in albums

restriction_reason If non-empty, contains a human-readable description of the reason why access to this message must be restricted

content Content of the message

reply_markup Reply markup for the message; may be null

Constructors

Message

Fields

reply_markup :: Maybe ReplyMarkup

content :: Maybe MessageContent

restriction_reason :: Maybe String

media_album_id :: Maybe Int

author_signature :: Maybe String

via_bot_user_id :: Maybe Int

ttl_expires_in :: Maybe Float

ttl :: Maybe Int

message_thread_id :: Maybe Int

reply_to_message_id :: Maybe Int

reply_in_chat_id :: Maybe Int

unread_reactions :: Maybe [UnreadReaction]

interaction_info :: Maybe MessageInteractionInfo

forward_info :: Maybe MessageForwardInfo

edit_date :: Maybe Int

date :: Maybe Int

contains_unread_mention :: Maybe Bool

is_channel_post :: Maybe Bool

has_timestamped_media :: Maybe Bool

can_get_media_timestamp_links :: Maybe Bool

can_get_viewers :: Maybe Bool

can_get_message_thread :: Maybe Bool

can_get_statistics :: Maybe Bool

can_get_added_reactions :: Maybe Bool

can_be_deleted_for_all_users :: Maybe Bool

can_be_deleted_only_for_self :: Maybe Bool

can_be_saved :: Maybe Bool

can_be_forwarded :: Maybe Bool

can_be_edited :: Maybe Bool

is_pinned :: Maybe Bool

is_outgoing :: Maybe Bool

scheduling_state :: Maybe MessageSchedulingState

sending_state :: Maybe MessageSendingState

chat_id :: Maybe Int

sender_id :: Maybe MessageSender

_id :: Maybe Int