TD.Data.AttachmentMenuBot

Description

Documentation

data AttachmentMenuBot

Constructors

AttachmentMenuBot
Represents a bot added to attachment menu

Fields

web_app_placeholder :: Maybe File

Default placeholder for opened Web Apps in SVG format; may be null

icon_color :: Maybe AttachmentMenuBotColor

Color to highlight selected icon of the bot if appropriate; may be null

macos_icon :: Maybe File

Attachment menu icon for the bot in TGS format for the official native macOS app; may be null

android_icon :: Maybe File

Attachment menu icon for the bot in TGS format for the official Android app; may be null

ios_animated_icon :: Maybe File

Attachment menu icon for the bot in TGS format for the official iOS app; may be null

ios_static_icon :: Maybe File

Attachment menu icon for the bot in SVG format for the official iOS app; may be null

default_icon :: Maybe File

Default attachment menu icon for the bot in SVG format; may be null

name_color :: Maybe AttachmentMenuBotColor

Color to highlight selected name of the bot if appropriate; may be null

name :: Maybe String

Name for the bot in attachment menu

supports_settings :: Maybe Bool

True, if the bot supports "settings_button_pressed" event

supports_channel_chats :: Maybe Bool

True, if the bot supports opening from attachment menu in channel chats

supports_group_chats :: Maybe Bool

True, if the bot supports opening from attachment menu in basic group and supergroup chats

supports_bot_chats :: Maybe Bool

True, if the bot supports opening from attachment menu in private chats with other bots

supports_user_chats :: Maybe Bool

True, if the bot supports opening from attachment menu in private chats with ordinary users

supports_self_chat :: Maybe Bool

True, if the bot supports opening from attachment menu in the chat with the bot

bot_user_id :: Maybe Int

User identifier of the bot added to attachment menu