TD.Data.AuthorizationState

Description

Documentation

data AuthorizationState

Represents the current authorization state of the TDLib client

Constructors

AuthorizationStateWaitTdlibParameters
Initializetion parameters are needed. Call setTdlibParameters to provide them

AuthorizationStateWaitPhoneNumber
TDLib needs the user's phone number to authorize. Call setAuthenticationPhoneNumber to provide the phone number, or use requestQrCodeAuthentication, or checkAuthenticationBotToken for other authentication options

AuthorizationStateWaitEmailAddress
TDLib needs the user's email address to authorize. Call setAuthenticationEmailAddress to provide the email address, or directly call checkAuthenticationEmailCode with Apple ID/Google ID token if allowed

Fields

allow_google_id :: Maybe Bool

allow_apple_id :: Maybe Bool

True, if authorization through Apple ID is allowed @allow_google_id True, if authorization through Google ID is allowed

AuthorizationStateWaitEmailCode
TDLib needs the user's authentication code sent to an email address to authorize. Call checkAuthenticationEmailCode to provide the code

Fields

next_phone_number_authorization_date :: Maybe Int

Point in time (Unix timestamp) when the user will be able to authorize with a code sent to the user's phone number; 0 if unknown

code_info :: Maybe EmailAddressAuthenticationCodeInfo

Information about the sent authentication code

allow_google_id :: Maybe Bool

allow_apple_id :: Maybe Bool

True, if authorization through Apple ID is allowed @allow_google_id True, if authorization through Google ID is allowed

AuthorizationStateWaitCode
TDLib needs the user's authentication code to authorize @code_info Information about the authorization code that was sent

Fields

_code_info :: Maybe AuthenticationCodeInfo

AuthorizationStateWaitOtherDeviceConfirmation
The user needs to confirm authorization on another logged in device by scanning a QR code with the provided link @link A tg:// URL for the QR code. The link will be updated frequently

Fields

link :: Maybe String

AuthorizationStateWaitRegistration
The user is unregistered and need to accept terms of service and enter their first name and last name to finish registration @terms_of_service Telegram terms of service

Fields

terms_of_service :: Maybe TermsOfService

AuthorizationStateWaitPassword
The user has been authorized, but needs to enter a 2-step verification password to start using the application password_hint Hint for the password; may be empty has_recovery_email_address True, if a recovery email address has been set up

Fields

recovery_email_address_pattern :: Maybe String

Pattern of the email address to which the recovery email was sent; empty until a recovery email has been sent

has_recovery_email_address :: Maybe Bool

password_hint :: Maybe String

AuthorizationStateReady
The user has been successfully authorized. TDLib is now ready to answer queries

AuthorizationStateLoggingOut
The user is currently logging out

AuthorizationStateClosing
TDLib is closing, all subsequent queries will be answered with the error 500. Note that closing TDLib can take a while. All resources will be freed only after authorizationStateClosed has been received

AuthorizationStateClosed
TDLib client is in its final state. All databases are closed and all resources are released. No other updates will be received after this. All queries will be responded to with error code 500. To continue working, one must create a new instance of the TDLib client