TD.Data.BotInfo

Description

Documentation

data BotInfo

Constructors

BotInfo
Contains information about a bot

Fields

default_channel_administrator_rights :: Maybe ChatAdministratorRights

Default administrator rights for adding the bot to channels; may be null

default_group_administrator_rights :: Maybe ChatAdministratorRights

Default administrator rights for adding the bot to basic group and supergroup chats; may be null

commands :: Maybe [BotCommand]

List of the bot commands

menu_button :: Maybe BotMenuButton

Information about a button to show instead of the bot commands menu button; may be null if ordinary bot commands menu must be shown

animation :: Maybe Animation

Animation shown in the chat with the bot if the chat is empty; may be null

photo :: Maybe Photo

Photo shown in the chat with the bot if the chat is empty; may be null

description :: Maybe String

short_description :: Maybe String

The text that is shown on the bot's profile page and is sent together with the link when users share the bot