TD.Data.CallState

Description

Documentation

data CallState

Describes the current call state

Constructors

CallStatePending
The call is pending, waiting to be accepted by a user is_created True, if the call has already been created by the server is_received True, if the call has already been received by the other party

Fields

is_received :: Maybe Bool

is_created :: Maybe Bool

CallStateExchangingKeys
The call has been answered and encryption keys are being exchanged

CallStateReady
The call is ready to use protocol Call protocols supported by the peer servers List of available call servers config A JSON-encoded call config encryption_key Call encryption key emojis Encryption key emojis fingerprint allow_p2p True, if peer-to-peer connection is allowed by users privacy settings

Fields

allow_p2p :: Maybe Bool

emojis :: Maybe [String]

encryption_key :: Maybe String

config :: Maybe String

servers :: Maybe [CallServer]

protocol :: Maybe CallProtocol

CallStateHangingUp
The call is hanging up after discardCall has been called

CallStateDiscarded
The call has ended successfully reason The reason, why the call has ended need_rating True, if the call rating must be sent to the server need_debug_information True, if the call debug information must be sent to the server need_log True, if the call log must be sent to the server

Fields

need_log :: Maybe Bool

need_debug_information :: Maybe Bool

need_rating :: Maybe Bool

reason :: Maybe CallDiscardReason

CallStateError
The call has ended with an error @error Error. An error with the code 4005000 will be returned if an outgoing call is missed because of an expired timeout

Fields

_error :: Maybe Error