TD.Data.CanBoostChatResult

Description

Documentation

data CanBoostChatResult

Represents result of checking whether the current user can boost the specific chat

Constructors

CanBoostChatResultOk
The chat can be boosted @currently_boosted_chat_id Identifier of the currently boosted chat from which boost will be removed; 0 if none

Fields

currently_boosted_chat_id :: Maybe Int

CanBoostChatResultInvalidChat
The chat can't be boosted

CanBoostChatResultAlreadyBoosted
The chat is already boosted by the user

CanBoostChatResultPremiumNeeded
The user must subscribe to Telegram Premium to be able to boost chats

CanBoostChatResultPremiumSubscriptionNeeded
The user must have Telegram Premium subscription instead of a gifted Telegram Premium

CanBoostChatResultWaitNeeded
The user must wait the specified time before the boost can be moved to another chat @retry_after Time left before the user can boost another chat

Fields

retry_after :: Maybe Int