TD.Data.ChatEventAction

Description

Documentation

data ChatEventAction

Represents a chat event

Constructors

ChatEventMessageEdited
A message was edited old_message The original message before the edit new_message The message after it was edited

Fields

new_message :: Maybe Message

old_message :: Maybe Message

ChatEventMessageDeleted
A message was deleted message Deleted message can_report_anti_spam_false_positive True, if the message deletion can be reported via reportSupergroupAntiSpamFalsePositive

Fields

can_report_anti_spam_false_positive :: Maybe Bool

message :: Maybe Message

ChatEventMessagePinned
A message was pinned @message Pinned message

Fields

message :: Maybe Message

ChatEventMessageUnpinned
A message was unpinned @message Unpinned message

Fields

message :: Maybe Message

ChatEventPollStopped
A poll in a message was stopped @message The message with the poll

Fields

message :: Maybe Message

ChatEventMemberJoined
A new member joined the chat

ChatEventMemberJoinedByInviteLink
A new member joined the chat via an invite link @invite_link Invite link used to join the chat

Fields

invite_link :: Maybe ChatInviteLink

ChatEventMemberJoinedByRequest
A new member was accepted to the chat by an administrator approver_user_id User identifier of the chat administrator, approved user join request invite_link Invite link used to join the chat; may be null

Fields

invite_link :: Maybe ChatInviteLink

approver_user_id :: Maybe Int

ChatEventMemberInvited
A new chat member was invited user_id New member user identifier status New member status

Fields

status :: Maybe ChatMemberStatus

user_id :: Maybe Int

ChatEventMemberLeft
A member left the chat

ChatEventMemberPromoted
A chat member has gained/lost administrator status, or the list of their administrator privileges has changed user_id Affected chat member user identifier old_status Previous status of the chat member @new_status New status of the chat member

Fields

new_status :: Maybe ChatMemberStatus

old_status :: Maybe ChatMemberStatus

user_id :: Maybe Int

ChatEventMemberRestricted
A chat member was restrictedunrestricted or bannedunbanned, or the list of their restrictions has changed member_id Affected chat member identifier old_status Previous status of the chat member @new_status New status of the chat member

Fields

new_status :: Maybe ChatMemberStatus

old_status :: Maybe ChatMemberStatus

member_id :: Maybe MessageSender

ChatEventAvailableReactionsChanged
The chat available reactions were changed old_available_reactions Previous chat available reactions new_available_reactions New chat available reactions

Fields

new_available_reactions :: Maybe ChatAvailableReactions

old_available_reactions :: Maybe ChatAvailableReactions

ChatEventDescriptionChanged
The chat description was changed old_description Previous chat description new_description New chat description

Fields

new_description :: Maybe String

old_description :: Maybe String

ChatEventLinkedChatChanged
The linked chat of a supergroup was changed old_linked_chat_id Previous supergroup linked chat identifier new_linked_chat_id New supergroup linked chat identifier

Fields

new_linked_chat_id :: Maybe Int

old_linked_chat_id :: Maybe Int

ChatEventLocationChanged
The supergroup location was changed old_location Previous location; may be null new_location New location; may be null

Fields

new_location :: Maybe ChatLocation

old_location :: Maybe ChatLocation

ChatEventMessageAutoDeleteTimeChanged
The message auto-delete timer was changed old_message_auto_delete_time Previous value of message_auto_delete_time new_message_auto_delete_time New value of message_auto_delete_time

Fields

new_message_auto_delete_time :: Maybe Int

old_message_auto_delete_time :: Maybe Int

ChatEventPermissionsChanged
The chat permissions was changed old_permissions Previous chat permissions new_permissions New chat permissions

Fields

new_permissions :: Maybe ChatPermissions

old_permissions :: Maybe ChatPermissions

ChatEventPhotoChanged
The chat photo was changed old_photo Previous chat photo value; may be null new_photo New chat photo value; may be null

Fields

new_photo :: Maybe ChatPhoto

old_photo :: Maybe ChatPhoto

ChatEventSlowModeDelayChanged
The slow_mode_delay setting of a supergroup was changed old_slow_mode_delay Previous value of slow_mode_delay, in seconds new_slow_mode_delay New value of slow_mode_delay, in seconds

Fields

new_slow_mode_delay :: Maybe Int

old_slow_mode_delay :: Maybe Int

ChatEventStickerSetChanged
The supergroup sticker set was changed old_sticker_set_id Previous identifier of the chat sticker set; 0 if none new_sticker_set_id New identifier of the chat sticker set; 0 if none

Fields

new_sticker_set_id :: Maybe Int

old_sticker_set_id :: Maybe Int

ChatEventTitleChanged
The chat title was changed old_title Previous chat title new_title New chat title

Fields

new_title :: Maybe String

old_title :: Maybe String

ChatEventUsernameChanged
The chat editable username was changed old_username Previous chat username new_username New chat username

Fields

new_username :: Maybe String

old_username :: Maybe String

ChatEventActiveUsernamesChanged
The chat active usernames were changed old_usernames Previous list of active usernames new_usernames New list of active usernames

Fields

new_usernames :: Maybe [String]

old_usernames :: Maybe [String]

ChatEventHasProtectedContentToggled
The has_protected_content setting of a channel was toggled @has_protected_content New value of has_protected_content

Fields

has_protected_content :: Maybe Bool

ChatEventInvitesToggled
The can_invite_users permission of a supergroup chat was toggled @can_invite_users New value of can_invite_users permission

Fields

can_invite_users :: Maybe Bool

ChatEventIsAllHistoryAvailableToggled
The is_all_history_available setting of a supergroup was toggled @is_all_history_available New value of is_all_history_available

Fields

is_all_history_available :: Maybe Bool

ChatEventHasAggressiveAntiSpamEnabledToggled
The has_aggressive_anti_spam_enabled setting of a supergroup was toggled @has_aggressive_anti_spam_enabled New value of has_aggressive_anti_spam_enabled

Fields

has_aggressive_anti_spam_enabled :: Maybe Bool

ChatEventSignMessagesToggled
The sign_messages setting of a channel was toggled @sign_messages New value of sign_messages

Fields

sign_messages :: Maybe Bool

ChatEventInviteLinkEdited
A chat invite link was edited old_invite_link Previous information about the invite link new_invite_link New information about the invite link

Fields

new_invite_link :: Maybe ChatInviteLink

old_invite_link :: Maybe ChatInviteLink

ChatEventInviteLinkRevoked
A chat invite link was revoked @invite_link The invite link

Fields

invite_link :: Maybe ChatInviteLink

ChatEventInviteLinkDeleted
A revoked chat invite link was deleted @invite_link The invite link

Fields

invite_link :: Maybe ChatInviteLink

ChatEventVideoChatCreated
A video chat was created @group_call_id Identifier of the video chat. The video chat can be received through the method getGroupCall

Fields

group_call_id :: Maybe Int

ChatEventVideoChatEnded
A video chat was ended @group_call_id Identifier of the video chat. The video chat can be received through the method getGroupCall

Fields

group_call_id :: Maybe Int

ChatEventVideoChatMuteNewParticipantsToggled
The mute_new_participants setting of a video chat was toggled @mute_new_participants New value of the mute_new_participants setting

Fields

mute_new_participants :: Maybe Bool

ChatEventVideoChatParticipantIsMutedToggled
A video chat participant was muted or unmuted participant_id Identifier of the affected group call participant is_muted New value of is_muted

Fields

is_muted :: Maybe Bool

participant_id :: Maybe MessageSender

ChatEventVideoChatParticipantVolumeLevelChanged
A video chat participant volume level was changed participant_id Identifier of the affected group call participant volume_level New value of volume_level; 1-20000 in hundreds of percents

Fields

volume_level :: Maybe Int

participant_id :: Maybe MessageSender

ChatEventIsForumToggled
The is_forum setting of a channel was toggled @is_forum New value of is_forum

Fields

is_forum :: Maybe Bool

ChatEventForumTopicCreated
A new forum topic was created @topic_info Information about the topic

Fields

topic_info :: Maybe ForumTopicInfo

ChatEventForumTopicEdited
A forum topic was edited old_topic_info Old information about the topic new_topic_info New information about the topic

Fields

new_topic_info :: Maybe ForumTopicInfo

old_topic_info :: Maybe ForumTopicInfo

ChatEventForumTopicToggleIsClosed
A forum topic was closed or reopened @topic_info New information about the topic

Fields

topic_info :: Maybe ForumTopicInfo

ChatEventForumTopicToggleIsHidden
The General forum topic was hidden or unhidden @topic_info New information about the topic

Fields

topic_info :: Maybe ForumTopicInfo

ChatEventForumTopicDeleted
A forum topic was deleted @topic_info Information about the topic

Fields

topic_info :: Maybe ForumTopicInfo

ChatEventForumTopicPinned
A pinned forum topic was changed old_topic_info Information about the old pinned topic; may be null new_topic_info Information about the new pinned topic; may be null

Fields

new_topic_info :: Maybe ForumTopicInfo

old_topic_info :: Maybe ForumTopicInfo