TD.Data.ChatPermissions

Description

Documentation

data ChatPermissions

Constructors

ChatPermissions
Describes actions that a user is allowed to take in a chat

Fields

can_manage_topics :: Maybe Bool

True, if the user can manage topics

can_pin_messages :: Maybe Bool

True, if the user can pin messages

can_invite_users :: Maybe Bool

True, if the user can invite new users to the chat

can_change_info :: Maybe Bool

True, if the user can change the chat title, photo, and other settings

can_add_web_page_previews :: Maybe Bool

True, if the user may add a web page preview to their messages. Implies can_send_messages permissions

can_send_other_messages :: Maybe Bool

True, if the user can send animations, games, stickers, and dice and use inline bots. Implies can_send_messages permissions

can_send_polls :: Maybe Bool

True, if the user can send polls. Implies can_send_messages permissions

can_send_media_messages :: Maybe Bool

True, if the user can send audio files, documents, photos, videos, video notes, and voice notes. Implies can_send_messages permissions

can_send_messages :: Maybe Bool

True, if the user can send text messages, contacts, locations, and venues