TD.Data.ConnectedWebsite

Description

Documentation

data ConnectedWebsite

Constructors

ConnectedWebsite
Contains information about one website the current user is logged in with Telegram

Fields

location :: Maybe String

Human-readable description of a country and a region from which the user was logged in, based on the IP address

ip_address :: Maybe String

IP address from which the user was logged in, in human-readable format

last_active_date :: Maybe Int

Point in time (Unix timestamp) when obtained authorization was last used

log_in_date :: Maybe Int

Point in time (Unix timestamp) when the user was logged in

platform :: Maybe String

Operating system the browser is running on

browser :: Maybe String

The version of a browser used to log in

bot_user_id :: Maybe Int

User identifier of a bot linked with the website

domain_name :: Maybe String

The domain name of the website

_id :: Maybe Int

Website identifier