TD.Data.InputInvoice

Description

Documentation

data InputInvoice

Describe an invoice to process

Constructors

InputInvoiceMessage
An invoice from a message of the type messageInvoice chat_id Chat identifier of the message message_id Message identifier

Fields

message_id :: Maybe Int

chat_id :: Maybe Int

InputInvoiceName
An invoice from a link of the type internalLinkTypeInvoice @name Name of the invoice

Fields

name :: Maybe String