TD.Data.InputMessageContent

Documentation

data InputMessageContent

The content of a message to send

Constructors

InputMessageText
A text message @text Formatted text to be sent; 1-GetOption("message_text_length_max") characters. Only Bold, Italic, Underline, Strikethrough, Spoiler, Code, Pre, PreCode, TextUrl and MentionName entities are allowed to be specified manually

Fields

clear_draft :: Maybe Bool

disable_web_page_preview :: Maybe Bool

True, if rich web page previews for URLs in the message text must be disabled @clear_draft True, if a chat message draft must be deleted

text :: Maybe FormattedText

InputMessageAnimation
An animation message (GIF-style). animation Animation file to be sent thumbnail Animation thumbnail; pass null to skip thumbnail uploading @added_sticker_file_ids File identifiers of the stickers added to the animation, if applicable

Fields

caption :: Maybe FormattedText

height :: Maybe Int

width :: Maybe Int

duration :: Maybe Int

Duration of the animation, in seconds width Width of the animation; may be replaced by the server height Height of the animation; may be replaced by the server @caption Animation caption; pass null to use an empty caption; 0-GetOption("message_caption_length_max") characters

added_sticker_file_ids :: Maybe [Int]

thumbnail :: Maybe InputThumbnail

animation :: Maybe InputFile

InputMessageAudio
An audio message audio Audio file to be sent album_cover_thumbnail Thumbnail of the cover for the album; pass null to skip thumbnail uploading duration Duration of the audio, in seconds; may be replaced by the server title Title of the audio; 0-64 characters; may be replaced by the server

Fields

caption :: Maybe FormattedText

performer :: Maybe String

Performer of the audio; 0-64 characters, may be replaced by the server @caption Audio caption; pass null to use an empty caption; 0-GetOption("message_caption_length_max") characters

title :: Maybe String

duration :: Maybe Int

Duration of the animation, in seconds width Width of the animation; may be replaced by the server height Height of the animation; may be replaced by the server @caption Animation caption; pass null to use an empty caption; 0-GetOption("message_caption_length_max") characters

album_cover_thumbnail :: Maybe InputThumbnail

audio :: Maybe InputFile

InputMessageDocument
A document message (general file) document Document to be sent thumbnail Document thumbnail; pass null to skip thumbnail uploading disable_content_type_detection If true, automatic file type detection will be disabled and the document will be always sent as file. Always true for files sent to secret chats caption Document caption; pass null to use an empty caption; 0-GetOption("message_caption_length_max") characters

Fields

caption :: Maybe FormattedText

disable_content_type_detection :: Maybe Bool

thumbnail :: Maybe InputThumbnail

document :: Maybe InputFile

InputMessagePhoto
A photo message photo Photo to send thumbnail Photo thumbnail to be sent; pass null to skip thumbnail uploading. The thumbnail is sent to the other party only in secret chats added_sticker_file_ids File identifiers of the stickers added to the photo, if applicable width Photo width height Photo height caption Photo caption; pass null to use an empty caption; 0-GetOption("message_caption_length_max") characters

Fields

ttl :: Maybe Int

Photo TTL (Time To Live), in seconds (0-60). A non-zero TTL can be specified only in private chats

caption :: Maybe FormattedText

height :: Maybe Int

width :: Maybe Int

added_sticker_file_ids :: Maybe [Int]

thumbnail :: Maybe InputThumbnail

photo :: Maybe InputFile

InputMessageSticker
A sticker message sticker Sticker to be sent thumbnail Sticker thumbnail; pass null to skip thumbnail uploading width Sticker width height Sticker height @emoji Emoji used to choose the sticker

Fields

emoji :: Maybe String

height :: Maybe Int

width :: Maybe Int

thumbnail :: Maybe InputThumbnail

sticker :: Maybe InputFile

InputMessageVideo
A video message video Video to be sent thumbnail Video thumbnail; pass null to skip thumbnail uploading @added_sticker_file_ids File identifiers of the stickers added to the video, if applicable

Fields

ttl :: Maybe Int

Photo TTL (Time To Live), in seconds (0-60). A non-zero TTL can be specified only in private chats

caption :: Maybe FormattedText

supports_streaming :: Maybe Bool

height :: Maybe Int

width :: Maybe Int

duration :: Maybe Int

Duration of the animation, in seconds width Width of the animation; may be replaced by the server height Height of the animation; may be replaced by the server @caption Animation caption; pass null to use an empty caption; 0-GetOption("message_caption_length_max") characters

added_sticker_file_ids :: Maybe [Int]

thumbnail :: Maybe InputThumbnail

video :: Maybe InputFile

InputMessageVideoNote
A video note message video_note Video note to be sent thumbnail Video thumbnail; pass null to skip thumbnail uploading duration Duration of the video, in seconds length Video width and height; must be positive and not greater than 640

Fields

_length :: Maybe Int

duration :: Maybe Int

Duration of the animation, in seconds width Width of the animation; may be replaced by the server height Height of the animation; may be replaced by the server @caption Animation caption; pass null to use an empty caption; 0-GetOption("message_caption_length_max") characters

thumbnail :: Maybe InputThumbnail

video_note :: Maybe InputFile

InputMessageVoiceNote
A voice note message voice_note Voice note to be sent duration Duration of the voice note, in seconds waveform Waveform representation of the voice note, in 5-bit format caption Voice note caption; pass null to use an empty caption; 0-GetOption("message_caption_length_max") characters

Fields

caption :: Maybe FormattedText

waveform :: Maybe String

duration :: Maybe Int

Duration of the animation, in seconds width Width of the animation; may be replaced by the server height Height of the animation; may be replaced by the server @caption Animation caption; pass null to use an empty caption; 0-GetOption("message_caption_length_max") characters

voice_note :: Maybe InputFile

InputMessageLocation
A message with a location location Location to be sent live_period Period for which the location can be updated, in seconds; must be between 60 and 86400 for a live location and 0 otherwise

Fields

proximity_alert_radius :: Maybe Int

For live locations, a maximum distance to another chat member for proximity alerts, in meters (0-100000). Pass 0 if the notification is disabled. Can't be enabled in channels and Saved Messages

heading :: Maybe Int

For live locations, a direction in which the location moves, in degrees; 1-360. Pass 0 if unknown

live_period :: Maybe Int

location :: Maybe Location

InputMessageVenue
A message with information about a venue @venue Venue to send

Fields

venue :: Maybe Venue

InputMessageContact
A message containing a user contact @contact Contact to send

Fields

contact :: Maybe Contact

InputMessageDice
A dice message emoji Emoji on which the dice throw animation is based clear_draft True, if the chat message draft must be deleted

Fields

clear_draft :: Maybe Bool

emoji :: Maybe String

InputMessageGame
A message with a game; not supported for channels or secret chats bot_user_id User identifier of the bot that owns the game game_short_name Short name of the game

Fields

game_short_name :: Maybe String

bot_user_id :: Maybe Int

InputMessageInvoice
A message with an invoice; can be used only by bots invoice Invoice title Product title; 1-32 characters @param_description Product description; 0-255 characters

Fields

start_parameter :: Maybe String

Unique invoice bot deep link parameter for the generation of this invoice. If empty, it would be possible to pay directly from forwards of the invoice message

provider_data :: Maybe String

provider_token :: Maybe String

payload :: Maybe String

The invoice payload provider_token Payment provider token provider_data JSON-encoded data about the invoice, which will be shared with the payment provider

photo_height :: Maybe Int

photo_width :: Maybe Int

photo_size :: Maybe Int

photo_url :: Maybe String

Product photo URL; optional photo_size Product photo size photo_width Product photo width @photo_height Product photo height

description :: Maybe String

title :: Maybe String

invoice :: Maybe Invoice

InputMessagePoll
A message with a poll. Polls can't be sent to secret chats. Polls can be sent only to a private chat with a bot question Poll question; 1-255 characters (up to 300 characters for bots) options List of poll answer options, 2-10 strings 1-100 characters each

Fields

is_closed :: Maybe Bool

True, if the poll needs to be sent already closed; for bots only

close_date :: Maybe Int

Point in time (Unix timestamp) when the poll will automatically be closed; for bots only

open_period :: Maybe Int

Amount of time the poll will be active after creation, in seconds; for bots only

_type :: Maybe PollType

is_anonymous :: Maybe Bool

True, if the poll voters are anonymous. Non-anonymous polls can't be sent or forwarded to channels @type Type of the poll

options :: Maybe [String]

question :: Maybe String

InputMessageForwarded
A forwarded message from_chat_id Identifier for the chat this forwarded message came from message_id Identifier of the message to forward

Fields

copy_options :: Maybe MessageCopyOptions

Options to be used to copy content of the message without reference to the original sender; pass null to forward the message as usual

in_game_share :: Maybe Bool

True, if a game message is being shared from a launched game; applies only to game messages

message_id :: Maybe Int

from_chat_id :: Maybe Int