TD.Data.Invoice

Description

Documentation

data Invoice

Constructors

Invoice
Product invoice

Fields

is_flexible :: Maybe Bool

True, if the total price depends on the shipping method

send_email_address_to_provider :: Maybe Bool

True, if the user's email address will be sent to the provider

send_phone_number_to_provider :: Maybe Bool

True, if the user's phone number will be sent to the provider

need_shipping_address :: Maybe Bool

True, if the user's shipping address is needed for payment

need_email_address :: Maybe Bool

True, if the user's email address is needed for payment

need_phone_number :: Maybe Bool

True, if the user's phone number is needed for payment

need_name :: Maybe Bool

True, if the user's name is needed for payment

is_test :: Maybe Bool

True, if the payment is a test payment

recurring_payment_terms_of_service_url :: Maybe String

An HTTP URL with terms of service for recurring payments. If non-empty, the invoice payment will result in recurring payments and the user must accept the terms of service before allowed to pay

suggested_tip_amounts :: Maybe [Int]

Suggested amounts of tip in the smallest units of the currency

max_tip_amount :: Maybe Int

The maximum allowed amount of tip in the smallest units of the currency

price_parts :: Maybe [LabeledPricePart]

A list of objects used to calculate the total price of the product

currency :: Maybe String

ISO 4217 currency code