TD.Data.Message

Description

Documentation

data Message

Constructors

Message
Describes a message

Fields

reply_markup :: Maybe ReplyMarkup

Reply markup for the message; may be null

content :: Maybe MessageContent

Content of the message

restriction_reason :: Maybe String

If non-empty, contains a human-readable description of the reason why access to this message must be restricted

media_album_id :: Maybe Int

Unique identifier of an album this message belongs to. Only audios, documents, photos and videos can be grouped together in albums

author_signature :: Maybe String

For channel posts and anonymous group messages, optional author signature

via_bot_user_id :: Maybe Int

If non-zero, the user identifier of the bot through which this message was sent

ttl_expires_in :: Maybe Float

Time left before the message expires, in seconds. If the TTL timer isn't started yet, equals to the value of the ttl field

ttl :: Maybe Int

For self-destructing messages, the message's TTL (Time To Live), in seconds; 0 if none. TDLib will send updateDeleteMessages or updateMessageContent once the TTL expires

message_thread_id :: Maybe Int

If non-zero, the identifier of the message thread the message belongs to; unique within the chat to which the message belongs

reply_to_message_id :: Maybe Int

If non-zero, the identifier of the message this message is replying to; can be the identifier of a deleted message

reply_in_chat_id :: Maybe Int

If non-zero, the identifier of the chat to which the replied message belongs; Currently, only messages in the Replies chat can have different reply_in_chat_id and chat_id

unread_reactions :: Maybe [UnreadReaction]

Information about unread reactions added to the message

interaction_info :: Maybe MessageInteractionInfo

Information about interactions with the message; may be null

forward_info :: Maybe MessageForwardInfo

Information about the initial message sender; may be null

edit_date :: Maybe Int

Point in time (Unix timestamp) when the message was last edited

date :: Maybe Int

Point in time (Unix timestamp) when the message was sent

contains_unread_mention :: Maybe Bool

True, if the message contains an unread mention for the current user

is_topic_message :: Maybe Bool

True, if the message is a forum topic message

is_channel_post :: Maybe Bool

True, if the message is a channel post. All messages to channels are channel posts, all other messages are not channel posts

has_timestamped_media :: Maybe Bool

True, if media timestamp entities refers to a media in this message as opposed to a media in the replied message

can_report_reactions :: Maybe Bool

True, if reactions on the message can be reported through reportMessageReactions

can_get_media_timestamp_links :: Maybe Bool

True, if media timestamp links can be generated for media timestamp entities in the message text, caption or web page description through getMessageLink

can_get_viewers :: Maybe Bool

True, if chat members already viewed the message can be received through getMessageViewers

can_get_message_thread :: Maybe Bool

True, if information about the message thread is available through getMessageThread and getMessageThreadHistory

can_get_statistics :: Maybe Bool

True, if the message statistics are available through getMessageStatistics

can_get_added_reactions :: Maybe Bool

True, if the list of added reactions is available through getMessageAddedReactions

can_be_deleted_for_all_users :: Maybe Bool

True, if the message can be deleted for all users

can_be_deleted_only_for_self :: Maybe Bool

True, if the message can be deleted only for the current user while other users will continue to see it

can_be_saved :: Maybe Bool

True, if content of the message can be saved locally or copied

can_be_forwarded :: Maybe Bool

True, if the message can be forwarded

can_be_edited :: Maybe Bool

True, if the message can be edited. For live location and poll messages this fields shows whether editMessageLiveLocation or stopPoll can be used with this message by the application

is_pinned :: Maybe Bool

True, if the message is pinned

is_outgoing :: Maybe Bool

True, if the message is outgoing

scheduling_state :: Maybe MessageSchedulingState

The scheduling state of the message; may be null

sending_state :: Maybe MessageSendingState

The sending state of the message; may be null

chat_id :: Maybe Int

Chat identifier

sender_id :: Maybe MessageSender

Identifier of the sender of the message

_id :: Maybe Int

Message identifier; unique for the chat to which the message belongs