TD.Data.MessageExtendedMedia

Description

Documentation

data MessageExtendedMedia

Describes a media, which is attached to an invoice

Constructors

MessageExtendedMediaPreview
The media is hidden until the invoice is paid

Fields

caption :: Maybe FormattedText

Media caption

minithumbnail :: Maybe Minithumbnail

Media minithumbnail; may be null

duration :: Maybe Int

Media duration; 0 if unknown

height :: Maybe Int

Media height; 0 if unknown

width :: Maybe Int

Media width; 0 if unknown

MessageExtendedMediaPhoto
The media is a photo photo The photo caption Photo caption

Fields

caption :: Maybe FormattedText

Media caption

photo :: Maybe Photo

MessageExtendedMediaVideo
The media is a video video The video caption Photo caption

Fields

caption :: Maybe FormattedText

Media caption

video :: Maybe Video

MessageExtendedMediaUnsupported
The media is unuspported @caption Media caption

Fields

caption :: Maybe FormattedText

Media caption