TD.Data.MessageSponsorType

Description

Documentation

data MessageSponsorType

Describes type of a message sponsor

Constructors

MessageSponsorTypeBot
The sponsor is a bot bot_user_id User identifier of the bot link An internal link to be opened when the sponsored message is clicked

Fields

link :: Maybe InternalLinkType

bot_user_id :: Maybe Int

MessageSponsorTypePublicChannel
The sponsor is a public channel chat chat_id Sponsor chat identifier link An internal link to be opened when the sponsored message is clicked; may be null if the sponsor chat needs to be opened instead

Fields

link :: Maybe InternalLinkType

chat_id :: Maybe Int

MessageSponsorTypePrivateChannel
The sponsor is a private channel chat title Title of the chat invite_link Invite link for the channel

Fields

invite_link :: Maybe String

title :: Maybe String

MessageSponsorTypeWebsite
The sponsor is a website url URL of the website name Name of the website

Fields

name :: Maybe String

url :: Maybe String