TD.Data.NotificationType

Description

Documentation

data NotificationType

Contains detailed information about a notification

Constructors

NotificationTypeNewMessage
New message was received message The message show_preview True, if message content must be displayed in notifications

Fields

show_preview :: Maybe Bool

message :: Maybe Message

NotificationTypeNewSecretChat
New secret chat was created

NotificationTypeNewCall
New call was received @call_id Call identifier

Fields

call_id :: Maybe Int

NotificationTypeNewPushMessage
New message was received through a push notification

Fields

content :: Maybe PushMessageContent

Push message content

is_outgoing :: Maybe Bool

True, if the message is outgoing

sender_name :: Maybe String

Name of the sender

sender_id :: Maybe MessageSender

Identifier of the sender of the message. Corresponding user or chat may be inaccessible

message_id :: Maybe Int

The message identifier. The message will not be available in the chat history, but the ID can be used in viewMessages, or as reply_to_message_id