TD.Data.PageBlock

Description

Documentation

data PageBlock

Describes a block of an instant view web page

Constructors

PageBlockTitle
The title of a page @title Title

Fields

title :: Maybe RichText

PageBlockSubtitle
The subtitle of a page @subtitle Subtitle

Fields

subtitle :: Maybe RichText

PageBlockAuthorDate
The author and publishing date of a page author Author publish_date Point in time (Unix timestamp) when the article was published; 0 if unknown

Fields

publish_date :: Maybe Int

author :: Maybe RichText

PageBlockHeader
A header @header Header

Fields

header :: Maybe RichText

PageBlockSubheader
A subheader @subheader Subheader

Fields

subheader :: Maybe RichText

PageBlockKicker
A kicker @kicker Kicker

Fields

kicker :: Maybe RichText

PageBlockParagraph
A text paragraph @text Paragraph text

Fields

text :: Maybe RichText

PageBlockPreformatted
A preformatted text paragraph text Paragraph text language Programming language for which the text needs to be formatted

Fields

language :: Maybe String

text :: Maybe RichText

PageBlockFooter
The footer of a page @footer Footer

Fields

footer :: Maybe RichText

PageBlockDivider
An empty block separating a page

PageBlockAnchor
An invisible anchor on a page, which can be used in a URL to open the page from the specified anchor @name Name of the anchor

Fields

name :: Maybe String

PageBlockList
A list of data blocks @items The items of the list

Fields

items :: Maybe [PageBlockListItem]

PageBlockBlockQuote
A block quote

Fields

credit :: Maybe RichText

Quote credit

text :: Maybe RichText

PageBlockPullQuote
A pull quote

Fields

credit :: Maybe RichText

Quote credit

text :: Maybe RichText

PageBlockAnimation
An animation

Fields

need_autoplay :: Maybe Bool

True, if the animation must be played automatically

caption :: Maybe PageBlockCaption

Animation caption

animation :: Maybe Animation

Animation file; may be null

PageBlockAudio
An audio file

Fields

caption :: Maybe PageBlockCaption

Animation caption

audio :: Maybe Audio

Audio file; may be null

PageBlockPhoto
A photo

Fields

url :: Maybe String

URL that needs to be opened when the photo is clicked

caption :: Maybe PageBlockCaption

Animation caption

photo :: Maybe Photo

Photo file; may be null

PageBlockVideo
A video

Fields

is_looped :: Maybe Bool

True, if the video must be looped

need_autoplay :: Maybe Bool

True, if the animation must be played automatically

caption :: Maybe PageBlockCaption

Animation caption

video :: Maybe Video

Video file; may be null

PageBlockVoiceNote
A voice note

Fields

caption :: Maybe PageBlockCaption

Animation caption

voice_note :: Maybe VoiceNote

Voice note; may be null

PageBlockCover
A page cover

Fields

cover :: Maybe PageBlock

Cover

PageBlockEmbedded
An embedded web page

Fields

allow_scrolling :: Maybe Bool

True, if scrolling needs to be allowed

is_full_width :: Maybe Bool

True, if the block must be full width

caption :: Maybe PageBlockCaption

Animation caption

height :: Maybe Int

Block height; 0 if unknown

width :: Maybe Int

Block width; 0 if unknown

poster_photo :: Maybe Photo

Poster photo, if available; may be null

html :: Maybe String

HTML-markup of the embedded page

url :: Maybe String

URL that needs to be opened when the photo is clicked

PageBlockEmbeddedPost
An embedded post

Fields

caption :: Maybe PageBlockCaption

Animation caption

page_blocks :: Maybe [PageBlock]

Post content

date :: Maybe Int

Point in time (Unix timestamp) when the post was created; 0 if unknown

author_photo :: Maybe Photo

Post author photo; may be null

_author :: Maybe String

Post author

url :: Maybe String

URL that needs to be opened when the photo is clicked

PageBlockCollage
A collage

Fields

caption :: Maybe PageBlockCaption

Animation caption

page_blocks :: Maybe [PageBlock]

Post content

PageBlockSlideshow
A slideshow

Fields

caption :: Maybe PageBlockCaption

Animation caption

page_blocks :: Maybe [PageBlock]

Post content

PageBlockChatLink
A link to a chat

Fields

username :: Maybe String

Chat username by which all other information about the chat can be resolved

_photo :: Maybe ChatPhotoInfo

Chat photo; may be null

_title :: Maybe String

Chat title

PageBlockTable
A table

Fields

is_striped :: Maybe Bool

True, if the table is striped

is_bordered :: Maybe Bool

True, if the table is bordered

cells :: Maybe [[PageBlockTableCell]]

Table cells

_caption :: Maybe RichText

Table caption

PageBlockDetails
A collapsible block

Fields

is_open :: Maybe Bool

True, if the block is open by default

page_blocks :: Maybe [PageBlock]

Post content

header :: Maybe RichText

PageBlockRelatedArticles
Related articles

Fields

articles :: Maybe [PageBlockRelatedArticle]

List of related articles

header :: Maybe RichText

PageBlockMap
A map

Fields

caption :: Maybe PageBlockCaption

Animation caption

height :: Maybe Int

Block height; 0 if unknown

width :: Maybe Int

Block width; 0 if unknown

zoom :: Maybe Int

Map zoom level

location :: Maybe Location

Location of the map center