TD.Data.PremiumSource

Description

Documentation

data PremiumSource

Describes a source from which the Premium features screen is opened

Constructors

PremiumSourceLimitExceeded
A limit was exceeded @limit_type Type of the exceeded limit

Fields

limit_type :: Maybe PremiumLimitType

PremiumSourceFeature
A user tried to use a Premium feature @feature The used feature

Fields

feature :: Maybe PremiumFeature

PremiumSourceLink
A user opened an internal link of the type internalLinkTypePremiumFeatures @referrer The referrer from the link

Fields

referrer :: Maybe String

PremiumSourceSettings
A user opened the Premium features screen from settings