TD.Data.PushMessageContent

Description

Documentation

data PushMessageContent

Contains content of a push message notification

Constructors

PushMessageContentHidden
A general message with hidden content @is_pinned True, if the message is a pinned message with the specified content

Fields

is_pinned :: Maybe Bool

PushMessageContentAnimation
An animation message (GIF-style). animation Message content; may be null caption Animation caption @is_pinned True, if the message is a pinned message with the specified content

Fields

is_pinned :: Maybe Bool

caption :: Maybe String

animation :: Maybe Animation

PushMessageContentAudio
An audio message audio Message content; may be null is_pinned True, if the message is a pinned message with the specified content

Fields

is_pinned :: Maybe Bool

audio :: Maybe Audio

PushMessageContentContact
A message with a user contact name Contact's name is_pinned True, if the message is a pinned message with the specified content

Fields

is_pinned :: Maybe Bool

name :: Maybe String

PushMessageContentContactRegistered
A contact has registered with Telegram

PushMessageContentDocument
A document message (a general file) document Message content; may be null is_pinned True, if the message is a pinned message with the specified content

Fields

is_pinned :: Maybe Bool

document :: Maybe Document

PushMessageContentGame
A message with a game title Game title, empty for pinned game message is_pinned True, if the message is a pinned message with the specified content

Fields

is_pinned :: Maybe Bool

title :: Maybe String

PushMessageContentGameScore
A new high score was achieved in a game title Game title, empty for pinned message score New score, 0 for pinned message @is_pinned True, if the message is a pinned message with the specified content

Fields

is_pinned :: Maybe Bool

score :: Maybe Int

title :: Maybe String

PushMessageContentInvoice
A message with an invoice from a bot price Product price is_pinned True, if the message is a pinned message with the specified content

Fields

is_pinned :: Maybe Bool

price :: Maybe String

PushMessageContentLocation
A message with a location is_live True, if the location is live is_pinned True, if the message is a pinned message with the specified content

Fields

is_pinned :: Maybe Bool

is_live :: Maybe Bool

PushMessageContentPhoto
A photo message

Fields

is_pinned :: Maybe Bool

is_secret :: Maybe Bool

True, if the photo is secret

caption :: Maybe String

photo :: Maybe Photo

Message content; may be null

PushMessageContentPoll
A message with a poll

Fields

is_pinned :: Maybe Bool

is_regular :: Maybe Bool

True, if the poll is regular and not in quiz mode

question :: Maybe String

Poll question

PushMessageContentScreenshotTaken
A screenshot of a message in the chat has been taken

PushMessageContentSticker
A message with a sticker

Fields

is_pinned :: Maybe Bool

emoji :: Maybe String

Emoji corresponding to the sticker; may be empty

sticker :: Maybe Sticker

Message content; may be null

PushMessageContentStory
A message with a story @is_pinned True, if the message is a pinned message with the specified content

Fields

is_pinned :: Maybe Bool

PushMessageContentText
A text message text Message text is_pinned True, if the message is a pinned message with the specified content

Fields

is_pinned :: Maybe Bool

text :: Maybe String

PushMessageContentVideo
A video message

Fields

is_pinned :: Maybe Bool

is_secret :: Maybe Bool

True, if the photo is secret

caption :: Maybe String

video :: Maybe Video

Message content; may be null

PushMessageContentVideoNote
A video note message video_note Message content; may be null is_pinned True, if the message is a pinned message with the specified content

Fields

is_pinned :: Maybe Bool

video_note :: Maybe VideoNote

PushMessageContentVoiceNote
A voice note message voice_note Message content; may be null is_pinned True, if the message is a pinned message with the specified content

Fields

is_pinned :: Maybe Bool

voice_note :: Maybe VoiceNote

PushMessageContentBasicGroupChatCreate
A newly created basic group

PushMessageContentChatAddMembers
New chat members were invited to a group

Fields

is_returned :: Maybe Bool

True, if the user has returned to the group themselves

is_current_user :: Maybe Bool

True, if the current user was added to the group

member_name :: Maybe String

Name of the added member

PushMessageContentChatChangePhoto
A chat photo was edited

PushMessageContentChatChangeTitle
A chat title was edited @title New chat title

Fields

title :: Maybe String

PushMessageContentChatSetBackground
A chat background was edited @is_same True, if the set background is the same as the background of the current user

Fields

is_same :: Maybe Bool

PushMessageContentChatSetTheme
A chat theme was edited @theme_name If non-empty, name of a new theme, set for the chat. Otherwise, the chat theme was reset to the default one

Fields

theme_name :: Maybe String

PushMessageContentChatDeleteMember
A chat member was deleted

Fields

is_left :: Maybe Bool

True, if the user has left the group themselves

is_current_user :: Maybe Bool

True, if the current user was added to the group

member_name :: Maybe String

Name of the added member

PushMessageContentChatJoinByLink
A new member joined the chat via an invite link

PushMessageContentChatJoinByRequest
A new member was accepted to the chat by an administrator

PushMessageContentRecurringPayment
A new recurring payment was made by the current user @amount The paid amount

Fields

amount :: Maybe String

PushMessageContentSuggestProfilePhoto
A profile photo was suggested to the user

PushMessageContentMessageForwards
A forwarded messages @total_count Number of forwarded messages

Fields

total_count :: Maybe Int

PushMessageContentMediaAlbum
A media album

Fields

has_documents :: Maybe Bool

True, if the album has at least one document

has_audios :: Maybe Bool

True, if the album has at least one audio file

has_videos :: Maybe Bool

True, if the album has at least one video file

has_photos :: Maybe Bool

True, if the album has at least one photo

total_count :: Maybe Int