TD.Data.RichText

Description

Documentation

data RichText

Describes a text object inside an instant-view web page

Constructors

RichTextPlain
A plain text @text Text

Fields

text :: Maybe String

RichTextBold
A bold rich text @text Text

Fields

_text :: Maybe RichText

RichTextItalic
An italicized rich text @text Text

Fields

_text :: Maybe RichText

RichTextUnderline
An underlined rich text @text Text

Fields

_text :: Maybe RichText

RichTextStrikethrough
A strikethrough rich text @text Text

Fields

_text :: Maybe RichText

RichTextFixed
A fixed-width rich text @text Text

Fields

_text :: Maybe RichText

RichTextUrl
A rich text URL link text Text url URL @is_cached True, if the URL has cached instant view server-side

Fields

is_cached :: Maybe Bool

url :: Maybe String

_text :: Maybe RichText

RichTextEmailAddress
A rich text email link text Text email_address Email address

Fields

email_address :: Maybe String

_text :: Maybe RichText

RichTextSubscript
A subscript rich text @text Text

Fields

_text :: Maybe RichText

RichTextSuperscript
A superscript rich text @text Text

Fields

_text :: Maybe RichText

RichTextMarked
A marked rich text @text Text

Fields

_text :: Maybe RichText

RichTextPhoneNumber
A rich text phone number text Text phone_number Phone number

Fields

phone_number :: Maybe String

_text :: Maybe RichText

RichTextIcon
A small image inside the text @document The image represented as a document. The image can be in GIF, JPEG or PNG format

Fields

height :: Maybe Int

Height of a bounding box in which the image must be shown; 0 if unknown

width :: Maybe Int

Width of a bounding box in which the image must be shown; 0 if unknown

document :: Maybe Document

RichTextReference
A reference to a richTexts object on the same web page text The text anchor_name The name of a richTextAnchor object, which is the first element of the target richTexts object @url An HTTP URL, opening the reference

Fields

url :: Maybe String

anchor_name :: Maybe String

_text :: Maybe RichText

RichTextAnchor
An anchor @name Anchor name

Fields

name :: Maybe String

RichTextAnchorLink
A link to an anchor on the same web page text The link text anchor_name The anchor name. If the name is empty, the link must bring back to top @url An HTTP URL, opening the anchor

Fields

url :: Maybe String

anchor_name :: Maybe String

_text :: Maybe RichText

RichTexts
A concatenation of rich texts @texts Texts

Fields

texts :: Maybe [RichText]