TD.Data.RichText

Documentation

data RichText

Describes a text object inside an instant-view web page

Constructors

RichTextPlain
A plain text

Fields

text :: Maybe Text

Text

RichTextBold
A bold rich text

Fields

_text :: Maybe RichText

Text

RichTextItalic
An italicized rich text

Fields

_text :: Maybe RichText

Text

RichTextUnderline
An underlined rich text

Fields

_text :: Maybe RichText

Text

RichTextStrikethrough
A strikethrough rich text

Fields

_text :: Maybe RichText

Text

RichTextFixed
A fixed-width rich text

Fields

_text :: Maybe RichText

Text

RichTextUrl
A rich text URL link

Fields

_text :: Maybe RichText

Text

url :: Maybe Text

URL

is_cached :: Maybe Bool

True, if the URL has cached instant view server-side

RichTextEmailAddress
A rich text email link

Fields

_text :: Maybe RichText

Text

email_address :: Maybe Text

Email address

RichTextSubscript
A subscript rich text

Fields

_text :: Maybe RichText

Text

RichTextSuperscript
A superscript rich text

Fields

_text :: Maybe RichText

Text

RichTextMarked
A marked rich text

Fields

_text :: Maybe RichText

Text

RichTextPhoneNumber
A rich text phone number

Fields

_text :: Maybe RichText

Text

phone_number :: Maybe Text

Phone number

RichTextIcon
A small image inside the text

Fields

document :: Maybe Document

The image represented as a document. The image can be in GIF, JPEG or PNG format

width :: Maybe Int

Width of a bounding box in which the image must be shown; 0 if unknown

height :: Maybe Int

Height of a bounding box in which the image must be shown; 0 if unknown

RichTextReference
A reference to a richTexts object on the same web page

Fields

_text :: Maybe RichText

Text

anchor_name :: Maybe Text

The name of a richTextAnchor object, which is the first element of the target richTexts object

url :: Maybe Text

URL

RichTextAnchor
An anchor

Fields

name :: Maybe Text

Anchor name

RichTextAnchorLink
A link to an anchor on the same web page

Fields

_text :: Maybe RichText

Text

anchor_name :: Maybe Text

The name of a richTextAnchor object, which is the first element of the target richTexts object

url :: Maybe Text

URL

RichTexts
A concatenation of rich texts

Fields

texts :: Maybe [RichText]

Texts

defaultRichTextPlain :: RichText

defaultRichTextBold :: RichText

defaultRichTextItalic :: RichText

defaultRichTextUnderline :: RichText

defaultRichTextStrikethrough :: RichText

defaultRichTextFixed :: RichText

defaultRichTextUrl :: RichText

defaultRichTextEmailAddress :: RichText

defaultRichTextSubscript :: RichText

defaultRichTextSuperscript :: RichText

defaultRichTextMarked :: RichText

defaultRichTextPhoneNumber :: RichText

defaultRichTextIcon :: RichText

defaultRichTextReference :: RichText

defaultRichTextAnchor :: RichText

defaultRichTextAnchorLink :: RichText

defaultRichTexts :: RichText