TD.Data.Session

Description

Documentation

data Session

Constructors

Session
Contains information about one session in a Telegram application used by the current user. Sessions must be shown to the user in the returned order

Fields

region :: Maybe String

Region code from which the session was created, based on the IP address

country :: Maybe String

A two-letter country code for the country from which the session was created, based on the IP address

ip :: Maybe String

IP address from which the session was created, in human-readable format

last_active_date :: Maybe Int

Point in time (Unix timestamp) when the session was last used

log_in_date :: Maybe Int

Point in time (Unix timestamp) when the user has logged in

system_version :: Maybe String

Version of the operating system the application has been run or is running on, as provided by the application

platform :: Maybe String

Operating system the application has been run or is running on, as provided by the application

device_model :: Maybe String

Model of the device the application has been run or is running on, as provided by the application

is_official_application :: Maybe Bool

True, if the application is an official application or uses the api_id of an official application

application_version :: Maybe String

The version of the application, as provided by the application

application_name :: Maybe String

Name of the application, as provided by the application

api_id :: Maybe Int

Telegram API identifier, as provided by the application

_type :: Maybe SessionType

Session type based on the system and application version, which can be used to display a corresponding icon

can_accept_calls :: Maybe Bool

True, if incoming calls can be accepted by the session

can_accept_secret_chats :: Maybe Bool

True, if incoming secret chats can be accepted by the session

is_password_pending :: Maybe Bool

True, if a 2-step verification password is needed to complete authorization of the session

is_current :: Maybe Bool

True, if this session is the current session

_id :: Maybe Int

Session identifier