TD.Data.Sticker

Description

Documentation

data Sticker

Constructors

Sticker
Describes a sticker set_id The identifier of the sticker set to which the sticker belongs; 0 if none width Sticker width; as defined by the sender @height Sticker height; as defined by the sender

Fields

sticker :: Maybe File

premium_animation :: Maybe File

is_premium :: Maybe Bool

thumbnail :: Maybe Thumbnail

Sticker thumbnail in WEBP or JPEG format; may be null is_premium True, if only Premium users can use the sticker premium_animation Premium animation of the sticker; may be null @sticker File containing the sticker

outline :: Maybe [ClosedVectorPath]

Sticker's outline represented as a list of closed vector paths; may be empty. The coordinate system origin is in the upper-left corner

custom_emoji_id :: Maybe Int

Identifier of the emoji if the sticker is a custom emoji

mask_position :: Maybe MaskPosition

_type :: Maybe StickerType

format :: Maybe StickerFormat

emoji :: Maybe String

Emoji corresponding to the sticker format Sticker format type Sticker type @mask_position Position where the mask is placed; may be null even the sticker is a mask

height :: Maybe Int

width :: Maybe Int

set_id :: Maybe Int