TD.Data.StickerSet

Description

Documentation

data StickerSet

Constructors

StickerSet
Represents a sticker set

Fields

emojis :: Maybe [Emojis]

A list of emoji corresponding to the stickers in the same order. The list is only for informational purposes, because a sticker is always sent with a fixed emoji from the corresponding Sticker object

stickers :: Maybe [Sticker]

List of stickers in this set

is_viewed :: Maybe Bool

True for already viewed trending sticker sets

sticker_type :: Maybe StickerType

Type of the stickers in the set

sticker_format :: Maybe StickerFormat

Format of the stickers in the set

is_official :: Maybe Bool

True, if the sticker set is official

is_archived :: Maybe Bool

True, if the sticker set has been archived. A sticker set can't be installed and archived simultaneously

is_installed :: Maybe Bool

True, if the sticker set has been installed by the current user

thumbnail_outline :: Maybe [ClosedVectorPath]

Sticker set thumbnail's outline represented as a list of closed vector paths; may be empty. The coordinate system origin is in the upper-left corner

thumbnail :: Maybe Thumbnail

Sticker set thumbnail in WEBP, TGS, or WEBM format with width and height 100; may be null. The file can be downloaded only before the thumbnail is changed

name :: Maybe String

Name of the sticker set

title :: Maybe String

Title of the sticker set

_id :: Maybe Int

Identifier of the sticker set