TD.Data.StorePaymentPurpose

Description

Documentation

data StorePaymentPurpose

Describes a purpose of an in-store payment

Constructors

StorePaymentPurposePremiumSubscription
The user subscribed to Telegram Premium @is_restore Pass true if this is a restore of a Telegram Premium purchase; only for App Store

Fields

is_restore :: Maybe Bool

StorePaymentPurposeGiftedPremium
The user gifted Telegram Premium to another user user_id Identifier of the user for which Premium was gifted currency ISO 4217 currency code of the payment currency @amount Paid amount, in the smallest units of the currency

Fields

amount :: Maybe Int

currency :: Maybe String

user_id :: Maybe Int