TD.Data.Story

Description

Documentation

data Story

Constructors

Story
Represents a story

Fields

caption :: Maybe FormattedText

Caption of the story

content :: Maybe StoryContent

Content of the story

privacy_settings :: Maybe StoryPrivacySettings

Privacy rules affecting story visibility; may be approximate for non-owned stories

interaction_info :: Maybe StoryInteractionInfo

Information about interactions with the story; may be null if the story isn't owned or there were no interactions

has_expired_viewers :: Maybe Bool

True, if users viewed the story can't be received, because the story has expired more than getOption("story_viewers_expiration_delay") seconds ago

can_get_viewers :: Maybe Bool

True, if users viewed the story can be received through getStoryViewers

can_be_replied :: Maybe Bool

True, if the story can be replied in the chat with the story sender

can_be_forwarded :: Maybe Bool

True, if the story can be forwarded as a message. Otherwise, screenshots and saving of the story content must be also forbidden

is_visible_only_for_self :: Maybe Bool

True, if the story is visible only for the current user

is_pinned :: Maybe Bool

True, if the story is saved in the sender's profile and will be available there after expiration

is_edited :: Maybe Bool

True, if the story was edited

is_being_edited :: Maybe Bool

True, if the story is being edited by the current user

date :: Maybe Int

Point in time (Unix timestamp) when the story was published

sender_chat_id :: Maybe Int

Identifier of the chat that posted the story

_id :: Maybe Int

Unique story identifier among stories of the given sender