TD.Data.Supergroup

Description

Documentation

data Supergroup

Constructors

Supergroup
Represents a supergroup or channel with zero or more members (subscribers in the case of channels). From the point of view of the system, a channel is a special kind of a supergroup: only administrators can post and see the list of members, and posts from all administrators use the name and photo of the channel instead of individual names and profile photos. Unlike supergroups, channels can have an unlimited number of subscribers

Fields

is_fake :: Maybe Bool

True, if many users reported this supergroup or channel as a fake account

is_scam :: Maybe Bool

True, if many users reported this supergroup or channel as a scam

restriction_reason :: Maybe String

If non-empty, contains a human-readable description of the reason why access to this supergroup or channel must be restricted

is_verified :: Maybe Bool

True, if the supergroup or channel is verified

is_broadcast_group :: Maybe Bool

True, if the supergroup is a broadcast group, i.e. only administrators can send messages and there is no limit on the number of members

is_channel :: Maybe Bool

True, if the supergroup is a channel

is_slow_mode_enabled :: Maybe Bool

True, if the slow mode is enabled in the supergroup

join_by_request :: Maybe Bool

True, if all users directly joining the supergroup need to be approved by supergroup administrators. Always false for channels and supergroups without username, location, or a linked chat

join_to_send_messages :: Maybe Bool

True, if users need to join the supergroup before they can send messages. Always true for channels and non-discussion supergroups

sign_messages :: Maybe Bool

True, if messages sent to the channel need to contain information about the sender. This field is only applicable to channels

has_location :: Maybe Bool

True, if the supergroup is connected to a location, i.e. the supergroup is a location-based supergroup

has_linked_chat :: Maybe Bool

True, if the channel has a discussion group, or the supergroup is the designated discussion group for a channel

member_count :: Maybe Int

Number of members in the supergroup or channel; 0 if unknown. Currently, it is guaranteed to be known only if the supergroup or channel was received through searchPublicChats, searchChatsNearby, getInactiveSupergroupChats, getSuitableDiscussionChats, getGroupsInCommon, or getUserPrivacySettingRules

status :: Maybe ChatMemberStatus

Status of the current user in the supergroup or channel; custom title will be always empty

date :: Maybe Int

Point in time (Unix timestamp) when the current user joined, or the point in time when the supergroup or channel was created, in case the user is not a member

username :: Maybe String

Username of the supergroup or channel; empty for private supergroups or channels

_id :: Maybe Int

Supergroup or channel identifier