TD.Data.UnconfirmedSession

Description

Documentation

data UnconfirmedSession

Constructors

UnconfirmedSession
Contains information about an unconfirmed session

Fields

location :: Maybe String

A human-readable description of the location from which the session was created, based on the IP address

device_model :: Maybe String

Model of the device that was used for the session creation, as provided by the application

log_in_date :: Maybe Int

Point in time (Unix timestamp) when the user has logged in

_id :: Maybe Int

Session identifier