TD.Data.User

Description

Documentation

data User

Constructors

User
Represents a user

Fields

added_to_attachment_menu :: Maybe Bool

True, if the user added the current bot to attachment menu; only available to bots

language_code :: Maybe String

IETF language tag of the user's language; only available to bots

_type :: Maybe UserType

Type of the user

have_access :: Maybe Bool

If false, the user is inaccessible, and the only information known about the user is inside this class. Identifier of the user can't be passed to any method except GetUser

is_fake :: Maybe Bool

True, if many users reported this user as a fake account

is_scam :: Maybe Bool

True, if many users reported this user as a scam

restriction_reason :: Maybe String

If non-empty, it contains a human-readable description of the reason why access to this user must be restricted

is_support :: Maybe Bool

True, if the user is Telegram support account

is_premium :: Maybe Bool

True, if the user is a Telegram Premium user

is_verified :: Maybe Bool

True, if the user is verified

is_mutual_contact :: Maybe Bool

The user is a contact of the current user and the current user is a contact of the user

is_contact :: Maybe Bool

The user is a contact of the current user

profile_photo :: Maybe ProfilePhoto

Profile photo of the user; may be null

status :: Maybe UserStatus

Current online status of the user

phone_number :: Maybe String

Phone number of the user

username :: Maybe String

Username of the user

last_name :: Maybe String

Last name of the user

first_name :: Maybe String

First name of the user

_id :: Maybe Int

User identifier