TD.Data.UserFullInfo

Description

Documentation

data UserFullInfo

Constructors

UserFullInfo
Contains full information about a user

Fields

bot_info :: Maybe BotInfo

For bots, information about the bot; may be null

group_in_common_count :: Maybe Int

Number of group chats where both the other user and the current user are a member; 0 for the current user

premium_gift_options :: Maybe [PremiumPaymentOption]

The list of available options for gifting Telegram Premium to the user

bio :: Maybe FormattedText

A short user bio; may be null for bots

need_phone_number_privacy_exception :: Maybe Bool

True, if the current user needs to explicitly allow to share their phone number with the user when the method addContact is used

has_restricted_voice_and_video_note_messages :: Maybe Bool

True, if voice and video notes can't be sent or forwarded to the user

has_private_forwards :: Maybe Bool

True, if the user can't be linked in forwarded messages due to their privacy settings

has_private_calls :: Maybe Bool

True, if the user can't be called due to their privacy settings

supports_video_calls :: Maybe Bool

True, if a video call can be created with the user

can_be_called :: Maybe Bool

True, if the user can be called

is_blocked :: Maybe Bool

True, if the user is blocked by the current user

photo :: Maybe ChatPhoto

User profile photo; may be null if empty or unknown. If non-null, then it is the same photo as in user.profile_photo and chat.photo