TD.Data.Video

Description

Documentation

data Video

Constructors

Video
Describes a video file

Fields

video :: Maybe File

File containing the video

thumbnail :: Maybe Thumbnail

Video thumbnail in JPEG or MPEG4 format; as defined by the sender; may be null

minithumbnail :: Maybe Minithumbnail

Video minithumbnail; may be null

supports_streaming :: Maybe Bool

True, if the video is supposed to be streamed

has_stickers :: Maybe Bool

True, if stickers were added to the video. The list of corresponding sticker sets can be received using getAttachedStickerSets

mime_type :: Maybe String

MIME type of the file; as defined by the sender

file_name :: Maybe String

Original name of the file; as defined by the sender

height :: Maybe Int

Video height; as defined by the sender

width :: Maybe Int

Video width; as defined by the sender

duration :: Maybe Int

Duration of the video, in seconds; as defined by the sender