TD.Data.WebPageInstantView

Description

Documentation

data WebPageInstantView

Constructors

WebPageInstantView
Describes an instant view page for a web page

Fields

feedback_link :: Maybe InternalLinkType

An internal link to be opened to leave feedback about the instant view

is_full :: Maybe Bool

True, if the instant view contains the full page. A network request might be needed to get the full web page instant view

is_rtl :: Maybe Bool

True, if the instant view must be shown from right to left

version :: Maybe Int

Version of the instant view; currently, can be 1 or 2

view_count :: Maybe Int

Number of the instant view views; 0 if unknown

page_blocks :: Maybe [PageBlock]

Content of the web page