TD.Query.CreateNewStickerSet

Description

Documentation

data CreateNewStickerSet

Creates a new sticker set. Returns the newly created sticker set

Constructors

CreateNewStickerSet

Fields

source :: Maybe String

Source of the sticker set; may be empty if unknown

stickers :: Maybe [InputSticker]

List of stickers to be added to the set; must be non-empty. All stickers must have the same format. For TGS stickers, uploadStickerFile must be used before the sticker is shown

sticker_type :: Maybe StickerType

Type of the stickers in the set

name :: Maybe String

Sticker set name. Can contain only English letters, digits and underscores. Must end with *"_by_username"* (*bot_username* is case insensitive) for bots; 1-64 characters

title :: Maybe String

Sticker set title; 1-64 characters

user_id :: Maybe Int

Sticker set owner; ignored for regular users