TD.Query.SendInlineQueryResultMessage

Description

Documentation

data SendInlineQueryResultMessage

Sends the result of an inline query as a message. Returns the sent message. Always clears a chat draft message

Constructors

SendInlineQueryResultMessage

Fields

hide_via_bot :: Maybe Bool

Pass true to hide the bot, via which the message is sent. Can be used only for bots GetOption("animation_search_bot_username"), GetOption("photo_search_bot_username"), and GetOption("venue_search_bot_username")

result_id :: Maybe String

Identifier of the inline result

query_id :: Maybe Int

Identifier of the inline query

options :: Maybe MessageSendOptions

Options to be used to send the message; pass null to use default options

reply_to_message_id :: Maybe Int

Identifier of a replied message; 0 if none

message_thread_id :: Maybe Int

If not 0, a message thread identifier in which the message will be sent

chat_id :: Maybe Int

Target chat