TD.Query.SendMessageAlbum

Description

Documentation

data SendMessageAlbum

Sends 2-10 messages grouped together into an album. Currently, only audio, document, photo and video messages can be grouped into an album. Documents and audio files can be only grouped in an album with messages of the same type. Returns sent messages

Constructors

SendMessageAlbum

Fields

only_preview :: Maybe Bool

Pass true to get fake messages instead of actually sending them

input_message_contents :: Maybe [InputMessageContent]

Contents of messages to be sent. At most 10 messages can be added to an album

options :: Maybe MessageSendOptions

Options to be used to send the messages; pass null to use default options

reply_to_message_id :: Maybe Int

Identifier of a replied message; 0 if none

message_thread_id :: Maybe Int

If not 0, a message thread identifier in which the messages will be sent

chat_id :: Maybe Int

Target chat