TD.Query.SetChatBackground

Description

Documentation

data SetChatBackground

Changes the background in a specific chat. Supported only in private and secret chats with non-deleted users

Constructors

SetChatBackground

Fields

dark_theme_dimming :: Maybe Int

Dimming of the background in dark themes, as a percentage; 0-100

_type :: Maybe BackgroundType

Background type; pass null to remove the current background

background :: Maybe InputBackground

The input background to use; pass null to create a new filled background or to remove the current background

chat_id :: Maybe Int

Chat identifier